refactor:
* split to multi emoji
