--- src/main/kotlin/net/purefunc/emoji/Emoji.kt
@@ -1,3635 +1,0 @@
-package net.purefunc.emoji
-
-enum class Emoji(
-    private val code: Int,
-) {
-
-    GRINNING_FACE(0x1F600),
-    GRINNING_FACE_WITH_BIG_EYES(0x1F603),
-    GRINNING_FACE_WITH_SMILING_EYES(0x1F604),
-    BEAMING_FACE_WITH_SMILING_EYES(0x1F601),
-    GRINNING_SQUINTING_FACE(0x1F606),
-    GRINNING_FACE_WITH_SWEAT(0x1F605),
-    ROLLING_ON_THE_FLOOR_LAUGHING(0x1F923),
-    FACE_WITH_TEARS_OF_JOY(0x1F602),
-    SLIGHTLY_SMILING_FACE(0x1F642),
-    UPSIDE_DOWN_FACE(0x1F643),
-    MELTING_FACE(0x1FAE0),
-    WINKING_FACE(0x1F609),
-    SMILING_FACE_WITH_SMILING_EYES(0x1F60A),
-    SMILING_FACE_WITH_HALO(0x1F607),
-    SMILING_FACE_WITH_HEARTS(0x1F970),
-    SMILING_FACE_WITH_HEART_EYES(0x1F60D),
-    STAR_STRUCK(0x1F929),
-    FACE_BLOWING_A_KISS(0x1F618),
-    KISSING_FACE(0x1F617),
-    SMILING_FACE(0x263A),
-    KISSING_FACE_WITH_CLOSED_EYES(0x1F61A),
-    KISSING_FACE_WITH_SMILING_EYES(0x1F619),
-    SMILING_FACE_WITH_TEAR(0x1F972),
-    FACE_SAVORING_FOOD(0x1F60B),
-    FACE_WITH_TONGUE(0x1F61B),
-    WINKING_FACE_WITH_TONGUE(0x1F61C),
-    ZANY_FACE(0x1F92A),
-    SQUINTING_FACE_WITH_TONGUE(0x1F61D),
-    MONEY_MOUTH_FACE(0x1F911),
-    SMILING_FACE_WITH_OPEN_HANDS(0x1F917),
-    FACE_WITH_HAND_OVER_MOUTH(0x1F92D),
-    FACE_WITH_OPEN_EYES_AND_HAND_OVER_MOUTH(0x1FAE2),
-    FACE_WITH_PEEKING_EYE(0x1FAE3),
-    SHUSHING_FACE(0x1F92B),
-    THINKING_FACE(0x1F914),
-    SALUTING_FACE(0x1FAE1),
-    ZIPPER_MOUTH_FACE(0x1F910),
-    FACE_WITH_RAISED_EYEBROW(0x1F928),
-    NEUTRAL_FACE(0x1F610),
-    EXPRESSIONLESS_FACE(0x1F611),
-    FACE_WITHOUT_MOUTH(0x1F636),
-    DOTTED_LINE_FACE(0x1FAE5),
-    FACE_IN_CLOUDS(0x1F636),
-    SMIRKING_FACE(0x1F60F),
-    UNAMUSED_FACE(0x1F612),
-    FACE_WITH_ROLLING_EYES(0x1F644),
-    GRIMACING_FACE(0x1F62C),
-    FACE_EXHALING(0x1F62E),
-    LYING_FACE(0x1F925),
-    RELIEVED_FACE(0x1F60C),
-    PENSIVE_FACE(0x1F614),
-    SLEEPY_FACE(0x1F62A),
-    DROOLING_FACE(0x1F924),
-    SLEEPING_FACE(0x1F634),
-    FACE_WITH_MEDICAL_MASK(0x1F637),
-    FACE_WITH_THERMOMETER(0x1F912),
-    FACE_WITH_HEAD_BANDAGE(0x1F915),
-    NAUSEATED_FACE(0x1F922),
-    FACE_VOMITING(0x1F92E),
-    SNEEZING_FACE(0x1F927),
-    HOT_FACE(0x1F975),
-    COLD_FACE(0x1F976),
-    WOOZY_FACE(0x1F974),
-    FACE_WITH_CROSSED_OUT_EYES(0x1F635),
-    FACE_WITH_SPIRAL_EYES(0x1F635),
-    EXPLODING_HEAD(0x1F92F),
-    COWBOY_HAT_FACE(0x1F920),
-    PARTYING_FACE(0x1F973),
-    DISGUISED_FACE(0x1F978),
-    SMILING_FACE_WITH_SUNGLASSES(0x1F60E),
-    NERD_FACE(0x1F913),
-    FACE_WITH_MONOCLE(0x1F9D0),
-    CONFUSED_FACE(0x1F615),
-    FACE_WITH_DIAGONAL_MOUTH(0x1FAE4),
-    WORRIED_FACE(0x1F61F),
-    SLIGHTLY_FROWNING_FACE(0x1F641),
-    FROWNING_FACE(0x2639),
-    FACE_WITH_OPEN_MOUTH(0x1F62E),
-    HUSHED_FACE(0x1F62F),
-    ASTONISHED_FACE(0x1F632),
-    FLUSHED_FACE(0x1F633),
-    PLEADING_FACE(0x1F97A),
-    FACE_HOLDING_BACK_TEARS(0x1F979),
-    FROWNING_FACE_WITH_OPEN_MOUTH(0x1F626),
-    ANGUISHED_FACE(0x1F627),
-    FEARFUL_FACE(0x1F628),
-    ANXIOUS_FACE_WITH_SWEAT(0x1F630),
-    SAD_BUT_RELIEVED_FACE(0x1F625),
-    CRYING_FACE(0x1F622),
-    LOUDLY_CRYING_FACE(0x1F62D),
-    FACE_SCREAMING_IN_FEAR(0x1F631),
-    CONFOUNDED_FACE(0x1F616),
-    PERSEVERING_FACE(0x1F623),
-    DISAPPOINTED_FACE(0x1F61E),
-    DOWNCAST_FACE_WITH_SWEAT(0x1F613),
-    WEARY_FACE(0x1F629),
-    TIRED_FACE(0x1F62B),
-    YAWNING_FACE(0x1F971),
-    FACE_WITH_STEAM_FROM_NOSE(0x1F624),
-    POUTING_FACE(0x1F621),
-    ANGRY_FACE(0x1F620),
-    FACE_WITH_SYMBOLS_ON_MOUTH(0x1F92C),
-    SMILING_FACE_WITH_HORNS(0x1F608),
-    ANGRY_FACE_WITH_HORNS(0x1F47F),
-    SKULL(0x1F480),
-    SKULL_AND_CROSSBONES(0x2620),
-    PILE_OF_POO(0x1F4A9),
-    CLOWN_FACE(0x1F921),
-    OGRE(0x1F479),
-    GOBLIN(0x1F47A),
-    GHOST(0x1F47B),
-    ALIEN(0x1F47D),
-    ALIEN_MONSTER(0x1F47E),
-    ROBOT(0x1F916),
-    GRINNING_CAT(0x1F63A),
-    GRINNING_CAT_WITH_SMILING_EYES(0x1F638),
-    CAT_WITH_TEARS_OF_JOY(0x1F639),
-    SMILING_CAT_WITH_HEART_EYES(0x1F63B),
-    CAT_WITH_WRY_SMILE(0x1F63C),
-    KISSING_CAT(0x1F63D),
-    WEARY_CAT(0x1F640),
-    CRYING_CAT(0x1F63F),
-    POUTING_CAT(0x1F63E),
-    SEE_NO_EVIL_MONKEY(0x1F648),
-    HEAR_NO_EVIL_MONKEY(0x1F649),
-    SPEAK_NO_EVIL_MONKEY(0x1F64A),
-    KISS_MARK(0x1F48B),
-    LOVE_LETTER(0x1F48C),
-    HEART_WITH_ARROW(0x1F498),
-    HEART_WITH_RIBBON(0x1F49D),
-    SPARKLING_HEART(0x1F496),
-    GROWING_HEART(0x1F497),
-    BEATING_HEART(0x1F493),
-    REVOLVING_HEARTS(0x1F49E),
-    TWO_HEARTS(0x1F495),
-    HEART_DECORATION(0x1F49F),
-    HEART_EXCLAMATION(0x2763),
-    BROKEN_HEART(0x1F494),
-    HEART_ON_FIRE(0x2764),
-    MENDING_HEART(0x2764),
-    RED_HEART(0x2764),
-    ORANGE_HEART(0x1F9E1),
-    YELLOW_HEART(0x1F49B),
-    GREEN_HEART(0x1F49A),
-    BLUE_HEART(0x1F499),
-    PURPLE_HEART(0x1F49C),
-    BROWN_HEART(0x1F90E),
-    BLACK_HEART(0x1F5A4),
-    WHITE_HEART(0x1F90D),
-    HUNDRED_POINTS(0x1F4AF),
-    ANGER_SYMBOL(0x1F4A2),
-    COLLISION(0x1F4A5),
-    DIZZY(0x1F4AB),
-    SWEAT_DROPLETS(0x1F4A6),
-    DASHING_AWAY(0x1F4A8),
-    HOLE(0x1F573),
-    BOMB(0x1F4A3),
-    SPEECH_BALLOON(0x1F4AC),
-    EYE_IN_SPEECH_BUBBLE(0x1F441),
-    LEFT_SPEECH_BUBBLE(0x1F5E8),
-    RIGHT_ANGER_BUBBLE(0x1F5EF),
-    THOUGHT_BALLOON(0x1F4AD),
-    ZZZ(0x1F4A4),
-    WAVING_HAND(0x1F44B),
-    WAVING_HAND_LIGHT_SKIN_TONE(0x1F44B),
-    WAVING_HAND_MEDIUM_LIGHT_SKIN_TONE(0x1F44B),
-    WAVING_HAND_MEDIUM_SKIN_TONE(0x1F44B),
-    WAVING_HAND_MEDIUM_DARK_SKIN_TONE(0x1F44B),
-    WAVING_HAND_DARK_SKIN_TONE(0x1F44B),
-    RAISED_BACK_OF_HAND(0x1F91A),
-    RAISED_BACK_OF_HAND_LIGHT_SKIN_TONE(0x1F91A),
-    RAISED_BACK_OF_HAND_MEDIUM_LIGHT_SKIN_TONE(0x1F91A),
-    RAISED_BACK_OF_HAND_MEDIUM_SKIN_TONE(0x1F91A),
-    RAISED_BACK_OF_HAND_MEDIUM_DARK_SKIN_TONE(0x1F91A),
-    RAISED_BACK_OF_HAND_DARK_SKIN_TONE(0x1F91A),
-    HAND_WITH_FINGERS_SPLAYED(0x1F590),
-    HAND_WITH_FINGERS_SPLAYED_LIGHT_SKIN_TONE(0x1F590),
-    HAND_WITH_FINGERS_SPLAYED_MEDIUM_LIGHT_SKIN_TONE(0x1F590),
-    HAND_WITH_FINGERS_SPLAYED_MEDIUM_SKIN_TONE(0x1F590),
-    HAND_WITH_FINGERS_SPLAYED_MEDIUM_DARK_SKIN_TONE(0x1F590),
-    HAND_WITH_FINGERS_SPLAYED_DARK_SKIN_TONE(0x1F590),
-    RAISED_HAND(0x270B),
-    RAISED_HAND_LIGHT_SKIN_TONE(0x270B),
-    RAISED_HAND_MEDIUM_LIGHT_SKIN_TONE(0x270B),
-    RAISED_HAND_MEDIUM_SKIN_TONE(0x270B),
-    RAISED_HAND_MEDIUM_DARK_SKIN_TONE(0x270B),
-    RAISED_HAND_DARK_SKIN_TONE(0x270B),
-    VULCAN_SALUTE(0x1F596),
-    VULCAN_SALUTE_LIGHT_SKIN_TONE(0x1F596),
-    VULCAN_SALUTE_MEDIUM_LIGHT_SKIN_TONE(0x1F596),
-    VULCAN_SALUTE_MEDIUM_SKIN_TONE(0x1F596),
-    VULCAN_SALUTE_MEDIUM_DARK_SKIN_TONE(0x1F596),
-    VULCAN_SALUTE_DARK_SKIN_TONE(0x1F596),
-    RIGHTWARDS_HAND(0x1FAF1),
-    RIGHTWARDS_HAND_LIGHT_SKIN_TONE(0x1FAF1),
-    RIGHTWARDS_HAND_MEDIUM_LIGHT_SKIN_TONE(0x1FAF1),
-    RIGHTWARDS_HAND_MEDIUM_SKIN_TONE(0x1FAF1),
-    RIGHTWARDS_HAND_MEDIUM_DARK_SKIN_TONE(0x1FAF1),
-    RIGHTWARDS_HAND_DARK_SKIN_TONE(0x1FAF1),
-    LEFTWARDS_HAND(0x1FAF2),
-    LEFTWARDS_HAND_LIGHT_SKIN_TONE(0x1FAF2),
-    LEFTWARDS_HAND_MEDIUM_LIGHT_SKIN_TONE(0x1FAF2),
-    LEFTWARDS_HAND_MEDIUM_SKIN_TONE(0x1FAF2),
-    LEFTWARDS_HAND_MEDIUM_DARK_SKIN_TONE(0x1FAF2),
-    LEFTWARDS_HAND_DARK_SKIN_TONE(0x1FAF2),
-    PALM_DOWN_HAND(0x1FAF3),
-    PALM_DOWN_HAND_LIGHT_SKIN_TONE(0x1FAF3),
-    PALM_DOWN_HAND_MEDIUM_LIGHT_SKIN_TONE(0x1FAF3),
-    PALM_DOWN_HAND_MEDIUM_SKIN_TONE(0x1FAF3),
-    PALM_DOWN_HAND_MEDIUM_DARK_SKIN_TONE(0x1FAF3),
-    PALM_DOWN_HAND_DARK_SKIN_TONE(0x1FAF3),
-    PALM_UP_HAND(0x1FAF4),
-    PALM_UP_HAND_LIGHT_SKIN_TONE(0x1FAF4),
-    PALM_UP_HAND_MEDIUM_LIGHT_SKIN_TONE(0x1FAF4),
-    PALM_UP_HAND_MEDIUM_SKIN_TONE(0x1FAF4),
-    PALM_UP_HAND_MEDIUM_DARK_SKIN_TONE(0x1FAF4),
-    PALM_UP_HAND_DARK_SKIN_TONE(0x1FAF4),
-    OK_HAND(0x1F44C),
-    OK_HAND_LIGHT_SKIN_TONE(0x1F44C),
-    OK_HAND_MEDIUM_LIGHT_SKIN_TONE(0x1F44C),
-    OK_HAND_MEDIUM_SKIN_TONE(0x1F44C),
-    OK_HAND_MEDIUM_DARK_SKIN_TONE(0x1F44C),
-    OK_HAND_DARK_SKIN_TONE(0x1F44C),
-    PINCHED_FINGERS(0x1F90C),
-    PINCHED_FINGERS_LIGHT_SKIN_TONE(0x1F90C),
-    PINCHED_FINGERS_MEDIUM_LIGHT_SKIN_TONE(0x1F90C),
-    PINCHED_FINGERS_MEDIUM_SKIN_TONE(0x1F90C),
-    PINCHED_FINGERS_MEDIUM_DARK_SKIN_TONE(0x1F90C),
-    PINCHED_FINGERS_DARK_SKIN_TONE(0x1F90C),
-    PINCHING_HAND(0x1F90F),
-    PINCHING_HAND_LIGHT_SKIN_TONE(0x1F90F),
-    PINCHING_HAND_MEDIUM_LIGHT_SKIN_TONE(0x1F90F),
-    PINCHING_HAND_MEDIUM_SKIN_TONE(0x1F90F),
-    PINCHING_HAND_MEDIUM_DARK_SKIN_TONE(0x1F90F),
-    PINCHING_HAND_DARK_SKIN_TONE(0x1F90F),
-    VICTORY_HAND(0x270C),
-    VICTORY_HAND_LIGHT_SKIN_TONE(0x270C),
-    VICTORY_HAND_MEDIUM_LIGHT_SKIN_TONE(0x270C),
-    VICTORY_HAND_MEDIUM_SKIN_TONE(0x270C),
-    VICTORY_HAND_MEDIUM_DARK_SKIN_TONE(0x270C),
-    VICTORY_HAND_DARK_SKIN_TONE(0x270C),
-    CROSSED_FINGERS(0x1F91E),
-    CROSSED_FINGERS_LIGHT_SKIN_TONE(0x1F91E),
-    CROSSED_FINGERS_MEDIUM_LIGHT_SKIN_TONE(0x1F91E),
-    CROSSED_FINGERS_MEDIUM_SKIN_TONE(0x1F91E),
-    CROSSED_FINGERS_MEDIUM_DARK_SKIN_TONE(0x1F91E),
-    CROSSED_FINGERS_DARK_SKIN_TONE(0x1F91E),
-    HAND_WITH_INDEX_FINGER_AND_THUMB_CROSSED(0x1FAF0),
-    HAND_WITH_INDEX_FINGER_AND_THUMB_CROSSED_LIGHT_SKIN_TONE(0x1FAF0),
-    HAND_WITH_INDEX_FINGER_AND_THUMB_CROSSED_MEDIUM_LIGHT_SKIN_TONE(0x1FAF0),
-    HAND_WITH_INDEX_FINGER_AND_THUMB_CROSSED_MEDIUM_SKIN_TONE(0x1FAF0),
-    HAND_WITH_INDEX_FINGER_AND_THUMB_CROSSED_MEDIUM_DARK_SKIN_TONE(0x1FAF0),
-    HAND_WITH_INDEX_FINGER_AND_THUMB_CROSSED_DARK_SKIN_TONE(0x1FAF0),
-    LOVE_YOU_GESTURE(0x1F91F),
-    LOVE_YOU_GESTURE_LIGHT_SKIN_TONE(0x1F91F),
-    LOVE_YOU_GESTURE_MEDIUM_LIGHT_SKIN_TONE(0x1F91F),
-    LOVE_YOU_GESTURE_MEDIUM_SKIN_TONE(0x1F91F),
-    LOVE_YOU_GESTURE_MEDIUM_DARK_SKIN_TONE(0x1F91F),
-    LOVE_YOU_GESTURE_DARK_SKIN_TONE(0x1F91F),
-    SIGN_OF_THE_HORNS(0x1F918),
-    SIGN_OF_THE_HORNS_LIGHT_SKIN_TONE(0x1F918),
-    SIGN_OF_THE_HORNS_MEDIUM_LIGHT_SKIN_TONE(0x1F918),
-    SIGN_OF_THE_HORNS_MEDIUM_SKIN_TONE(0x1F918),
-    SIGN_OF_THE_HORNS_MEDIUM_DARK_SKIN_TONE(0x1F918),
-    SIGN_OF_THE_HORNS_DARK_SKIN_TONE(0x1F918),
-    CALL_ME_HAND(0x1F919),
-    CALL_ME_HAND_LIGHT_SKIN_TONE(0x1F919),
-    CALL_ME_HAND_MEDIUM_LIGHT_SKIN_TONE(0x1F919),
-    CALL_ME_HAND_MEDIUM_SKIN_TONE(0x1F919),
-    CALL_ME_HAND_MEDIUM_DARK_SKIN_TONE(0x1F919),
-    CALL_ME_HAND_DARK_SKIN_TONE(0x1F919),
-    BACKHAND_INDEX_POINTING_LEFT(0x1F448),
-    BACKHAND_INDEX_POINTING_LEFT_LIGHT_SKIN_TONE(0x1F448),
-    BACKHAND_INDEX_POINTING_LEFT_MEDIUM_LIGHT_SKIN_TONE(0x1F448),
-    BACKHAND_INDEX_POINTING_LEFT_MEDIUM_SKIN_TONE(0x1F448),
-    BACKHAND_INDEX_POINTING_LEFT_MEDIUM_DARK_SKIN_TONE(0x1F448),
-    BACKHAND_INDEX_POINTING_LEFT_DARK_SKIN_TONE(0x1F448),
-    BACKHAND_INDEX_POINTING_RIGHT(0x1F449),
-    BACKHAND_INDEX_POINTING_RIGHT_LIGHT_SKIN_TONE(0x1F449),
-    BACKHAND_INDEX_POINTING_RIGHT_MEDIUM_LIGHT_SKIN_TONE(0x1F449),
-    BACKHAND_INDEX_POINTING_RIGHT_MEDIUM_SKIN_TONE(0x1F449),
-    BACKHAND_INDEX_POINTING_RIGHT_MEDIUM_DARK_SKIN_TONE(0x1F449),
-    BACKHAND_INDEX_POINTING_RIGHT_DARK_SKIN_TONE(0x1F449),
-    BACKHAND_INDEX_POINTING_UP(0x1F446),
-    BACKHAND_INDEX_POINTING_UP_LIGHT_SKIN_TONE(0x1F446),
-    BACKHAND_INDEX_POINTING_UP_MEDIUM_LIGHT_SKIN_TONE(0x1F446),
-    BACKHAND_INDEX_POINTING_UP_MEDIUM_SKIN_TONE(0x1F446),
-    BACKHAND_INDEX_POINTING_UP_MEDIUM_DARK_SKIN_TONE(0x1F446),
-    BACKHAND_INDEX_POINTING_UP_DARK_SKIN_TONE(0x1F446),
-    MIDDLE_FINGER(0x1F595),
-    MIDDLE_FINGER_LIGHT_SKIN_TONE(0x1F595),
-    MIDDLE_FINGER_MEDIUM_LIGHT_SKIN_TONE(0x1F595),
-    MIDDLE_FINGER_MEDIUM_SKIN_TONE(0x1F595),
-    MIDDLE_FINGER_MEDIUM_DARK_SKIN_TONE(0x1F595),
-    MIDDLE_FINGER_DARK_SKIN_TONE(0x1F595),
-    BACKHAND_INDEX_POINTING_DOWN(0x1F447),
-    BACKHAND_INDEX_POINTING_DOWN_LIGHT_SKIN_TONE(0x1F447),
-    BACKHAND_INDEX_POINTING_DOWN_MEDIUM_LIGHT_SKIN_TONE(0x1F447),
-    BACKHAND_INDEX_POINTING_DOWN_MEDIUM_SKIN_TONE(0x1F447),
-    BACKHAND_INDEX_POINTING_DOWN_MEDIUM_DARK_SKIN_TONE(0x1F447),
-    BACKHAND_INDEX_POINTING_DOWN_DARK_SKIN_TONE(0x1F447),
-    INDEX_POINTING_UP(0x261D),
-    INDEX_POINTING_UP_LIGHT_SKIN_TONE(0x261D),
-    INDEX_POINTING_UP_MEDIUM_LIGHT_SKIN_TONE(0x261D),
-    INDEX_POINTING_UP_MEDIUM_SKIN_TONE(0x261D),
-    INDEX_POINTING_UP_MEDIUM_DARK_SKIN_TONE(0x261D),
-    INDEX_POINTING_UP_DARK_SKIN_TONE(0x261D),
-    INDEX_POINTING_AT_THE_VIEWER(0x1FAF5),
-    INDEX_POINTING_AT_THE_VIEWER_LIGHT_SKIN_TONE(0x1FAF5),
-    INDEX_POINTING_AT_THE_VIEWER_MEDIUM_LIGHT_SKIN_TONE(0x1FAF5),
-    INDEX_POINTING_AT_THE_VIEWER_MEDIUM_SKIN_TONE(0x1FAF5),
-    INDEX_POINTING_AT_THE_VIEWER_MEDIUM_DARK_SKIN_TONE(0x1FAF5),
-    INDEX_POINTING_AT_THE_VIEWER_DARK_SKIN_TONE(0x1FAF5),
-    THUMBS_UP(0x1F44D),
-    THUMBS_UP_LIGHT_SKIN_TONE(0x1F44D),
-    THUMBS_UP_MEDIUM_LIGHT_SKIN_TONE(0x1F44D),
-    THUMBS_UP_MEDIUM_SKIN_TONE(0x1F44D),
-    THUMBS_UP_MEDIUM_DARK_SKIN_TONE(0x1F44D),
-    THUMBS_UP_DARK_SKIN_TONE(0x1F44D),
-    THUMBS_DOWN(0x1F44E),
-    THUMBS_DOWN_LIGHT_SKIN_TONE(0x1F44E),
-    THUMBS_DOWN_MEDIUM_LIGHT_SKIN_TONE(0x1F44E),
-    THUMBS_DOWN_MEDIUM_SKIN_TONE(0x1F44E),
-    THUMBS_DOWN_MEDIUM_DARK_SKIN_TONE(0x1F44E),
-    THUMBS_DOWN_DARK_SKIN_TONE(0x1F44E),
-    RAISED_FIST(0x270A),
-    RAISED_FIST_LIGHT_SKIN_TONE(0x270A),
-    RAISED_FIST_MEDIUM_LIGHT_SKIN_TONE(0x270A),
-    RAISED_FIST_MEDIUM_SKIN_TONE(0x270A),
-    RAISED_FIST_MEDIUM_DARK_SKIN_TONE(0x270A),
-    RAISED_FIST_DARK_SKIN_TONE(0x270A),
-    ONCOMING_FIST(0x1F44A),
-    ONCOMING_FIST_LIGHT_SKIN_TONE(0x1F44A),
-    ONCOMING_FIST_MEDIUM_LIGHT_SKIN_TONE(0x1F44A),
-    ONCOMING_FIST_MEDIUM_SKIN_TONE(0x1F44A),
-    ONCOMING_FIST_MEDIUM_DARK_SKIN_TONE(0x1F44A),
-    ONCOMING_FIST_DARK_SKIN_TONE(0x1F44A),
-    LEFT_FACING_FIST(0x1F91B),
-    LEFT_FACING_FIST_LIGHT_SKIN_TONE(0x1F91B),
-    LEFT_FACING_FIST_MEDIUM_LIGHT_SKIN_TONE(0x1F91B),
-    LEFT_FACING_FIST_MEDIUM_SKIN_TONE(0x1F91B),
-    LEFT_FACING_FIST_MEDIUM_DARK_SKIN_TONE(0x1F91B),
-    LEFT_FACING_FIST_DARK_SKIN_TONE(0x1F91B),
-    RIGHT_FACING_FIST(0x1F91C),
-    RIGHT_FACING_FIST_LIGHT_SKIN_TONE(0x1F91C),
-    RIGHT_FACING_FIST_MEDIUM_LIGHT_SKIN_TONE(0x1F91C),
-    RIGHT_FACING_FIST_MEDIUM_SKIN_TONE(0x1F91C),
-    RIGHT_FACING_FIST_MEDIUM_DARK_SKIN_TONE(0x1F91C),
-    RIGHT_FACING_FIST_DARK_SKIN_TONE(0x1F91C),
-    CLAPPING_HANDS(0x1F44F),
-    CLAPPING_HANDS_LIGHT_SKIN_TONE(0x1F44F),
-    CLAPPING_HANDS_MEDIUM_LIGHT_SKIN_TONE(0x1F44F),
-    CLAPPING_HANDS_MEDIUM_SKIN_TONE(0x1F44F),
-    CLAPPING_HANDS_MEDIUM_DARK_SKIN_TONE(0x1F44F),
-    CLAPPING_HANDS_DARK_SKIN_TONE(0x1F44F),
-    RAISING_HANDS(0x1F64C),
-    RAISING_HANDS_LIGHT_SKIN_TONE(0x1F64C),
-    RAISING_HANDS_MEDIUM_LIGHT_SKIN_TONE(0x1F64C),
-    RAISING_HANDS_MEDIUM_SKIN_TONE(0x1F64C),
-    RAISING_HANDS_MEDIUM_DARK_SKIN_TONE(0x1F64C),
-    RAISING_HANDS_DARK_SKIN_TONE(0x1F64C),
-    HEART_HANDS(0x1FAF6),
-    HEART_HANDS_LIGHT_SKIN_TONE(0x1FAF6),
-    HEART_HANDS_MEDIUM_LIGHT_SKIN_TONE(0x1FAF6),
-    HEART_HANDS_MEDIUM_SKIN_TONE(0x1FAF6),
-    HEART_HANDS_MEDIUM_DARK_SKIN_TONE(0x1FAF6),
-    HEART_HANDS_DARK_SKIN_TONE(0x1FAF6),
-    OPEN_HANDS(0x1F450),
-    OPEN_HANDS_LIGHT_SKIN_TONE(0x1F450),
-    OPEN_HANDS_MEDIUM_LIGHT_SKIN_TONE(0x1F450),
-    OPEN_HANDS_MEDIUM_SKIN_TONE(0x1F450),
-    OPEN_HANDS_MEDIUM_DARK_SKIN_TONE(0x1F450),
-    OPEN_HANDS_DARK_SKIN_TONE(0x1F450),
-    PALMS_UP_TOGETHER(0x1F932),
-    PALMS_UP_TOGETHER_LIGHT_SKIN_TONE(0x1F932),
-    PALMS_UP_TOGETHER_MEDIUM_LIGHT_SKIN_TONE(0x1F932),
-    PALMS_UP_TOGETHER_MEDIUM_SKIN_TONE(0x1F932),
-    PALMS_UP_TOGETHER_MEDIUM_DARK_SKIN_TONE(0x1F932),
-    PALMS_UP_TOGETHER_DARK_SKIN_TONE(0x1F932),
-    HANDSHAKE(0x1F91D),
-    HANDSHAKE_LIGHT_SKIN_TONE(0x1F91D),
-    HANDSHAKE_MEDIUM_LIGHT_SKIN_TONE(0x1F91D),
-    HANDSHAKE_MEDIUM_SKIN_TONE(0x1F91D),
-    HANDSHAKE_MEDIUM_DARK_SKIN_TONE(0x1F91D),
-    HANDSHAKE_DARK_SKIN_TONE(0x1F91D),
-    HANDSHAKE_LIGHT_SKIN_TONE_MEDIUM_LIGHT_SKIN_TONE(0x1FAF1),
-    HANDSHAKE_LIGHT_SKIN_TONE_MEDIUM_SKIN_TONE(0x1FAF1),
-    HANDSHAKE_LIGHT_SKIN_TONE_MEDIUM_DARK_SKIN_TONE(0x1FAF1),
-    HANDSHAKE_LIGHT_SKIN_TONE_DARK_SKIN_TONE(0x1FAF1),
-    HANDSHAKE_MEDIUM_LIGHT_SKIN_TONE_LIGHT_SKIN_TONE(0x1FAF1),
-    HANDSHAKE_MEDIUM_LIGHT_SKIN_TONE_MEDIUM_SKIN_TONE(0x1FAF1),
-    HANDSHAKE_MEDIUM_LIGHT_SKIN_TONE_MEDIUM_DARK_SKIN_TONE(0x1FAF1),
-    HANDSHAKE_MEDIUM_LIGHT_SKIN_TONE_DARK_SKIN_TONE(0x1FAF1),
-    HANDSHAKE_MEDIUM_SKIN_TONE_LIGHT_SKIN_TONE(0x1FAF1),
-    HANDSHAKE_MEDIUM_SKIN_TONE_MEDIUM_LIGHT_SKIN_TONE(0x1FAF1),
-    HANDSHAKE_MEDIUM_SKIN_TONE_MEDIUM_DARK_SKIN_TONE(0x1FAF1),
-    HANDSHAKE_MEDIUM_SKIN_TONE_DARK_SKIN_TONE(0x1FAF1),
-    HANDSHAKE_MEDIUM_DARK_SKIN_TONE_LIGHT_SKIN_TONE(0x1FAF1),
-    HANDSHAKE_MEDIUM_DARK_SKIN_TONE_MEDIUM_LIGHT_SKIN_TONE(0x1FAF1),
-    HANDSHAKE_MEDIUM_DARK_SKIN_TONE_MEDIUM_SKIN_TONE(0x1FAF1),
-    HANDSHAKE_MEDIUM_DARK_SKIN_TONE_DARK_SKIN_TONE(0x1FAF1),
-    HANDSHAKE_DARK_SKIN_TONE_LIGHT_SKIN_TONE(0x1FAF1),
-    HANDSHAKE_DARK_SKIN_TONE_MEDIUM_LIGHT_SKIN_TONE(0x1FAF1),
-    HANDSHAKE_DARK_SKIN_TONE_MEDIUM_SKIN_TONE(0x1FAF1),
-    HANDSHAKE_DARK_SKIN_TONE_MEDIUM_DARK_SKIN_TONE(0x1FAF1),
-    FOLDED_HANDS(0x1F64F),
-    FOLDED_HANDS_LIGHT_SKIN_TONE(0x1F64F),
-    FOLDED_HANDS_MEDIUM_LIGHT_SKIN_TONE(0x1F64F),
-    FOLDED_HANDS_MEDIUM_SKIN_TONE(0x1F64F),
-    FOLDED_HANDS_MEDIUM_DARK_SKIN_TONE(0x1F64F),
-    FOLDED_HANDS_DARK_SKIN_TONE(0x1F64F),
-    WRITING_HAND(0x270D),
-    WRITING_HAND_LIGHT_SKIN_TONE(0x270D),
-    WRITING_HAND_MEDIUM_LIGHT_SKIN_TONE(0x270D),
-    WRITING_HAND_MEDIUM_SKIN_TONE(0x270D),
-    WRITING_HAND_MEDIUM_DARK_SKIN_TONE(0x270D),
-    WRITING_HAND_DARK_SKIN_TONE(0x270D),
-    NAIL_POLISH(0x1F485),
-    NAIL_POLISH_LIGHT_SKIN_TONE(0x1F485),
-    NAIL_POLISH_MEDIUM_LIGHT_SKIN_TONE(0x1F485),
-    NAIL_POLISH_MEDIUM_SKIN_TONE(0x1F485),
-    NAIL_POLISH_MEDIUM_DARK_SKIN_TONE(0x1F485),
-    NAIL_POLISH_DARK_SKIN_TONE(0x1F485),
-    SELFIE(0x1F933),
-    SELFIE_LIGHT_SKIN_TONE(0x1F933),
-    SELFIE_MEDIUM_LIGHT_SKIN_TONE(0x1F933),
-    SELFIE_MEDIUM_SKIN_TONE(0x1F933),
-    SELFIE_MEDIUM_DARK_SKIN_TONE(0x1F933),
-    SELFIE_DARK_SKIN_TONE(0x1F933),
-    FLEXED_BICEPS(0x1F4AA),
-    FLEXED_BICEPS_LIGHT_SKIN_TONE(0x1F4AA),
-    FLEXED_BICEPS_MEDIUM_LIGHT_SKIN_TONE(0x1F4AA),
-    FLEXED_BICEPS_MEDIUM_SKIN_TONE(0x1F4AA),
-    FLEXED_BICEPS_MEDIUM_DARK_SKIN_TONE(0x1F4AA),
-    FLEXED_BICEPS_DARK_SKIN_TONE(0x1F4AA),
-    MECHANICAL_ARM(0x1F9BE),
-    MECHANICAL_LEG(0x1F9BF),
-    LEG(0x1F9B5),
-    LEG_LIGHT_SKIN_TONE(0x1F9B5),
-    LEG_MEDIUM_LIGHT_SKIN_TONE(0x1F9B5),
-    LEG_MEDIUM_SKIN_TONE(0x1F9B5),
-    LEG_MEDIUM_DARK_SKIN_TONE(0x1F9B5),
-    LEG_DARK_SKIN_TONE(0x1F9B5),
-    FOOT(0x1F9B6),
-    FOOT_LIGHT_SKIN_TONE(0x1F9B6),
-    FOOT_MEDIUM_LIGHT_SKIN_TONE(0x1F9B6),
-    FOOT_MEDIUM_SKIN_TONE(0x1F9B6),
-    FOOT_MEDIUM_DARK_SKIN_TONE(0x1F9B6),
-    FOOT_DARK_SKIN_TONE(0x1F9B6),
-    EAR(0x1F442),
-    EAR_LIGHT_SKIN_TONE(0x1F442),
-    EAR_MEDIUM_LIGHT_SKIN_TONE(0x1F442),
-    EAR_MEDIUM_SKIN_TONE(0x1F442),
-    EAR_MEDIUM_DARK_SKIN_TONE(0x1F442),
-    EAR_DARK_SKIN_TONE(0x1F442),
-    EAR_WITH_HEARING_AID(0x1F9BB),
-    EAR_WITH_HEARING_AID_LIGHT_SKIN_TONE(0x1F9BB),
-    EAR_WITH_HEARING_AID_MEDIUM_LIGHT_SKIN_TONE(0x1F9BB),
-    EAR_WITH_HEARING_AID_MEDIUM_SKIN_TONE(0x1F9BB),
-    EAR_WITH_HEARING_AID_MEDIUM_DARK_SKIN_TONE(0x1F9BB),
-    EAR_WITH_HEARING_AID_DARK_SKIN_TONE(0x1F9BB),
-    NOSE(0x1F443),
-    NOSE_LIGHT_SKIN_TONE(0x1F443),
-    NOSE_MEDIUM_LIGHT_SKIN_TONE(0x1F443),
-    NOSE_MEDIUM_SKIN_TONE(0x1F443),
-    NOSE_MEDIUM_DARK_SKIN_TONE(0x1F443),
-    NOSE_DARK_SKIN_TONE(0x1F443),
-    BRAIN(0x1F9E0),
-    ANATOMICAL_HEART(0x1FAC0),
-    LUNGS(0x1FAC1),
-    TOOTH(0x1F9B7),
-    BONE(0x1F9B4),
-    EYES(0x1F440),
-    EYE(0x1F441),
-    TONGUE(0x1F445),
-    MOUTH(0x1F444),
-    BITING_LIP(0x1FAE6),
-    BABY(0x1F476),
-    BABY_LIGHT_SKIN_TONE(0x1F476),
-    BABY_MEDIUM_LIGHT_SKIN_TONE(0x1F476),
-    BABY_MEDIUM_SKIN_TONE(0x1F476),
-    BABY_MEDIUM_DARK_SKIN_TONE(0x1F476),
-    BABY_DARK_SKIN_TONE(0x1F476),
-    CHILD(0x1F9D2),
-    CHILD_LIGHT_SKIN_TONE(0x1F9D2),
-    CHILD_MEDIUM_LIGHT_SKIN_TONE(0x1F9D2),
-    CHILD_MEDIUM_SKIN_TONE(0x1F9D2),
-    CHILD_MEDIUM_DARK_SKIN_TONE(0x1F9D2),
-    CHILD_DARK_SKIN_TONE(0x1F9D2),
-    BOY(0x1F466),
-    BOY_LIGHT_SKIN_TONE(0x1F466),
-    BOY_MEDIUM_LIGHT_SKIN_TONE(0x1F466),
-    BOY_MEDIUM_SKIN_TONE(0x1F466),
-    BOY_MEDIUM_DARK_SKIN_TONE(0x1F466),
-    BOY_DARK_SKIN_TONE(0x1F466),
-    GIRL(0x1F467),
-    GIRL_LIGHT_SKIN_TONE(0x1F467),
-    GIRL_MEDIUM_LIGHT_SKIN_TONE(0x1F467),
-    GIRL_MEDIUM_SKIN_TONE(0x1F467),
-    GIRL_MEDIUM_DARK_SKIN_TONE(0x1F467),
-    GIRL_DARK_SKIN_TONE(0x1F467),
-    PERSON(0x1F9D1),
-    PERSON_LIGHT_SKIN_TONE(0x1F9D1),
-    PERSON_MEDIUM_LIGHT_SKIN_TONE(0x1F9D1),
-    PERSON_MEDIUM_SKIN_TONE(0x1F9D1),
-    PERSON_MEDIUM_DARK_SKIN_TONE(0x1F9D1),
-    PERSON_DARK_SKIN_TONE(0x1F9D1),
-    PERSON_BLOND_HAIR(0x1F471),
-    PERSON_LIGHT_SKIN_TONE_BLOND_HAIR(0x1F471),
-    PERSON_MEDIUM_LIGHT_SKIN_TONE_BLOND_HAIR(0x1F471),
-    PERSON_MEDIUM_SKIN_TONE_BLOND_HAIR(0x1F471),
-    PERSON_MEDIUM_DARK_SKIN_TONE_BLOND_HAIR(0x1F471),
-    PERSON_DARK_SKIN_TONE_BLOND_HAIR(0x1F471),
-    MAN(0x1F468),
-    MAN_LIGHT_SKIN_TONE(0x1F468),
-    MAN_MEDIUM_LIGHT_SKIN_TONE(0x1F468),
-    MAN_MEDIUM_SKIN_TONE(0x1F468),
-    MAN_MEDIUM_DARK_SKIN_TONE(0x1F468),
-    MAN_DARK_SKIN_TONE(0x1F468),
-    PERSON_BEARD(0x1F9D4),
-    PERSON_LIGHT_SKIN_TONE_BEARD(0x1F9D4),
-    PERSON_MEDIUM_LIGHT_SKIN_TONE_BEARD(0x1F9D4),
-    PERSON_MEDIUM_SKIN_TONE_BEARD(0x1F9D4),
-    PERSON_MEDIUM_DARK_SKIN_TONE_BEARD(0x1F9D4),
-    PERSON_DARK_SKIN_TONE_BEARD(0x1F9D4),
-    MAN_BEARD(0x1F9D4),
-    MAN_LIGHT_SKIN_TONE_BEARD(0x1F9D4),
-    MAN_MEDIUM_LIGHT_SKIN_TONE_BEARD(0x1F9D4),
-    MAN_MEDIUM_SKIN_TONE_BEARD(0x1F9D4),
-    MAN_MEDIUM_DARK_SKIN_TONE_BEARD(0x1F9D4),
-    MAN_DARK_SKIN_TONE_BEARD(0x1F9D4),
-    WOMAN_BEARD(0x1F9D4),
-    WOMAN_LIGHT_SKIN_TONE_BEARD(0x1F9D4),
-    WOMAN_MEDIUM_LIGHT_SKIN_TONE_BEARD(0x1F9D4),
-    WOMAN_MEDIUM_SKIN_TONE_BEARD(0x1F9D4),
-    WOMAN_MEDIUM_DARK_SKIN_TONE_BEARD(0x1F9D4),
-    WOMAN_DARK_SKIN_TONE_BEARD(0x1F9D4),
-    MAN_RED_HAIR(0x1F468),
-    MAN_LIGHT_SKIN_TONE_RED_HAIR(0x1F468),
-    MAN_MEDIUM_LIGHT_SKIN_TONE_RED_HAIR(0x1F468),
-    MAN_MEDIUM_SKIN_TONE_RED_HAIR(0x1F468),
-    MAN_MEDIUM_DARK_SKIN_TONE_RED_HAIR(0x1F468),
-    MAN_DARK_SKIN_TONE_RED_HAIR(0x1F468),
-    MAN_CURLY_HAIR(0x1F468),
-    MAN_LIGHT_SKIN_TONE_CURLY_HAIR(0x1F468),
-    MAN_MEDIUM_LIGHT_SKIN_TONE_CURLY_HAIR(0x1F468),
-    MAN_MEDIUM_SKIN_TONE_CURLY_HAIR(0x1F468),
-    MAN_MEDIUM_DARK_SKIN_TONE_CURLY_HAIR(0x1F468),
-    MAN_DARK_SKIN_TONE_CURLY_HAIR(0x1F468),
-    MAN_WHITE_HAIR(0x1F468),
-    MAN_LIGHT_SKIN_TONE_WHITE_HAIR(0x1F468),
-    MAN_MEDIUM_LIGHT_SKIN_TONE_WHITE_HAIR(0x1F468),
-    MAN_MEDIUM_SKIN_TONE_WHITE_HAIR(0x1F468),
-    MAN_MEDIUM_DARK_SKIN_TONE_WHITE_HAIR(0x1F468),
-    MAN_DARK_SKIN_TONE_WHITE_HAIR(0x1F468),
-    MAN_BALD(0x1F468),
-    MAN_LIGHT_SKIN_TONE_BALD(0x1F468),
-    MAN_MEDIUM_LIGHT_SKIN_TONE_BALD(0x1F468),
-    MAN_MEDIUM_SKIN_TONE_BALD(0x1F468),
-    MAN_MEDIUM_DARK_SKIN_TONE_BALD(0x1F468),
-    MAN_DARK_SKIN_TONE_BALD(0x1F468),
-    WOMAN(0x1F469),
-    WOMAN_LIGHT_SKIN_TONE(0x1F469),
-    WOMAN_MEDIUM_LIGHT_SKIN_TONE(0x1F469),
-    WOMAN_MEDIUM_SKIN_TONE(0x1F469),
-    WOMAN_MEDIUM_DARK_SKIN_TONE(0x1F469),
-    WOMAN_DARK_SKIN_TONE(0x1F469),
-    WOMAN_RED_HAIR(0x1F469),
-    WOMAN_LIGHT_SKIN_TONE_RED_HAIR(0x1F469),
-    WOMAN_MEDIUM_LIGHT_SKIN_TONE_RED_HAIR(0x1F469),
-    WOMAN_MEDIUM_SKIN_TONE_RED_HAIR(0x1F469),
-    WOMAN_MEDIUM_DARK_SKIN_TONE_RED_HAIR(0x1F469),
-    WOMAN_DARK_SKIN_TONE_RED_HAIR(0x1F469),
-    PERSON_RED_HAIR(0x1F9D1),
-    PERSON_LIGHT_SKIN_TONE_RED_HAIR(0x1F9D1),
-    PERSON_MEDIUM_LIGHT_SKIN_TONE_RED_HAIR(0x1F9D1),
-    PERSON_MEDIUM_SKIN_TONE_RED_HAIR(0x1F9D1),
-    PERSON_MEDIUM_DARK_SKIN_TONE_RED_HAIR(0x1F9D1),
-    PERSON_DARK_SKIN_TONE_RED_HAIR(0x1F9D1),
-    WOMAN_CURLY_HAIR(0x1F469),
-    WOMAN_LIGHT_SKIN_TONE_CURLY_HAIR(0x1F469),
-    WOMAN_MEDIUM_LIGHT_SKIN_TONE_CURLY_HAIR(0x1F469),
-    WOMAN_MEDIUM_SKIN_TONE_CURLY_HAIR(0x1F469),
-    WOMAN_MEDIUM_DARK_SKIN_TONE_CURLY_HAIR(0x1F469),
-    WOMAN_DARK_SKIN_TONE_CURLY_HAIR(0x1F469),
-    PERSON_CURLY_HAIR(0x1F9D1),
-    PERSON_LIGHT_SKIN_TONE_CURLY_HAIR(0x1F9D1),
-    PERSON_MEDIUM_LIGHT_SKIN_TONE_CURLY_HAIR(0x1F9D1),
-    PERSON_MEDIUM_SKIN_TONE_CURLY_HAIR(0x1F9D1),
-    PERSON_MEDIUM_DARK_SKIN_TONE_CURLY_HAIR(0x1F9D1),
-    PERSON_DARK_SKIN_TONE_CURLY_HAIR(0x1F9D1),
-    WOMAN_WHITE_HAIR(0x1F469),
-    WOMAN_LIGHT_SKIN_TONE_WHITE_HAIR(0x1F469),
-    WOMAN_MEDIUM_LIGHT_SKIN_TONE_WHITE_HAIR(0x1F469),
-    WOMAN_MEDIUM_SKIN_TONE_WHITE_HAIR(0x1F469),
-    WOMAN_MEDIUM_DARK_SKIN_TONE_WHITE_HAIR(0x1F469),
-    WOMAN_DARK_SKIN_TONE_WHITE_HAIR(0x1F469),
-    PERSON_WHITE_HAIR(0x1F9D1),
-    PERSON_LIGHT_SKIN_TONE_WHITE_HAIR(0x1F9D1),
-    PERSON_MEDIUM_LIGHT_SKIN_TONE_WHITE_HAIR(0x1F9D1),
-    PERSON_MEDIUM_SKIN_TONE_WHITE_HAIR(0x1F9D1),
-    PERSON_MEDIUM_DARK_SKIN_TONE_WHITE_HAIR(0x1F9D1),
-    PERSON_DARK_SKIN_TONE_WHITE_HAIR(0x1F9D1),
-    WOMAN_BALD(0x1F469),
-    WOMAN_LIGHT_SKIN_TONE_BALD(0x1F469),
-    WOMAN_MEDIUM_LIGHT_SKIN_TONE_BALD(0x1F469),
-    WOMAN_MEDIUM_SKIN_TONE_BALD(0x1F469),
-    WOMAN_MEDIUM_DARK_SKIN_TONE_BALD(0x1F469),
-    WOMAN_DARK_SKIN_TONE_BALD(0x1F469),
-    PERSON_BALD(0x1F9D1),
-    PERSON_LIGHT_SKIN_TONE_BALD(0x1F9D1),
-    PERSON_MEDIUM_LIGHT_SKIN_TONE_BALD(0x1F9D1),
-    PERSON_MEDIUM_SKIN_TONE_BALD(0x1F9D1),
-    PERSON_MEDIUM_DARK_SKIN_TONE_BALD(0x1F9D1),
-    PERSON_DARK_SKIN_TONE_BALD(0x1F9D1),
-    WOMAN_BLOND_HAIR(0x1F471),
-    WOMAN_LIGHT_SKIN_TONE_BLOND_HAIR(0x1F471),
-    WOMAN_MEDIUM_LIGHT_SKIN_TONE_BLOND_HAIR(0x1F471),
-    WOMAN_MEDIUM_SKIN_TONE_BLOND_HAIR(0x1F471),
-    WOMAN_MEDIUM_DARK_SKIN_TONE_BLOND_HAIR(0x1F471),
-    WOMAN_DARK_SKIN_TONE_BLOND_HAIR(0x1F471),
-    MAN_BLOND_HAIR(0x1F471),
-    MAN_LIGHT_SKIN_TONE_BLOND_HAIR(0x1F471),
-    MAN_MEDIUM_LIGHT_SKIN_TONE_BLOND_HAIR(0x1F471),
-    MAN_MEDIUM_SKIN_TONE_BLOND_HAIR(0x1F471),
-    MAN_MEDIUM_DARK_SKIN_TONE_BLOND_HAIR(0x1F471),
-    MAN_DARK_SKIN_TONE_BLOND_HAIR(0x1F471),
-    OLDER_PERSON(0x1F9D3),
-    OLDER_PERSON_LIGHT_SKIN_TONE(0x1F9D3),
-    OLDER_PERSON_MEDIUM_LIGHT_SKIN_TONE(0x1F9D3),
-    OLDER_PERSON_MEDIUM_SKIN_TONE(0x1F9D3),
-    OLDER_PERSON_MEDIUM_DARK_SKIN_TONE(0x1F9D3),
-    OLDER_PERSON_DARK_SKIN_TONE(0x1F9D3),
-    OLD_MAN(0x1F474),
-    OLD_MAN_LIGHT_SKIN_TONE(0x1F474),
-    OLD_MAN_MEDIUM_LIGHT_SKIN_TONE(0x1F474),
-    OLD_MAN_MEDIUM_SKIN_TONE(0x1F474),
-    OLD_MAN_MEDIUM_DARK_SKIN_TONE(0x1F474),
-    OLD_MAN_DARK_SKIN_TONE(0x1F474),
-    OLD_WOMAN(0x1F475),
-    OLD_WOMAN_LIGHT_SKIN_TONE(0x1F475),
-    OLD_WOMAN_MEDIUM_LIGHT_SKIN_TONE(0x1F475),
-    OLD_WOMAN_MEDIUM_SKIN_TONE(0x1F475),
-    OLD_WOMAN_MEDIUM_DARK_SKIN_TONE(0x1F475),
-    OLD_WOMAN_DARK_SKIN_TONE(0x1F475),
-    PERSON_FROWNING(0x1F64D),
-    PERSON_FROWNING_LIGHT_SKIN_TONE(0x1F64D),
-    PERSON_FROWNING_MEDIUM_LIGHT_SKIN_TONE(0x1F64D),
-    PERSON_FROWNING_MEDIUM_SKIN_TONE(0x1F64D),
-    PERSON_FROWNING_MEDIUM_DARK_SKIN_TONE(0x1F64D),
-    PERSON_FROWNING_DARK_SKIN_TONE(0x1F64D),
-    MAN_FROWNING(0x1F64D),
-    MAN_FROWNING_LIGHT_SKIN_TONE(0x1F64D),
-    MAN_FROWNING_MEDIUM_LIGHT_SKIN_TONE(0x1F64D),
-    MAN_FROWNING_MEDIUM_SKIN_TONE(0x1F64D),
-    MAN_FROWNING_MEDIUM_DARK_SKIN_TONE(0x1F64D),
-    MAN_FROWNING_DARK_SKIN_TONE(0x1F64D),
-    WOMAN_FROWNING(0x1F64D),
-    WOMAN_FROWNING_LIGHT_SKIN_TONE(0x1F64D),
-    WOMAN_FROWNING_MEDIUM_LIGHT_SKIN_TONE(0x1F64D),
-    WOMAN_FROWNING_MEDIUM_SKIN_TONE(0x1F64D),
-    WOMAN_FROWNING_MEDIUM_DARK_SKIN_TONE(0x1F64D),
-    WOMAN_FROWNING_DARK_SKIN_TONE(0x1F64D),
-    PERSON_POUTING(0x1F64E),
-    PERSON_POUTING_LIGHT_SKIN_TONE(0x1F64E),
-    PERSON_POUTING_MEDIUM_LIGHT_SKIN_TONE(0x1F64E),
-    PERSON_POUTING_MEDIUM_SKIN_TONE(0x1F64E),
-    PERSON_POUTING_MEDIUM_DARK_SKIN_TONE(0x1F64E),
-    PERSON_POUTING_DARK_SKIN_TONE(0x1F64E),
-    MAN_POUTING(0x1F64E),
-    MAN_POUTING_LIGHT_SKIN_TONE(0x1F64E),
-    MAN_POUTING_MEDIUM_LIGHT_SKIN_TONE(0x1F64E),
-    MAN_POUTING_MEDIUM_SKIN_TONE(0x1F64E),
-    MAN_POUTING_MEDIUM_DARK_SKIN_TONE(0x1F64E),
-    MAN_POUTING_DARK_SKIN_TONE(0x1F64E),
-    WOMAN_POUTING(0x1F64E),
-    WOMAN_POUTING_LIGHT_SKIN_TONE(0x1F64E),
-    WOMAN_POUTING_MEDIUM_LIGHT_SKIN_TONE(0x1F64E),
-    WOMAN_POUTING_MEDIUM_SKIN_TONE(0x1F64E),
-    WOMAN_POUTING_MEDIUM_DARK_SKIN_TONE(0x1F64E),
-    WOMAN_POUTING_DARK_SKIN_TONE(0x1F64E),
-    PERSON_GESTURING_NO(0x1F645),
-    PERSON_GESTURING_NO_LIGHT_SKIN_TONE(0x1F645),
-    PERSON_GESTURING_NO_MEDIUM_LIGHT_SKIN_TONE(0x1F645),
-    PERSON_GESTURING_NO_MEDIUM_SKIN_TONE(0x1F645),
-    PERSON_GESTURING_NO_MEDIUM_DARK_SKIN_TONE(0x1F645),
-    PERSON_GESTURING_NO_DARK_SKIN_TONE(0x1F645),
-    MAN_GESTURING_NO(0x1F645),
-    MAN_GESTURING_NO_LIGHT_SKIN_TONE(0x1F645),
-    MAN_GESTURING_NO_MEDIUM_LIGHT_SKIN_TONE(0x1F645),
-    MAN_GESTURING_NO_MEDIUM_SKIN_TONE(0x1F645),
-    MAN_GESTURING_NO_MEDIUM_DARK_SKIN_TONE(0x1F645),
-    MAN_GESTURING_NO_DARK_SKIN_TONE(0x1F645),
-    WOMAN_GESTURING_NO(0x1F645),
-    WOMAN_GESTURING_NO_LIGHT_SKIN_TONE(0x1F645),
-    WOMAN_GESTURING_NO_MEDIUM_LIGHT_SKIN_TONE(0x1F645),
-    WOMAN_GESTURING_NO_MEDIUM_SKIN_TONE(0x1F645),
-    WOMAN_GESTURING_NO_MEDIUM_DARK_SKIN_TONE(0x1F645),
-    WOMAN_GESTURING_NO_DARK_SKIN_TONE(0x1F645),
-    PERSON_GESTURING_OK(0x1F646),
-    PERSON_GESTURING_OK_LIGHT_SKIN_TONE(0x1F646),
-    PERSON_GESTURING_OK_MEDIUM_LIGHT_SKIN_TONE(0x1F646),
-    PERSON_GESTURING_OK_MEDIUM_SKIN_TONE(0x1F646),
-    PERSON_GESTURING_OK_MEDIUM_DARK_SKIN_TONE(0x1F646),
-    PERSON_GESTURING_OK_DARK_SKIN_TONE(0x1F646),
-    MAN_GESTURING_OK(0x1F646),
-    MAN_GESTURING_OK_LIGHT_SKIN_TONE(0x1F646),
-    MAN_GESTURING_OK_MEDIUM_LIGHT_SKIN_TONE(0x1F646),
-    MAN_GESTURING_OK_MEDIUM_SKIN_TONE(0x1F646),
-    MAN_GESTURING_OK_MEDIUM_DARK_SKIN_TONE(0x1F646),
-    MAN_GESTURING_OK_DARK_SKIN_TONE(0x1F646),
-    WOMAN_GESTURING_OK(0x1F646),
-    WOMAN_GESTURING_OK_LIGHT_SKIN_TONE(0x1F646),
-    WOMAN_GESTURING_OK_MEDIUM_LIGHT_SKIN_TONE(0x1F646),
-    WOMAN_GESTURING_OK_MEDIUM_SKIN_TONE(0x1F646),
-    WOMAN_GESTURING_OK_MEDIUM_DARK_SKIN_TONE(0x1F646),
-    WOMAN_GESTURING_OK_DARK_SKIN_TONE(0x1F646),
-    PERSON_TIPPING_HAND(0x1F481),
-    PERSON_TIPPING_HAND_LIGHT_SKIN_TONE(0x1F481),
-    PERSON_TIPPING_HAND_MEDIUM_LIGHT_SKIN_TONE(0x1F481),
-    PERSON_TIPPING_HAND_MEDIUM_SKIN_TONE(0x1F481),
-    PERSON_TIPPING_HAND_MEDIUM_DARK_SKIN_TONE(0x1F481),
-    PERSON_TIPPING_HAND_DARK_SKIN_TONE(0x1F481),
-    MAN_TIPPING_HAND(0x1F481),
-    MAN_TIPPING_HAND_LIGHT_SKIN_TONE(0x1F481),
-    MAN_TIPPING_HAND_MEDIUM_LIGHT_SKIN_TONE(0x1F481),
-    MAN_TIPPING_HAND_MEDIUM_SKIN_TONE(0x1F481),
-    MAN_TIPPING_HAND_MEDIUM_DARK_SKIN_TONE(0x1F481),
-    MAN_TIPPING_HAND_DARK_SKIN_TONE(0x1F481),
-    WOMAN_TIPPING_HAND(0x1F481),
-    WOMAN_TIPPING_HAND_LIGHT_SKIN_TONE(0x1F481),
-    WOMAN_TIPPING_HAND_MEDIUM_LIGHT_SKIN_TONE(0x1F481),
-    WOMAN_TIPPING_HAND_MEDIUM_SKIN_TONE(0x1F481),
-    WOMAN_TIPPING_HAND_MEDIUM_DARK_SKIN_TONE(0x1F481),
-    WOMAN_TIPPING_HAND_DARK_SKIN_TONE(0x1F481),
-    PERSON_RAISING_HAND(0x1F64B),
-    PERSON_RAISING_HAND_LIGHT_SKIN_TONE(0x1F64B),
-    PERSON_RAISING_HAND_MEDIUM_LIGHT_SKIN_TONE(0x1F64B),
-    PERSON_RAISING_HAND_MEDIUM_SKIN_TONE(0x1F64B),
-    PERSON_RAISING_HAND_MEDIUM_DARK_SKIN_TONE(0x1F64B),
-    PERSON_RAISING_HAND_DARK_SKIN_TONE(0x1F64B),
-    MAN_RAISING_HAND(0x1F64B),
-    MAN_RAISING_HAND_LIGHT_SKIN_TONE(0x1F64B),
-    MAN_RAISING_HAND_MEDIUM_LIGHT_SKIN_TONE(0x1F64B),
-    MAN_RAISING_HAND_MEDIUM_SKIN_TONE(0x1F64B),
-    MAN_RAISING_HAND_MEDIUM_DARK_SKIN_TONE(0x1F64B),
-    MAN_RAISING_HAND_DARK_SKIN_TONE(0x1F64B),
-    WOMAN_RAISING_HAND(0x1F64B),
-    WOMAN_RAISING_HAND_LIGHT_SKIN_TONE(0x1F64B),
-    WOMAN_RAISING_HAND_MEDIUM_LIGHT_SKIN_TONE(0x1F64B),
-    WOMAN_RAISING_HAND_MEDIUM_SKIN_TONE(0x1F64B),
-    WOMAN_RAISING_HAND_MEDIUM_DARK_SKIN_TONE(0x1F64B),
-    WOMAN_RAISING_HAND_DARK_SKIN_TONE(0x1F64B),
-    DEAF_PERSON(0x1F9CF),
-    DEAF_PERSON_LIGHT_SKIN_TONE(0x1F9CF),
-    DEAF_PERSON_MEDIUM_LIGHT_SKIN_TONE(0x1F9CF),
-    DEAF_PERSON_MEDIUM_SKIN_TONE(0x1F9CF),
-    DEAF_PERSON_MEDIUM_DARK_SKIN_TONE(0x1F9CF),
-    DEAF_PERSON_DARK_SKIN_TONE(0x1F9CF),
-    DEAF_MAN(0x1F9CF),
-    DEAF_MAN_LIGHT_SKIN_TONE(0x1F9CF),
-    DEAF_MAN_MEDIUM_LIGHT_SKIN_TONE(0x1F9CF),
-    DEAF_MAN_MEDIUM_SKIN_TONE(0x1F9CF),
-    DEAF_MAN_MEDIUM_DARK_SKIN_TONE(0x1F9CF),
-    DEAF_MAN_DARK_SKIN_TONE(0x1F9CF),
-    DEAF_WOMAN(0x1F9CF),
-    DEAF_WOMAN_LIGHT_SKIN_TONE(0x1F9CF),
-    DEAF_WOMAN_MEDIUM_LIGHT_SKIN_TONE(0x1F9CF),
-    DEAF_WOMAN_MEDIUM_SKIN_TONE(0x1F9CF),
-    DEAF_WOMAN_MEDIUM_DARK_SKIN_TONE(0x1F9CF),
-    DEAF_WOMAN_DARK_SKIN_TONE(0x1F9CF),
-    PERSON_BOWING(0x1F647),
-    PERSON_BOWING_LIGHT_SKIN_TONE(0x1F647),
-    PERSON_BOWING_MEDIUM_LIGHT_SKIN_TONE(0x1F647),
-    PERSON_BOWING_MEDIUM_SKIN_TONE(0x1F647),
-    PERSON_BOWING_MEDIUM_DARK_SKIN_TONE(0x1F647),
-    PERSON_BOWING_DARK_SKIN_TONE(0x1F647),
-    MAN_BOWING(0x1F647),
-    MAN_BOWING_LIGHT_SKIN_TONE(0x1F647),
-    MAN_BOWING_MEDIUM_LIGHT_SKIN_TONE(0x1F647),
-    MAN_BOWING_MEDIUM_SKIN_TONE(0x1F647),
-    MAN_BOWING_MEDIUM_DARK_SKIN_TONE(0x1F647),
-    MAN_BOWING_DARK_SKIN_TONE(0x1F647),
-    WOMAN_BOWING(0x1F647),
-    WOMAN_BOWING_LIGHT_SKIN_TONE(0x1F647),
-    WOMAN_BOWING_MEDIUM_LIGHT_SKIN_TONE(0x1F647),
-    WOMAN_BOWING_MEDIUM_SKIN_TONE(0x1F647),
-    WOMAN_BOWING_MEDIUM_DARK_SKIN_TONE(0x1F647),
-    WOMAN_BOWING_DARK_SKIN_TONE(0x1F647),
-    PERSON_FACEPALMING(0x1F926),
-    PERSON_FACEPALMING_LIGHT_SKIN_TONE(0x1F926),
-    PERSON_FACEPALMING_MEDIUM_LIGHT_SKIN_TONE(0x1F926),
-    PERSON_FACEPALMING_MEDIUM_SKIN_TONE(0x1F926),
-    PERSON_FACEPALMING_MEDIUM_DARK_SKIN_TONE(0x1F926),
-    PERSON_FACEPALMING_DARK_SKIN_TONE(0x1F926),
-    MAN_FACEPALMING(0x1F926),
-    MAN_FACEPALMING_LIGHT_SKIN_TONE(0x1F926),
-    MAN_FACEPALMING_MEDIUM_LIGHT_SKIN_TONE(0x1F926),
-    MAN_FACEPALMING_MEDIUM_SKIN_TONE(0x1F926),
-    MAN_FACEPALMING_MEDIUM_DARK_SKIN_TONE(0x1F926),
-    MAN_FACEPALMING_DARK_SKIN_TONE(0x1F926),
-    WOMAN_FACEPALMING(0x1F926),
-    WOMAN_FACEPALMING_LIGHT_SKIN_TONE(0x1F926),
-    WOMAN_FACEPALMING_MEDIUM_LIGHT_SKIN_TONE(0x1F926),
-    WOMAN_FACEPALMING_MEDIUM_SKIN_TONE(0x1F926),
-    WOMAN_FACEPALMING_MEDIUM_DARK_SKIN_TONE(0x1F926),
-    WOMAN_FACEPALMING_DARK_SKIN_TONE(0x1F926),
-    PERSON_SHRUGGING(0x1F937),
-    PERSON_SHRUGGING_LIGHT_SKIN_TONE(0x1F937),
-    PERSON_SHRUGGING_MEDIUM_LIGHT_SKIN_TONE(0x1F937),
-    PERSON_SHRUGGING_MEDIUM_SKIN_TONE(0x1F937),
-    PERSON_SHRUGGING_MEDIUM_DARK_SKIN_TONE(0x1F937),
-    PERSON_SHRUGGING_DARK_SKIN_TONE(0x1F937),
-    MAN_SHRUGGING(0x1F937),
-    MAN_SHRUGGING_LIGHT_SKIN_TONE(0x1F937),
-    MAN_SHRUGGING_MEDIUM_LIGHT_SKIN_TONE(0x1F937),
-    MAN_SHRUGGING_MEDIUM_SKIN_TONE(0x1F937),
-    MAN_SHRUGGING_MEDIUM_DARK_SKIN_TONE(0x1F937),
-    MAN_SHRUGGING_DARK_SKIN_TONE(0x1F937),
-    WOMAN_SHRUGGING(0x1F937),
-    WOMAN_SHRUGGING_LIGHT_SKIN_TONE(0x1F937),
-    WOMAN_SHRUGGING_MEDIUM_LIGHT_SKIN_TONE(0x1F937),
-    WOMAN_SHRUGGING_MEDIUM_SKIN_TONE(0x1F937),
-    WOMAN_SHRUGGING_MEDIUM_DARK_SKIN_TONE(0x1F937),
-    WOMAN_SHRUGGING_DARK_SKIN_TONE(0x1F937),
-    HEALTH_WORKER(0x1F9D1),
-    HEALTH_WORKER_LIGHT_SKIN_TONE(0x1F9D1),
-    HEALTH_WORKER_MEDIUM_LIGHT_SKIN_TONE(0x1F9D1),
-    HEALTH_WORKER_MEDIUM_SKIN_TONE(0x1F9D1),
-    HEALTH_WORKER_MEDIUM_DARK_SKIN_TONE(0x1F9D1),
-    HEALTH_WORKER_DARK_SKIN_TONE(0x1F9D1),
-    MAN_HEALTH_WORKER(0x1F468),
-    MAN_HEALTH_WORKER_LIGHT_SKIN_TONE(0x1F468),
-    MAN_HEALTH_WORKER_MEDIUM_LIGHT_SKIN_TONE(0x1F468),
-    MAN_HEALTH_WORKER_MEDIUM_SKIN_TONE(0x1F468),
-    MAN_HEALTH_WORKER_MEDIUM_DARK_SKIN_TONE(0x1F468),
-    MAN_HEALTH_WORKER_DARK_SKIN_TONE(0x1F468),
-    WOMAN_HEALTH_WORKER(0x1F469),
-    WOMAN_HEALTH_WORKER_LIGHT_SKIN_TONE(0x1F469),
-    WOMAN_HEALTH_WORKER_MEDIUM_LIGHT_SKIN_TONE(0x1F469),
-    WOMAN_HEALTH_WORKER_MEDIUM_SKIN_TONE(0x1F469),
-    WOMAN_HEALTH_WORKER_MEDIUM_DARK_SKIN_TONE(0x1F469),
-    WOMAN_HEALTH_WORKER_DARK_SKIN_TONE(0x1F469),
-    STUDENT(0x1F9D1),
-    STUDENT_LIGHT_SKIN_TONE(0x1F9D1),
-    STUDENT_MEDIUM_LIGHT_SKIN_TONE(0x1F9D1),
-    STUDENT_MEDIUM_SKIN_TONE(0x1F9D1),
-    STUDENT_MEDIUM_DARK_SKIN_TONE(0x1F9D1),
-    STUDENT_DARK_SKIN_TONE(0x1F9D1),
-    MAN_STUDENT(0x1F468),
-    MAN_STUDENT_LIGHT_SKIN_TONE(0x1F468),
-    MAN_STUDENT_MEDIUM_LIGHT_SKIN_TONE(0x1F468),
-    MAN_STUDENT_MEDIUM_SKIN_TONE(0x1F468),
-    MAN_STUDENT_MEDIUM_DARK_SKIN_TONE(0x1F468),
-    MAN_STUDENT_DARK_SKIN_TONE(0x1F468),
-    WOMAN_STUDENT(0x1F469),
-    WOMAN_STUDENT_LIGHT_SKIN_TONE(0x1F469),
-    WOMAN_STUDENT_MEDIUM_LIGHT_SKIN_TONE(0x1F469),
-    WOMAN_STUDENT_MEDIUM_SKIN_TONE(0x1F469),
-    WOMAN_STUDENT_MEDIUM_DARK_SKIN_TONE(0x1F469),
-    WOMAN_STUDENT_DARK_SKIN_TONE(0x1F469),
-    TEACHER(0x1F9D1),
-    TEACHER_LIGHT_SKIN_TONE(0x1F9D1),
-    TEACHER_MEDIUM_LIGHT_SKIN_TONE(0x1F9D1),
-    TEACHER_MEDIUM_SKIN_TONE(0x1F9D1),
-    TEACHER_MEDIUM_DARK_SKIN_TONE(0x1F9D1),
-    TEACHER_DARK_SKIN_TONE(0x1F9D1),
-    MAN_TEACHER(0x1F468),
-    MAN_TEACHER_LIGHT_SKIN_TONE(0x1F468),
-    MAN_TEACHER_MEDIUM_LIGHT_SKIN_TONE(0x1F468),
-    MAN_TEACHER_MEDIUM_SKIN_TONE(0x1F468),
-    MAN_TEACHER_MEDIUM_DARK_SKIN_TONE(0x1F468),
-    MAN_TEACHER_DARK_SKIN_TONE(0x1F468),
-    WOMAN_TEACHER(0x1F469),
-    WOMAN_TEACHER_LIGHT_SKIN_TONE(0x1F469),
-    WOMAN_TEACHER_MEDIUM_LIGHT_SKIN_TONE(0x1F469),
-    WOMAN_TEACHER_MEDIUM_SKIN_TONE(0x1F469),
-    WOMAN_TEACHER_MEDIUM_DARK_SKIN_TONE(0x1F469),
-    WOMAN_TEACHER_DARK_SKIN_TONE(0x1F469),
-    JUDGE(0x1F9D1),
-    JUDGE_LIGHT_SKIN_TONE(0x1F9D1),
-    JUDGE_MEDIUM_LIGHT_SKIN_TONE(0x1F9D1),
-    JUDGE_MEDIUM_SKIN_TONE(0x1F9D1),
-    JUDGE_MEDIUM_DARK_SKIN_TONE(0x1F9D1),
-    JUDGE_DARK_SKIN_TONE(0x1F9D1),
-    MAN_JUDGE(0x1F468),
-    MAN_JUDGE_LIGHT_SKIN_TONE(0x1F468),
-    MAN_JUDGE_MEDIUM_LIGHT_SKIN_TONE(0x1F468),
-    MAN_JUDGE_MEDIUM_SKIN_TONE(0x1F468),
-    MAN_JUDGE_MEDIUM_DARK_SKIN_TONE(0x1F468),
-    MAN_JUDGE_DARK_SKIN_TONE(0x1F468),
-    WOMAN_JUDGE(0x1F469),
-    WOMAN_JUDGE_LIGHT_SKIN_TONE(0x1F469),
-    WOMAN_JUDGE_MEDIUM_LIGHT_SKIN_TONE(0x1F469),
-    WOMAN_JUDGE_MEDIUM_SKIN_TONE(0x1F469),
-    WOMAN_JUDGE_MEDIUM_DARK_SKIN_TONE(0x1F469),
-    WOMAN_JUDGE_DARK_SKIN_TONE(0x1F469),
-    FARMER(0x1F9D1),
-    FARMER_LIGHT_SKIN_TONE(0x1F9D1),
-    FARMER_MEDIUM_LIGHT_SKIN_TONE(0x1F9D1),
-    FARMER_MEDIUM_SKIN_TONE(0x1F9D1),
-    FARMER_MEDIUM_DARK_SKIN_TONE(0x1F9D1),
-    FARMER_DARK_SKIN_TONE(0x1F9D1),
-    MAN_FARMER(0x1F468),
-    MAN_FARMER_LIGHT_SKIN_TONE(0x1F468),
-    MAN_FARMER_MEDIUM_LIGHT_SKIN_TONE(0x1F468),
-    MAN_FARMER_MEDIUM_SKIN_TONE(0x1F468),
-    MAN_FARMER_MEDIUM_DARK_SKIN_TONE(0x1F468),
-    MAN_FARMER_DARK_SKIN_TONE(0x1F468),
-    WOMAN_FARMER(0x1F469),
-    WOMAN_FARMER_LIGHT_SKIN_TONE(0x1F469),
-    WOMAN_FARMER_MEDIUM_LIGHT_SKIN_TONE(0x1F469),
-    WOMAN_FARMER_MEDIUM_SKIN_TONE(0x1F469),
-    WOMAN_FARMER_MEDIUM_DARK_SKIN_TONE(0x1F469),
-    WOMAN_FARMER_DARK_SKIN_TONE(0x1F469),
-    COOK(0x1F9D1),
-    COOK_LIGHT_SKIN_TONE(0x1F9D1),
-    COOK_MEDIUM_LIGHT_SKIN_TONE(0x1F9D1),
-    COOK_MEDIUM_SKIN_TONE(0x1F9D1),
-    COOK_MEDIUM_DARK_SKIN_TONE(0x1F9D1),
-    COOK_DARK_SKIN_TONE(0x1F9D1),
-    MAN_COOK(0x1F468),
-    MAN_COOK_LIGHT_SKIN_TONE(0x1F468),
-    MAN_COOK_MEDIUM_LIGHT_SKIN_TONE(0x1F468),
-    MAN_COOK_MEDIUM_SKIN_TONE(0x1F468),
-    MAN_COOK_MEDIUM_DARK_SKIN_TONE(0x1F468),
-    MAN_COOK_DARK_SKIN_TONE(0x1F468),
-    WOMAN_COOK(0x1F469),
-    WOMAN_COOK_LIGHT_SKIN_TONE(0x1F469),
-    WOMAN_COOK_MEDIUM_LIGHT_SKIN_TONE(0x1F469),
-    WOMAN_COOK_MEDIUM_SKIN_TONE(0x1F469),
-    WOMAN_COOK_MEDIUM_DARK_SKIN_TONE(0x1F469),
-    WOMAN_COOK_DARK_SKIN_TONE(0x1F469),
-    MECHANIC(0x1F9D1),
-    MECHANIC_LIGHT_SKIN_TONE(0x1F9D1),
-    MECHANIC_MEDIUM_LIGHT_SKIN_TONE(0x1F9D1),
-    MECHANIC_MEDIUM_SKIN_TONE(0x1F9D1),
-    MECHANIC_MEDIUM_DARK_SKIN_TONE(0x1F9D1),
-    MECHANIC_DARK_SKIN_TONE(0x1F9D1),
-    MAN_MECHANIC(0x1F468),
-    MAN_MECHANIC_LIGHT_SKIN_TONE(0x1F468),
-    MAN_MECHANIC_MEDIUM_LIGHT_SKIN_TONE(0x1F468),
-    MAN_MECHANIC_MEDIUM_SKIN_TONE(0x1F468),
-    MAN_MECHANIC_MEDIUM_DARK_SKIN_TONE(0x1F468),
-    MAN_MECHANIC_DARK_SKIN_TONE(0x1F468),
-    WOMAN_MECHANIC(0x1F469),
-    WOMAN_MECHANIC_LIGHT_SKIN_TONE(0x1F469),
-    WOMAN_MECHANIC_MEDIUM_LIGHT_SKIN_TONE(0x1F469),
-    WOMAN_MECHANIC_MEDIUM_SKIN_TONE(0x1F469),
-    WOMAN_MECHANIC_MEDIUM_DARK_SKIN_TONE(0x1F469),
-    WOMAN_MECHANIC_DARK_SKIN_TONE(0x1F469),
-    FACTORY_WORKER(0x1F9D1),
-    FACTORY_WORKER_LIGHT_SKIN_TONE(0x1F9D1),
-    FACTORY_WORKER_MEDIUM_LIGHT_SKIN_TONE(0x1F9D1),
-    FACTORY_WORKER_MEDIUM_SKIN_TONE(0x1F9D1),
-    FACTORY_WORKER_MEDIUM_DARK_SKIN_TONE(0x1F9D1),
-    FACTORY_WORKER_DARK_SKIN_TONE(0x1F9D1),
-    MAN_FACTORY_WORKER(0x1F468),
-    MAN_FACTORY_WORKER_LIGHT_SKIN_TONE(0x1F468),
-    MAN_FACTORY_WORKER_MEDIUM_LIGHT_SKIN_TONE(0x1F468),
-    MAN_FACTORY_WORKER_MEDIUM_SKIN_TONE(0x1F468),
-    MAN_FACTORY_WORKER_MEDIUM_DARK_SKIN_TONE(0x1F468),
-    MAN_FACTORY_WORKER_DARK_SKIN_TONE(0x1F468),
-    WOMAN_FACTORY_WORKER(0x1F469),
-    WOMAN_FACTORY_WORKER_LIGHT_SKIN_TONE(0x1F469),
-    WOMAN_FACTORY_WORKER_MEDIUM_LIGHT_SKIN_TONE(0x1F469),
-    WOMAN_FACTORY_WORKER_MEDIUM_SKIN_TONE(0x1F469),
-    WOMAN_FACTORY_WORKER_MEDIUM_DARK_SKIN_TONE(0x1F469),
-    WOMAN_FACTORY_WORKER_DARK_SKIN_TONE(0x1F469),
-    OFFICE_WORKER(0x1F9D1),
-    OFFICE_WORKER_LIGHT_SKIN_TONE(0x1F9D1),
-    OFFICE_WORKER_MEDIUM_LIGHT_SKIN_TONE(0x1F9D1),
-    OFFICE_WORKER_MEDIUM_SKIN_TONE(0x1F9D1),
-    OFFICE_WORKER_MEDIUM_DARK_SKIN_TONE(0x1F9D1),
-    OFFICE_WORKER_DARK_SKIN_TONE(0x1F9D1),
-    MAN_OFFICE_WORKER(0x1F468),
-    MAN_OFFICE_WORKER_LIGHT_SKIN_TONE(0x1F468),
-    MAN_OFFICE_WORKER_MEDIUM_LIGHT_SKIN_TONE(0x1F468),
-    MAN_OFFICE_WORKER_MEDIUM_SKIN_TONE(0x1F468),
-    MAN_OFFICE_WORKER_MEDIUM_DARK_SKIN_TONE(0x1F468),
-    MAN_OFFICE_WORKER_DARK_SKIN_TONE(0x1F468),
-    WOMAN_OFFICE_WORKER(0x1F469),
-    WOMAN_OFFICE_WORKER_LIGHT_SKIN_TONE(0x1F469),
-    WOMAN_OFFICE_WORKER_MEDIUM_LIGHT_SKIN_TONE(0x1F469),
-    WOMAN_OFFICE_WORKER_MEDIUM_SKIN_TONE(0x1F469),
-    WOMAN_OFFICE_WORKER_MEDIUM_DARK_SKIN_TONE(0x1F469),
-    WOMAN_OFFICE_WORKER_DARK_SKIN_TONE(0x1F469),
-    SCIENTIST(0x1F9D1),
-    SCIENTIST_LIGHT_SKIN_TONE(0x1F9D1),
-    SCIENTIST_MEDIUM_LIGHT_SKIN_TONE(0x1F9D1),
-    SCIENTIST_MEDIUM_SKIN_TONE(0x1F9D1),
-    SCIENTIST_MEDIUM_DARK_SKIN_TONE(0x1F9D1),
-    SCIENTIST_DARK_SKIN_TONE(0x1F9D1),
-    MAN_SCIENTIST(0x1F468),
-    MAN_SCIENTIST_LIGHT_SKIN_TONE(0x1F468),
-    MAN_SCIENTIST_MEDIUM_LIGHT_SKIN_TONE(0x1F468),
-    MAN_SCIENTIST_MEDIUM_SKIN_TONE(0x1F468),
-    MAN_SCIENTIST_MEDIUM_DARK_SKIN_TONE(0x1F468),
-    MAN_SCIENTIST_DARK_SKIN_TONE(0x1F468),
-    WOMAN_SCIENTIST(0x1F469),
-    WOMAN_SCIENTIST_LIGHT_SKIN_TONE(0x1F469),
-    WOMAN_SCIENTIST_MEDIUM_LIGHT_SKIN_TONE(0x1F469),
-    WOMAN_SCIENTIST_MEDIUM_SKIN_TONE(0x1F469),
-    WOMAN_SCIENTIST_MEDIUM_DARK_SKIN_TONE(0x1F469),
-    WOMAN_SCIENTIST_DARK_SKIN_TONE(0x1F469),
-    TECHNOLOGIST(0x1F9D1),
-    TECHNOLOGIST_LIGHT_SKIN_TONE(0x1F9D1),
-    TECHNOLOGIST_MEDIUM_LIGHT_SKIN_TONE(0x1F9D1),
-    TECHNOLOGIST_MEDIUM_SKIN_TONE(0x1F9D1),
-    TECHNOLOGIST_MEDIUM_DARK_SKIN_TONE(0x1F9D1),
-    TECHNOLOGIST_DARK_SKIN_TONE(0x1F9D1),
-    MAN_TECHNOLOGIST(0x1F468),
-    MAN_TECHNOLOGIST_LIGHT_SKIN_TONE(0x1F468),
-    MAN_TECHNOLOGIST_MEDIUM_LIGHT_SKIN_TONE(0x1F468),
-    MAN_TECHNOLOGIST_MEDIUM_SKIN_TONE(0x1F468),
-    MAN_TECHNOLOGIST_MEDIUM_DARK_SKIN_TONE(0x1F468),
-    MAN_TECHNOLOGIST_DARK_SKIN_TONE(0x1F468),
-    WOMAN_TECHNOLOGIST(0x1F469),
-    WOMAN_TECHNOLOGIST_LIGHT_SKIN_TONE(0x1F469),
-    WOMAN_TECHNOLOGIST_MEDIUM_LIGHT_SKIN_TONE(0x1F469),
-    WOMAN_TECHNOLOGIST_MEDIUM_SKIN_TONE(0x1F469),
-    WOMAN_TECHNOLOGIST_MEDIUM_DARK_SKIN_TONE(0x1F469),
-    WOMAN_TECHNOLOGIST_DARK_SKIN_TONE(0x1F469),
-    SINGER(0x1F9D1),
-    SINGER_LIGHT_SKIN_TONE(0x1F9D1),
-    SINGER_MEDIUM_LIGHT_SKIN_TONE(0x1F9D1),
-    SINGER_MEDIUM_SKIN_TONE(0x1F9D1),
-    SINGER_MEDIUM_DARK_SKIN_TONE(0x1F9D1),
-    SINGER_DARK_SKIN_TONE(0x1F9D1),
-    MAN_SINGER(0x1F468),
-    MAN_SINGER_LIGHT_SKIN_TONE(0x1F468),
-    MAN_SINGER_MEDIUM_LIGHT_SKIN_TONE(0x1F468),
-    MAN_SINGER_MEDIUM_SKIN_TONE(0x1F468),
-    MAN_SINGER_MEDIUM_DARK_SKIN_TONE(0x1F468),
-    MAN_SINGER_DARK_SKIN_TONE(0x1F468),
-    WOMAN_SINGER(0x1F469),
-    WOMAN_SINGER_LIGHT_SKIN_TONE(0x1F469),
-    WOMAN_SINGER_MEDIUM_LIGHT_SKIN_TONE(0x1F469),
-    WOMAN_SINGER_MEDIUM_SKIN_TONE(0x1F469),
-    WOMAN_SINGER_MEDIUM_DARK_SKIN_TONE(0x1F469),
-    WOMAN_SINGER_DARK_SKIN_TONE(0x1F469),
-    ARTIST(0x1F9D1),
-    ARTIST_LIGHT_SKIN_TONE(0x1F9D1),
-    ARTIST_MEDIUM_LIGHT_SKIN_TONE(0x1F9D1),
-    ARTIST_MEDIUM_SKIN_TONE(0x1F9D1),
-    ARTIST_MEDIUM_DARK_SKIN_TONE(0x1F9D1),
-    ARTIST_DARK_SKIN_TONE(0x1F9D1),
-    MAN_ARTIST(0x1F468),
-    MAN_ARTIST_LIGHT_SKIN_TONE(0x1F468),
-    MAN_ARTIST_MEDIUM_LIGHT_SKIN_TONE(0x1F468),
-    MAN_ARTIST_MEDIUM_SKIN_TONE(0x1F468),
-    MAN_ARTIST_MEDIUM_DARK_SKIN_TONE(0x1F468),
-    MAN_ARTIST_DARK_SKIN_TONE(0x1F468),
-    WOMAN_ARTIST(0x1F469),
-    WOMAN_ARTIST_LIGHT_SKIN_TONE(0x1F469),
-    WOMAN_ARTIST_MEDIUM_LIGHT_SKIN_TONE(0x1F469),
-    WOMAN_ARTIST_MEDIUM_SKIN_TONE(0x1F469),
-    WOMAN_ARTIST_MEDIUM_DARK_SKIN_TONE(0x1F469),
-    WOMAN_ARTIST_DARK_SKIN_TONE(0x1F469),
-    PILOT(0x1F9D1),
-    PILOT_LIGHT_SKIN_TONE(0x1F9D1),
-    PILOT_MEDIUM_LIGHT_SKIN_TONE(0x1F9D1),
-    PILOT_MEDIUM_SKIN_TONE(0x1F9D1),
-    PILOT_MEDIUM_DARK_SKIN_TONE(0x1F9D1),
-    PILOT_DARK_SKIN_TONE(0x1F9D1),
-    MAN_PILOT(0x1F468),
-    MAN_PILOT_LIGHT_SKIN_TONE(0x1F468),
-    MAN_PILOT_MEDIUM_LIGHT_SKIN_TONE(0x1F468),
-    MAN_PILOT_MEDIUM_SKIN_TONE(0x1F468),
-    MAN_PILOT_MEDIUM_DARK_SKIN_TONE(0x1F468),
-    MAN_PILOT_DARK_SKIN_TONE(0x1F468),
-    WOMAN_PILOT(0x1F469),
-    WOMAN_PILOT_LIGHT_SKIN_TONE(0x1F469),
-    WOMAN_PILOT_MEDIUM_LIGHT_SKIN_TONE(0x1F469),
-    WOMAN_PILOT_MEDIUM_SKIN_TONE(0x1F469),
-    WOMAN_PILOT_MEDIUM_DARK_SKIN_TONE(0x1F469),
-    WOMAN_PILOT_DARK_SKIN_TONE(0x1F469),
-    ASTRONAUT(0x1F9D1),
-    ASTRONAUT_LIGHT_SKIN_TONE(0x1F9D1),
-    ASTRONAUT_MEDIUM_LIGHT_SKIN_TONE(0x1F9D1),
-    ASTRONAUT_MEDIUM_SKIN_TONE(0x1F9D1),
-    ASTRONAUT_MEDIUM_DARK_SKIN_TONE(0x1F9D1),
-    ASTRONAUT_DARK_SKIN_TONE(0x1F9D1),
-    MAN_ASTRONAUT(0x1F468),
-    MAN_ASTRONAUT_LIGHT_SKIN_TONE(0x1F468),
-    MAN_ASTRONAUT_MEDIUM_LIGHT_SKIN_TONE(0x1F468),
-    MAN_ASTRONAUT_MEDIUM_SKIN_TONE(0x1F468),
-    MAN_ASTRONAUT_MEDIUM_DARK_SKIN_TONE(0x1F468),
-    MAN_ASTRONAUT_DARK_SKIN_TONE(0x1F468),
-    WOMAN_ASTRONAUT(0x1F469),
-    WOMAN_ASTRONAUT_LIGHT_SKIN_TONE(0x1F469),
-    WOMAN_ASTRONAUT_MEDIUM_LIGHT_SKIN_TONE(0x1F469),
-    WOMAN_ASTRONAUT_MEDIUM_SKIN_TONE(0x1F469),
-    WOMAN_ASTRONAUT_MEDIUM_DARK_SKIN_TONE(0x1F469),
-    WOMAN_ASTRONAUT_DARK_SKIN_TONE(0x1F469),
-    FIREFIGHTER(0x1F9D1),
-    FIREFIGHTER_LIGHT_SKIN_TONE(0x1F9D1),
-    FIREFIGHTER_MEDIUM_LIGHT_SKIN_TONE(0x1F9D1),
-    FIREFIGHTER_MEDIUM_SKIN_TONE(0x1F9D1),
-    FIREFIGHTER_MEDIUM_DARK_SKIN_TONE(0x1F9D1),
-    FIREFIGHTER_DARK_SKIN_TONE(0x1F9D1),
-    MAN_FIREFIGHTER(0x1F468),
-    MAN_FIREFIGHTER_LIGHT_SKIN_TONE(0x1F468),
-    MAN_FIREFIGHTER_MEDIUM_LIGHT_SKIN_TONE(0x1F468),
-    MAN_FIREFIGHTER_MEDIUM_SKIN_TONE(0x1F468),
-    MAN_FIREFIGHTER_MEDIUM_DARK_SKIN_TONE(0x1F468),
-    MAN_FIREFIGHTER_DARK_SKIN_TONE(0x1F468),
-    WOMAN_FIREFIGHTER(0x1F469),
-    WOMAN_FIREFIGHTER_LIGHT_SKIN_TONE(0x1F469),
-    WOMAN_FIREFIGHTER_MEDIUM_LIGHT_SKIN_TONE(0x1F469),
-    WOMAN_FIREFIGHTER_MEDIUM_SKIN_TONE(0x1F469),
-    WOMAN_FIREFIGHTER_MEDIUM_DARK_SKIN_TONE(0x1F469),
-    WOMAN_FIREFIGHTER_DARK_SKIN_TONE(0x1F469),
-    POLICE_OFFICER(0x1F46E),
-    POLICE_OFFICER_LIGHT_SKIN_TONE(0x1F46E),
-    POLICE_OFFICER_MEDIUM_LIGHT_SKIN_TONE(0x1F46E),
-    POLICE_OFFICER_MEDIUM_SKIN_TONE(0x1F46E),
-    POLICE_OFFICER_MEDIUM_DARK_SKIN_TONE(0x1F46E),
-    POLICE_OFFICER_DARK_SKIN_TONE(0x1F46E),
-    MAN_POLICE_OFFICER(0x1F46E),
-    MAN_POLICE_OFFICER_LIGHT_SKIN_TONE(0x1F46E),
-    MAN_POLICE_OFFICER_MEDIUM_LIGHT_SKIN_TONE(0x1F46E),
-    MAN_POLICE_OFFICER_MEDIUM_SKIN_TONE(0x1F46E),
-    MAN_POLICE_OFFICER_MEDIUM_DARK_SKIN_TONE(0x1F46E),
-    MAN_POLICE_OFFICER_DARK_SKIN_TONE(0x1F46E),
-    WOMAN_POLICE_OFFICER(0x1F46E),
-    WOMAN_POLICE_OFFICER_LIGHT_SKIN_TONE(0x1F46E),
-    WOMAN_POLICE_OFFICER_MEDIUM_LIGHT_SKIN_TONE(0x1F46E),
-    WOMAN_POLICE_OFFICER_MEDIUM_SKIN_TONE(0x1F46E),
-    WOMAN_POLICE_OFFICER_MEDIUM_DARK_SKIN_TONE(0x1F46E),
-    WOMAN_POLICE_OFFICER_DARK_SKIN_TONE(0x1F46E),
-    DETECTIVE(0x1F575),
-    DETECTIVE_LIGHT_SKIN_TONE(0x1F575),
-    DETECTIVE_MEDIUM_LIGHT_SKIN_TONE(0x1F575),
-    DETECTIVE_MEDIUM_SKIN_TONE(0x1F575),
-    DETECTIVE_MEDIUM_DARK_SKIN_TONE(0x1F575),
-    DETECTIVE_DARK_SKIN_TONE(0x1F575),
-    MAN_DETECTIVE(0x1F575),
-    MAN_DETECTIVE_LIGHT_SKIN_TONE(0x1F575),
-    MAN_DETECTIVE_MEDIUM_LIGHT_SKIN_TONE(0x1F575),
-    MAN_DETECTIVE_MEDIUM_SKIN_TONE(0x1F575),
-    MAN_DETECTIVE_MEDIUM_DARK_SKIN_TONE(0x1F575),
-    MAN_DETECTIVE_DARK_SKIN_TONE(0x1F575),
-    WOMAN_DETECTIVE(0x1F575),
-    WOMAN_DETECTIVE_LIGHT_SKIN_TONE(0x1F575),
-    WOMAN_DETECTIVE_MEDIUM_LIGHT_SKIN_TONE(0x1F575),
-    WOMAN_DETECTIVE_MEDIUM_SKIN_TONE(0x1F575),
-    WOMAN_DETECTIVE_MEDIUM_DARK_SKIN_TONE(0x1F575),
-    WOMAN_DETECTIVE_DARK_SKIN_TONE(0x1F575),
-    GUARD(0x1F482),
-    GUARD_LIGHT_SKIN_TONE(0x1F482),
-    GUARD_MEDIUM_LIGHT_SKIN_TONE(0x1F482),
-    GUARD_MEDIUM_SKIN_TONE(0x1F482),
-    GUARD_MEDIUM_DARK_SKIN_TONE(0x1F482),
-    GUARD_DARK_SKIN_TONE(0x1F482),
-    MAN_GUARD(0x1F482),
-    MAN_GUARD_LIGHT_SKIN_TONE(0x1F482),
-    MAN_GUARD_MEDIUM_LIGHT_SKIN_TONE(0x1F482),
-    MAN_GUARD_MEDIUM_SKIN_TONE(0x1F482),
-    MAN_GUARD_MEDIUM_DARK_SKIN_TONE(0x1F482),
-    MAN_GUARD_DARK_SKIN_TONE(0x1F482),
-    WOMAN_GUARD(0x1F482),
-    WOMAN_GUARD_LIGHT_SKIN_TONE(0x1F482),
-    WOMAN_GUARD_MEDIUM_LIGHT_SKIN_TONE(0x1F482),
-    WOMAN_GUARD_MEDIUM_SKIN_TONE(0x1F482),
-    WOMAN_GUARD_MEDIUM_DARK_SKIN_TONE(0x1F482),
-    WOMAN_GUARD_DARK_SKIN_TONE(0x1F482),
-    NINJA(0x1F977),
-    NINJA_LIGHT_SKIN_TONE(0x1F977),
-    NINJA_MEDIUM_LIGHT_SKIN_TONE(0x1F977),
-    NINJA_MEDIUM_SKIN_TONE(0x1F977),
-    NINJA_MEDIUM_DARK_SKIN_TONE(0x1F977),
-    NINJA_DARK_SKIN_TONE(0x1F977),
-    CONSTRUCTION_WORKER(0x1F477),
-    CONSTRUCTION_WORKER_LIGHT_SKIN_TONE(0x1F477),
-    CONSTRUCTION_WORKER_MEDIUM_LIGHT_SKIN_TONE(0x1F477),
-    CONSTRUCTION_WORKER_MEDIUM_SKIN_TONE(0x1F477),
-    CONSTRUCTION_WORKER_MEDIUM_DARK_SKIN_TONE(0x1F477),
-    CONSTRUCTION_WORKER_DARK_SKIN_TONE(0x1F477),
-    MAN_CONSTRUCTION_WORKER(0x1F477),
-    MAN_CONSTRUCTION_WORKER_LIGHT_SKIN_TONE(0x1F477),
-    MAN_CONSTRUCTION_WORKER_MEDIUM_LIGHT_SKIN_TONE(0x1F477),
-    MAN_CONSTRUCTION_WORKER_MEDIUM_SKIN_TONE(0x1F477),
-    MAN_CONSTRUCTION_WORKER_MEDIUM_DARK_SKIN_TONE(0x1F477),
-    MAN_CONSTRUCTION_WORKER_DARK_SKIN_TONE(0x1F477),
-    WOMAN_CONSTRUCTION_WORKER(0x1F477),
-    WOMAN_CONSTRUCTION_WORKER_LIGHT_SKIN_TONE(0x1F477),
-    WOMAN_CONSTRUCTION_WORKER_MEDIUM_LIGHT_SKIN_TONE(0x1F477),
-    WOMAN_CONSTRUCTION_WORKER_MEDIUM_SKIN_TONE(0x1F477),
-    WOMAN_CONSTRUCTION_WORKER_MEDIUM_DARK_SKIN_TONE(0x1F477),
-    WOMAN_CONSTRUCTION_WORKER_DARK_SKIN_TONE(0x1F477),
-    PERSON_WITH_CROWN(0x1FAC5),
-    PERSON_WITH_CROWN_LIGHT_SKIN_TONE(0x1FAC5),
-    PERSON_WITH_CROWN_MEDIUM_LIGHT_SKIN_TONE(0x1FAC5),
-    PERSON_WITH_CROWN_MEDIUM_SKIN_TONE(0x1FAC5),
-    PERSON_WITH_CROWN_MEDIUM_DARK_SKIN_TONE(0x1FAC5),
-    PERSON_WITH_CROWN_DARK_SKIN_TONE(0x1FAC5),
-    PRINCE(0x1F934),
-    PRINCE_LIGHT_SKIN_TONE(0x1F934),
-    PRINCE_MEDIUM_LIGHT_SKIN_TONE(0x1F934),
-    PRINCE_MEDIUM_SKIN_TONE(0x1F934),
-    PRINCE_MEDIUM_DARK_SKIN_TONE(0x1F934),
-    PRINCE_DARK_SKIN_TONE(0x1F934),
-    PRINCESS(0x1F478),
-    PRINCESS_LIGHT_SKIN_TONE(0x1F478),
-    PRINCESS_MEDIUM_LIGHT_SKIN_TONE(0x1F478),
-    PRINCESS_MEDIUM_SKIN_TONE(0x1F478),
-    PRINCESS_MEDIUM_DARK_SKIN_TONE(0x1F478),
-    PRINCESS_DARK_SKIN_TONE(0x1F478),
-    PERSON_WEARING_TURBAN(0x1F473),
-    PERSON_WEARING_TURBAN_LIGHT_SKIN_TONE(0x1F473),
-    PERSON_WEARING_TURBAN_MEDIUM_LIGHT_SKIN_TONE(0x1F473),
-    PERSON_WEARING_TURBAN_MEDIUM_SKIN_TONE(0x1F473),
-    PERSON_WEARING_TURBAN_MEDIUM_DARK_SKIN_TONE(0x1F473),
-    PERSON_WEARING_TURBAN_DARK_SKIN_TONE(0x1F473),
-    MAN_WEARING_TURBAN(0x1F473),
-    MAN_WEARING_TURBAN_LIGHT_SKIN_TONE(0x1F473),
-    MAN_WEARING_TURBAN_MEDIUM_LIGHT_SKIN_TONE(0x1F473),
-    MAN_WEARING_TURBAN_MEDIUM_SKIN_TONE(0x1F473),
-    MAN_WEARING_TURBAN_MEDIUM_DARK_SKIN_TONE(0x1F473),
-    MAN_WEARING_TURBAN_DARK_SKIN_TONE(0x1F473),
-    WOMAN_WEARING_TURBAN(0x1F473),
-    WOMAN_WEARING_TURBAN_LIGHT_SKIN_TONE(0x1F473),
-    WOMAN_WEARING_TURBAN_MEDIUM_LIGHT_SKIN_TONE(0x1F473),
-    WOMAN_WEARING_TURBAN_MEDIUM_SKIN_TONE(0x1F473),
-    WOMAN_WEARING_TURBAN_MEDIUM_DARK_SKIN_TONE(0x1F473),
-    WOMAN_WEARING_TURBAN_DARK_SKIN_TONE(0x1F473),
-    PERSON_WITH_SKULLCAP(0x1F472),
-    PERSON_WITH_SKULLCAP_LIGHT_SKIN_TONE(0x1F472),
-    PERSON_WITH_SKULLCAP_MEDIUM_LIGHT_SKIN_TONE(0x1F472),
-    PERSON_WITH_SKULLCAP_MEDIUM_SKIN_TONE(0x1F472),
-    PERSON_WITH_SKULLCAP_MEDIUM_DARK_SKIN_TONE(0x1F472),
-    PERSON_WITH_SKULLCAP_DARK_SKIN_TONE(0x1F472),
-    WOMAN_WITH_HEADSCARF(0x1F9D5),
-    WOMAN_WITH_HEADSCARF_LIGHT_SKIN_TONE(0x1F9D5),
-    WOMAN_WITH_HEADSCARF_MEDIUM_LIGHT_SKIN_TONE(0x1F9D5),
-    WOMAN_WITH_HEADSCARF_MEDIUM_SKIN_TONE(0x1F9D5),
-    WOMAN_WITH_HEADSCARF_MEDIUM_DARK_SKIN_TONE(0x1F9D5),
-    WOMAN_WITH_HEADSCARF_DARK_SKIN_TONE(0x1F9D5),
-    PERSON_IN_TUXEDO(0x1F935),
-    PERSON_IN_TUXEDO_LIGHT_SKIN_TONE(0x1F935),
-    PERSON_IN_TUXEDO_MEDIUM_LIGHT_SKIN_TONE(0x1F935),
-    PERSON_IN_TUXEDO_MEDIUM_SKIN_TONE(0x1F935),
-    PERSON_IN_TUXEDO_MEDIUM_DARK_SKIN_TONE(0x1F935),
-    PERSON_IN_TUXEDO_DARK_SKIN_TONE(0x1F935),
-    MAN_IN_TUXEDO(0x1F935),
-    MAN_IN_TUXEDO_LIGHT_SKIN_TONE(0x1F935),
-    MAN_IN_TUXEDO_MEDIUM_LIGHT_SKIN_TONE(0x1F935),
-    MAN_IN_TUXEDO_MEDIUM_SKIN_TONE(0x1F935),
-    MAN_IN_TUXEDO_MEDIUM_DARK_SKIN_TONE(0x1F935),
-    MAN_IN_TUXEDO_DARK_SKIN_TONE(0x1F935),
-    WOMAN_IN_TUXEDO(0x1F935),
-    WOMAN_IN_TUXEDO_LIGHT_SKIN_TONE(0x1F935),
-    WOMAN_IN_TUXEDO_MEDIUM_LIGHT_SKIN_TONE(0x1F935),
-    WOMAN_IN_TUXEDO_MEDIUM_SKIN_TONE(0x1F935),
-    WOMAN_IN_TUXEDO_MEDIUM_DARK_SKIN_TONE(0x1F935),
-    WOMAN_IN_TUXEDO_DARK_SKIN_TONE(0x1F935),
-    PERSON_WITH_VEIL(0x1F470),
-    PERSON_WITH_VEIL_LIGHT_SKIN_TONE(0x1F470),
-    PERSON_WITH_VEIL_MEDIUM_LIGHT_SKIN_TONE(0x1F470),
-    PERSON_WITH_VEIL_MEDIUM_SKIN_TONE(0x1F470),
-    PERSON_WITH_VEIL_MEDIUM_DARK_SKIN_TONE(0x1F470),
-    PERSON_WITH_VEIL_DARK_SKIN_TONE(0x1F470),
-    MAN_WITH_VEIL(0x1F470),
-    MAN_WITH_VEIL_LIGHT_SKIN_TONE(0x1F470),
-    MAN_WITH_VEIL_MEDIUM_LIGHT_SKIN_TONE(0x1F470),
-    MAN_WITH_VEIL_MEDIUM_SKIN_TONE(0x1F470),
-    MAN_WITH_VEIL_MEDIUM_DARK_SKIN_TONE(0x1F470),
-    MAN_WITH_VEIL_DARK_SKIN_TONE(0x1F470),
-    WOMAN_WITH_VEIL(0x1F470),
-    WOMAN_WITH_VEIL_LIGHT_SKIN_TONE(0x1F470),
-    WOMAN_WITH_VEIL_MEDIUM_LIGHT_SKIN_TONE(0x1F470),
-    WOMAN_WITH_VEIL_MEDIUM_SKIN_TONE(0x1F470),
-    WOMAN_WITH_VEIL_MEDIUM_DARK_SKIN_TONE(0x1F470),
-    WOMAN_WITH_VEIL_DARK_SKIN_TONE(0x1F470),
-    PREGNANT_WOMAN(0x1F930),
-    PREGNANT_WOMAN_LIGHT_SKIN_TONE(0x1F930),
-    PREGNANT_WOMAN_MEDIUM_LIGHT_SKIN_TONE(0x1F930),
-    PREGNANT_WOMAN_MEDIUM_SKIN_TONE(0x1F930),
-    PREGNANT_WOMAN_MEDIUM_DARK_SKIN_TONE(0x1F930),
-    PREGNANT_WOMAN_DARK_SKIN_TONE(0x1F930),
-    PREGNANT_MAN(0x1FAC3),
-    PREGNANT_MAN_LIGHT_SKIN_TONE(0x1FAC3),
-    PREGNANT_MAN_MEDIUM_LIGHT_SKIN_TONE(0x1FAC3),
-    PREGNANT_MAN_MEDIUM_SKIN_TONE(0x1FAC3),
-    PREGNANT_MAN_MEDIUM_DARK_SKIN_TONE(0x1FAC3),
-    PREGNANT_MAN_DARK_SKIN_TONE(0x1FAC3),
-    PREGNANT_PERSON(0x1FAC4),
-    PREGNANT_PERSON_LIGHT_SKIN_TONE(0x1FAC4),
-    PREGNANT_PERSON_MEDIUM_LIGHT_SKIN_TONE(0x1FAC4),
-    PREGNANT_PERSON_MEDIUM_SKIN_TONE(0x1FAC4),
-    PREGNANT_PERSON_MEDIUM_DARK_SKIN_TONE(0x1FAC4),
-    PREGNANT_PERSON_DARK_SKIN_TONE(0x1FAC4),
-    BREAST_FEEDING(0x1F931),
-    BREAST_FEEDING_LIGHT_SKIN_TONE(0x1F931),
-    BREAST_FEEDING_MEDIUM_LIGHT_SKIN_TONE(0x1F931),
-    BREAST_FEEDING_MEDIUM_SKIN_TONE(0x1F931),
-    BREAST_FEEDING_MEDIUM_DARK_SKIN_TONE(0x1F931),
-    BREAST_FEEDING_DARK_SKIN_TONE(0x1F931),
-    WOMAN_FEEDING_BABY(0x1F469),
-    WOMAN_FEEDING_BABY_LIGHT_SKIN_TONE(0x1F469),
-    WOMAN_FEEDING_BABY_MEDIUM_LIGHT_SKIN_TONE(0x1F469),
-    WOMAN_FEEDING_BABY_MEDIUM_SKIN_TONE(0x1F469),
-    WOMAN_FEEDING_BABY_MEDIUM_DARK_SKIN_TONE(0x1F469),
-    WOMAN_FEEDING_BABY_DARK_SKIN_TONE(0x1F469),
-    MAN_FEEDING_BABY(0x1F468),
-    MAN_FEEDING_BABY_LIGHT_SKIN_TONE(0x1F468),
-    MAN_FEEDING_BABY_MEDIUM_LIGHT_SKIN_TONE(0x1F468),
-    MAN_FEEDING_BABY_MEDIUM_SKIN_TONE(0x1F468),
-    MAN_FEEDING_BABY_MEDIUM_DARK_SKIN_TONE(0x1F468),
-    MAN_FEEDING_BABY_DARK_SKIN_TONE(0x1F468),
-    PERSON_FEEDING_BABY(0x1F9D1),
-    PERSON_FEEDING_BABY_LIGHT_SKIN_TONE(0x1F9D1),
-    PERSON_FEEDING_BABY_MEDIUM_LIGHT_SKIN_TONE(0x1F9D1),
-    PERSON_FEEDING_BABY_MEDIUM_SKIN_TONE(0x1F9D1),
-    PERSON_FEEDING_BABY_MEDIUM_DARK_SKIN_TONE(0x1F9D1),
-    PERSON_FEEDING_BABY_DARK_SKIN_TONE(0x1F9D1),
-    BABY_ANGEL(0x1F47C),
-    BABY_ANGEL_LIGHT_SKIN_TONE(0x1F47C),
-    BABY_ANGEL_MEDIUM_LIGHT_SKIN_TONE(0x1F47C),
-    BABY_ANGEL_MEDIUM_SKIN_TONE(0x1F47C),
-    BABY_ANGEL_MEDIUM_DARK_SKIN_TONE(0x1F47C),
-    BABY_ANGEL_DARK_SKIN_TONE(0x1F47C),
-    SANTA_CLAUS(0x1F385),
-    SANTA_CLAUS_LIGHT_SKIN_TONE(0x1F385),
-    SANTA_CLAUS_MEDIUM_LIGHT_SKIN_TONE(0x1F385),
-    SANTA_CLAUS_MEDIUM_SKIN_TONE(0x1F385),
-    SANTA_CLAUS_MEDIUM_DARK_SKIN_TONE(0x1F385),
-    SANTA_CLAUS_DARK_SKIN_TONE(0x1F385),
-    MRS_CLAUS(0x1F936),
-    MRS_CLAUS_LIGHT_SKIN_TONE(0x1F936),
-    MRS_CLAUS_MEDIUM_LIGHT_SKIN_TONE(0x1F936),
-    MRS_CLAUS_MEDIUM_SKIN_TONE(0x1F936),
-    MRS_CLAUS_MEDIUM_DARK_SKIN_TONE(0x1F936),
-    MRS_CLAUS_DARK_SKIN_TONE(0x1F936),
-    MX_CLAUS(0x1F9D1),
-    MX_CLAUS_LIGHT_SKIN_TONE(0x1F9D1),
-    MX_CLAUS_MEDIUM_LIGHT_SKIN_TONE(0x1F9D1),
-    MX_CLAUS_MEDIUM_SKIN_TONE(0x1F9D1),
-    MX_CLAUS_MEDIUM_DARK_SKIN_TONE(0x1F9D1),
-    MX_CLAUS_DARK_SKIN_TONE(0x1F9D1),
-    SUPERHERO(0x1F9B8),
-    SUPERHERO_LIGHT_SKIN_TONE(0x1F9B8),
-    SUPERHERO_MEDIUM_LIGHT_SKIN_TONE(0x1F9B8),
-    SUPERHERO_MEDIUM_SKIN_TONE(0x1F9B8),
-    SUPERHERO_MEDIUM_DARK_SKIN_TONE(0x1F9B8),
-    SUPERHERO_DARK_SKIN_TONE(0x1F9B8),
-    MAN_SUPERHERO(0x1F9B8),
-    MAN_SUPERHERO_LIGHT_SKIN_TONE(0x1F9B8),
-    MAN_SUPERHERO_MEDIUM_LIGHT_SKIN_TONE(0x1F9B8),
-    MAN_SUPERHERO_MEDIUM_SKIN_TONE(0x1F9B8),
-    MAN_SUPERHERO_MEDIUM_DARK_SKIN_TONE(0x1F9B8),
-    MAN_SUPERHERO_DARK_SKIN_TONE(0x1F9B8),
-    WOMAN_SUPERHERO(0x1F9B8),
-    WOMAN_SUPERHERO_LIGHT_SKIN_TONE(0x1F9B8),
-    WOMAN_SUPERHERO_MEDIUM_LIGHT_SKIN_TONE(0x1F9B8),
-    WOMAN_SUPERHERO_MEDIUM_SKIN_TONE(0x1F9B8),
-    WOMAN_SUPERHERO_MEDIUM_DARK_SKIN_TONE(0x1F9B8),
-    WOMAN_SUPERHERO_DARK_SKIN_TONE(0x1F9B8),
-    SUPERVILLAIN(0x1F9B9),
-    SUPERVILLAIN_LIGHT_SKIN_TONE(0x1F9B9),
-    SUPERVILLAIN_MEDIUM_LIGHT_SKIN_TONE(0x1F9B9),
-    SUPERVILLAIN_MEDIUM_SKIN_TONE(0x1F9B9),
-    SUPERVILLAIN_MEDIUM_DARK_SKIN_TONE(0x1F9B9),
-    SUPERVILLAIN_DARK_SKIN_TONE(0x1F9B9),
-    MAN_SUPERVILLAIN(0x1F9B9),
-    MAN_SUPERVILLAIN_LIGHT_SKIN_TONE(0x1F9B9),
-    MAN_SUPERVILLAIN_MEDIUM_LIGHT_SKIN_TONE(0x1F9B9),
-    MAN_SUPERVILLAIN_MEDIUM_SKIN_TONE(0x1F9B9),
-    MAN_SUPERVILLAIN_MEDIUM_DARK_SKIN_TONE(0x1F9B9),
-    MAN_SUPERVILLAIN_DARK_SKIN_TONE(0x1F9B9),
-    WOMAN_SUPERVILLAIN(0x1F9B9),
-    WOMAN_SUPERVILLAIN_LIGHT_SKIN_TONE(0x1F9B9),
-    WOMAN_SUPERVILLAIN_MEDIUM_LIGHT_SKIN_TONE(0x1F9B9),
-    WOMAN_SUPERVILLAIN_MEDIUM_SKIN_TONE(0x1F9B9),
-    WOMAN_SUPERVILLAIN_MEDIUM_DARK_SKIN_TONE(0x1F9B9),
-    WOMAN_SUPERVILLAIN_DARK_SKIN_TONE(0x1F9B9),
-    MAGE(0x1F9D9),
-    MAGE_LIGHT_SKIN_TONE(0x1F9D9),
-    MAGE_MEDIUM_LIGHT_SKIN_TONE(0x1F9D9),
-    MAGE_MEDIUM_SKIN_TONE(0x1F9D9),
-    MAGE_MEDIUM_DARK_SKIN_TONE(0x1F9D9),
-    MAGE_DARK_SKIN_TONE(0x1F9D9),
-    MAN_MAGE(0x1F9D9),
-    MAN_MAGE_LIGHT_SKIN_TONE(0x1F9D9),
-    MAN_MAGE_MEDIUM_LIGHT_SKIN_TONE(0x1F9D9),
-    MAN_MAGE_MEDIUM_SKIN_TONE(0x1F9D9),
-    MAN_MAGE_MEDIUM_DARK_SKIN_TONE(0x1F9D9),
-    MAN_MAGE_DARK_SKIN_TONE(0x1F9D9),
-    WOMAN_MAGE(0x1F9D9),
-    WOMAN_MAGE_LIGHT_SKIN_TONE(0x1F9D9),
-    WOMAN_MAGE_MEDIUM_LIGHT_SKIN_TONE(0x1F9D9),
-    WOMAN_MAGE_MEDIUM_SKIN_TONE(0x1F9D9),
-    WOMAN_MAGE_MEDIUM_DARK_SKIN_TONE(0x1F9D9),
-    WOMAN_MAGE_DARK_SKIN_TONE(0x1F9D9),
-    FAIRY(0x1F9DA),
-    FAIRY_LIGHT_SKIN_TONE(0x1F9DA),
-    FAIRY_MEDIUM_LIGHT_SKIN_TONE(0x1F9DA),
-    FAIRY_MEDIUM_SKIN_TONE(0x1F9DA),
-    FAIRY_MEDIUM_DARK_SKIN_TONE(0x1F9DA),
-    FAIRY_DARK_SKIN_TONE(0x1F9DA),
-    MAN_FAIRY(0x1F9DA),
-    MAN_FAIRY_LIGHT_SKIN_TONE(0x1F9DA),
-    MAN_FAIRY_MEDIUM_LIGHT_SKIN_TONE(0x1F9DA),
-    MAN_FAIRY_MEDIUM_SKIN_TONE(0x1F9DA),
-    MAN_FAIRY_MEDIUM_DARK_SKIN_TONE(0x1F9DA),
-    MAN_FAIRY_DARK_SKIN_TONE(0x1F9DA),
-    WOMAN_FAIRY(0x1F9DA),
-    WOMAN_FAIRY_LIGHT_SKIN_TONE(0x1F9DA),
-    WOMAN_FAIRY_MEDIUM_LIGHT_SKIN_TONE(0x1F9DA),
-    WOMAN_FAIRY_MEDIUM_SKIN_TONE(0x1F9DA),
-    WOMAN_FAIRY_MEDIUM_DARK_SKIN_TONE(0x1F9DA),
-    WOMAN_FAIRY_DARK_SKIN_TONE(0x1F9DA),
-    VAMPIRE(0x1F9DB),
-    VAMPIRE_LIGHT_SKIN_TONE(0x1F9DB),
-    VAMPIRE_MEDIUM_LIGHT_SKIN_TONE(0x1F9DB),
-    VAMPIRE_MEDIUM_SKIN_TONE(0x1F9DB),
-    VAMPIRE_MEDIUM_DARK_SKIN_TONE(0x1F9DB),
-    VAMPIRE_DARK_SKIN_TONE(0x1F9DB),
-    MAN_VAMPIRE(0x1F9DB),
-    MAN_VAMPIRE_LIGHT_SKIN_TONE(0x1F9DB),
-    MAN_VAMPIRE_MEDIUM_LIGHT_SKIN_TONE(0x1F9DB),
-    MAN_VAMPIRE_MEDIUM_SKIN_TONE(0x1F9DB),
-    MAN_VAMPIRE_MEDIUM_DARK_SKIN_TONE(0x1F9DB),
-    MAN_VAMPIRE_DARK_SKIN_TONE(0x1F9DB),
-    WOMAN_VAMPIRE(0x1F9DB),
-    WOMAN_VAMPIRE_LIGHT_SKIN_TONE(0x1F9DB),
-    WOMAN_VAMPIRE_MEDIUM_LIGHT_SKIN_TONE(0x1F9DB),
-    WOMAN_VAMPIRE_MEDIUM_SKIN_TONE(0x1F9DB),
-    WOMAN_VAMPIRE_MEDIUM_DARK_SKIN_TONE(0x1F9DB),
-    WOMAN_VAMPIRE_DARK_SKIN_TONE(0x1F9DB),
-    MERPERSON(0x1F9DC),
-    MERPERSON_LIGHT_SKIN_TONE(0x1F9DC),
-    MERPERSON_MEDIUM_LIGHT_SKIN_TONE(0x1F9DC),
-    MERPERSON_MEDIUM_SKIN_TONE(0x1F9DC),
-    MERPERSON_MEDIUM_DARK_SKIN_TONE(0x1F9DC),
-    MERPERSON_DARK_SKIN_TONE(0x1F9DC),
-    MERMAN(0x1F9DC),
-    MERMAN_LIGHT_SKIN_TONE(0x1F9DC),
-    MERMAN_MEDIUM_LIGHT_SKIN_TONE(0x1F9DC),
-    MERMAN_MEDIUM_SKIN_TONE(0x1F9DC),
-    MERMAN_MEDIUM_DARK_SKIN_TONE(0x1F9DC),
-    MERMAN_DARK_SKIN_TONE(0x1F9DC),
-    MERMAID(0x1F9DC),
-    MERMAID_LIGHT_SKIN_TONE(0x1F9DC),
-    MERMAID_MEDIUM_LIGHT_SKIN_TONE(0x1F9DC),
-    MERMAID_MEDIUM_SKIN_TONE(0x1F9DC),
-    MERMAID_MEDIUM_DARK_SKIN_TONE(0x1F9DC),
-    MERMAID_DARK_SKIN_TONE(0x1F9DC),
-    ELF(0x1F9DD),
-    ELF_LIGHT_SKIN_TONE(0x1F9DD),
-    ELF_MEDIUM_LIGHT_SKIN_TONE(0x1F9DD),
-    ELF_MEDIUM_SKIN_TONE(0x1F9DD),
-    ELF_MEDIUM_DARK_SKIN_TONE(0x1F9DD),
-    ELF_DARK_SKIN_TONE(0x1F9DD),
-    MAN_ELF(0x1F9DD),
-    MAN_ELF_LIGHT_SKIN_TONE(0x1F9DD),
-    MAN_ELF_MEDIUM_LIGHT_SKIN_TONE(0x1F9DD),
-    MAN_ELF_MEDIUM_SKIN_TONE(0x1F9DD),
-    MAN_ELF_MEDIUM_DARK_SKIN_TONE(0x1F9DD),
-    MAN_ELF_DARK_SKIN_TONE(0x1F9DD),
-    WOMAN_ELF(0x1F9DD),
-    WOMAN_ELF_LIGHT_SKIN_TONE(0x1F9DD),
-    WOMAN_ELF_MEDIUM_LIGHT_SKIN_TONE(0x1F9DD),
-    WOMAN_ELF_MEDIUM_SKIN_TONE(0x1F9DD),
-    WOMAN_ELF_MEDIUM_DARK_SKIN_TONE(0x1F9DD),
-    WOMAN_ELF_DARK_SKIN_TONE(0x1F9DD),
-    GENIE(0x1F9DE),
-    MAN_GENIE(0x1F9DE),
-    WOMAN_GENIE(0x1F9DE),
-    ZOMBIE(0x1F9DF),
-    MAN_ZOMBIE(0x1F9DF),
-    WOMAN_ZOMBIE(0x1F9DF),
-    TROLL(0x1F9CC),
-    PERSON_GETTING_MASSAGE(0x1F486),
-    PERSON_GETTING_MASSAGE_LIGHT_SKIN_TONE(0x1F486),
-    PERSON_GETTING_MASSAGE_MEDIUM_LIGHT_SKIN_TONE(0x1F486),
-    PERSON_GETTING_MASSAGE_MEDIUM_SKIN_TONE(0x1F486),
-    PERSON_GETTING_MASSAGE_MEDIUM_DARK_SKIN_TONE(0x1F486),
-    PERSON_GETTING_MASSAGE_DARK_SKIN_TONE(0x1F486),
-    MAN_GETTING_MASSAGE(0x1F486),
-    MAN_GETTING_MASSAGE_LIGHT_SKIN_TONE(0x1F486),
-    MAN_GETTING_MASSAGE_MEDIUM_LIGHT_SKIN_TONE(0x1F486),
-    MAN_GETTING_MASSAGE_MEDIUM_SKIN_TONE(0x1F486),
-    MAN_GETTING_MASSAGE_MEDIUM_DARK_SKIN_TONE(0x1F486),
-    MAN_GETTING_MASSAGE_DARK_SKIN_TONE(0x1F486),
-    WOMAN_GETTING_MASSAGE(0x1F486),
-    WOMAN_GETTING_MASSAGE_LIGHT_SKIN_TONE(0x1F486),
-    WOMAN_GETTING_MASSAGE_MEDIUM_LIGHT_SKIN_TONE(0x1F486),
-    WOMAN_GETTING_MASSAGE_MEDIUM_SKIN_TONE(0x1F486),
-    WOMAN_GETTING_MASSAGE_MEDIUM_DARK_SKIN_TONE(0x1F486),
-    WOMAN_GETTING_MASSAGE_DARK_SKIN_TONE(0x1F486),
-    PERSON_GETTING_HAIRCUT(0x1F487),
-    PERSON_GETTING_HAIRCUT_LIGHT_SKIN_TONE(0x1F487),
-    PERSON_GETTING_HAIRCUT_MEDIUM_LIGHT_SKIN_TONE(0x1F487),
-    PERSON_GETTING_HAIRCUT_MEDIUM_SKIN_TONE(0x1F487),
-    PERSON_GETTING_HAIRCUT_MEDIUM_DARK_SKIN_TONE(0x1F487),
-    PERSON_GETTING_HAIRCUT_DARK_SKIN_TONE(0x1F487),
-    MAN_GETTING_HAIRCUT(0x1F487),
-    MAN_GETTING_HAIRCUT_LIGHT_SKIN_TONE(0x1F487),
-    MAN_GETTING_HAIRCUT_MEDIUM_LIGHT_SKIN_TONE(0x1F487),
-    MAN_GETTING_HAIRCUT_MEDIUM_SKIN_TONE(0x1F487),
-    MAN_GETTING_HAIRCUT_MEDIUM_DARK_SKIN_TONE(0x1F487),
-    MAN_GETTING_HAIRCUT_DARK_SKIN_TONE(0x1F487),
-    WOMAN_GETTING_HAIRCUT(0x1F487),
-    WOMAN_GETTING_HAIRCUT_LIGHT_SKIN_TONE(0x1F487),
-    WOMAN_GETTING_HAIRCUT_MEDIUM_LIGHT_SKIN_TONE(0x1F487),
-    WOMAN_GETTING_HAIRCUT_MEDIUM_SKIN_TONE(0x1F487),
-    WOMAN_GETTING_HAIRCUT_MEDIUM_DARK_SKIN_TONE(0x1F487),
-    WOMAN_GETTING_HAIRCUT_DARK_SKIN_TONE(0x1F487),
-    PERSON_WALKING(0x1F6B6),
-    PERSON_WALKING_LIGHT_SKIN_TONE(0x1F6B6),
-    PERSON_WALKING_MEDIUM_LIGHT_SKIN_TONE(0x1F6B6),
-    PERSON_WALKING_MEDIUM_SKIN_TONE(0x1F6B6),
-    PERSON_WALKING_MEDIUM_DARK_SKIN_TONE(0x1F6B6),
-    PERSON_WALKING_DARK_SKIN_TONE(0x1F6B6),
-    MAN_WALKING(0x1F6B6),
-    MAN_WALKING_LIGHT_SKIN_TONE(0x1F6B6),
-    MAN_WALKING_MEDIUM_LIGHT_SKIN_TONE(0x1F6B6),
-    MAN_WALKING_MEDIUM_SKIN_TONE(0x1F6B6),
-    MAN_WALKING_MEDIUM_DARK_SKIN_TONE(0x1F6B6),
-    MAN_WALKING_DARK_SKIN_TONE(0x1F6B6),
-    WOMAN_WALKING(0x1F6B6),
-    WOMAN_WALKING_LIGHT_SKIN_TONE(0x1F6B6),
-    WOMAN_WALKING_MEDIUM_LIGHT_SKIN_TONE(0x1F6B6),
-    WOMAN_WALKING_MEDIUM_SKIN_TONE(0x1F6B6),
-    WOMAN_WALKING_MEDIUM_DARK_SKIN_TONE(0x1F6B6),
-    WOMAN_WALKING_DARK_SKIN_TONE(0x1F6B6),
-    PERSON_STANDING(0x1F9CD),
-    PERSON_STANDING_LIGHT_SKIN_TONE(0x1F9CD),
-    PERSON_STANDING_MEDIUM_LIGHT_SKIN_TONE(0x1F9CD),
-    PERSON_STANDING_MEDIUM_SKIN_TONE(0x1F9CD),
-    PERSON_STANDING_MEDIUM_DARK_SKIN_TONE(0x1F9CD),
-    PERSON_STANDING_DARK_SKIN_TONE(0x1F9CD),
-    MAN_STANDING(0x1F9CD),
-    MAN_STANDING_LIGHT_SKIN_TONE(0x1F9CD),
-    MAN_STANDING_MEDIUM_LIGHT_SKIN_TONE(0x1F9CD),
-    MAN_STANDING_MEDIUM_SKIN_TONE(0x1F9CD),
-    MAN_STANDING_MEDIUM_DARK_SKIN_TONE(0x1F9CD),
-    MAN_STANDING_DARK_SKIN_TONE(0x1F9CD),
-    WOMAN_STANDING(0x1F9CD),
-    WOMAN_STANDING_LIGHT_SKIN_TONE(0x1F9CD),
-    WOMAN_STANDING_MEDIUM_LIGHT_SKIN_TONE(0x1F9CD),
-    WOMAN_STANDING_MEDIUM_SKIN_TONE(0x1F9CD),
-    WOMAN_STANDING_MEDIUM_DARK_SKIN_TONE(0x1F9CD),
-    WOMAN_STANDING_DARK_SKIN_TONE(0x1F9CD),
-    PERSON_KNEELING(0x1F9CE),
-    PERSON_KNEELING_LIGHT_SKIN_TONE(0x1F9CE),
-    PERSON_KNEELING_MEDIUM_LIGHT_SKIN_TONE(0x1F9CE),
-    PERSON_KNEELING_MEDIUM_SKIN_TONE(0x1F9CE),
-    PERSON_KNEELING_MEDIUM_DARK_SKIN_TONE(0x1F9CE),
-    PERSON_KNEELING_DARK_SKIN_TONE(0x1F9CE),
-    MAN_KNEELING(0x1F9CE),
-    MAN_KNEELING_LIGHT_SKIN_TONE(0x1F9CE),
-    MAN_KNEELING_MEDIUM_LIGHT_SKIN_TONE(0x1F9CE),
-    MAN_KNEELING_MEDIUM_SKIN_TONE(0x1F9CE),
-    MAN_KNEELING_MEDIUM_DARK_SKIN_TONE(0x1F9CE),
-    MAN_KNEELING_DARK_SKIN_TONE(0x1F9CE),
-    WOMAN_KNEELING(0x1F9CE),
-    WOMAN_KNEELING_LIGHT_SKIN_TONE(0x1F9CE),
-    WOMAN_KNEELING_MEDIUM_LIGHT_SKIN_TONE(0x1F9CE),
-    WOMAN_KNEELING_MEDIUM_SKIN_TONE(0x1F9CE),
-    WOMAN_KNEELING_MEDIUM_DARK_SKIN_TONE(0x1F9CE),
-    WOMAN_KNEELING_DARK_SKIN_TONE(0x1F9CE),
-    PERSON_WITH_WHITE_CANE(0x1F9D1),
-    PERSON_WITH_WHITE_CANE_LIGHT_SKIN_TONE(0x1F9D1),
-    PERSON_WITH_WHITE_CANE_MEDIUM_LIGHT_SKIN_TONE(0x1F9D1),
-    PERSON_WITH_WHITE_CANE_MEDIUM_SKIN_TONE(0x1F9D1),
-    PERSON_WITH_WHITE_CANE_MEDIUM_DARK_SKIN_TONE(0x1F9D1),
-    PERSON_WITH_WHITE_CANE_DARK_SKIN_TONE(0x1F9D1),
-    MAN_WITH_WHITE_CANE(0x1F468),
-    MAN_WITH_WHITE_CANE_LIGHT_SKIN_TONE(0x1F468),
-    MAN_WITH_WHITE_CANE_MEDIUM_LIGHT_SKIN_TONE(0x1F468),
-    MAN_WITH_WHITE_CANE_MEDIUM_SKIN_TONE(0x1F468),
-    MAN_WITH_WHITE_CANE_MEDIUM_DARK_SKIN_TONE(0x1F468),
-    MAN_WITH_WHITE_CANE_DARK_SKIN_TONE(0x1F468),
-    WOMAN_WITH_WHITE_CANE(0x1F469),
-    WOMAN_WITH_WHITE_CANE_LIGHT_SKIN_TONE(0x1F469),
-    WOMAN_WITH_WHITE_CANE_MEDIUM_LIGHT_SKIN_TONE(0x1F469),
-    WOMAN_WITH_WHITE_CANE_MEDIUM_SKIN_TONE(0x1F469),
-    WOMAN_WITH_WHITE_CANE_MEDIUM_DARK_SKIN_TONE(0x1F469),
-    WOMAN_WITH_WHITE_CANE_DARK_SKIN_TONE(0x1F469),
-    PERSON_IN_MOTORIZED_WHEELCHAIR(0x1F9D1),
-    PERSON_IN_MOTORIZED_WHEELCHAIR_LIGHT_SKIN_TONE(0x1F9D1),
-    PERSON_IN_MOTORIZED_WHEELCHAIR_MEDIUM_LIGHT_SKIN_TONE(0x1F9D1),
-    PERSON_IN_MOTORIZED_WHEELCHAIR_MEDIUM_SKIN_TONE(0x1F9D1),
-    PERSON_IN_MOTORIZED_WHEELCHAIR_MEDIUM_DARK_SKIN_TONE(0x1F9D1),
-    PERSON_IN_MOTORIZED_WHEELCHAIR_DARK_SKIN_TONE(0x1F9D1),
-    MAN_IN_MOTORIZED_WHEELCHAIR(0x1F468),
-    MAN_IN_MOTORIZED_WHEELCHAIR_LIGHT_SKIN_TONE(0x1F468),
-    MAN_IN_MOTORIZED_WHEELCHAIR_MEDIUM_LIGHT_SKIN_TONE(0x1F468),
-    MAN_IN_MOTORIZED_WHEELCHAIR_MEDIUM_SKIN_TONE(0x1F468),
-    MAN_IN_MOTORIZED_WHEELCHAIR_MEDIUM_DARK_SKIN_TONE(0x1F468),
-    MAN_IN_MOTORIZED_WHEELCHAIR_DARK_SKIN_TONE(0x1F468),
-    WOMAN_IN_MOTORIZED_WHEELCHAIR(0x1F469),
-    WOMAN_IN_MOTORIZED_WHEELCHAIR_LIGHT_SKIN_TONE(0x1F469),
-    WOMAN_IN_MOTORIZED_WHEELCHAIR_MEDIUM_LIGHT_SKIN_TONE(0x1F469),
-    WOMAN_IN_MOTORIZED_WHEELCHAIR_MEDIUM_SKIN_TONE(0x1F469),
-    WOMAN_IN_MOTORIZED_WHEELCHAIR_MEDIUM_DARK_SKIN_TONE(0x1F469),
-    WOMAN_IN_MOTORIZED_WHEELCHAIR_DARK_SKIN_TONE(0x1F469),
-    PERSON_IN_MANUAL_WHEELCHAIR(0x1F9D1),
-    PERSON_IN_MANUAL_WHEELCHAIR_LIGHT_SKIN_TONE(0x1F9D1),
-    PERSON_IN_MANUAL_WHEELCHAIR_MEDIUM_LIGHT_SKIN_TONE(0x1F9D1),
-    PERSON_IN_MANUAL_WHEELCHAIR_MEDIUM_SKIN_TONE(0x1F9D1),
-    PERSON_IN_MANUAL_WHEELCHAIR_MEDIUM_DARK_SKIN_TONE(0x1F9D1),
-    PERSON_IN_MANUAL_WHEELCHAIR_DARK_SKIN_TONE(0x1F9D1),
-    MAN_IN_MANUAL_WHEELCHAIR(0x1F468),
-    MAN_IN_MANUAL_WHEELCHAIR_LIGHT_SKIN_TONE(0x1F468),
-    MAN_IN_MANUAL_WHEELCHAIR_MEDIUM_LIGHT_SKIN_TONE(0x1F468),
-    MAN_IN_MANUAL_WHEELCHAIR_MEDIUM_SKIN_TONE(0x1F468),
-    MAN_IN_MANUAL_WHEELCHAIR_MEDIUM_DARK_SKIN_TONE(0x1F468),
-    MAN_IN_MANUAL_WHEELCHAIR_DARK_SKIN_TONE(0x1F468),
-    WOMAN_IN_MANUAL_WHEELCHAIR(0x1F469),
-    WOMAN_IN_MANUAL_WHEELCHAIR_LIGHT_SKIN_TONE(0x1F469),
-    WOMAN_IN_MANUAL_WHEELCHAIR_MEDIUM_LIGHT_SKIN_TONE(0x1F469),
-    WOMAN_IN_MANUAL_WHEELCHAIR_MEDIUM_SKIN_TONE(0x1F469),
-    WOMAN_IN_MANUAL_WHEELCHAIR_MEDIUM_DARK_SKIN_TONE(0x1F469),
-    WOMAN_IN_MANUAL_WHEELCHAIR_DARK_SKIN_TONE(0x1F469),
-    PERSON_RUNNING(0x1F3C3),
-    PERSON_RUNNING_LIGHT_SKIN_TONE(0x1F3C3),
-    PERSON_RUNNING_MEDIUM_LIGHT_SKIN_TONE(0x1F3C3),
-    PERSON_RUNNING_MEDIUM_SKIN_TONE(0x1F3C3),
-    PERSON_RUNNING_MEDIUM_DARK_SKIN_TONE(0x1F3C3),
-    PERSON_RUNNING_DARK_SKIN_TONE(0x1F3C3),
-    MAN_RUNNING(0x1F3C3),
-    MAN_RUNNING_LIGHT_SKIN_TONE(0x1F3C3),
-    MAN_RUNNING_MEDIUM_LIGHT_SKIN_TONE(0x1F3C3),
-    MAN_RUNNING_MEDIUM_SKIN_TONE(0x1F3C3),
-    MAN_RUNNING_MEDIUM_DARK_SKIN_TONE(0x1F3C3),
-    MAN_RUNNING_DARK_SKIN_TONE(0x1F3C3),
-    WOMAN_RUNNING(0x1F3C3),
-    WOMAN_RUNNING_LIGHT_SKIN_TONE(0x1F3C3),
-    WOMAN_RUNNING_MEDIUM_LIGHT_SKIN_TONE(0x1F3C3),
-    WOMAN_RUNNING_MEDIUM_SKIN_TONE(0x1F3C3),
-    WOMAN_RUNNING_MEDIUM_DARK_SKIN_TONE(0x1F3C3),
-    WOMAN_RUNNING_DARK_SKIN_TONE(0x1F3C3),
-    WOMAN_DANCING(0x1F483),
-    WOMAN_DANCING_LIGHT_SKIN_TONE(0x1F483),
-    WOMAN_DANCING_MEDIUM_LIGHT_SKIN_TONE(0x1F483),
-    WOMAN_DANCING_MEDIUM_SKIN_TONE(0x1F483),
-    WOMAN_DANCING_MEDIUM_DARK_SKIN_TONE(0x1F483),
-    WOMAN_DANCING_DARK_SKIN_TONE(0x1F483),
-    MAN_DANCING(0x1F57A),
-    MAN_DANCING_LIGHT_SKIN_TONE(0x1F57A),
-    MAN_DANCING_MEDIUM_LIGHT_SKIN_TONE(0x1F57A),
-    MAN_DANCING_MEDIUM_SKIN_TONE(0x1F57A),
-    MAN_DANCING_MEDIUM_DARK_SKIN_TONE(0x1F57A),
-    MAN_DANCING_DARK_SKIN_TONE(0x1F57A),
-    PERSON_IN_SUIT_LEVITATING(0x1F574),
-    PERSON_IN_SUIT_LEVITATING_LIGHT_SKIN_TONE(0x1F574),
-    PERSON_IN_SUIT_LEVITATING_MEDIUM_LIGHT_SKIN_TONE(0x1F574),
-    PERSON_IN_SUIT_LEVITATING_MEDIUM_SKIN_TONE(0x1F574),
-    PERSON_IN_SUIT_LEVITATING_MEDIUM_DARK_SKIN_TONE(0x1F574),
-    PERSON_IN_SUIT_LEVITATING_DARK_SKIN_TONE(0x1F574),
-    PEOPLE_WITH_BUNNY_EARS(0x1F46F),
-    MEN_WITH_BUNNY_EARS(0x1F46F),
-    WOMEN_WITH_BUNNY_EARS(0x1F46F),
-    PERSON_IN_STEAMY_ROOM(0x1F9D6),
-    PERSON_IN_STEAMY_ROOM_LIGHT_SKIN_TONE(0x1F9D6),
-    PERSON_IN_STEAMY_ROOM_MEDIUM_LIGHT_SKIN_TONE(0x1F9D6),
-    PERSON_IN_STEAMY_ROOM_MEDIUM_SKIN_TONE(0x1F9D6),
-    PERSON_IN_STEAMY_ROOM_MEDIUM_DARK_SKIN_TONE(0x1F9D6),
-    PERSON_IN_STEAMY_ROOM_DARK_SKIN_TONE(0x1F9D6),
-    MAN_IN_STEAMY_ROOM(0x1F9D6),
-    MAN_IN_STEAMY_ROOM_LIGHT_SKIN_TONE(0x1F9D6),
-    MAN_IN_STEAMY_ROOM_MEDIUM_LIGHT_SKIN_TONE(0x1F9D6),
-    MAN_IN_STEAMY_ROOM_MEDIUM_SKIN_TONE(0x1F9D6),
-    MAN_IN_STEAMY_ROOM_MEDIUM_DARK_SKIN_TONE(0x1F9D6),
-    MAN_IN_STEAMY_ROOM_DARK_SKIN_TONE(0x1F9D6),
-    WOMAN_IN_STEAMY_ROOM(0x1F9D6),
-    WOMAN_IN_STEAMY_ROOM_LIGHT_SKIN_TONE(0x1F9D6),
-    WOMAN_IN_STEAMY_ROOM_MEDIUM_LIGHT_SKIN_TONE(0x1F9D6),
-    WOMAN_IN_STEAMY_ROOM_MEDIUM_SKIN_TONE(0x1F9D6),
-    WOMAN_IN_STEAMY_ROOM_MEDIUM_DARK_SKIN_TONE(0x1F9D6),
-    WOMAN_IN_STEAMY_ROOM_DARK_SKIN_TONE(0x1F9D6),
-    PERSON_CLIMBING(0x1F9D7),
-    PERSON_CLIMBING_LIGHT_SKIN_TONE(0x1F9D7),
-    PERSON_CLIMBING_MEDIUM_LIGHT_SKIN_TONE(0x1F9D7),
-    PERSON_CLIMBING_MEDIUM_SKIN_TONE(0x1F9D7),
-    PERSON_CLIMBING_MEDIUM_DARK_SKIN_TONE(0x1F9D7),
-    PERSON_CLIMBING_DARK_SKIN_TONE(0x1F9D7),
-    MAN_CLIMBING(0x1F9D7),
-    MAN_CLIMBING_LIGHT_SKIN_TONE(0x1F9D7),
-    MAN_CLIMBING_MEDIUM_LIGHT_SKIN_TONE(0x1F9D7),
-    MAN_CLIMBING_MEDIUM_SKIN_TONE(0x1F9D7),
-    MAN_CLIMBING_MEDIUM_DARK_SKIN_TONE(0x1F9D7),
-    MAN_CLIMBING_DARK_SKIN_TONE(0x1F9D7),
-    WOMAN_CLIMBING(0x1F9D7),
-    WOMAN_CLIMBING_LIGHT_SKIN_TONE(0x1F9D7),
-    WOMAN_CLIMBING_MEDIUM_LIGHT_SKIN_TONE(0x1F9D7),
-    WOMAN_CLIMBING_MEDIUM_SKIN_TONE(0x1F9D7),
-    WOMAN_CLIMBING_MEDIUM_DARK_SKIN_TONE(0x1F9D7),
-    WOMAN_CLIMBING_DARK_SKIN_TONE(0x1F9D7),
-    PERSON_FENCING(0x1F93A),
-    HORSE_RACING(0x1F3C7),
-    HORSE_RACING_LIGHT_SKIN_TONE(0x1F3C7),
-    HORSE_RACING_MEDIUM_LIGHT_SKIN_TONE(0x1F3C7),
-    HORSE_RACING_MEDIUM_SKIN_TONE(0x1F3C7),
-    HORSE_RACING_MEDIUM_DARK_SKIN_TONE(0x1F3C7),
-    HORSE_RACING_DARK_SKIN_TONE(0x1F3C7),
-    SKIER(0x26F7),
-    SNOWBOARDER(0x1F3C2),
-    SNOWBOARDER_LIGHT_SKIN_TONE(0x1F3C2),
-    SNOWBOARDER_MEDIUM_LIGHT_SKIN_TONE(0x1F3C2),
-    SNOWBOARDER_MEDIUM_SKIN_TONE(0x1F3C2),
-    SNOWBOARDER_MEDIUM_DARK_SKIN_TONE(0x1F3C2),
-    SNOWBOARDER_DARK_SKIN_TONE(0x1F3C2),
-    PERSON_GOLFING(0x1F3CC),
-    PERSON_GOLFING_LIGHT_SKIN_TONE(0x1F3CC),
-    PERSON_GOLFING_MEDIUM_LIGHT_SKIN_TONE(0x1F3CC),
-    PERSON_GOLFING_MEDIUM_SKIN_TONE(0x1F3CC),
-    PERSON_GOLFING_MEDIUM_DARK_SKIN_TONE(0x1F3CC),
-    PERSON_GOLFING_DARK_SKIN_TONE(0x1F3CC),
-    MAN_GOLFING(0x1F3CC),
-    MAN_GOLFING_LIGHT_SKIN_TONE(0x1F3CC),
-    MAN_GOLFING_MEDIUM_LIGHT_SKIN_TONE(0x1F3CC),
-    MAN_GOLFING_MEDIUM_SKIN_TONE(0x1F3CC),
-    MAN_GOLFING_MEDIUM_DARK_SKIN_TONE(0x1F3CC),
-    MAN_GOLFING_DARK_SKIN_TONE(0x1F3CC),
-    WOMAN_GOLFING(0x1F3CC),
-    WOMAN_GOLFING_LIGHT_SKIN_TONE(0x1F3CC),
-    WOMAN_GOLFING_MEDIUM_LIGHT_SKIN_TONE(0x1F3CC),
-    WOMAN_GOLFING_MEDIUM_SKIN_TONE(0x1F3CC),
-    WOMAN_GOLFING_MEDIUM_DARK_SKIN_TONE(0x1F3CC),
-    WOMAN_GOLFING_DARK_SKIN_TONE(0x1F3CC),
-    PERSON_SURFING(0x1F3C4),
-    PERSON_SURFING_LIGHT_SKIN_TONE(0x1F3C4),
-    PERSON_SURFING_MEDIUM_LIGHT_SKIN_TONE(0x1F3C4),
-    PERSON_SURFING_MEDIUM_SKIN_TONE(0x1F3C4),
-    PERSON_SURFING_MEDIUM_DARK_SKIN_TONE(0x1F3C4),
-    PERSON_SURFING_DARK_SKIN_TONE(0x1F3C4),
-    MAN_SURFING(0x1F3C4),
-    MAN_SURFING_LIGHT_SKIN_TONE(0x1F3C4),
-    MAN_SURFING_MEDIUM_LIGHT_SKIN_TONE(0x1F3C4),
-    MAN_SURFING_MEDIUM_SKIN_TONE(0x1F3C4),
-    MAN_SURFING_MEDIUM_DARK_SKIN_TONE(0x1F3C4),
-    MAN_SURFING_DARK_SKIN_TONE(0x1F3C4),
-    WOMAN_SURFING(0x1F3C4),
-    WOMAN_SURFING_LIGHT_SKIN_TONE(0x1F3C4),
-    WOMAN_SURFING_MEDIUM_LIGHT_SKIN_TONE(0x1F3C4),
-    WOMAN_SURFING_MEDIUM_SKIN_TONE(0x1F3C4),
-    WOMAN_SURFING_MEDIUM_DARK_SKIN_TONE(0x1F3C4),
-    WOMAN_SURFING_DARK_SKIN_TONE(0x1F3C4),
-    PERSON_ROWING_BOAT(0x1F6A3),
-    PERSON_ROWING_BOAT_LIGHT_SKIN_TONE(0x1F6A3),
-    PERSON_ROWING_BOAT_MEDIUM_LIGHT_SKIN_TONE(0x1F6A3),
-    PERSON_ROWING_BOAT_MEDIUM_SKIN_TONE(0x1F6A3),
-    PERSON_ROWING_BOAT_MEDIUM_DARK_SKIN_TONE(0x1F6A3),
-    PERSON_ROWING_BOAT_DARK_SKIN_TONE(0x1F6A3),
-    MAN_ROWING_BOAT(0x1F6A3),
-    MAN_ROWING_BOAT_LIGHT_SKIN_TONE(0x1F6A3),
-    MAN_ROWING_BOAT_MEDIUM_LIGHT_SKIN_TONE(0x1F6A3),
-    MAN_ROWING_BOAT_MEDIUM_SKIN_TONE(0x1F6A3),
-    MAN_ROWING_BOAT_MEDIUM_DARK_SKIN_TONE(0x1F6A3),
-    MAN_ROWING_BOAT_DARK_SKIN_TONE(0x1F6A3),
-    WOMAN_ROWING_BOAT(0x1F6A3),
-    WOMAN_ROWING_BOAT_LIGHT_SKIN_TONE(0x1F6A3),
-    WOMAN_ROWING_BOAT_MEDIUM_LIGHT_SKIN_TONE(0x1F6A3),
-    WOMAN_ROWING_BOAT_MEDIUM_SKIN_TONE(0x1F6A3),
-    WOMAN_ROWING_BOAT_MEDIUM_DARK_SKIN_TONE(0x1F6A3),
-    WOMAN_ROWING_BOAT_DARK_SKIN_TONE(0x1F6A3),
-    PERSON_SWIMMING(0x1F3CA),
-    PERSON_SWIMMING_LIGHT_SKIN_TONE(0x1F3CA),
-    PERSON_SWIMMING_MEDIUM_LIGHT_SKIN_TONE(0x1F3CA),
-    PERSON_SWIMMING_MEDIUM_SKIN_TONE(0x1F3CA),
-    PERSON_SWIMMING_MEDIUM_DARK_SKIN_TONE(0x1F3CA),
-    PERSON_SWIMMING_DARK_SKIN_TONE(0x1F3CA),
-    MAN_SWIMMING(0x1F3CA),
-    MAN_SWIMMING_LIGHT_SKIN_TONE(0x1F3CA),
-    MAN_SWIMMING_MEDIUM_LIGHT_SKIN_TONE(0x1F3CA),
-    MAN_SWIMMING_MEDIUM_SKIN_TONE(0x1F3CA),
-    MAN_SWIMMING_MEDIUM_DARK_SKIN_TONE(0x1F3CA),
-    MAN_SWIMMING_DARK_SKIN_TONE(0x1F3CA),
-    WOMAN_SWIMMING(0x1F3CA),
-    WOMAN_SWIMMING_LIGHT_SKIN_TONE(0x1F3CA),
-    WOMAN_SWIMMING_MEDIUM_LIGHT_SKIN_TONE(0x1F3CA),
-    WOMAN_SWIMMING_MEDIUM_SKIN_TONE(0x1F3CA),
-    WOMAN_SWIMMING_MEDIUM_DARK_SKIN_TONE(0x1F3CA),
-    WOMAN_SWIMMING_DARK_SKIN_TONE(0x1F3CA),
-    PERSON_BOUNCING_BALL(0x26F9),
-    PERSON_BOUNCING_BALL_LIGHT_SKIN_TONE(0x26F9),
-    PERSON_BOUNCING_BALL_MEDIUM_LIGHT_SKIN_TONE(0x26F9),
-    PERSON_BOUNCING_BALL_MEDIUM_SKIN_TONE(0x26F9),
-    PERSON_BOUNCING_BALL_MEDIUM_DARK_SKIN_TONE(0x26F9),
-    PERSON_BOUNCING_BALL_DARK_SKIN_TONE(0x26F9),
-    MAN_BOUNCING_BALL(0x26F9),
-    MAN_BOUNCING_BALL_LIGHT_SKIN_TONE(0x26F9),
-    MAN_BOUNCING_BALL_MEDIUM_LIGHT_SKIN_TONE(0x26F9),
-    MAN_BOUNCING_BALL_MEDIUM_SKIN_TONE(0x26F9),
-    MAN_BOUNCING_BALL_MEDIUM_DARK_SKIN_TONE(0x26F9),
-    MAN_BOUNCING_BALL_DARK_SKIN_TONE(0x26F9),
-    WOMAN_BOUNCING_BALL(0x26F9),
-    WOMAN_BOUNCING_BALL_LIGHT_SKIN_TONE(0x26F9),
-    WOMAN_BOUNCING_BALL_MEDIUM_LIGHT_SKIN_TONE(0x26F9),
-    WOMAN_BOUNCING_BALL_MEDIUM_SKIN_TONE(0x26F9),
-    WOMAN_BOUNCING_BALL_MEDIUM_DARK_SKIN_TONE(0x26F9),
-    WOMAN_BOUNCING_BALL_DARK_SKIN_TONE(0x26F9),
-    PERSON_LIFTING_WEIGHTS(0x1F3CB),
-    PERSON_LIFTING_WEIGHTS_LIGHT_SKIN_TONE(0x1F3CB),
-    PERSON_LIFTING_WEIGHTS_MEDIUM_LIGHT_SKIN_TONE(0x1F3CB),
-    PERSON_LIFTING_WEIGHTS_MEDIUM_SKIN_TONE(0x1F3CB),
-    PERSON_LIFTING_WEIGHTS_MEDIUM_DARK_SKIN_TONE(0x1F3CB),
-    PERSON_LIFTING_WEIGHTS_DARK_SKIN_TONE(0x1F3CB),
-    MAN_LIFTING_WEIGHTS(0x1F3CB),
-    MAN_LIFTING_WEIGHTS_LIGHT_SKIN_TONE(0x1F3CB),
-    MAN_LIFTING_WEIGHTS_MEDIUM_LIGHT_SKIN_TONE(0x1F3CB),
-    MAN_LIFTING_WEIGHTS_MEDIUM_SKIN_TONE(0x1F3CB),
-    MAN_LIFTING_WEIGHTS_MEDIUM_DARK_SKIN_TONE(0x1F3CB),
-    MAN_LIFTING_WEIGHTS_DARK_SKIN_TONE(0x1F3CB),
-    WOMAN_LIFTING_WEIGHTS(0x1F3CB),
-    WOMAN_LIFTING_WEIGHTS_LIGHT_SKIN_TONE(0x1F3CB),
-    WOMAN_LIFTING_WEIGHTS_MEDIUM_LIGHT_SKIN_TONE(0x1F3CB),
-    WOMAN_LIFTING_WEIGHTS_MEDIUM_SKIN_TONE(0x1F3CB),
-    WOMAN_LIFTING_WEIGHTS_MEDIUM_DARK_SKIN_TONE(0x1F3CB),
-    WOMAN_LIFTING_WEIGHTS_DARK_SKIN_TONE(0x1F3CB),
-    PERSON_BIKING(0x1F6B4),
-    PERSON_BIKING_LIGHT_SKIN_TONE(0x1F6B4),
-    PERSON_BIKING_MEDIUM_LIGHT_SKIN_TONE(0x1F6B4),
-    PERSON_BIKING_MEDIUM_SKIN_TONE(0x1F6B4),
-    PERSON_BIKING_MEDIUM_DARK_SKIN_TONE(0x1F6B4),
-    PERSON_BIKING_DARK_SKIN_TONE(0x1F6B4),
-    MAN_BIKING(0x1F6B4),
-    MAN_BIKING_LIGHT_SKIN_TONE(0x1F6B4),
-    MAN_BIKING_MEDIUM_LIGHT_SKIN_TONE(0x1F6B4),
-    MAN_BIKING_MEDIUM_SKIN_TONE(0x1F6B4),
-    MAN_BIKING_MEDIUM_DARK_SKIN_TONE(0x1F6B4),
-    MAN_BIKING_DARK_SKIN_TONE(0x1F6B4),
-    WOMAN_BIKING(0x1F6B4),
-    WOMAN_BIKING_LIGHT_SKIN_TONE(0x1F6B4),
-    WOMAN_BIKING_MEDIUM_LIGHT_SKIN_TONE(0x1F6B4),
-    WOMAN_BIKING_MEDIUM_SKIN_TONE(0x1F6B4),
-    WOMAN_BIKING_MEDIUM_DARK_SKIN_TONE(0x1F6B4),
-    WOMAN_BIKING_DARK_SKIN_TONE(0x1F6B4),
-    PERSON_MOUNTAIN_BIKING(0x1F6B5),
-    PERSON_MOUNTAIN_BIKING_LIGHT_SKIN_TONE(0x1F6B5),
-    PERSON_MOUNTAIN_BIKING_MEDIUM_LIGHT_SKIN_TONE(0x1F6B5),
-    PERSON_MOUNTAIN_BIKING_MEDIUM_SKIN_TONE(0x1F6B5),
-    PERSON_MOUNTAIN_BIKING_MEDIUM_DARK_SKIN_TONE(0x1F6B5),
-    PERSON_MOUNTAIN_BIKING_DARK_SKIN_TONE(0x1F6B5),
-    MAN_MOUNTAIN_BIKING(0x1F6B5),
-    MAN_MOUNTAIN_BIKING_LIGHT_SKIN_TONE(0x1F6B5),
-    MAN_MOUNTAIN_BIKING_MEDIUM_LIGHT_SKIN_TONE(0x1F6B5),
-    MAN_MOUNTAIN_BIKING_MEDIUM_SKIN_TONE(0x1F6B5),
-    MAN_MOUNTAIN_BIKING_MEDIUM_DARK_SKIN_TONE(0x1F6B5),
-    MAN_MOUNTAIN_BIKING_DARK_SKIN_TONE(0x1F6B5),
-    WOMAN_MOUNTAIN_BIKING(0x1F6B5),
-    WOMAN_MOUNTAIN_BIKING_LIGHT_SKIN_TONE(0x1F6B5),
-    WOMAN_MOUNTAIN_BIKING_MEDIUM_LIGHT_SKIN_TONE(0x1F6B5),
-    WOMAN_MOUNTAIN_BIKING_MEDIUM_SKIN_TONE(0x1F6B5),
-    WOMAN_MOUNTAIN_BIKING_MEDIUM_DARK_SKIN_TONE(0x1F6B5),
-    WOMAN_MOUNTAIN_BIKING_DARK_SKIN_TONE(0x1F6B5),
-    PERSON_CARTWHEELING(0x1F938),
-    PERSON_CARTWHEELING_LIGHT_SKIN_TONE(0x1F938),
-    PERSON_CARTWHEELING_MEDIUM_LIGHT_SKIN_TONE(0x1F938),
-    PERSON_CARTWHEELING_MEDIUM_SKIN_TONE(0x1F938),
-    PERSON_CARTWHEELING_MEDIUM_DARK_SKIN_TONE(0x1F938),
-    PERSON_CARTWHEELING_DARK_SKIN_TONE(0x1F938),
-    MAN_CARTWHEELING(0x1F938),
-    MAN_CARTWHEELING_LIGHT_SKIN_TONE(0x1F938),
-    MAN_CARTWHEELING_MEDIUM_LIGHT_SKIN_TONE(0x1F938),
-    MAN_CARTWHEELING_MEDIUM_SKIN_TONE(0x1F938),
-    MAN_CARTWHEELING_MEDIUM_DARK_SKIN_TONE(0x1F938),
-    MAN_CARTWHEELING_DARK_SKIN_TONE(0x1F938),
-    WOMAN_CARTWHEELING(0x1F938),
-    WOMAN_CARTWHEELING_LIGHT_SKIN_TONE(0x1F938),
-    WOMAN_CARTWHEELING_MEDIUM_LIGHT_SKIN_TONE(0x1F938),
-    WOMAN_CARTWHEELING_MEDIUM_SKIN_TONE(0x1F938),
-    WOMAN_CARTWHEELING_MEDIUM_DARK_SKIN_TONE(0x1F938),
-    WOMAN_CARTWHEELING_DARK_SKIN_TONE(0x1F938),
-    PEOPLE_WRESTLING(0x1F93C),
-    MEN_WRESTLING(0x1F93C),
-    WOMEN_WRESTLING(0x1F93C),
-    PERSON_PLAYING_WATER_POLO(0x1F93D),
-    PERSON_PLAYING_WATER_POLO_LIGHT_SKIN_TONE(0x1F93D),
-    PERSON_PLAYING_WATER_POLO_MEDIUM_LIGHT_SKIN_TONE(0x1F93D),
-    PERSON_PLAYING_WATER_POLO_MEDIUM_SKIN_TONE(0x1F93D),
-    PERSON_PLAYING_WATER_POLO_MEDIUM_DARK_SKIN_TONE(0x1F93D),
-    PERSON_PLAYING_WATER_POLO_DARK_SKIN_TONE(0x1F93D),
-    MAN_PLAYING_WATER_POLO(0x1F93D),
-    MAN_PLAYING_WATER_POLO_LIGHT_SKIN_TONE(0x1F93D),
-    MAN_PLAYING_WATER_POLO_MEDIUM_LIGHT_SKIN_TONE(0x1F93D),
-    MAN_PLAYING_WATER_POLO_MEDIUM_SKIN_TONE(0x1F93D),
-    MAN_PLAYING_WATER_POLO_MEDIUM_DARK_SKIN_TONE(0x1F93D),
-    MAN_PLAYING_WATER_POLO_DARK_SKIN_TONE(0x1F93D),
-    WOMAN_PLAYING_WATER_POLO(0x1F93D),
-    WOMAN_PLAYING_WATER_POLO_LIGHT_SKIN_TONE(0x1F93D),
-    WOMAN_PLAYING_WATER_POLO_MEDIUM_LIGHT_SKIN_TONE(0x1F93D),
-    WOMAN_PLAYING_WATER_POLO_MEDIUM_SKIN_TONE(0x1F93D),
-    WOMAN_PLAYING_WATER_POLO_MEDIUM_DARK_SKIN_TONE(0x1F93D),
-    WOMAN_PLAYING_WATER_POLO_DARK_SKIN_TONE(0x1F93D),
-    PERSON_PLAYING_HANDBALL(0x1F93E),
-    PERSON_PLAYING_HANDBALL_LIGHT_SKIN_TONE(0x1F93E),
-    PERSON_PLAYING_HANDBALL_MEDIUM_LIGHT_SKIN_TONE(0x1F93E),
-    PERSON_PLAYING_HANDBALL_MEDIUM_SKIN_TONE(0x1F93E),
-    PERSON_PLAYING_HANDBALL_MEDIUM_DARK_SKIN_TONE(0x1F93E),
-    PERSON_PLAYING_HANDBALL_DARK_SKIN_TONE(0x1F93E),
-    MAN_PLAYING_HANDBALL(0x1F93E),
-    MAN_PLAYING_HANDBALL_LIGHT_SKIN_TONE(0x1F93E),
-    MAN_PLAYING_HANDBALL_MEDIUM_LIGHT_SKIN_TONE(0x1F93E),
-    MAN_PLAYING_HANDBALL_MEDIUM_SKIN_TONE(0x1F93E),
-    MAN_PLAYING_HANDBALL_MEDIUM_DARK_SKIN_TONE(0x1F93E),
-    MAN_PLAYING_HANDBALL_DARK_SKIN_TONE(0x1F93E),
-    WOMAN_PLAYING_HANDBALL(0x1F93E),
-    WOMAN_PLAYING_HANDBALL_LIGHT_SKIN_TONE(0x1F93E),
-    WOMAN_PLAYING_HANDBALL_MEDIUM_LIGHT_SKIN_TONE(0x1F93E),
-    WOMAN_PLAYING_HANDBALL_MEDIUM_SKIN_TONE(0x1F93E),
-    WOMAN_PLAYING_HANDBALL_MEDIUM_DARK_SKIN_TONE(0x1F93E),
-    WOMAN_PLAYING_HANDBALL_DARK_SKIN_TONE(0x1F93E),
-    PERSON_JUGGLING(0x1F939),
-    PERSON_JUGGLING_LIGHT_SKIN_TONE(0x1F939),
-    PERSON_JUGGLING_MEDIUM_LIGHT_SKIN_TONE(0x1F939),
-    PERSON_JUGGLING_MEDIUM_SKIN_TONE(0x1F939),
-    PERSON_JUGGLING_MEDIUM_DARK_SKIN_TONE(0x1F939),
-    PERSON_JUGGLING_DARK_SKIN_TONE(0x1F939),
-    MAN_JUGGLING(0x1F939),
-    MAN_JUGGLING_LIGHT_SKIN_TONE(0x1F939),
-    MAN_JUGGLING_MEDIUM_LIGHT_SKIN_TONE(0x1F939),
-    MAN_JUGGLING_MEDIUM_SKIN_TONE(0x1F939),
-    MAN_JUGGLING_MEDIUM_DARK_SKIN_TONE(0x1F939),
-    MAN_JUGGLING_DARK_SKIN_TONE(0x1F939),
-    WOMAN_JUGGLING(0x1F939),
-    WOMAN_JUGGLING_LIGHT_SKIN_TONE(0x1F939),
-    WOMAN_JUGGLING_MEDIUM_LIGHT_SKIN_TONE(0x1F939),
-    WOMAN_JUGGLING_MEDIUM_SKIN_TONE(0x1F939),
-    WOMAN_JUGGLING_MEDIUM_DARK_SKIN_TONE(0x1F939),
-    WOMAN_JUGGLING_DARK_SKIN_TONE(0x1F939),
-    PERSON_IN_LOTUS_POSITION(0x1F9D8),
-    PERSON_IN_LOTUS_POSITION_LIGHT_SKIN_TONE(0x1F9D8),
-    PERSON_IN_LOTUS_POSITION_MEDIUM_LIGHT_SKIN_TONE(0x1F9D8),
-    PERSON_IN_LOTUS_POSITION_MEDIUM_SKIN_TONE(0x1F9D8),
-    PERSON_IN_LOTUS_POSITION_MEDIUM_DARK_SKIN_TONE(0x1F9D8),
-    PERSON_IN_LOTUS_POSITION_DARK_SKIN_TONE(0x1F9D8),
-    MAN_IN_LOTUS_POSITION(0x1F9D8),
-    MAN_IN_LOTUS_POSITION_LIGHT_SKIN_TONE(0x1F9D8),
-    MAN_IN_LOTUS_POSITION_MEDIUM_LIGHT_SKIN_TONE(0x1F9D8),
-    MAN_IN_LOTUS_POSITION_MEDIUM_SKIN_TONE(0x1F9D8),
-    MAN_IN_LOTUS_POSITION_MEDIUM_DARK_SKIN_TONE(0x1F9D8),
-    MAN_IN_LOTUS_POSITION_DARK_SKIN_TONE(0x1F9D8),
-    WOMAN_IN_LOTUS_POSITION(0x1F9D8),
-    WOMAN_IN_LOTUS_POSITION_LIGHT_SKIN_TONE(0x1F9D8),
-    WOMAN_IN_LOTUS_POSITION_MEDIUM_LIGHT_SKIN_TONE(0x1F9D8),
-    WOMAN_IN_LOTUS_POSITION_MEDIUM_SKIN_TONE(0x1F9D8),
-    WOMAN_IN_LOTUS_POSITION_MEDIUM_DARK_SKIN_TONE(0x1F9D8),
-    WOMAN_IN_LOTUS_POSITION_DARK_SKIN_TONE(0x1F9D8),
-    PERSON_TAKING_BATH(0x1F6C0),
-    PERSON_TAKING_BATH_LIGHT_SKIN_TONE(0x1F6C0),
-    PERSON_TAKING_BATH_MEDIUM_LIGHT_SKIN_TONE(0x1F6C0),
-    PERSON_TAKING_BATH_MEDIUM_SKIN_TONE(0x1F6C0),
-    PERSON_TAKING_BATH_MEDIUM_DARK_SKIN_TONE(0x1F6C0),
-    PERSON_TAKING_BATH_DARK_SKIN_TONE(0x1F6C0),
-    PERSON_IN_BED(0x1F6CC),
-    PERSON_IN_BED_LIGHT_SKIN_TONE(0x1F6CC),
-    PERSON_IN_BED_MEDIUM_LIGHT_SKIN_TONE(0x1F6CC),
-    PERSON_IN_BED_MEDIUM_SKIN_TONE(0x1F6CC),
-    PERSON_IN_BED_MEDIUM_DARK_SKIN_TONE(0x1F6CC),
-    PERSON_IN_BED_DARK_SKIN_TONE(0x1F6CC),
-    PEOPLE_HOLDING_HANDS(0x1F9D1),
-    PEOPLE_HOLDING_HANDS_LIGHT_SKIN_TONE(0x1F9D1),
-    PEOPLE_HOLDING_HANDS_LIGHT_SKIN_TONE_MEDIUM_LIGHT_SKIN_TONE(0x1F9D1),
-    PEOPLE_HOLDING_HANDS_LIGHT_SKIN_TONE_MEDIUM_SKIN_TONE(0x1F9D1),
-    PEOPLE_HOLDING_HANDS_LIGHT_SKIN_TONE_MEDIUM_DARK_SKIN_TONE(0x1F9D1),
-    PEOPLE_HOLDING_HANDS_LIGHT_SKIN_TONE_DARK_SKIN_TONE(0x1F9D1),
-    PEOPLE_HOLDING_HANDS_MEDIUM_LIGHT_SKIN_TONE_LIGHT_SKIN_TONE(0x1F9D1),
-    PEOPLE_HOLDING_HANDS_MEDIUM_LIGHT_SKIN_TONE(0x1F9D1),
-    PEOPLE_HOLDING_HANDS_MEDIUM_LIGHT_SKIN_TONE_MEDIUM_SKIN_TONE(0x1F9D1),
-    PEOPLE_HOLDING_HANDS_MEDIUM_LIGHT_SKIN_TONE_MEDIUM_DARK_SKIN_TONE(0x1F9D1),
-    PEOPLE_HOLDING_HANDS_MEDIUM_LIGHT_SKIN_TONE_DARK_SKIN_TONE(0x1F9D1),
-    PEOPLE_HOLDING_HANDS_MEDIUM_SKIN_TONE_LIGHT_SKIN_TONE(0x1F9D1),
-    PEOPLE_HOLDING_HANDS_MEDIUM_SKIN_TONE_MEDIUM_LIGHT_SKIN_TONE(0x1F9D1),
-    PEOPLE_HOLDING_HANDS_MEDIUM_SKIN_TONE(0x1F9D1),
-    PEOPLE_HOLDING_HANDS_MEDIUM_SKIN_TONE_MEDIUM_DARK_SKIN_TONE(0x1F9D1),
-    PEOPLE_HOLDING_HANDS_MEDIUM_SKIN_TONE_DARK_SKIN_TONE(0x1F9D1),
-    PEOPLE_HOLDING_HANDS_MEDIUM_DARK_SKIN_TONE_LIGHT_SKIN_TONE(0x1F9D1),
-    PEOPLE_HOLDING_HANDS_MEDIUM_DARK_SKIN_TONE_MEDIUM_LIGHT_SKIN_TONE(0x1F9D1),
-    PEOPLE_HOLDING_HANDS_MEDIUM_DARK_SKIN_TONE_MEDIUM_SKIN_TONE(0x1F9D1),
-    PEOPLE_HOLDING_HANDS_MEDIUM_DARK_SKIN_TONE(0x1F9D1),
-    PEOPLE_HOLDING_HANDS_MEDIUM_DARK_SKIN_TONE_DARK_SKIN_TONE(0x1F9D1),
-    PEOPLE_HOLDING_HANDS_DARK_SKIN_TONE_LIGHT_SKIN_TONE(0x1F9D1),
-    PEOPLE_HOLDING_HANDS_DARK_SKIN_TONE_MEDIUM_LIGHT_SKIN_TONE(0x1F9D1),
-    PEOPLE_HOLDING_HANDS_DARK_SKIN_TONE_MEDIUM_SKIN_TONE(0x1F9D1),
-    PEOPLE_HOLDING_HANDS_DARK_SKIN_TONE_MEDIUM_DARK_SKIN_TONE(0x1F9D1),
-    PEOPLE_HOLDING_HANDS_DARK_SKIN_TONE(0x1F9D1),
-    WOMEN_HOLDING_HANDS(0x1F46D),
-    WOMEN_HOLDING_HANDS_LIGHT_SKIN_TONE(0x1F46D),
-    WOMEN_HOLDING_HANDS_LIGHT_SKIN_TONE_MEDIUM_LIGHT_SKIN_TONE(0x1F469),
-    WOMEN_HOLDING_HANDS_LIGHT_SKIN_TONE_MEDIUM_SKIN_TONE(0x1F469),
-    WOMEN_HOLDING_HANDS_LIGHT_SKIN_TONE_MEDIUM_DARK_SKIN_TONE(0x1F469),
-    WOMEN_HOLDING_HANDS_LIGHT_SKIN_TONE_DARK_SKIN_TONE(0x1F469),
-    WOMEN_HOLDING_HANDS_MEDIUM_LIGHT_SKIN_TONE_LIGHT_SKIN_TONE(0x1F469),
-    WOMEN_HOLDING_HANDS_MEDIUM_LIGHT_SKIN_TONE(0x1F46D),
-    WOMEN_HOLDING_HANDS_MEDIUM_LIGHT_SKIN_TONE_MEDIUM_SKIN_TONE(0x1F469),
-    WOMEN_HOLDING_HANDS_MEDIUM_LIGHT_SKIN_TONE_MEDIUM_DARK_SKIN_TONE(0x1F469),
-    WOMEN_HOLDING_HANDS_MEDIUM_LIGHT_SKIN_TONE_DARK_SKIN_TONE(0x1F469),
-    WOMEN_HOLDING_HANDS_MEDIUM_SKIN_TONE_LIGHT_SKIN_TONE(0x1F469),
-    WOMEN_HOLDING_HANDS_MEDIUM_SKIN_TONE_MEDIUM_LIGHT_SKIN_TONE(0x1F469),
-    WOMEN_HOLDING_HANDS_MEDIUM_SKIN_TONE(0x1F46D),
-    WOMEN_HOLDING_HANDS_MEDIUM_SKIN_TONE_MEDIUM_DARK_SKIN_TONE(0x1F469),
-    WOMEN_HOLDING_HANDS_MEDIUM_SKIN_TONE_DARK_SKIN_TONE(0x1F469),
-    WOMEN_HOLDING_HANDS_MEDIUM_DARK_SKIN_TONE_LIGHT_SKIN_TONE(0x1F469),
-    WOMEN_HOLDING_HANDS_MEDIUM_DARK_SKIN_TONE_MEDIUM_LIGHT_SKIN_TONE(0x1F469),
-    WOMEN_HOLDING_HANDS_MEDIUM_DARK_SKIN_TONE_MEDIUM_SKIN_TONE(0x1F469),
-    WOMEN_HOLDING_HANDS_MEDIUM_DARK_SKIN_TONE(0x1F46D),
-    WOMEN_HOLDING_HANDS_MEDIUM_DARK_SKIN_TONE_DARK_SKIN_TONE(0x1F469),
-    WOMEN_HOLDING_HANDS_DARK_SKIN_TONE_LIGHT_SKIN_TONE(0x1F469),
-    WOMEN_HOLDING_HANDS_DARK_SKIN_TONE_MEDIUM_LIGHT_SKIN_TONE(0x1F469),
-    WOMEN_HOLDING_HANDS_DARK_SKIN_TONE_MEDIUM_SKIN_TONE(0x1F469),
-    WOMEN_HOLDING_HANDS_DARK_SKIN_TONE_MEDIUM_DARK_SKIN_TONE(0x1F469),
-    WOMEN_HOLDING_HANDS_DARK_SKIN_TONE(0x1F46D),
-    WOMAN_AND_MAN_HOLDING_HANDS(0x1F46B),
-    WOMAN_AND_MAN_HOLDING_HANDS_LIGHT_SKIN_TONE(0x1F46B),
-    WOMAN_AND_MAN_HOLDING_HANDS_LIGHT_SKIN_TONE_MEDIUM_LIGHT_SKIN_TONE(0x1F469),
-    WOMAN_AND_MAN_HOLDING_HANDS_LIGHT_SKIN_TONE_MEDIUM_SKIN_TONE(0x1F469),
-    WOMAN_AND_MAN_HOLDING_HANDS_LIGHT_SKIN_TONE_MEDIUM_DARK_SKIN_TONE(0x1F469),
-    WOMAN_AND_MAN_HOLDING_HANDS_LIGHT_SKIN_TONE_DARK_SKIN_TONE(0x1F469),
-    WOMAN_AND_MAN_HOLDING_HANDS_MEDIUM_LIGHT_SKIN_TONE_LIGHT_SKIN_TONE(0x1F469),
-    WOMAN_AND_MAN_HOLDING_HANDS_MEDIUM_LIGHT_SKIN_TONE(0x1F46B),
-    WOMAN_AND_MAN_HOLDING_HANDS_MEDIUM_LIGHT_SKIN_TONE_MEDIUM_SKIN_TONE(0x1F469),
-    WOMAN_AND_MAN_HOLDING_HANDS_MEDIUM_LIGHT_SKIN_TONE_MEDIUM_DARK_SKIN_TONE(0x1F469),
-    WOMAN_AND_MAN_HOLDING_HANDS_MEDIUM_LIGHT_SKIN_TONE_DARK_SKIN_TONE(0x1F469),
-    WOMAN_AND_MAN_HOLDING_HANDS_MEDIUM_SKIN_TONE_LIGHT_SKIN_TONE(0x1F469),
-    WOMAN_AND_MAN_HOLDING_HANDS_MEDIUM_SKIN_TONE_MEDIUM_LIGHT_SKIN_TONE(0x1F469),
-    WOMAN_AND_MAN_HOLDING_HANDS_MEDIUM_SKIN_TONE(0x1F46B),
-    WOMAN_AND_MAN_HOLDING_HANDS_MEDIUM_SKIN_TONE_MEDIUM_DARK_SKIN_TONE(0x1F469),
-    WOMAN_AND_MAN_HOLDING_HANDS_MEDIUM_SKIN_TONE_DARK_SKIN_TONE(0x1F469),
-    WOMAN_AND_MAN_HOLDING_HANDS_MEDIUM_DARK_SKIN_TONE_LIGHT_SKIN_TONE(0x1F469),
-    WOMAN_AND_MAN_HOLDING_HANDS_MEDIUM_DARK_SKIN_TONE_MEDIUM_LIGHT_SKIN_TONE(0x1F469),
-    WOMAN_AND_MAN_HOLDING_HANDS_MEDIUM_DARK_SKIN_TONE_MEDIUM_SKIN_TONE(0x1F469),
-    WOMAN_AND_MAN_HOLDING_HANDS_MEDIUM_DARK_SKIN_TONE(0x1F46B),
-    WOMAN_AND_MAN_HOLDING_HANDS_MEDIUM_DARK_SKIN_TONE_DARK_SKIN_TONE(0x1F469),
-    WOMAN_AND_MAN_HOLDING_HANDS_DARK_SKIN_TONE_LIGHT_SKIN_TONE(0x1F469),
-    WOMAN_AND_MAN_HOLDING_HANDS_DARK_SKIN_TONE_MEDIUM_LIGHT_SKIN_TONE(0x1F469),
-    WOMAN_AND_MAN_HOLDING_HANDS_DARK_SKIN_TONE_MEDIUM_SKIN_TONE(0x1F469),
-    WOMAN_AND_MAN_HOLDING_HANDS_DARK_SKIN_TONE_MEDIUM_DARK_SKIN_TONE(0x1F469),
-    WOMAN_AND_MAN_HOLDING_HANDS_DARK_SKIN_TONE(0x1F46B),
-    MEN_HOLDING_HANDS(0x1F46C),
-    MEN_HOLDING_HANDS_LIGHT_SKIN_TONE(0x1F46C),
-    MEN_HOLDING_HANDS_LIGHT_SKIN_TONE_MEDIUM_LIGHT_SKIN_TONE(0x1F468),
-    MEN_HOLDING_HANDS_LIGHT_SKIN_TONE_MEDIUM_SKIN_TONE(0x1F468),
-    MEN_HOLDING_HANDS_LIGHT_SKIN_TONE_MEDIUM_DARK_SKIN_TONE(0x1F468),
-    MEN_HOLDING_HANDS_LIGHT_SKIN_TONE_DARK_SKIN_TONE(0x1F468),
-    MEN_HOLDING_HANDS_MEDIUM_LIGHT_SKIN_TONE_LIGHT_SKIN_TONE(0x1F468),
-    MEN_HOLDING_HANDS_MEDIUM_LIGHT_SKIN_TONE(0x1F46C),
-    MEN_HOLDING_HANDS_MEDIUM_LIGHT_SKIN_TONE_MEDIUM_SKIN_TONE(0x1F468),
-    MEN_HOLDING_HANDS_MEDIUM_LIGHT_SKIN_TONE_MEDIUM_DARK_SKIN_TONE(0x1F468),
-    MEN_HOLDING_HANDS_MEDIUM_LIGHT_SKIN_TONE_DARK_SKIN_TONE(0x1F468),
-    MEN_HOLDING_HANDS_MEDIUM_SKIN_TONE_LIGHT_SKIN_TONE(0x1F468),
-    MEN_HOLDING_HANDS_MEDIUM_SKIN_TONE_MEDIUM_LIGHT_SKIN_TONE(0x1F468),
-    MEN_HOLDING_HANDS_MEDIUM_SKIN_TONE(0x1F46C),
-    MEN_HOLDING_HANDS_MEDIUM_SKIN_TONE_MEDIUM_DARK_SKIN_TONE(0x1F468),
-    MEN_HOLDING_HANDS_MEDIUM_SKIN_TONE_DARK_SKIN_TONE(0x1F468),
-    MEN_HOLDING_HANDS_MEDIUM_DARK_SKIN_TONE_LIGHT_SKIN_TONE(0x1F468),
-    MEN_HOLDING_HANDS_MEDIUM_DARK_SKIN_TONE_MEDIUM_LIGHT_SKIN_TONE(0x1F468),
-    MEN_HOLDING_HANDS_MEDIUM_DARK_SKIN_TONE_MEDIUM_SKIN_TONE(0x1F468),
-    MEN_HOLDING_HANDS_MEDIUM_DARK_SKIN_TONE(0x1F46C),
-    MEN_HOLDING_HANDS_MEDIUM_DARK_SKIN_TONE_DARK_SKIN_TONE(0x1F468),
-    MEN_HOLDING_HANDS_DARK_SKIN_TONE_LIGHT_SKIN_TONE(0x1F468),
-    MEN_HOLDING_HANDS_DARK_SKIN_TONE_MEDIUM_LIGHT_SKIN_TONE(0x1F468),
-    MEN_HOLDING_HANDS_DARK_SKIN_TONE_MEDIUM_SKIN_TONE(0x1F468),
-    MEN_HOLDING_HANDS_DARK_SKIN_TONE_MEDIUM_DARK_SKIN_TONE(0x1F468),
-    MEN_HOLDING_HANDS_DARK_SKIN_TONE(0x1F46C),
-    KISS(0x1F48F),
-    KISS_LIGHT_SKIN_TONE(0x1F48F),
-    KISS_MEDIUM_LIGHT_SKIN_TONE(0x1F48F),
-    KISS_MEDIUM_SKIN_TONE(0x1F48F),
-    KISS_MEDIUM_DARK_SKIN_TONE(0x1F48F),
-    KISS_DARK_SKIN_TONE(0x1F48F),
-    KISS_PERSON_PERSON_LIGHT_SKIN_TONE_MEDIUM_LIGHT_SKIN_TONE(0x1F9D1),
-    KISS_PERSON_PERSON_LIGHT_SKIN_TONE_MEDIUM_SKIN_TONE(0x1F9D1),
-    KISS_PERSON_PERSON_LIGHT_SKIN_TONE_MEDIUM_DARK_SKIN_TONE(0x1F9D1),
-    KISS_PERSON_PERSON_LIGHT_SKIN_TONE_DARK_SKIN_TONE(0x1F9D1),
-    KISS_PERSON_PERSON_MEDIUM_LIGHT_SKIN_TONE_LIGHT_SKIN_TONE(0x1F9D1),
-    KISS_PERSON_PERSON_MEDIUM_LIGHT_SKIN_TONE_MEDIUM_SKIN_TONE(0x1F9D1),
-    KISS_PERSON_PERSON_MEDIUM_LIGHT_SKIN_TONE_MEDIUM_DARK_SKIN_TONE(0x1F9D1),
-    KISS_PERSON_PERSON_MEDIUM_LIGHT_SKIN_TONE_DARK_SKIN_TONE(0x1F9D1),
-    KISS_PERSON_PERSON_MEDIUM_SKIN_TONE_LIGHT_SKIN_TONE(0x1F9D1),
-    KISS_PERSON_PERSON_MEDIUM_SKIN_TONE_MEDIUM_LIGHT_SKIN_TONE(0x1F9D1),
-    KISS_PERSON_PERSON_MEDIUM_SKIN_TONE_MEDIUM_DARK_SKIN_TONE(0x1F9D1),
-    KISS_PERSON_PERSON_MEDIUM_SKIN_TONE_DARK_SKIN_TONE(0x1F9D1),
-    KISS_PERSON_PERSON_MEDIUM_DARK_SKIN_TONE_LIGHT_SKIN_TONE(0x1F9D1),
-    KISS_PERSON_PERSON_MEDIUM_DARK_SKIN_TONE_MEDIUM_LIGHT_SKIN_TONE(0x1F9D1),
-    KISS_PERSON_PERSON_MEDIUM_DARK_SKIN_TONE_MEDIUM_SKIN_TONE(0x1F9D1),
-    KISS_PERSON_PERSON_MEDIUM_DARK_SKIN_TONE_DARK_SKIN_TONE(0x1F9D1),
-    KISS_PERSON_PERSON_DARK_SKIN_TONE_LIGHT_SKIN_TONE(0x1F9D1),
-    KISS_PERSON_PERSON_DARK_SKIN_TONE_MEDIUM_LIGHT_SKIN_TONE(0x1F9D1),
-    KISS_PERSON_PERSON_DARK_SKIN_TONE_MEDIUM_SKIN_TONE(0x1F9D1),
-    KISS_PERSON_PERSON_DARK_SKIN_TONE_MEDIUM_DARK_SKIN_TONE(0x1F9D1),
-    KISS_WOMAN_MAN(0x1F469),
-    KISS_WOMAN_MAN_LIGHT_SKIN_TONE(0x1F469),
-    KISS_WOMAN_MAN_LIGHT_SKIN_TONE_MEDIUM_LIGHT_SKIN_TONE(0x1F469),
-    KISS_WOMAN_MAN_LIGHT_SKIN_TONE_MEDIUM_SKIN_TONE(0x1F469),
-    KISS_WOMAN_MAN_LIGHT_SKIN_TONE_MEDIUM_DARK_SKIN_TONE(0x1F469),
-    KISS_WOMAN_MAN_LIGHT_SKIN_TONE_DARK_SKIN_TONE(0x1F469),
-    KISS_WOMAN_MAN_MEDIUM_LIGHT_SKIN_TONE_LIGHT_SKIN_TONE(0x1F469),
-    KISS_WOMAN_MAN_MEDIUM_LIGHT_SKIN_TONE(0x1F469),
-    KISS_WOMAN_MAN_MEDIUM_LIGHT_SKIN_TONE_MEDIUM_SKIN_TONE(0x1F469),
-    KISS_WOMAN_MAN_MEDIUM_LIGHT_SKIN_TONE_MEDIUM_DARK_SKIN_TONE(0x1F469),
-    KISS_WOMAN_MAN_MEDIUM_LIGHT_SKIN_TONE_DARK_SKIN_TONE(0x1F469),
-    KISS_WOMAN_MAN_MEDIUM_SKIN_TONE_LIGHT_SKIN_TONE(0x1F469),
-    KISS_WOMAN_MAN_MEDIUM_SKIN_TONE_MEDIUM_LIGHT_SKIN_TONE(0x1F469),
-    KISS_WOMAN_MAN_MEDIUM_SKIN_TONE(0x1F469),
-    KISS_WOMAN_MAN_MEDIUM_SKIN_TONE_MEDIUM_DARK_SKIN_TONE(0x1F469),
-    KISS_WOMAN_MAN_MEDIUM_SKIN_TONE_DARK_SKIN_TONE(0x1F469),
-    KISS_WOMAN_MAN_MEDIUM_DARK_SKIN_TONE_LIGHT_SKIN_TONE(0x1F469),
-    KISS_WOMAN_MAN_MEDIUM_DARK_SKIN_TONE_MEDIUM_LIGHT_SKIN_TONE(0x1F469),
-    KISS_WOMAN_MAN_MEDIUM_DARK_SKIN_TONE_MEDIUM_SKIN_TONE(0x1F469),
-    KISS_WOMAN_MAN_MEDIUM_DARK_SKIN_TONE(0x1F469),
-    KISS_WOMAN_MAN_MEDIUM_DARK_SKIN_TONE_DARK_SKIN_TONE(0x1F469),
-    KISS_WOMAN_MAN_DARK_SKIN_TONE_LIGHT_SKIN_TONE(0x1F469),
-    KISS_WOMAN_MAN_DARK_SKIN_TONE_MEDIUM_LIGHT_SKIN_TONE(0x1F469),
-    KISS_WOMAN_MAN_DARK_SKIN_TONE_MEDIUM_SKIN_TONE(0x1F469),
-    KISS_WOMAN_MAN_DARK_SKIN_TONE_MEDIUM_DARK_SKIN_TONE(0x1F469),
-    KISS_WOMAN_MAN_DARK_SKIN_TONE(0x1F469),
-    KISS_MAN_MAN(0x1F468),
-    KISS_MAN_MAN_LIGHT_SKIN_TONE(0x1F468),
-    KISS_MAN_MAN_LIGHT_SKIN_TONE_MEDIUM_LIGHT_SKIN_TONE(0x1F468),
-    KISS_MAN_MAN_LIGHT_SKIN_TONE_MEDIUM_SKIN_TONE(0x1F468),
-    KISS_MAN_MAN_LIGHT_SKIN_TONE_MEDIUM_DARK_SKIN_TONE(0x1F468),
-    KISS_MAN_MAN_LIGHT_SKIN_TONE_DARK_SKIN_TONE(0x1F468),
-    KISS_MAN_MAN_MEDIUM_LIGHT_SKIN_TONE_LIGHT_SKIN_TONE(0x1F468),
-    KISS_MAN_MAN_MEDIUM_LIGHT_SKIN_TONE(0x1F468),
-    KISS_MAN_MAN_MEDIUM_LIGHT_SKIN_TONE_MEDIUM_SKIN_TONE(0x1F468),
-    KISS_MAN_MAN_MEDIUM_LIGHT_SKIN_TONE_MEDIUM_DARK_SKIN_TONE(0x1F468),
-    KISS_MAN_MAN_MEDIUM_LIGHT_SKIN_TONE_DARK_SKIN_TONE(0x1F468),
-    KISS_MAN_MAN_MEDIUM_SKIN_TONE_LIGHT_SKIN_TONE(0x1F468),
-    KISS_MAN_MAN_MEDIUM_SKIN_TONE_MEDIUM_LIGHT_SKIN_TONE(0x1F468),
-    KISS_MAN_MAN_MEDIUM_SKIN_TONE(0x1F468),
-    KISS_MAN_MAN_MEDIUM_SKIN_TONE_MEDIUM_DARK_SKIN_TONE(0x1F468),
-    KISS_MAN_MAN_MEDIUM_SKIN_TONE_DARK_SKIN_TONE(0x1F468),
-    KISS_MAN_MAN_MEDIUM_DARK_SKIN_TONE_LIGHT_SKIN_TONE(0x1F468),
-    KISS_MAN_MAN_MEDIUM_DARK_SKIN_TONE_MEDIUM_LIGHT_SKIN_TONE(0x1F468),
-    KISS_MAN_MAN_MEDIUM_DARK_SKIN_TONE_MEDIUM_SKIN_TONE(0x1F468),
-    KISS_MAN_MAN_MEDIUM_DARK_SKIN_TONE(0x1F468),
-    KISS_MAN_MAN_MEDIUM_DARK_SKIN_TONE_DARK_SKIN_TONE(0x1F468),
-    KISS_MAN_MAN_DARK_SKIN_TONE_LIGHT_SKIN_TONE(0x1F468),
-    KISS_MAN_MAN_DARK_SKIN_TONE_MEDIUM_LIGHT_SKIN_TONE(0x1F468),
-    KISS_MAN_MAN_DARK_SKIN_TONE_MEDIUM_SKIN_TONE(0x1F468),
-    KISS_MAN_MAN_DARK_SKIN_TONE_MEDIUM_DARK_SKIN_TONE(0x1F468),
-    KISS_MAN_MAN_DARK_SKIN_TONE(0x1F468),
-    KISS_WOMAN_WOMAN(0x1F469),
-    KISS_WOMAN_WOMAN_LIGHT_SKIN_TONE(0x1F469),
-    KISS_WOMAN_WOMAN_LIGHT_SKIN_TONE_MEDIUM_LIGHT_SKIN_TONE(0x1F469),
-    KISS_WOMAN_WOMAN_LIGHT_SKIN_TONE_MEDIUM_SKIN_TONE(0x1F469),
-    KISS_WOMAN_WOMAN_LIGHT_SKIN_TONE_MEDIUM_DARK_SKIN_TONE(0x1F469),
-    KISS_WOMAN_WOMAN_LIGHT_SKIN_TONE_DARK_SKIN_TONE(0x1F469),
-    KISS_WOMAN_WOMAN_MEDIUM_LIGHT_SKIN_TONE_LIGHT_SKIN_TONE(0x1F469),
-    KISS_WOMAN_WOMAN_MEDIUM_LIGHT_SKIN_TONE(0x1F469),
-    KISS_WOMAN_WOMAN_MEDIUM_LIGHT_SKIN_TONE_MEDIUM_SKIN_TONE(0x1F469),
-    KISS_WOMAN_WOMAN_MEDIUM_LIGHT_SKIN_TONE_MEDIUM_DARK_SKIN_TONE(0x1F469),
-    KISS_WOMAN_WOMAN_MEDIUM_LIGHT_SKIN_TONE_DARK_SKIN_TONE(0x1F469),
-    KISS_WOMAN_WOMAN_MEDIUM_SKIN_TONE_LIGHT_SKIN_TONE(0x1F469),
-    KISS_WOMAN_WOMAN_MEDIUM_SKIN_TONE_MEDIUM_LIGHT_SKIN_TONE(0x1F469),
-    KISS_WOMAN_WOMAN_MEDIUM_SKIN_TONE(0x1F469),
-    KISS_WOMAN_WOMAN_MEDIUM_SKIN_TONE_MEDIUM_DARK_SKIN_TONE(0x1F469),
-    KISS_WOMAN_WOMAN_MEDIUM_SKIN_TONE_DARK_SKIN_TONE(0x1F469),
-    KISS_WOMAN_WOMAN_MEDIUM_DARK_SKIN_TONE_LIGHT_SKIN_TONE(0x1F469),
-    KISS_WOMAN_WOMAN_MEDIUM_DARK_SKIN_TONE_MEDIUM_LIGHT_SKIN_TONE(0x1F469),
-    KISS_WOMAN_WOMAN_MEDIUM_DARK_SKIN_TONE_MEDIUM_SKIN_TONE(0x1F469),
-    KISS_WOMAN_WOMAN_MEDIUM_DARK_SKIN_TONE(0x1F469),
-    KISS_WOMAN_WOMAN_MEDIUM_DARK_SKIN_TONE_DARK_SKIN_TONE(0x1F469),
-    KISS_WOMAN_WOMAN_DARK_SKIN_TONE_LIGHT_SKIN_TONE(0x1F469),
-    KISS_WOMAN_WOMAN_DARK_SKIN_TONE_MEDIUM_LIGHT_SKIN_TONE(0x1F469),
-    KISS_WOMAN_WOMAN_DARK_SKIN_TONE_MEDIUM_SKIN_TONE(0x1F469),
-    KISS_WOMAN_WOMAN_DARK_SKIN_TONE_MEDIUM_DARK_SKIN_TONE(0x1F469),
-    KISS_WOMAN_WOMAN_DARK_SKIN_TONE(0x1F469),
-    COUPLE_WITH_HEART(0x1F491),
-    COUPLE_WITH_HEART_LIGHT_SKIN_TONE(0x1F491),
-    COUPLE_WITH_HEART_MEDIUM_LIGHT_SKIN_TONE(0x1F491),
-    COUPLE_WITH_HEART_MEDIUM_SKIN_TONE(0x1F491),
-    COUPLE_WITH_HEART_MEDIUM_DARK_SKIN_TONE(0x1F491),
-    COUPLE_WITH_HEART_DARK_SKIN_TONE(0x1F491),
-    COUPLE_WITH_HEART_PERSON_PERSON_LIGHT_SKIN_TONE_MEDIUM_LIGHT_SKIN_TONE(0x1F9D1),
-    COUPLE_WITH_HEART_PERSON_PERSON_LIGHT_SKIN_TONE_MEDIUM_SKIN_TONE(0x1F9D1),
-    COUPLE_WITH_HEART_PERSON_PERSON_LIGHT_SKIN_TONE_MEDIUM_DARK_SKIN_TONE(0x1F9D1),
-    COUPLE_WITH_HEART_PERSON_PERSON_LIGHT_SKIN_TONE_DARK_SKIN_TONE(0x1F9D1),
-    COUPLE_WITH_HEART_PERSON_PERSON_MEDIUM_LIGHT_SKIN_TONE_LIGHT_SKIN_TONE(0x1F9D1),
-    COUPLE_WITH_HEART_PERSON_PERSON_MEDIUM_LIGHT_SKIN_TONE_MEDIUM_SKIN_TONE(0x1F9D1),
-    COUPLE_WITH_HEART_PERSON_PERSON_MEDIUM_LIGHT_SKIN_TONE_MEDIUM_DARK_SKIN_TONE(0x1F9D1),
-    COUPLE_WITH_HEART_PERSON_PERSON_MEDIUM_LIGHT_SKIN_TONE_DARK_SKIN_TONE(0x1F9D1),
-    COUPLE_WITH_HEART_PERSON_PERSON_MEDIUM_SKIN_TONE_LIGHT_SKIN_TONE(0x1F9D1),
-    COUPLE_WITH_HEART_PERSON_PERSON_MEDIUM_SKIN_TONE_MEDIUM_LIGHT_SKIN_TONE(0x1F9D1),
-    COUPLE_WITH_HEART_PERSON_PERSON_MEDIUM_SKIN_TONE_MEDIUM_DARK_SKIN_TONE(0x1F9D1),
-    COUPLE_WITH_HEART_PERSON_PERSON_MEDIUM_SKIN_TONE_DARK_SKIN_TONE(0x1F9D1),
-    COUPLE_WITH_HEART_PERSON_PERSON_MEDIUM_DARK_SKIN_TONE_LIGHT_SKIN_TONE(0x1F9D1),
-    COUPLE_WITH_HEART_PERSON_PERSON_MEDIUM_DARK_SKIN_TONE_MEDIUM_LIGHT_SKIN_TONE(0x1F9D1),
-    COUPLE_WITH_HEART_PERSON_PERSON_MEDIUM_DARK_SKIN_TONE_MEDIUM_SKIN_TONE(0x1F9D1),
-    COUPLE_WITH_HEART_PERSON_PERSON_MEDIUM_DARK_SKIN_TONE_DARK_SKIN_TONE(0x1F9D1),
-    COUPLE_WITH_HEART_PERSON_PERSON_DARK_SKIN_TONE_LIGHT_SKIN_TONE(0x1F9D1),
-    COUPLE_WITH_HEART_PERSON_PERSON_DARK_SKIN_TONE_MEDIUM_LIGHT_SKIN_TONE(0x1F9D1),
-    COUPLE_WITH_HEART_PERSON_PERSON_DARK_SKIN_TONE_MEDIUM_SKIN_TONE(0x1F9D1),
-    COUPLE_WITH_HEART_PERSON_PERSON_DARK_SKIN_TONE_MEDIUM_DARK_SKIN_TONE(0x1F9D1),
-    COUPLE_WITH_HEART_WOMAN_MAN(0x1F469),
-    COUPLE_WITH_HEART_WOMAN_MAN_LIGHT_SKIN_TONE(0x1F469),
-    COUPLE_WITH_HEART_WOMAN_MAN_LIGHT_SKIN_TONE_MEDIUM_LIGHT_SKIN_TONE(0x1F469),
-    COUPLE_WITH_HEART_WOMAN_MAN_LIGHT_SKIN_TONE_MEDIUM_SKIN_TONE(0x1F469),
-    COUPLE_WITH_HEART_WOMAN_MAN_LIGHT_SKIN_TONE_MEDIUM_DARK_SKIN_TONE(0x1F469),
-    COUPLE_WITH_HEART_WOMAN_MAN_LIGHT_SKIN_TONE_DARK_SKIN_TONE(0x1F469),
-    COUPLE_WITH_HEART_WOMAN_MAN_MEDIUM_LIGHT_SKIN_TONE_LIGHT_SKIN_TONE(0x1F469),
-    COUPLE_WITH_HEART_WOMAN_MAN_MEDIUM_LIGHT_SKIN_TONE(0x1F469),
-    COUPLE_WITH_HEART_WOMAN_MAN_MEDIUM_LIGHT_SKIN_TONE_MEDIUM_SKIN_TONE(0x1F469),
-    COUPLE_WITH_HEART_WOMAN_MAN_MEDIUM_LIGHT_SKIN_TONE_MEDIUM_DARK_SKIN_TONE(0x1F469),
-    COUPLE_WITH_HEART_WOMAN_MAN_MEDIUM_LIGHT_SKIN_TONE_DARK_SKIN_TONE(0x1F469),
-    COUPLE_WITH_HEART_WOMAN_MAN_MEDIUM_SKIN_TONE_LIGHT_SKIN_TONE(0x1F469),
-    COUPLE_WITH_HEART_WOMAN_MAN_MEDIUM_SKIN_TONE_MEDIUM_LIGHT_SKIN_TONE(0x1F469),
-    COUPLE_WITH_HEART_WOMAN_MAN_MEDIUM_SKIN_TONE(0x1F469),
-    COUPLE_WITH_HEART_WOMAN_MAN_MEDIUM_SKIN_TONE_MEDIUM_DARK_SKIN_TONE(0x1F469),
-    COUPLE_WITH_HEART_WOMAN_MAN_MEDIUM_SKIN_TONE_DARK_SKIN_TONE(0x1F469),
-    COUPLE_WITH_HEART_WOMAN_MAN_MEDIUM_DARK_SKIN_TONE_LIGHT_SKIN_TONE(0x1F469),
-    COUPLE_WITH_HEART_WOMAN_MAN_MEDIUM_DARK_SKIN_TONE_MEDIUM_LIGHT_SKIN_TONE(0x1F469),
-    COUPLE_WITH_HEART_WOMAN_MAN_MEDIUM_DARK_SKIN_TONE_MEDIUM_SKIN_TONE(0x1F469),
-    COUPLE_WITH_HEART_WOMAN_MAN_MEDIUM_DARK_SKIN_TONE(0x1F469),
-    COUPLE_WITH_HEART_WOMAN_MAN_MEDIUM_DARK_SKIN_TONE_DARK_SKIN_TONE(0x1F469),
-    COUPLE_WITH_HEART_WOMAN_MAN_DARK_SKIN_TONE_LIGHT_SKIN_TONE(0x1F469),
-    COUPLE_WITH_HEART_WOMAN_MAN_DARK_SKIN_TONE_MEDIUM_LIGHT_SKIN_TONE(0x1F469),
-    COUPLE_WITH_HEART_WOMAN_MAN_DARK_SKIN_TONE_MEDIUM_SKIN_TONE(0x1F469),
-    COUPLE_WITH_HEART_WOMAN_MAN_DARK_SKIN_TONE_MEDIUM_DARK_SKIN_TONE(0x1F469),
-    COUPLE_WITH_HEART_WOMAN_MAN_DARK_SKIN_TONE(0x1F469),
-    COUPLE_WITH_HEART_MAN_MAN(0x1F468),
-    COUPLE_WITH_HEART_MAN_MAN_LIGHT_SKIN_TONE(0x1F468),
-    COUPLE_WITH_HEART_MAN_MAN_LIGHT_SKIN_TONE_MEDIUM_LIGHT_SKIN_TONE(0x1F468),
-    COUPLE_WITH_HEART_MAN_MAN_LIGHT_SKIN_TONE_MEDIUM_SKIN_TONE(0x1F468),
-    COUPLE_WITH_HEART_MAN_MAN_LIGHT_SKIN_TONE_MEDIUM_DARK_SKIN_TONE(0x1F468),
-    COUPLE_WITH_HEART_MAN_MAN_LIGHT_SKIN_TONE_DARK_SKIN_TONE(0x1F468),
-    COUPLE_WITH_HEART_MAN_MAN_MEDIUM_LIGHT_SKIN_TONE_LIGHT_SKIN_TONE(0x1F468),
-    COUPLE_WITH_HEART_MAN_MAN_MEDIUM_LIGHT_SKIN_TONE(0x1F468),
-    COUPLE_WITH_HEART_MAN_MAN_MEDIUM_LIGHT_SKIN_TONE_MEDIUM_SKIN_TONE(0x1F468),
-    COUPLE_WITH_HEART_MAN_MAN_MEDIUM_LIGHT_SKIN_TONE_MEDIUM_DARK_SKIN_TONE(0x1F468),
-    COUPLE_WITH_HEART_MAN_MAN_MEDIUM_LIGHT_SKIN_TONE_DARK_SKIN_TONE(0x1F468),
-    COUPLE_WITH_HEART_MAN_MAN_MEDIUM_SKIN_TONE_LIGHT_SKIN_TONE(0x1F468),
-    COUPLE_WITH_HEART_MAN_MAN_MEDIUM_SKIN_TONE_MEDIUM_LIGHT_SKIN_TONE(0x1F468),
-    COUPLE_WITH_HEART_MAN_MAN_MEDIUM_SKIN_TONE(0x1F468),
-    COUPLE_WITH_HEART_MAN_MAN_MEDIUM_SKIN_TONE_MEDIUM_DARK_SKIN_TONE(0x1F468),
-    COUPLE_WITH_HEART_MAN_MAN_MEDIUM_SKIN_TONE_DARK_SKIN_TONE(0x1F468),
-    COUPLE_WITH_HEART_MAN_MAN_MEDIUM_DARK_SKIN_TONE_LIGHT_SKIN_TONE(0x1F468),
-    COUPLE_WITH_HEART_MAN_MAN_MEDIUM_DARK_SKIN_TONE_MEDIUM_LIGHT_SKIN_TONE(0x1F468),
-    COUPLE_WITH_HEART_MAN_MAN_MEDIUM_DARK_SKIN_TONE_MEDIUM_SKIN_TONE(0x1F468),
-    COUPLE_WITH_HEART_MAN_MAN_MEDIUM_DARK_SKIN_TONE(0x1F468),
-    COUPLE_WITH_HEART_MAN_MAN_MEDIUM_DARK_SKIN_TONE_DARK_SKIN_TONE(0x1F468),
-    COUPLE_WITH_HEART_MAN_MAN_DARK_SKIN_TONE_LIGHT_SKIN_TONE(0x1F468),
-    COUPLE_WITH_HEART_MAN_MAN_DARK_SKIN_TONE_MEDIUM_LIGHT_SKIN_TONE(0x1F468),
-    COUPLE_WITH_HEART_MAN_MAN_DARK_SKIN_TONE_MEDIUM_SKIN_TONE(0x1F468),
-    COUPLE_WITH_HEART_MAN_MAN_DARK_SKIN_TONE_MEDIUM_DARK_SKIN_TONE(0x1F468),
-    COUPLE_WITH_HEART_MAN_MAN_DARK_SKIN_TONE(0x1F468),
-    COUPLE_WITH_HEART_WOMAN_WOMAN(0x1F469),
-    COUPLE_WITH_HEART_WOMAN_WOMAN_LIGHT_SKIN_TONE(0x1F469),
-    COUPLE_WITH_HEART_WOMAN_WOMAN_LIGHT_SKIN_TONE_MEDIUM_LIGHT_SKIN_TONE(0x1F469),
-    COUPLE_WITH_HEART_WOMAN_WOMAN_LIGHT_SKIN_TONE_MEDIUM_SKIN_TONE(0x1F469),
-    COUPLE_WITH_HEART_WOMAN_WOMAN_LIGHT_SKIN_TONE_MEDIUM_DARK_SKIN_TONE(0x1F469),
-    COUPLE_WITH_HEART_WOMAN_WOMAN_LIGHT_SKIN_TONE_DARK_SKIN_TONE(0x1F469),
-    COUPLE_WITH_HEART_WOMAN_WOMAN_MEDIUM_LIGHT_SKIN_TONE_LIGHT_SKIN_TONE(0x1F469),
-    COUPLE_WITH_HEART_WOMAN_WOMAN_MEDIUM_LIGHT_SKIN_TONE(0x1F469),
-    COUPLE_WITH_HEART_WOMAN_WOMAN_MEDIUM_LIGHT_SKIN_TONE_MEDIUM_SKIN_TONE(0x1F469),
-    COUPLE_WITH_HEART_WOMAN_WOMAN_MEDIUM_LIGHT_SKIN_TONE_MEDIUM_DARK_SKIN_TONE(0x1F469),
-    COUPLE_WITH_HEART_WOMAN_WOMAN_MEDIUM_LIGHT_SKIN_TONE_DARK_SKIN_TONE(0x1F469),
-    COUPLE_WITH_HEART_WOMAN_WOMAN_MEDIUM_SKIN_TONE_LIGHT_SKIN_TONE(0x1F469),
-    COUPLE_WITH_HEART_WOMAN_WOMAN_MEDIUM_SKIN_TONE_MEDIUM_LIGHT_SKIN_TONE(0x1F469),
-    COUPLE_WITH_HEART_WOMAN_WOMAN_MEDIUM_SKIN_TONE(0x1F469),
-    COUPLE_WITH_HEART_WOMAN_WOMAN_MEDIUM_SKIN_TONE_MEDIUM_DARK_SKIN_TONE(0x1F469),
-    COUPLE_WITH_HEART_WOMAN_WOMAN_MEDIUM_SKIN_TONE_DARK_SKIN_TONE(0x1F469),
-    COUPLE_WITH_HEART_WOMAN_WOMAN_MEDIUM_DARK_SKIN_TONE_LIGHT_SKIN_TONE(0x1F469),
-    COUPLE_WITH_HEART_WOMAN_WOMAN_MEDIUM_DARK_SKIN_TONE_MEDIUM_LIGHT_SKIN_TONE(0x1F469),
-    COUPLE_WITH_HEART_WOMAN_WOMAN_MEDIUM_DARK_SKIN_TONE_MEDIUM_SKIN_TONE(0x1F469),
-    COUPLE_WITH_HEART_WOMAN_WOMAN_MEDIUM_DARK_SKIN_TONE(0x1F469),
-    COUPLE_WITH_HEART_WOMAN_WOMAN_MEDIUM_DARK_SKIN_TONE_DARK_SKIN_TONE(0x1F469),
-    COUPLE_WITH_HEART_WOMAN_WOMAN_DARK_SKIN_TONE_LIGHT_SKIN_TONE(0x1F469),
-    COUPLE_WITH_HEART_WOMAN_WOMAN_DARK_SKIN_TONE_MEDIUM_LIGHT_SKIN_TONE(0x1F469),
-    COUPLE_WITH_HEART_WOMAN_WOMAN_DARK_SKIN_TONE_MEDIUM_SKIN_TONE(0x1F469),
-    COUPLE_WITH_HEART_WOMAN_WOMAN_DARK_SKIN_TONE_MEDIUM_DARK_SKIN_TONE(0x1F469),
-    COUPLE_WITH_HEART_WOMAN_WOMAN_DARK_SKIN_TONE(0x1F469),
-    FAMILY(0x1F46A),
-    FAMILY_MAN_WOMAN_BOY(0x1F468),
-    FAMILY_MAN_WOMAN_GIRL(0x1F468),
-    FAMILY_MAN_WOMAN_GIRL_BOY(0x1F468),
-    FAMILY_MAN_WOMAN_BOY_BOY(0x1F468),
-    FAMILY_MAN_WOMAN_GIRL_GIRL(0x1F468),
-    FAMILY_MAN_MAN_BOY(0x1F468),
-    FAMILY_MAN_MAN_GIRL(0x1F468),
-    FAMILY_MAN_MAN_GIRL_BOY(0x1F468),
-    FAMILY_MAN_MAN_BOY_BOY(0x1F468),
-    FAMILY_MAN_MAN_GIRL_GIRL(0x1F468),
-    FAMILY_WOMAN_WOMAN_BOY(0x1F469),
-    FAMILY_WOMAN_WOMAN_GIRL(0x1F469),
-    FAMILY_WOMAN_WOMAN_GIRL_BOY(0x1F469),
-    FAMILY_WOMAN_WOMAN_BOY_BOY(0x1F469),
-    FAMILY_WOMAN_WOMAN_GIRL_GIRL(0x1F469),
-    FAMILY_MAN_BOY(0x1F468),
-    FAMILY_MAN_BOY_BOY(0x1F468),
-    FAMILY_MAN_GIRL(0x1F468),
-    FAMILY_MAN_GIRL_BOY(0x1F468),
-    FAMILY_MAN_GIRL_GIRL(0x1F468),
-    FAMILY_WOMAN_BOY(0x1F469),
-    FAMILY_WOMAN_BOY_BOY(0x1F469),
-    FAMILY_WOMAN_GIRL(0x1F469),
-    FAMILY_WOMAN_GIRL_BOY(0x1F469),
-    FAMILY_WOMAN_GIRL_GIRL(0x1F469),
-    SPEAKING_HEAD(0x1F5E3),
-    BUST_IN_SILHOUETTE(0x1F464),
-    BUSTS_IN_SILHOUETTE(0x1F465),
-    PEOPLE_HUGGING(0x1FAC2),
-    FOOTPRINTS(0x1F463),
-    MONKEY_FACE(0x1F435),
-    MONKEY(0x1F412),
-    GORILLA(0x1F98D),
-    ORANGUTAN(0x1F9A7),
-    DOG_FACE(0x1F436),
-    DOG(0x1F415),
-    GUIDE_DOG(0x1F9AE),
-    SERVICE_DOG(0x1F415),
-    POODLE(0x1F429),
-    WOLF(0x1F43A),
-    FOX(0x1F98A),
-    RACCOON(0x1F99D),
-    CAT_FACE(0x1F431),
-    CAT(0x1F408),
-    BLACK_CAT(0x1F408),
-    LION(0x1F981),
-    TIGER_FACE(0x1F42F),
-    TIGER(0x1F405),
-    LEOPARD(0x1F406),
-    HORSE_FACE(0x1F434),
-    HORSE(0x1F40E),
-    UNICORN(0x1F984),
-    ZEBRA(0x1F993),
-    DEER(0x1F98C),
-    BISON(0x1F9AC),
-    COW_FACE(0x1F42E),
-    OX(0x1F402),
-    WATER_BUFFALO(0x1F403),
-    COW(0x1F404),
-    PIG_FACE(0x1F437),
-    PIG(0x1F416),
-    BOAR(0x1F417),
-    PIG_NOSE(0x1F43D),
-    RAM(0x1F40F),
-    EWE(0x1F411),
-    GOAT(0x1F410),
-    CAMEL(0x1F42A),
-    TWO_HUMP_CAMEL(0x1F42B),
-    LLAMA(0x1F999),
-    GIRAFFE(0x1F992),
-    ELEPHANT(0x1F418),
-    MAMMOTH(0x1F9A3),
-    RHINOCEROS(0x1F98F),
-    HIPPOPOTAMUS(0x1F99B),
-    MOUSE_FACE(0x1F42D),
-    MOUSE(0x1F401),
-    RAT(0x1F400),
-    HAMSTER(0x1F439),
-    RABBIT_FACE(0x1F430),
-    RABBIT(0x1F407),
-    CHIPMUNK(0x1F43F),
-    BEAVER(0x1F9AB),
-    HEDGEHOG(0x1F994),
-    BAT(0x1F987),
-    BEAR(0x1F43B),
-    POLAR_BEAR(0x1F43B),
-    KOALA(0x1F428),
-    PANDA(0x1F43C),
-    SLOTH(0x1F9A5),
-    OTTER(0x1F9A6),
-    SKUNK(0x1F9A8),
-    KANGAROO(0x1F998),
-    BADGER(0x1F9A1),
-    PAW_PRINTS(0x1F43E),
-    TURKEY(0x1F983),
-    CHICKEN(0x1F414),
-    ROOSTER(0x1F413),
-    HATCHING_CHICK(0x1F423),
-    BABY_CHICK(0x1F424),
-    FRONT_FACING_BABY_CHICK(0x1F425),
-    BIRD(0x1F426),
-    PENGUIN(0x1F427),
-    DOVE(0x1F54A),
-    EAGLE(0x1F985),
-    DUCK(0x1F986),
-    SWAN(0x1F9A2),
-    OWL(0x1F989),
-    DODO(0x1F9A4),
-    FEATHER(0x1FAB6),
-    FLAMINGO(0x1F9A9),
-    PEACOCK(0x1F99A),
-    PARROT(0x1F99C),
-    FROG(0x1F438),
-    CROCODILE(0x1F40A),
-    TURTLE(0x1F422),
-    LIZARD(0x1F98E),
-    SNAKE(0x1F40D),
-    DRAGON_FACE(0x1F432),
-    DRAGON(0x1F409),
-    SAUROPOD(0x1F995),
-    T_REX(0x1F996),
-    SPOUTING_WHALE(0x1F433),
-    WHALE(0x1F40B),
-    DOLPHIN(0x1F42C),
-    SEAL(0x1F9AD),
-    FISH(0x1F41F),
-    TROPICAL_FISH(0x1F420),
-    BLOWFISH(0x1F421),
-    SHARK(0x1F988),
-    OCTOPUS(0x1F419),
-    SPIRAL_SHELL(0x1F41A),
-    CORAL(0x1FAB8),
-    SNAIL(0x1F40C),
-    BUTTERFLY(0x1F98B),
-    BUG(0x1F41B),
-    ANT(0x1F41C),
-    HONEYBEE(0x1F41D),
-    BEETLE(0x1FAB2),
-    LADY_BEETLE(0x1F41E),
-    CRICKET(0x1F997),
-    COCKROACH(0x1FAB3),
-    SPIDER(0x1F577),
-    SPIDER_WEB(0x1F578),
-    SCORPION(0x1F982),
-    MOSQUITO(0x1F99F),
-    FLY(0x1FAB0),
-    WORM(0x1FAB1),
-    MICROBE(0x1F9A0),
-    BOUQUET(0x1F490),
-    CHERRY_BLOSSOM(0x1F338),
-    WHITE_FLOWER(0x1F4AE),
-    LOTUS(0x1FAB7),
-    ROSETTE(0x1F3F5),
-    ROSE(0x1F339),
-    WILTED_FLOWER(0x1F940),
-    HIBISCUS(0x1F33A),
-    SUNFLOWER(0x1F33B),
-    BLOSSOM(0x1F33C),
-    TULIP(0x1F337),
-    SEEDLING(0x1F331),
-    POTTED_PLANT(0x1FAB4),
-    EVERGREEN_TREE(0x1F332),
-    DECIDUOUS_TREE(0x1F333),
-    PALM_TREE(0x1F334),
-    CACTUS(0x1F335),
-    SHEAF_OF_RICE(0x1F33E),
-    HERB(0x1F33F),
-    SHAMROCK(0x2618),
-    FOUR_LEAF_CLOVER(0x1F340),
-    MAPLE_LEAF(0x1F341),
-    FALLEN_LEAF(0x1F342),
-    LEAF_FLUTTERING_IN_WIND(0x1F343),
-    EMPTY_NEST(0x1FAB9),
-    NEST_WITH_EGGS(0x1FABA),
-    GRAPES(0x1F347),
-    MELON(0x1F348),
-    WATERMELON(0x1F349),
-    TANGERINE(0x1F34A),
-    LEMON(0x1F34B),
-    BANANA(0x1F34C),
-    PINEAPPLE(0x1F34D),
-    MANGO(0x1F96D),
-    RED_APPLE(0x1F34E),
-    GREEN_APPLE(0x1F34F),
-    PEAR(0x1F350),
-    PEACH(0x1F351),
-    CHERRIES(0x1F352),
-    STRAWBERRY(0x1F353),
-    BLUEBERRIES(0x1FAD0),
-    KIWI_FRUIT(0x1F95D),
-    TOMATO(0x1F345),
-    OLIVE(0x1FAD2),
-    COCONUT(0x1F965),
-    AVOCADO(0x1F951),
-    EGGPLANT(0x1F346),
-    POTATO(0x1F954),
-    CARROT(0x1F955),
-    EAR_OF_CORN(0x1F33D),
-    HOT_PEPPER(0x1F336),
-    BELL_PEPPER(0x1FAD1),
-    CUCUMBER(0x1F952),
-    LEAFY_GREEN(0x1F96C),
-    BROCCOLI(0x1F966),
-    GARLIC(0x1F9C4),
-    ONION(0x1F9C5),
-    MUSHROOM(0x1F344),
-    PEANUTS(0x1F95C),
-    BEANS(0x1FAD8),
-    CHESTNUT(0x1F330),
-    BREAD(0x1F35E),
-    CROISSANT(0x1F950),
-    BAGUETTE_BREAD(0x1F956),
-    FLATBREAD(0x1FAD3),
-    PRETZEL(0x1F968),
-    BAGEL(0x1F96F),
-    PANCAKES(0x1F95E),
-    WAFFLE(0x1F9C7),
-    CHEESE_WEDGE(0x1F9C0),
-    MEAT_ON_BONE(0x1F356),
-    POULTRY_LEG(0x1F357),
-    CUT_OF_MEAT(0x1F969),
-    BACON(0x1F953),
-    HAMBURGER(0x1F354),
-    FRENCH_FRIES(0x1F35F),
-    PIZZA(0x1F355),
-    HOT_DOG(0x1F32D),
-    SANDWICH(0x1F96A),
-    TACO(0x1F32E),
-    BURRITO(0x1F32F),
-    TAMALE(0x1FAD4),
-    STUFFED_FLATBREAD(0x1F959),
-    FALAFEL(0x1F9C6),
-    EGG(0x1F95A),
-    COOKING(0x1F373),
-    SHALLOW_PAN_OF_FOOD(0x1F958),
-    POT_OF_FOOD(0x1F372),
-    FONDUE(0x1FAD5),
-    BOWL_WITH_SPOON(0x1F963),
-    GREEN_SALAD(0x1F957),
-    POPCORN(0x1F37F),
-    BUTTER(0x1F9C8),
-    SALT(0x1F9C2),
-    CANNED_FOOD(0x1F96B),
-    BENTO_BOX(0x1F371),
-    RICE_CRACKER(0x1F358),
-    RICE_BALL(0x1F359),
-    COOKED_RICE(0x1F35A),
-    CURRY_RICE(0x1F35B),
-    STEAMING_BOWL(0x1F35C),
-    SPAGHETTI(0x1F35D),
-    ROASTED_SWEET_POTATO(0x1F360),
-    ODEN(0x1F362),
-    SUSHI(0x1F363),
-    FRIED_SHRIMP(0x1F364),
-    FISH_CAKE_WITH_SWIRL(0x1F365),
-    MOON_CAKE(0x1F96E),
-    DANGO(0x1F361),
-    DUMPLING(0x1F95F),
-    FORTUNE_COOKIE(0x1F960),
-    TAKEOUT_BOX(0x1F961),
-    CRAB(0x1F980),
-    LOBSTER(0x1F99E),
-    SHRIMP(0x1F990),
-    SQUID(0x1F991),
-    OYSTER(0x1F9AA),
-    SOFT_ICE_CREAM(0x1F366),
-    SHAVED_ICE(0x1F367),
-    ICE_CREAM(0x1F368),
-    DOUGHNUT(0x1F369),
-    COOKIE(0x1F36A),
-    BIRTHDAY_CAKE(0x1F382),
-    SHORTCAKE(0x1F370),
-    CUPCAKE(0x1F9C1),
-    PIE(0x1F967),
-    CHOCOLATE_BAR(0x1F36B),
-    CANDY(0x1F36C),
-    LOLLIPOP(0x1F36D),
-    CUSTARD(0x1F36E),
-    HONEY_POT(0x1F36F),
-    BABY_BOTTLE(0x1F37C),
-    GLASS_OF_MILK(0x1F95B),
-    HOT_BEVERAGE(0x2615),
-    TEAPOT(0x1FAD6),
-    TEACUP_WITHOUT_HANDLE(0x1F375),
-    SAKE(0x1F376),
-    BOTTLE_WITH_POPPING_CORK(0x1F37E),
-    WINE_GLASS(0x1F377),
-    COCKTAIL_GLASS(0x1F378),
-    TROPICAL_DRINK(0x1F379),
-    BEER_MUG(0x1F37A),
-    CLINKING_BEER_MUGS(0x1F37B),
-    CLINKING_GLASSES(0x1F942),
-    TUMBLER_GLASS(0x1F943),
-    POURING_LIQUID(0x1FAD7),
-    CUP_WITH_STRAW(0x1F964),
-    BUBBLE_TEA(0x1F9CB),
-    BEVERAGE_BOX(0x1F9C3),
-    MATE(0x1F9C9),
-    ICE(0x1F9CA),
-    CHOPSTICKS(0x1F962),
-    FORK_AND_KNIFE_WITH_PLATE(0x1F37D),
-    FORK_AND_KNIFE(0x1F374),
-    SPOON(0x1F944),
-    KITCHEN_KNIFE(0x1F52A),
-    JAR(0x1FAD9),
-    AMPHORA(0x1F3FA),
-    GLOBE_SHOWING_EUROPE_AFRICA(0x1F30D),
-    GLOBE_SHOWING_AMERICAS(0x1F30E),
-    GLOBE_SHOWING_ASIA_AUSTRALIA(0x1F30F),
-    GLOBE_WITH_MERIDIANS(0x1F310),
-    WORLD_MAP(0x1F5FA),
-    MAP_OF_JAPAN(0x1F5FE),
-    COMPASS(0x1F9ED),
-    SNOW_CAPPED_MOUNTAIN(0x1F3D4),
-    MOUNTAIN(0x26F0),
-    VOLCANO(0x1F30B),
-    MOUNT_FUJI(0x1F5FB),
-    CAMPING(0x1F3D5),
-    BEACH_WITH_UMBRELLA(0x1F3D6),
-    DESERT(0x1F3DC),
-    DESERT_ISLAND(0x1F3DD),
-    NATIONAL_PARK(0x1F3DE),
-    STADIUM(0x1F3DF),
-    CLASSICAL_BUILDING(0x1F3DB),
-    BUILDING_CONSTRUCTION(0x1F3D7),
-    BRICK(0x1F9F1),
-    ROCK(0x1FAA8),
-    WOOD(0x1FAB5),
-    HUT(0x1F6D6),
-    HOUSES(0x1F3D8),
-    DERELICT_HOUSE(0x1F3DA),
-    HOUSE(0x1F3E0),
-    HOUSE_WITH_GARDEN(0x1F3E1),
-    OFFICE_BUILDING(0x1F3E2),
-    JAPANESE_POST_OFFICE(0x1F3E3),
-    POST_OFFICE(0x1F3E4),
-    HOSPITAL(0x1F3E5),
-    BANK(0x1F3E6),
-    HOTEL(0x1F3E8),
-    LOVE_HOTEL(0x1F3E9),
-    CONVENIENCE_STORE(0x1F3EA),
-    SCHOOL(0x1F3EB),
-    DEPARTMENT_STORE(0x1F3EC),
-    FACTORY(0x1F3ED),
-    JAPANESE_CASTLE(0x1F3EF),
-    CASTLE(0x1F3F0),
-    WEDDING(0x1F492),
-    TOKYO_TOWER(0x1F5FC),
-    STATUE_OF_LIBERTY(0x1F5FD),
-    CHURCH(0x26EA),
-    MOSQUE(0x1F54C),
-    HINDU_TEMPLE(0x1F6D5),
-    SYNAGOGUE(0x1F54D),
-    SHINTO_SHRINE(0x26E9),
-    KAABA(0x1F54B),
-    FOUNTAIN(0x26F2),
-    TENT(0x26FA),
-    FOGGY(0x1F301),
-    NIGHT_WITH_STARS(0x1F303),
-    CITYSCAPE(0x1F3D9),
-    SUNRISE_OVER_MOUNTAINS(0x1F304),
-    SUNRISE(0x1F305),
-    CITYSCAPE_AT_DUSK(0x1F306),
-    SUNSET(0x1F307),
-    BRIDGE_AT_NIGHT(0x1F309),
-    HOT_SPRINGS(0x2668),
-    CAROUSEL_HORSE(0x1F3A0),
-    PLAYGROUND_SLIDE(0x1F6DD),
-    FERRIS_WHEEL(0x1F3A1),
-    ROLLER_COASTER(0x1F3A2),
-    BARBER_POLE(0x1F488),
-    CIRCUS_TENT(0x1F3AA),
-    LOCOMOTIVE(0x1F682),
-    RAILWAY_CAR(0x1F683),
-    HIGH_SPEED_TRAIN(0x1F684),
-    BULLET_TRAIN(0x1F685),
-    TRAIN(0x1F686),
-    METRO(0x1F687),
-    LIGHT_RAIL(0x1F688),
-    STATION(0x1F689),
-    TRAM(0x1F68A),
-    MONORAIL(0x1F69D),
-    MOUNTAIN_RAILWAY(0x1F69E),
-    TRAM_CAR(0x1F68B),
-    BUS(0x1F68C),
-    ONCOMING_BUS(0x1F68D),
-    TROLLEYBUS(0x1F68E),
-    MINIBUS(0x1F690),
-    AMBULANCE(0x1F691),
-    FIRE_ENGINE(0x1F692),
-    POLICE_CAR(0x1F693),
-    ONCOMING_POLICE_CAR(0x1F694),
-    TAXI(0x1F695),
-    ONCOMING_TAXI(0x1F696),
-    AUTOMOBILE(0x1F697),
-    ONCOMING_AUTOMOBILE(0x1F698),
-    SPORT_UTILITY_VEHICLE(0x1F699),
-    PICKUP_TRUCK(0x1F6FB),
-    DELIVERY_TRUCK(0x1F69A),
-    ARTICULATED_LORRY(0x1F69B),
-    TRACTOR(0x1F69C),
-    RACING_CAR(0x1F3CE),
-    MOTORCYCLE(0x1F3CD),
-    MOTOR_SCOOTER(0x1F6F5),
-    MANUAL_WHEELCHAIR(0x1F9BD),
-    MOTORIZED_WHEELCHAIR(0x1F9BC),
-    AUTO_RICKSHAW(0x1F6FA),
-    BICYCLE(0x1F6B2),
-    KICK_SCOOTER(0x1F6F4),
-    SKATEBOARD(0x1F6F9),
-    ROLLER_SKATE(0x1F6FC),
-    BUS_STOP(0x1F68F),
-    MOTORWAY(0x1F6E3),
-    RAILWAY_TRACK(0x1F6E4),
-    OIL_DRUM(0x1F6E2),
-    FUEL_PUMP(0x26FD),
-    WHEEL(0x1F6DE),
-    POLICE_CAR_LIGHT(0x1F6A8),
-    HORIZONTAL_TRAFFIC_LIGHT(0x1F6A5),
-    VERTICAL_TRAFFIC_LIGHT(0x1F6A6),
-    STOP_SIGN(0x1F6D1),
-    CONSTRUCTION(0x1F6A7),
-    ANCHOR(0x2693),
-    RING_BUOY(0x1F6DF),
-    SAILBOAT(0x26F5),
-    CANOE(0x1F6F6),
-    SPEEDBOAT(0x1F6A4),
-    PASSENGER_SHIP(0x1F6F3),
-    FERRY(0x26F4),
-    MOTOR_BOAT(0x1F6E5),
-    SHIP(0x1F6A2),
-    AIRPLANE(0x2708),
-    SMALL_AIRPLANE(0x1F6E9),
-    AIRPLANE_DEPARTURE(0x1F6EB),
-    AIRPLANE_ARRIVAL(0x1F6EC),
-    PARACHUTE(0x1FA82),
-    SEAT(0x1F4BA),
-    HELICOPTER(0x1F681),
-    SUSPENSION_RAILWAY(0x1F69F),
-    MOUNTAIN_CABLEWAY(0x1F6A0),
-    AERIAL_TRAMWAY(0x1F6A1),
-    SATELLITE(0x1F6F0),
-    ROCKET(0x1F680),
-    FLYING_SAUCER(0x1F6F8),
-    BELLHOP_BELL(0x1F6CE),
-    LUGGAGE(0x1F9F3),
-    HOURGLASS_DONE(0x231B),
-    HOURGLASS_NOT_DONE(0x23F3),
-    WATCH(0x231A),
-    ALARM_CLOCK(0x23F0),
-    STOPWATCH(0x23F1),
-    TIMER_CLOCK(0x23F2),
-    MANTELPIECE_CLOCK(0x1F570),
-    TWELVE_OCLOCK(0x1F55B),
-    TWELVE_THIRTY(0x1F567),
-    ONE_OCLOCK(0x1F550),
-    ONE_THIRTY(0x1F55C),
-    TWO_OCLOCK(0x1F551),
-    TWO_THIRTY(0x1F55D),
-    THREE_OCLOCK(0x1F552),
-    THREE_THIRTY(0x1F55E),
-    FOUR_OCLOCK(0x1F553),
-    FOUR_THIRTY(0x1F55F),
-    FIVE_OCLOCK(0x1F554),
-    FIVE_THIRTY(0x1F560),
-    SIX_OCLOCK(0x1F555),
-    SIX_THIRTY(0x1F561),
-    SEVEN_OCLOCK(0x1F556),
-    SEVEN_THIRTY(0x1F562),
-    EIGHT_OCLOCK(0x1F557),
-    EIGHT_THIRTY(0x1F563),
-    NINE_OCLOCK(0x1F558),
-    NINE_THIRTY(0x1F564),
-    TEN_OCLOCK(0x1F559),
-    TEN_THIRTY(0x1F565),
-    ELEVEN_OCLOCK(0x1F55A),
-    ELEVEN_THIRTY(0x1F566),
-    NEW_MOON(0x1F311),
-    WAXING_CRESCENT_MOON(0x1F312),
-    FIRST_QUARTER_MOON(0x1F313),
-    WAXING_GIBBOUS_MOON(0x1F314),
-    FULL_MOON(0x1F315),
-    WANING_GIBBOUS_MOON(0x1F316),
-    LAST_QUARTER_MOON(0x1F317),
-    WANING_CRESCENT_MOON(0x1F318),
-    CRESCENT_MOON(0x1F319),
-    NEW_MOON_FACE(0x1F31A),
-    FIRST_QUARTER_MOON_FACE(0x1F31B),
-    LAST_QUARTER_MOON_FACE(0x1F31C),
-    THERMOMETER(0x1F321),
-    SUN(0x2600),
-    FULL_MOON_FACE(0x1F31D),
-    SUN_WITH_FACE(0x1F31E),
-    RINGED_PLANET(0x1FA90),
-    STAR(0x2B50),
-    GLOWING_STAR(0x1F31F),
-    SHOOTING_STAR(0x1F320),
-    MILKY_WAY(0x1F30C),
-    CLOUD(0x2601),
-    SUN_BEHIND_CLOUD(0x26C5),
-    CLOUD_WITH_LIGHTNING_AND_RAIN(0x26C8),
-    SUN_BEHIND_SMALL_CLOUD(0x1F324),
-    SUN_BEHIND_LARGE_CLOUD(0x1F325),
-    SUN_BEHIND_RAIN_CLOUD(0x1F326),
-    CLOUD_WITH_RAIN(0x1F327),
-    CLOUD_WITH_SNOW(0x1F328),
-    CLOUD_WITH_LIGHTNING(0x1F329),
-    TORNADO(0x1F32A),
-    FOG(0x1F32B),
-    WIND_FACE(0x1F32C),
-    CYCLONE(0x1F300),
-    RAINBOW(0x1F308),
-    CLOSED_UMBRELLA(0x1F302),
-    UMBRELLA(0x2602),
-    UMBRELLA_WITH_RAIN_DROPS(0x2614),
-    UMBRELLA_ON_GROUND(0x26F1),
-    HIGH_VOLTAGE(0x26A1),
-    SNOWFLAKE(0x2744),
-    SNOWMAN(0x2603),
-    SNOWMAN_WITHOUT_SNOW(0x26C4),
-    COMET(0x2604),
-    FIRE(0x1F525),
-    DROPLET(0x1F4A7),
-    WATER_WAVE(0x1F30A),
-    JACK_O_LANTERN(0x1F383),
-    CHRISTMAS_TREE(0x1F384),
-    FIREWORKS(0x1F386),
-    SPARKLER(0x1F387),
-    FIRECRACKER(0x1F9E8),
-    SPARKLES(0x2728),
-    BALLOON(0x1F388),
-    PARTY_POPPER(0x1F389),
-    CONFETTI_BALL(0x1F38A),
-    TANABATA_TREE(0x1F38B),
-    PINE_DECORATION(0x1F38D),
-    JAPANESE_DOLLS(0x1F38E),
-    CARP_STREAMER(0x1F38F),
-    WIND_CHIME(0x1F390),
-    MOON_VIEWING_CEREMONY(0x1F391),
-    RED_ENVELOPE(0x1F9E7),
-    RIBBON(0x1F380),
-    WRAPPED_GIFT(0x1F381),
-    REMINDER_RIBBON(0x1F397),
-    ADMISSION_TICKETS(0x1F39F),
-    TICKET(0x1F3AB),
-    MILITARY_MEDAL(0x1F396),
-    TROPHY(0x1F3C6),
-    SPORTS_MEDAL(0x1F3C5),
-    FIRST_PLACE_MEDAL(0x1F947),
-    SECOND_PLACE_MEDAL(0x1F948),
-    THIRD_PLACE_MEDAL(0x1F949),
-    SOCCER_BALL(0x26BD),
-    BASEBALL(0x26BE),
-    SOFTBALL(0x1F94E),
-    BASKETBALL(0x1F3C0),
-    VOLLEYBALL(0x1F3D0),
-    AMERICAN_FOOTBALL(0x1F3C8),
-    RUGBY_FOOTBALL(0x1F3C9),
-    TENNIS(0x1F3BE),
-    FLYING_DISC(0x1F94F),
-    BOWLING(0x1F3B3),
-    CRICKET_GAME(0x1F3CF),
-    FIELD_HOCKEY(0x1F3D1),
-    ICE_HOCKEY(0x1F3D2),
-    LACROSSE(0x1F94D),
-    PING_PONG(0x1F3D3),
-    BADMINTON(0x1F3F8),
-    BOXING_GLOVE(0x1F94A),
-    MARTIAL_ARTS_UNIFORM(0x1F94B),
-    GOAL_NET(0x1F945),
-    FLAG_IN_HOLE(0x26F3),
-    ICE_SKATE(0x26F8),
-    FISHING_POLE(0x1F3A3),
-    DIVING_MASK(0x1F93F),
-    RUNNING_SHIRT(0x1F3BD),
-    SKIS(0x1F3BF),
-    SLED(0x1F6F7),
-    CURLING_STONE(0x1F94C),
-    BULLSEYE(0x1F3AF),
-    YO_YO(0x1FA80),
-    KITE(0x1FA81),
-    POOL_8_BALL(0x1F3B1),
-    CRYSTAL_BALL(0x1F52E),
-    MAGIC_WAND(0x1FA84),
-    NAZAR_AMULET(0x1F9FF),
-    HAMSA(0x1FAAC),
-    VIDEO_GAME(0x1F3AE),
-    JOYSTICK(0x1F579),
-    SLOT_MACHINE(0x1F3B0),
-    GAME_DIE(0x1F3B2),
-    PUZZLE_PIECE(0x1F9E9),
-    TEDDY_BEAR(0x1F9F8),
-    PIÑATA(0x1FA85),
-    MIRROR_BALL(0x1FAA9),
-    NESTING_DOLLS(0x1FA86),
-    SPADE_SUIT(0x2660),
-    HEART_SUIT(0x2665),
-    DIAMOND_SUIT(0x2666),
-    CLUB_SUIT(0x2663),
-    CHESS_PAWN(0x265F),
-    JOKER(0x1F0CF),
-    MAHJONG_RED_DRAGON(0x1F004),
-    FLOWER_PLAYING_CARDS(0x1F3B4),
-    PERFORMING_ARTS(0x1F3AD),
-    FRAMED_PICTURE(0x1F5BC),
-    ARTIST_PALETTE(0x1F3A8),
-    THREAD(0x1F9F5),
-    SEWING_NEEDLE(0x1FAA1),
-    YARN(0x1F9F6),
-    KNOT(0x1FAA2),
-    GLASSES(0x1F453),
-    SUNGLASSES(0x1F576),
-    GOGGLES(0x1F97D),
-    LAB_COAT(0x1F97C),
-    SAFETY_VEST(0x1F9BA),
-    NECKTIE(0x1F454),
-    T_SHIRT(0x1F455),
-    JEANS(0x1F456),
-    SCARF(0x1F9E3),
-    GLOVES(0x1F9E4),
-    COAT(0x1F9E5),
-    SOCKS(0x1F9E6),
-    DRESS(0x1F457),
-    KIMONO(0x1F458),
-    SARI(0x1F97B),
-    ONE_PIECE_SWIMSUIT(0x1FA71),
-    BRIEFS(0x1FA72),
-    SHORTS(0x1FA73),
-    BIKINI(0x1F459),
-    WOMANS_CLOTHES(0x1F45A),
-    PURSE(0x1F45B),
-    HANDBAG(0x1F45C),
-    CLUTCH_BAG(0x1F45D),
-    SHOPPING_BAGS(0x1F6CD),
-    BACKPACK(0x1F392),
-    THONG_SANDAL(0x1FA74),
-    MANS_SHOE(0x1F45E),
-    RUNNING_SHOE(0x1F45F),
-    HIKING_BOOT(0x1F97E),
-    FLAT_SHOE(0x1F97F),
-    HIGH_HEELED_SHOE(0x1F460),
-    WOMANS_SANDAL(0x1F461),
-    BALLET_SHOES(0x1FA70),
-    WOMANS_BOOT(0x1F462),
-    CROWN(0x1F451),
-    WOMANS_HAT(0x1F452),
-    TOP_HAT(0x1F3A9),
-    GRADUATION_CAP(0x1F393),
-    BILLED_CAP(0x1F9E2),
-    MILITARY_HELMET(0x1FA96),
-    RESCUE_WORKERS_HELMET(0x26D1),
-    PRAYER_BEADS(0x1F4FF),
-    LIPSTICK(0x1F484),
-    RING(0x1F48D),
-    GEM_STONE(0x1F48E),
-    MUTED_SPEAKER(0x1F507),
-    SPEAKER_LOW_VOLUME(0x1F508),
-    SPEAKER_MEDIUM_VOLUME(0x1F509),
-    SPEAKER_HIGH_VOLUME(0x1F50A),
-    LOUDSPEAKER(0x1F4E2),
-    MEGAPHONE(0x1F4E3),
-    POSTAL_HORN(0x1F4EF),
-    BELL(0x1F514),
-    BELL_WITH_SLASH(0x1F515),
-    MUSICAL_SCORE(0x1F3BC),
-    MUSICAL_NOTE(0x1F3B5),
-    MUSICAL_NOTES(0x1F3B6),
-    STUDIO_MICROPHONE(0x1F399),
-    LEVEL_SLIDER(0x1F39A),
-    CONTROL_KNOBS(0x1F39B),
-    MICROPHONE(0x1F3A4),
-    HEADPHONE(0x1F3A7),
-    RADIO(0x1F4FB),
-    SAXOPHONE(0x1F3B7),
-    ACCORDION(0x1FA97),
-    GUITAR(0x1F3B8),
-    MUSICAL_KEYBOARD(0x1F3B9),
-    TRUMPET(0x1F3BA),
-    VIOLIN(0x1F3BB),
-    BANJO(0x1FA95),
-    DRUM(0x1F941),
-    LONG_DRUM(0x1FA98),
-    MOBILE_PHONE(0x1F4F1),
-    MOBILE_PHONE_WITH_ARROW(0x1F4F2),
-    TELEPHONE(0x260E),
-    TELEPHONE_RECEIVER(0x1F4DE),
-    PAGER(0x1F4DF),
-    FAX_MACHINE(0x1F4E0),
-    BATTERY(0x1F50B),
-    LOW_BATTERY(0x1FAAB),
-    ELECTRIC_PLUG(0x1F50C),
-    LAPTOP(0x1F4BB),
-    DESKTOP_COMPUTER(0x1F5A5),
-    PRINTER(0x1F5A8),
-    KEYBOARD(0x2328),
-    COMPUTER_MOUSE(0x1F5B1),
-    TRACKBALL(0x1F5B2),
-    COMPUTER_DISK(0x1F4BD),
-    FLOPPY_DISK(0x1F4BE),
-    OPTICAL_DISK(0x1F4BF),
-    DVD(0x1F4C0),
-    ABACUS(0x1F9EE),
-    MOVIE_CAMERA(0x1F3A5),
-    FILM_FRAMES(0x1F39E),
-    FILM_PROJECTOR(0x1F4FD),
-    CLAPPER_BOARD(0x1F3AC),
-    TELEVISION(0x1F4FA),
-    CAMERA(0x1F4F7),
-    CAMERA_WITH_FLASH(0x1F4F8),
-    VIDEO_CAMERA(0x1F4F9),
-    VIDEOCASSETTE(0x1F4FC),
-    MAGNIFYING_GLASS_TILTED_LEFT(0x1F50D),
-    MAGNIFYING_GLASS_TILTED_RIGHT(0x1F50E),
-    CANDLE(0x1F56F),
-    LIGHT_BULB(0x1F4A1),
-    FLASHLIGHT(0x1F526),
-    RED_PAPER_LANTERN(0x1F3EE),
-    DIYA_LAMP(0x1FA94),
-    NOTEBOOK_WITH_DECORATIVE_COVER(0x1F4D4),
-    CLOSED_BOOK(0x1F4D5),
-    OPEN_BOOK(0x1F4D6),
-    GREEN_BOOK(0x1F4D7),
-    BLUE_BOOK(0x1F4D8),
-    ORANGE_BOOK(0x1F4D9),
-    BOOKS(0x1F4DA),
-    NOTEBOOK(0x1F4D3),
-    LEDGER(0x1F4D2),
-    PAGE_WITH_CURL(0x1F4C3),
-    SCROLL(0x1F4DC),
-    PAGE_FACING_UP(0x1F4C4),
-    NEWSPAPER(0x1F4F0),
-    ROLLED_UP_NEWSPAPER(0x1F5DE),
-    BOOKMARK_TABS(0x1F4D1),
-    BOOKMARK(0x1F516),
-    LABEL(0x1F3F7),
-    MONEY_BAG(0x1F4B0),
-    COIN(0x1FA99),
-    YEN_BANKNOTE(0x1F4B4),
-    DOLLAR_BANKNOTE(0x1F4B5),
-    EURO_BANKNOTE(0x1F4B6),
-    POUND_BANKNOTE(0x1F4B7),
-    MONEY_WITH_WINGS(0x1F4B8),
-    CREDIT_CARD(0x1F4B3),
-    RECEIPT(0x1F9FE),
-    CHART_INCREASING_WITH_YEN(0x1F4B9),
-    ENVELOPE(0x2709),
-    E_MAIL(0x1F4E7),
-    INCOMING_ENVELOPE(0x1F4E8),
-    ENVELOPE_WITH_ARROW(0x1F4E9),
-    OUTBOX_TRAY(0x1F4E4),
-    INBOX_TRAY(0x1F4E5),
-    PACKAGES(0x1F4E6),
-    CLOSED_MAILBOX_WITH_RAISED_FLAG(0x1F4EB),
-    CLOSED_MAILBOX_WITH_LOWERED_FLAG(0x1F4EA),
-    OPEN_MAILBOX_WITH_RAISED_FLAG(0x1F4EC),
-    OPEN_MAILBOX_WITH_LOWERED_FLAG(0x1F4ED),
-    POSTBOX(0x1F4EE),
-    BALLOT_BOX_WITH_BALLOT(0x1F5F3),
-    PENCIL(0x270F),
-    BLACK_NIB(0x2712),
-    FOUNTAIN_PEN(0x1F58B),
-    PEN(0x1F58A),
-    PAINTBRUSH(0x1F58C),
-    CRAYON(0x1F58D),
-    MEMO(0x1F4DD),
-    BRIEFCASE(0x1F4BC),
-    FILE_FOLDER(0x1F4C1),
-    OPEN_FILE_FOLDER(0x1F4C2),
-    CARD_INDEX_DIVIDERS(0x1F5C2),
-    CALENDAR(0x1F4C5),
-    TEAR_OFF_CALENDAR(0x1F4C6),
-    SPIRAL_NOTEPAD(0x1F5D2),
-    SPIRAL_CALENDAR(0x1F5D3),
-    CARD_INDEX(0x1F4C7),
-    CHART_INCREASING(0x1F4C8),
-    CHART_DECREASING(0x1F4C9),
-    BAR_CHART(0x1F4CA),
-    CLIPBOARD(0x1F4CB),
-    PUSHPIN(0x1F4CC),
-    ROUND_PUSHPIN(0x1F4CD),
-    PAPERCLIP(0x1F4CE),
-    LINKED_PAPERCLIPS(0x1F587),
-    STRAIGHT_RULER(0x1F4CF),
-    TRIANGULAR_RULER(0x1F4D0),
-    SCISSORS(0x2702),
-    CARD_FILE_BOX(0x1F5C3),
-    FILE_CABINET(0x1F5C4),
-    WASTEBASKET(0x1F5D1),
-    LOCKED(0x1F512),
-    UNLOCKED(0x1F513),
-    LOCKED_WITH_PEN(0x1F50F),
-    LOCKED_WITH_KEY(0x1F510),
-    KEY(0x1F511),
-    OLD_KEY(0x1F5DD),
-    HAMMER(0x1F528),
-    AXE(0x1FA93),
-    PICK(0x26CF),
-    HAMMER_AND_PICK(0x2692),
-    HAMMER_AND_WRENCH(0x1F6E0),
-    DAGGER(0x1F5E1),
-    CROSSED_SWORDS(0x2694),
-    WATER_PISTOL(0x1F52B),
-    BOOMERANG(0x1FA83),
-    BOW_AND_ARROW(0x1F3F9),
-    SHIELD(0x1F6E1),
-    CARPENTRY_SAW(0x1FA9A),
-    WRENCH(0x1F527),
-    SCREWDRIVER(0x1FA9B),
-    NUT_AND_BOLT(0x1F529),
-    GEAR(0x2699),
-    CLAMP(0x1F5DC),
-    BALANCE_SCALE(0x2696),
-    WHITE_CANE(0x1F9AF),
-    LINK(0x1F517),
-    CHAINS(0x26D3),
-    HOOK(0x1FA9D),
-    TOOLBOX(0x1F9F0),
-    MAGNET(0x1F9F2),
-    LADDER(0x1FA9C),
-    ALEMBIC(0x2697),
-    TEST_TUBE(0x1F9EA),
-    PETRI_DISH(0x1F9EB),
-    DNA(0x1F9EC),
-    MICROSCOPE(0x1F52C),
-    TELESCOPE(0x1F52D),
-    SATELLITE_ANTENNA(0x1F4E1),
-    SYRINGE(0x1F489),
-    DROP_OF_BLOOD(0x1FA78),
-    PILL(0x1F48A),
-    ADHESIVE_BANDAGE(0x1FA79),
-    CRUTCH(0x1FA7C),
-    STETHOSCOPE(0x1FA7A),
-    X_RAY(0x1FA7B),
-    DOOR(0x1F6AA),
-    ELEVATOR(0x1F6D7),
-    MIRROR(0x1FA9E),
-    WINDOW(0x1FA9F),
-    BED(0x1F6CF),
-    COUCH_AND_LAMP(0x1F6CB),
-    CHAIR(0x1FA91),
-    TOILET(0x1F6BD),
-    PLUNGER(0x1FAA0),
-    SHOWER(0x1F6BF),
-    BATHTUB(0x1F6C1),
-    MOUSE_TRAP(0x1FAA4),
-    RAZOR(0x1FA92),
-    LOTION_BOTTLE(0x1F9F4),
-    SAFETY_PIN(0x1F9F7),
-    BROOM(0x1F9F9),
-    BASKET(0x1F9FA),
-    ROLL_OF_PAPER(0x1F9FB),
-    BUCKET(0x1FAA3),
-    SOAP(0x1F9FC),
-    BUBBLES(0x1FAE7),
-    TOOTHBRUSH(0x1FAA5),
-    SPONGE(0x1F9FD),
-    FIRE_EXTINGUISHER(0x1F9EF),
-    SHOPPING_CART(0x1F6D2),
-    CIGARETTE(0x1F6AC),
-    COFFIN(0x26B0),
-    HEADSTONE(0x1FAA6),
-    FUNERAL_URN(0x26B1),
-    MOAI(0x1F5FF),
-    PLACARD(0x1FAA7),
-    IDENTIFICATION_CARD(0x1FAAA),
-    ATM_SIGN(0x1F3E7),
-    LITTER_IN_BIN_SIGN(0x1F6AE),
-    POTABLE_WATER(0x1F6B0),
-    WHEELCHAIR_SYMBOL(0x267F),
-    MENS_ROOM(0x1F6B9),
-    WOMENS_ROOM(0x1F6BA),
-    RESTROOM(0x1F6BB),
-    BABY_SYMBOL(0x1F6BC),
-    WATER_CLOSET(0x1F6BE),
-    PASSPORT_CONTROL(0x1F6C2),
-    CUSTOMS(0x1F6C3),
-    BAGGAGE_CLAIM(0x1F6C4),
-    LEFT_LUGGAGE(0x1F6C5),
-    WARNING(0x26A0),
-    CHILDREN_CROSSING(0x1F6B8),
-    NO_ENTRY(0x26D4),
-    PROHIBITED(0x1F6AB),
-    NO_BICYCLES(0x1F6B3),
-    NO_SMOKING(0x1F6AD),
-    NO_LITTERING(0x1F6AF),
-    NON_POTABLE_WATER(0x1F6B1),
-    NO_PEDESTRIANS(0x1F6B7),
-    NO_MOBILE_PHONES(0x1F4F5),
-    NO_ONE_UNDER_EIGHTEEN(0x1F51E),
-    RADIOACTIVE(0x2622),
-    BIOHAZARD(0x2623),
-    UP_ARROW(0x2B06),
-    UP_RIGHT_ARROW(0x2197),
-    RIGHT_ARROW(0x27A1),
-    DOWN_RIGHT_ARROW(0x2198),
-    DOWN_ARROW(0x2B07),
-    DOWN_LEFT_ARROW(0x2199),
-    LEFT_ARROW(0x2B05),
-    UP_LEFT_ARROW(0x2196),
-    UP_DOWN_ARROW(0x2195),
-    LEFT_RIGHT_ARROW(0x2194),
-    RIGHT_ARROW_CURVING_LEFT(0x21A9),
-    LEFT_ARROW_CURVING_RIGHT(0x21AA),
-    RIGHT_ARROW_CURVING_UP(0x2934),
-    RIGHT_ARROW_CURVING_DOWN(0x2935),
-    CLOCKWISE_VERTICAL_ARROWS(0x1F503),
-    COUNTERCLOCKWISE_ARROWS_BUTTON(0x1F504),
-    BACK_ARROW(0x1F519),
-    END_ARROW(0x1F51A),
-    ON_ARROW(0x1F51B),
-    SOON_ARROW(0x1F51C),
-    TOP_ARROW(0x1F51D),
-    PLACE_OF_WORSHIP(0x1F6D0),
-    ATOM_SYMBOL(0x269B),
-    OM(0x1F549),
-    STAR_OF_DAVID(0x2721),
-    WHEEL_OF_DHARMA(0x2638),
-    YIN_YANG(0x262F),
-    LATIN_CROSS(0x271D),
-    ORTHODOX_CROSS(0x2626),
-    STAR_AND_CRESCENT(0x262A),
-    PEACE_SYMBOL(0x262E),
-    MENORAH(0x1F54E),
-    DOTTED_SIX_POINTED_STAR(0x1F52F),
-    ARIES(0x2648),
-    TAURUS(0x2649),
-    GEMINI(0x264A),
-    CANCER(0x264B),
-    LEO(0x264C),
-    VIRGO(0x264D),
-    LIBRA(0x264E),
-    SCORPIO(0x264F),
-    SAGITTARIUS(0x2650),
-    CAPRICORN(0x2651),
-    AQUARIUS(0x2652),
-    PISCES(0x2653),
-    OPHIUCHUS(0x26CE),
-    SHUFFLE_TRACKS_BUTTON(0x1F500),
-    REPEAT_BUTTON(0x1F501),
-    REPEAT_SINGLE_BUTTON(0x1F502),
-    PLAY_BUTTON(0x25B6),
-    FAST_FORWARD_BUTTON(0x23E9),
-    NEXT_TRACK_BUTTON(0x23ED),
-    PLAY_OR_PAUSE_BUTTON(0x23EF),
-    REVERSE_BUTTON(0x25C0),
-    FAST_REVERSE_BUTTON(0x23EA),
-    LAST_TRACK_BUTTON(0x23EE),
-    UPWARDS_BUTTON(0x1F53C),
-    FAST_UP_BUTTON(0x23EB),
-    DOWNWARDS_BUTTON(0x1F53D),
-    FAST_DOWN_BUTTON(0x23EC),
-    PAUSE_BUTTON(0x23F8),
-    STOP_BUTTON(0x23F9),
-    RECORD_BUTTON(0x23FA),
-    EJECT_BUTTON(0x23CF),
-    CINEMA(0x1F3A6),
-    DIM_BUTTON(0x1F505),
-    BRIGHT_BUTTON(0x1F506),
-    ANTENNA_BARS(0x1F4F6),
-    VIBRATION_MODE(0x1F4F3),
-    MOBILE_PHONE_OFF(0x1F4F4),
-    FEMALE_SIGN(0x2640),
-    MALE_SIGN(0x2642),
-    TRANSGENDER_SYMBOL(0x26A7),
-    MULTIPLY(0x2716),
-    PLUS(0x2795),
-    MINUS(0x2796),
-    DIVIDE(0x2797),
-    HEAVY_EQUALS_SIGN(0x1F7F0),
-    INFINITY(0x267E),
-    DOUBLE_EXCLAMATION_MARK(0x203C),
-    EXCLAMATION_QUESTION_MARK(0x2049),
-    RED_QUESTION_MARK(0x2753),
-    WHITE_QUESTION_MARK(0x2754),
-    WHITE_EXCLAMATION_MARK(0x2755),
-    RED_EXCLAMATION_MARK(0x2757),
-    WAVY_DASH(0x3030),
-    CURRENCY_EXCHANGE(0x1F4B1),
-    HEAVY_DOLLAR_SIGN(0x1F4B2),
-    MEDICAL_SYMBOL(0x2695),
-    RECYCLING_SYMBOL(0x267B),
-    FLEUR_DE_LIS(0x269C),
-    TRIDENT_EMBLEM(0x1F531),
-    NAME_BADGE(0x1F4DB),
-    JAPANESE_SYMBOL_FOR_BEGINNER(0x1F530),
-    HOLLOW_RED_CIRCLE(0x2B55),
-    CHECK_MARK_BUTTON(0x2705),
-    CHECK_BOX_WITH_CHECK(0x2611),
-    CHECK_MARK(0x2714),
-    CROSS_MARK(0x274C),
-    CROSS_MARK_BUTTON(0x274E),
-    CURLY_LOOP(0x27B0),
-    DOUBLE_CURLY_LOOP(0x27BF),
-    PART_ALTERNATION_MARK(0x303D),
-    EIGHT_SPOKED_ASTERISK(0x2733),
-    EIGHT_POINTED_STAR(0x2734),
-    SPARKLE(0x2747),
-    COPYRIGHT(0x00A9),
-    REGISTERED(0x00AE),
-    TRADE_MARK(0x2122),
-    KEYCAP_HASH(0x0023),
-    KEYCAP_STAR(0x002A),
-    KEYCAP_0(0x0030),
-    KEYCAP_1(0x0031),
-    KEYCAP_2(0x0032),
-    KEYCAP_3(0x0033),
-    KEYCAP_4(0x0034),
-    KEYCAP_5(0x0035),
-    KEYCAP_6(0x0036),
-    KEYCAP_7(0x0037),
-    KEYCAP_8(0x0038),
-    KEYCAP_9(0x0039),
-    KEYCAP_10(0x1F51F),
-    INPUT_LATIN_UPPERCASE(0x1F520),
-    INPUT_LATIN_LOWERCASE(0x1F521),
-    INPUT_NUMBERS(0x1F522),
-    INPUT_SYMBOLS(0x1F523),
-    INPUT_LATIN_LETTERS(0x1F524),
-    A_BUTTON_BLOOD_TYPE(0x1F170),
-    AB_BUTTON_BLOOD_TYPE(0x1F18E),
-    B_BUTTON_BLOOD_TYPE(0x1F171),
-    CL_BUTTON(0x1F191),
-    COOL_BUTTON(0x1F192),
-    FREE_BUTTON(0x1F193),
-    INFORMATION(0x2139),
-    ID_BUTTON(0x1F194),
-    CIRCLED_M(0x24C2),
-    NEW_BUTTON(0x1F195),
-    NG_BUTTON(0x1F196),
-    O_BUTTON_BLOOD_TYPE(0x1F17E),
-    OK_BUTTON(0x1F197),
-    P_BUTTON(0x1F17F),
-    SOS_BUTTON(0x1F198),
-    UP_BUTTON(0x1F199),
-    VS_BUTTON(0x1F19A),
-    JAPANESE_HERE_BUTTON(0x1F201),
-    JAPANESE_SERVICE_CHARGE_BUTTON(0x1F202),
-    JAPANESE_MONTHLY_AMOUNT_BUTTON(0x1F237),
-    JAPANESE_NOT_FREE_OF_CHARGE_BUTTON(0x1F236),
-    JAPANESE_RESERVED_BUTTON(0x1F22F),
-    JAPANESE_BARGAIN_BUTTON(0x1F250),
-    JAPANESE_DISCOUNT_BUTTON(0x1F239),
-    JAPANESE_FREE_OF_CHARGE_BUTTON(0x1F21A),
-    JAPANESE_PROHIBITED_BUTTON(0x1F232),
-    JAPANESE_ACCEPTABLE_BUTTON(0x1F251),
-    JAPANESE_APPLICATION_BUTTON(0x1F238),
-    JAPANESE_PASSING_GRADE_BUTTON(0x1F234),
-    JAPANESE_VACANCY_BUTTON(0x1F233),
-    JAPANESE_CONGRATULATIONS_BUTTON(0x3297),
-    JAPANESE_SECRET_BUTTON(0x3299),
-    JAPANESE_OPEN_FOR_BUSINESS_BUTTON(0x1F23A),
-    JAPANESE_NO_VACANCY_BUTTON(0x1F235),
-    RED_CIRCLE(0x1F534),
-    ORANGE_CIRCLE(0x1F7E0),
-    YELLOW_CIRCLE(0x1F7E1),
-    GREEN_CIRCLE(0x1F7E2),
-    BLUE_CIRCLE(0x1F535),
-    PURPLE_CIRCLE(0x1F7E3),
-    BROWN_CIRCLE(0x1F7E4),
-    BLACK_CIRCLE(0x26AB),
-    WHITE_CIRCLE(0x26AA),
-    RED_SQUARE(0x1F7E5),
-    ORANGE_SQUARE(0x1F7E7),
-    YELLOW_SQUARE(0x1F7E8),
-    GREEN_SQUARE(0x1F7E9),
-    BLUE_SQUARE(0x1F7E6),
-    PURPLE_SQUARE(0x1F7EA),
-    BROWN_SQUARE(0x1F7EB),
-    BLACK_LARGE_SQUARE(0x2B1B),
-    WHITE_LARGE_SQUARE(0x2B1C),
-    BLACK_MEDIUM_SQUARE(0x25FC),
-    WHITE_MEDIUM_SQUARE(0x25FB),
-    BLACK_MEDIUM_SMALL_SQUARE(0x25FE),
-    WHITE_MEDIUM_SMALL_SQUARE(0x25FD),
-    BLACK_SMALL_SQUARE(0x25AA),
-    WHITE_SMALL_SQUARE(0x25AB),
-    LARGE_ORANGE_DIAMOND(0x1F536),
-    LARGE_BLUE_DIAMOND(0x1F537),
-    SMALL_ORANGE_DIAMOND(0x1F538),
-    SMALL_BLUE_DIAMOND(0x1F539),
-    RED_TRIANGLE_POINTED_UP(0x1F53A),
-    RED_TRIANGLE_POINTED_DOWN(0x1F53B),
-    DIAMOND_WITH_A_DOT(0x1F4A0),
-    RADIO_BUTTON(0x1F518),
-    WHITE_SQUARE_BUTTON(0x1F533),
-    BLACK_SQUARE_BUTTON(0x1F532),
-    CHEQUERED_FLAG(0x1F3C1),
-    TRIANGULAR_FLAG(0x1F6A9),
-    CROSSED_FLAGS(0x1F38C),
-    BLACK_FLAG(0x1F3F4),
-    WHITE_FLAG(0x1F3F3),
-    RAINBOW_FLAG(0x1F3F3),
-    TRANSGENDER_FLAG(0x1F3F3),
-    PIRATE_FLAG(0x1F3F4),
-    FLAG_ASCENSION_ISLAND(0x1F1E6),
-    FLAG_ANDORRA(0x1F1E6),
-    FLAG_UNITED_ARAB_EMIRATES(0x1F1E6),
-    FLAG_AFGHANISTAN(0x1F1E6),
-    FLAG_ANTIGUA__BARBUDA(0x1F1E6),
-    FLAG_ANGUILLA(0x1F1E6),
-    FLAG_ALBANIA(0x1F1E6),
-    FLAG_ARMENIA(0x1F1E6),
-    FLAG_ANGOLA(0x1F1E6),
-    FLAG_ANTARCTICA(0x1F1E6),
-    FLAG_ARGENTINA(0x1F1E6),
-    FLAG_AMERICAN_SAMOA(0x1F1E6),
-    FLAG_AUSTRIA(0x1F1E6),
-    FLAG_AUSTRALIA(0x1F1E6),
-    FLAG_ARUBA(0x1F1E6),
-    FLAG_ÅLAND_ISLANDS(0x1F1E6),
-    FLAG_AZERBAIJAN(0x1F1E6),
-    FLAG_BOSNIA__HERZEGOVINA(0x1F1E7),
-    FLAG_BARBADOS(0x1F1E7),
-    FLAG_BANGLADESH(0x1F1E7),
-    FLAG_BELGIUM(0x1F1E7),
-    FLAG_BURKINA_FASO(0x1F1E7),
-    FLAG_BULGARIA(0x1F1E7),
-    FLAG_BAHRAIN(0x1F1E7),
-    FLAG_BURUNDI(0x1F1E7),
-    FLAG_BENIN(0x1F1E7),
-    FLAG_ST_BARTHÉLEMY(0x1F1E7),
-    FLAG_BERMUDA(0x1F1E7),
-    FLAG_BRUNEI(0x1F1E7),
-    FLAG_BOLIVIA(0x1F1E7),
-    FLAG_CARIBBEAN_NETHERLANDS(0x1F1E7),
-    FLAG_BRAZIL(0x1F1E7),
-    FLAG_BAHAMAS(0x1F1E7),
-    FLAG_BHUTAN(0x1F1E7),
-    FLAG_BOUVET_ISLAND(0x1F1E7),
-    FLAG_BOTSWANA(0x1F1E7),
-    FLAG_BELARUS(0x1F1E7),
-    FLAG_BELIZE(0x1F1E7),
-    FLAG_CANADA(0x1F1E8),
-    FLAG_COCOS_KEELING_ISLANDS(0x1F1E8),
-    FLAG_CONGO___KINSHASA(0x1F1E8),
-    FLAG_CENTRAL_AFRICAN_REPUBLIC(0x1F1E8),
-    FLAG_CONGO___BRAZZAVILLE(0x1F1E8),
-    FLAG_SWITZERLAND(0x1F1E8),
-    FLAG_CÔTE_DIVOIRE(0x1F1E8),
-    FLAG_COOK_ISLANDS(0x1F1E8),
-    FLAG_CHILE(0x1F1E8),
-    FLAG_CAMEROON(0x1F1E8),
-    FLAG_CHINA(0x1F1E8),
-    FLAG_COLOMBIA(0x1F1E8),
-    FLAG_CLIPPERTON_ISLAND(0x1F1E8),
-    FLAG_COSTA_RICA(0x1F1E8),
-    FLAG_CUBA(0x1F1E8),
-    FLAG_CAPE_VERDE(0x1F1E8),
-    FLAG_CURAÇAO(0x1F1E8),
-    FLAG_CHRISTMAS_ISLAND(0x1F1E8),
-    FLAG_CYPRUS(0x1F1E8),
-    FLAG_CZECHIA(0x1F1E8),
-    FLAG_GERMANY(0x1F1E9),
-    FLAG_DIEGO_GARCIA(0x1F1E9),
-    FLAG_DJIBOUTI(0x1F1E9),
-    FLAG_DENMARK(0x1F1E9),
-    FLAG_DOMINICA(0x1F1E9),
-    FLAG_DOMINICAN_REPUBLIC(0x1F1E9),
-    FLAG_ALGERIA(0x1F1E9),
-    FLAG_CEUTA__MELILLA(0x1F1EA),
-    FLAG_ECUADOR(0x1F1EA),
-    FLAG_ESTONIA(0x1F1EA),
-    FLAG_EGYPT(0x1F1EA),
-    FLAG_WESTERN_SAHARA(0x1F1EA),
-    FLAG_ERITREA(0x1F1EA),
-    FLAG_SPAIN(0x1F1EA),
-    FLAG_ETHIOPIA(0x1F1EA),
-    FLAG_EUROPEAN_UNION(0x1F1EA),
-    FLAG_FINLAND(0x1F1EB),
-    FLAG_FIJI(0x1F1EB),
-    FLAG_FALKLAND_ISLANDS(0x1F1EB),
-    FLAG_MICRONESIA(0x1F1EB),
-    FLAG_FAROE_ISLANDS(0x1F1EB),
-    FLAG_FRANCE(0x1F1EB),
-    FLAG_GABON(0x1F1EC),
-    FLAG_UNITED_KINGDOM(0x1F1EC),
-    FLAG_GRENADA(0x1F1EC),
-    FLAG_GEORGIA(0x1F1EC),
-    FLAG_FRENCH_GUIANA(0x1F1EC),
-    FLAG_GUERNSEY(0x1F1EC),
-    FLAG_GHANA(0x1F1EC),
-    FLAG_GIBRALTAR(0x1F1EC),
-    FLAG_GREENLAND(0x1F1EC),
-    FLAG_GAMBIA(0x1F1EC),
-    FLAG_GUINEA(0x1F1EC),
-    FLAG_GUADELOUPE(0x1F1EC),
-    FLAG_EQUATORIAL_GUINEA(0x1F1EC),
-    FLAG_GREECE(0x1F1EC),
-    FLAG_SOUTH_GEORGIA__SOUTH_SANDWICH_ISLANDS(0x1F1EC),
-    FLAG_GUATEMALA(0x1F1EC),
-    FLAG_GUAM(0x1F1EC),
-    FLAG_GUINEA_BISSAU(0x1F1EC),
-    FLAG_GUYANA(0x1F1EC),
-    FLAG_HONG_KONG_SAR_CHINA(0x1F1ED),
-    FLAG_HEARD__MCDONALD_ISLANDS(0x1F1ED),
-    FLAG_HONDURAS(0x1F1ED),
-    FLAG_CROATIA(0x1F1ED),
-    FLAG_HAITI(0x1F1ED),
-    FLAG_HUNGARY(0x1F1ED),
-    FLAG_CANARY_ISLANDS(0x1F1EE),
-    FLAG_INDONESIA(0x1F1EE),
-    FLAG_IRELAND(0x1F1EE),
-    FLAG_ISRAEL(0x1F1EE),
-    FLAG_ISLE_OF_MAN(0x1F1EE),
-    FLAG_INDIA(0x1F1EE),
-    FLAG_BRITISH_INDIAN_OCEAN_TERRITORY(0x1F1EE),
-    FLAG_IRAQ(0x1F1EE),
-    FLAG_IRAN(0x1F1EE),
-    FLAG_ICELAND(0x1F1EE),
-    FLAG_ITALY(0x1F1EE),
-    FLAG_JERSEY(0x1F1EF),
-    FLAG_JAMAICA(0x1F1EF),
-    FLAG_JORDAN(0x1F1EF),
-    FLAG_JAPAN(0x1F1EF),
-    FLAG_KENYA(0x1F1F0),
-    FLAG_KYRGYZSTAN(0x1F1F0),
-    FLAG_CAMBODIA(0x1F1F0),
-    FLAG_KIRIBATI(0x1F1F0),
-    FLAG_COMOROS(0x1F1F0),
-    FLAG_ST_KITTS__NEVIS(0x1F1F0),
-    FLAG_NORTH_KOREA(0x1F1F0),
-    FLAG_SOUTH_KOREA(0x1F1F0),
-    FLAG_KUWAIT(0x1F1F0),
-    FLAG_CAYMAN_ISLANDS(0x1F1F0),
-    FLAG_KAZAKHSTAN(0x1F1F0),
-    FLAG_LAOS(0x1F1F1),
-    FLAG_LEBANON(0x1F1F1),
-    FLAG_ST_LUCIA(0x1F1F1),
-    FLAG_LIECHTENSTEIN(0x1F1F1),
-    FLAG_SRI_LANKA(0x1F1F1),
-    FLAG_LIBERIA(0x1F1F1),
-    FLAG_LESOTHO(0x1F1F1),
-    FLAG_LITHUANIA(0x1F1F1),
-    FLAG_LUXEMBOURG(0x1F1F1),
-    FLAG_LATVIA(0x1F1F1),
-    FLAG_LIBYA(0x1F1F1),
-    FLAG_MOROCCO(0x1F1F2),
-    FLAG_MONACO(0x1F1F2),
-    FLAG_MOLDOVA(0x1F1F2),
-    FLAG_MONTENEGRO(0x1F1F2),
-    FLAG_ST_MARTIN(0x1F1F2),
-    FLAG_MADAGASCAR(0x1F1F2),
-    FLAG_MARSHALL_ISLANDS(0x1F1F2),
-    FLAG_NORTH_MACEDONIA(0x1F1F2),
-    FLAG_MALI(0x1F1F2),
-    FLAG_MYANMAR_BURMA(0x1F1F2),
-    FLAG_MONGOLIA(0x1F1F2),
-    FLAG_MACAO_SAR_CHINA(0x1F1F2),
-    FLAG_NORTHERN_MARIANA_ISLANDS(0x1F1F2),
-    FLAG_MARTINIQUE(0x1F1F2),
-    FLAG_MAURITANIA(0x1F1F2),
-    FLAG_MONTSERRAT(0x1F1F2),
-    FLAG_MALTA(0x1F1F2),
-    FLAG_MAURITIUS(0x1F1F2),
-    FLAG_MALDIVES(0x1F1F2),
-    FLAG_MALAWI(0x1F1F2),
-    FLAG_MEXICO(0x1F1F2),
-    FLAG_MALAYSIA(0x1F1F2),
-    FLAG_MOZAMBIQUE(0x1F1F2),
-    FLAG_NAMIBIA(0x1F1F3),
-    FLAG_NEW_CALEDONIA(0x1F1F3),
-    FLAG_NIGER(0x1F1F3),
-    FLAG_NORFOLK_ISLAND(0x1F1F3),
-    FLAG_NIGERIA(0x1F1F3),
-    FLAG_NICARAGUA(0x1F1F3),
-    FLAG_NETHERLANDS(0x1F1F3),
-    FLAG_NORWAY(0x1F1F3),
-    FLAG_NEPAL(0x1F1F3),
-    FLAG_NAURU(0x1F1F3),
-    FLAG_NIUE(0x1F1F3),
-    FLAG_NEW_ZEALAND(0x1F1F3),
-    FLAG_OMAN(0x1F1F4),
-    FLAG_PANAMA(0x1F1F5),
-    FLAG_PERU(0x1F1F5),
-    FLAG_FRENCH_POLYNESIA(0x1F1F5),
-    FLAG_PAPUA_NEW_GUINEA(0x1F1F5),
-    FLAG_PHILIPPINES(0x1F1F5),
-    FLAG_PAKISTAN(0x1F1F5),
-    FLAG_POLAND(0x1F1F5),
-    FLAG_ST_PIERRE__MIQUELON(0x1F1F5),
-    FLAG_PITCAIRN_ISLANDS(0x1F1F5),
-    FLAG_PUERTO_RICO(0x1F1F5),
-    FLAG_PALESTINIAN_TERRITORIES(0x1F1F5),
-    FLAG_PORTUGAL(0x1F1F5),
-    FLAG_PALAU(0x1F1F5),
-    FLAG_PARAGUAY(0x1F1F5),
-    FLAG_QATAR(0x1F1F6),
-    FLAG_RÉUNION(0x1F1F7),
-    FLAG_ROMANIA(0x1F1F7),
-    FLAG_SERBIA(0x1F1F7),
-    FLAG_RUSSIA(0x1F1F7),
-    FLAG_RWANDA(0x1F1F7),
-    FLAG_SAUDI_ARABIA(0x1F1F8),
-    FLAG_SOLOMON_ISLANDS(0x1F1F8),
-    FLAG_SEYCHELLES(0x1F1F8),
-    FLAG_SUDAN(0x1F1F8),
-    FLAG_SWEDEN(0x1F1F8),
-    FLAG_SINGAPORE(0x1F1F8),
-    FLAG_ST_HELENA(0x1F1F8),
-    FLAG_SLOVENIA(0x1F1F8),
-    FLAG_SVALBARD__JAN_MAYEN(0x1F1F8),
-    FLAG_SLOVAKIA(0x1F1F8),
-    FLAG_SIERRA_LEONE(0x1F1F8),
-    FLAG_SAN_MARINO(0x1F1F8),
-    FLAG_SENEGAL(0x1F1F8),
-    FLAG_SOMALIA(0x1F1F8),
-    FLAG_SURINAME(0x1F1F8),
-    FLAG_SOUTH_SUDAN(0x1F1F8),
-    FLAG_SÃO_TOMÉ__PRÍNCIPE(0x1F1F8),
-    FLAG_EL_SALVADOR(0x1F1F8),
-    FLAG_SINT_MAARTEN(0x1F1F8),
-    FLAG_SYRIA(0x1F1F8),
-    FLAG_ESWATINI(0x1F1F8),
-    FLAG_TRISTAN_DA_CUNHA(0x1F1F9),
-    FLAG_TURKS__CAICOS_ISLANDS(0x1F1F9),
-    FLAG_CHAD(0x1F1F9),
-    FLAG_FRENCH_SOUTHERN_TERRITORIES(0x1F1F9),
-    FLAG_TOGO(0x1F1F9),
-    FLAG_THAILAND(0x1F1F9),
-    FLAG_TAJIKISTAN(0x1F1F9),
-    FLAG_TOKELAU(0x1F1F9),
-    FLAG_TIMOR_LESTE(0x1F1F9),
-    FLAG_TURKMENISTAN(0x1F1F9),
-    FLAG_TUNISIA(0x1F1F9),
-    FLAG_TONGA(0x1F1F9),
-    FLAG_TURKEY(0x1F1F9),
-    FLAG_TRINIDAD__TOBAGO(0x1F1F9),
-    FLAG_TUVALU(0x1F1F9),
-    FLAG_TAIWAN(0x1F1F9),
-    FLAG_TANZANIA(0x1F1F9),
-    FLAG_UKRAINE(0x1F1FA),
-    FLAG_UGANDA(0x1F1FA),
-    FLAG_US_OUTLYING_ISLANDS(0x1F1FA),
-    FLAG_UNITED_NATIONS(0x1F1FA),
-    FLAG_UNITED_STATES(0x1F1FA),
-    FLAG_URUGUAY(0x1F1FA),
-    FLAG_UZBEKISTAN(0x1F1FA),
-    FLAG_VATICAN_CITY(0x1F1FB),
-    FLAG_ST_VINCENT__GRENADINES(0x1F1FB),
-    FLAG_VENEZUELA(0x1F1FB),
-    FLAG_BRITISH_VIRGIN_ISLANDS(0x1F1FB),
-    FLAG_US_VIRGIN_ISLANDS(0x1F1FB),
-    FLAG_VIETNAM(0x1F1FB),
-    FLAG_VANUATU(0x1F1FB),
-    FLAG_WALLIS__FUTUNA(0x1F1FC),
-    FLAG_SAMOA(0x1F1FC),
-    FLAG_KOSOVO(0x1F1FD),
-    FLAG_YEMEN(0x1F1FE),
-    FLAG_MAYOTTE(0x1F1FE),
-    FLAG_SOUTH_AFRICA(0x1F1FF),
-    FLAG_ZAMBIA(0x1F1FF),
-    FLAG_ZIMBABWE(0x1F1FF),
-    FLAG_ENGLAND(0x1F3F4),
-    FLAG_SCOTLAND(0x1F3F4),
-    FLAG_WALES(0x1F3F4);
-
-    override fun toString(): String {
-        return String(Character.toChars(code))
-    }
-}
