+++ src/main/kotlin/net/purefunc/emoji/Emoji0.kt
@@ -1,0 +1,1011 @@
+package net.purefunc.emoji
+
+enum class Emoji0(
+    private val code: Int,
+) {
+
+    GRINNING_FACE(0x1F600),
+    GRINNING_FACE_WITH_BIG_EYES(0x1F603),
+    GRINNING_FACE_WITH_SMILING_EYES(0x1F604),
+    BEAMING_FACE_WITH_SMILING_EYES(0x1F601),
+    GRINNING_SQUINTING_FACE(0x1F606),
+    GRINNING_FACE_WITH_SWEAT(0x1F605),
+    ROLLING_ON_THE_FLOOR_LAUGHING(0x1F923),
+    FACE_WITH_TEARS_OF_JOY(0x1F602),
+    SLIGHTLY_SMILING_FACE(0x1F642),
+    UPSIDE_DOWN_FACE(0x1F643),
+    MELTING_FACE(0x1FAE0),
+    WINKING_FACE(0x1F609),
+    SMILING_FACE_WITH_SMILING_EYES(0x1F60A),
+    SMILING_FACE_WITH_HALO(0x1F607),
+    SMILING_FACE_WITH_HEARTS(0x1F970),
+    SMILING_FACE_WITH_HEART_EYES(0x1F60D),
+    STAR_STRUCK(0x1F929),
+    FACE_BLOWING_A_KISS(0x1F618),
+    KISSING_FACE(0x1F617),
+    SMILING_FACE(0x263A),
+    KISSING_FACE_WITH_CLOSED_EYES(0x1F61A),
+    KISSING_FACE_WITH_SMILING_EYES(0x1F619),
+    SMILING_FACE_WITH_TEAR(0x1F972),
+    FACE_SAVORING_FOOD(0x1F60B),
+    FACE_WITH_TONGUE(0x1F61B),
+    WINKING_FACE_WITH_TONGUE(0x1F61C),
+    ZANY_FACE(0x1F92A),
+    SQUINTING_FACE_WITH_TONGUE(0x1F61D),
+    MONEY_MOUTH_FACE(0x1F911),
+    SMILING_FACE_WITH_OPEN_HANDS(0x1F917),
+    FACE_WITH_HAND_OVER_MOUTH(0x1F92D),
+    FACE_WITH_OPEN_EYES_AND_HAND_OVER_MOUTH(0x1FAE2),
+    FACE_WITH_PEEKING_EYE(0x1FAE3),
+    SHUSHING_FACE(0x1F92B),
+    THINKING_FACE(0x1F914),
+    SALUTING_FACE(0x1FAE1),
+    ZIPPER_MOUTH_FACE(0x1F910),
+    FACE_WITH_RAISED_EYEBROW(0x1F928),
+    NEUTRAL_FACE(0x1F610),
+    EXPRESSIONLESS_FACE(0x1F611),
+    FACE_WITHOUT_MOUTH(0x1F636),
+    DOTTED_LINE_FACE(0x1FAE5),
+    FACE_IN_CLOUDS(0x1F636),
+    SMIRKING_FACE(0x1F60F),
+    UNAMUSED_FACE(0x1F612),
+    FACE_WITH_ROLLING_EYES(0x1F644),
+    GRIMACING_FACE(0x1F62C),
+    FACE_EXHALING(0x1F62E),
+    LYING_FACE(0x1F925),
+    RELIEVED_FACE(0x1F60C),
+    PENSIVE_FACE(0x1F614),
+    SLEEPY_FACE(0x1F62A),
+    DROOLING_FACE(0x1F924),
+    SLEEPING_FACE(0x1F634),
+    FACE_WITH_MEDICAL_MASK(0x1F637),
+    FACE_WITH_THERMOMETER(0x1F912),
+    FACE_WITH_HEAD_BANDAGE(0x1F915),
+    NAUSEATED_FACE(0x1F922),
+    FACE_VOMITING(0x1F92E),
+    SNEEZING_FACE(0x1F927),
+    HOT_FACE(0x1F975),
+    COLD_FACE(0x1F976),
+    WOOZY_FACE(0x1F974),
+    FACE_WITH_CROSSED_OUT_EYES(0x1F635),
+    FACE_WITH_SPIRAL_EYES(0x1F635),
+    EXPLODING_HEAD(0x1F92F),
+    COWBOY_HAT_FACE(0x1F920),
+    PARTYING_FACE(0x1F973),
+    DISGUISED_FACE(0x1F978),
+    SMILING_FACE_WITH_SUNGLASSES(0x1F60E),
+    NERD_FACE(0x1F913),
+    FACE_WITH_MONOCLE(0x1F9D0),
+    CONFUSED_FACE(0x1F615),
+    FACE_WITH_DIAGONAL_MOUTH(0x1FAE4),
+    WORRIED_FACE(0x1F61F),
+    SLIGHTLY_FROWNING_FACE(0x1F641),
+    FROWNING_FACE(0x2639),
+    FACE_WITH_OPEN_MOUTH(0x1F62E),
+    HUSHED_FACE(0x1F62F),
+    ASTONISHED_FACE(0x1F632),
+    FLUSHED_FACE(0x1F633),
+    PLEADING_FACE(0x1F97A),
+    FACE_HOLDING_BACK_TEARS(0x1F979),
+    FROWNING_FACE_WITH_OPEN_MOUTH(0x1F626),
+    ANGUISHED_FACE(0x1F627),
+    FEARFUL_FACE(0x1F628),
+    ANXIOUS_FACE_WITH_SWEAT(0x1F630),
+    SAD_BUT_RELIEVED_FACE(0x1F625),
+    CRYING_FACE(0x1F622),
+    LOUDLY_CRYING_FACE(0x1F62D),
+    FACE_SCREAMING_IN_FEAR(0x1F631),
+    CONFOUNDED_FACE(0x1F616),
+    PERSEVERING_FACE(0x1F623),
+    DISAPPOINTED_FACE(0x1F61E),
+    DOWNCAST_FACE_WITH_SWEAT(0x1F613),
+    WEARY_FACE(0x1F629),
+    TIRED_FACE(0x1F62B),
+    YAWNING_FACE(0x1F971),
+    FACE_WITH_STEAM_FROM_NOSE(0x1F624),
+    POUTING_FACE(0x1F621),
+    ANGRY_FACE(0x1F620),
+    FACE_WITH_SYMBOLS_ON_MOUTH(0x1F92C),
+    SMILING_FACE_WITH_HORNS(0x1F608),
+    ANGRY_FACE_WITH_HORNS(0x1F47F),
+    SKULL(0x1F480),
+    SKULL_AND_CROSSBONES(0x2620),
+    PILE_OF_POO(0x1F4A9),
+    CLOWN_FACE(0x1F921),
+    OGRE(0x1F479),
+    GOBLIN(0x1F47A),
+    GHOST(0x1F47B),
+    ALIEN(0x1F47D),
+    ALIEN_MONSTER(0x1F47E),
+    ROBOT(0x1F916),
+    GRINNING_CAT(0x1F63A),
+    GRINNING_CAT_WITH_SMILING_EYES(0x1F638),
+    CAT_WITH_TEARS_OF_JOY(0x1F639),
+    SMILING_CAT_WITH_HEART_EYES(0x1F63B),
+    CAT_WITH_WRY_SMILE(0x1F63C),
+    KISSING_CAT(0x1F63D),
+    WEARY_CAT(0x1F640),
+    CRYING_CAT(0x1F63F),
+    POUTING_CAT(0x1F63E),
+    SEE_NO_EVIL_MONKEY(0x1F648),
+    HEAR_NO_EVIL_MONKEY(0x1F649),
+    SPEAK_NO_EVIL_MONKEY(0x1F64A),
+    KISS_MARK(0x1F48B),
+    LOVE_LETTER(0x1F48C),
+    HEART_WITH_ARROW(0x1F498),
+    HEART_WITH_RIBBON(0x1F49D),
+    SPARKLING_HEART(0x1F496),
+    GROWING_HEART(0x1F497),
+    BEATING_HEART(0x1F493),
+    REVOLVING_HEARTS(0x1F49E),
+    TWO_HEARTS(0x1F495),
+    HEART_DECORATION(0x1F49F),
+    HEART_EXCLAMATION(0x2763),
+    BROKEN_HEART(0x1F494),
+    HEART_ON_FIRE(0x2764),
+    MENDING_HEART(0x2764),
+    RED_HEART(0x2764),
+    ORANGE_HEART(0x1F9E1),
+    YELLOW_HEART(0x1F49B),
+    GREEN_HEART(0x1F49A),
+    BLUE_HEART(0x1F499),
+    PURPLE_HEART(0x1F49C),
+    BROWN_HEART(0x1F90E),
+    BLACK_HEART(0x1F5A4),
+    WHITE_HEART(0x1F90D),
+    HUNDRED_POINTS(0x1F4AF),
+    ANGER_SYMBOL(0x1F4A2),
+    COLLISION(0x1F4A5),
+    DIZZY(0x1F4AB),
+    SWEAT_DROPLETS(0x1F4A6),
+    DASHING_AWAY(0x1F4A8),
+    HOLE(0x1F573),
+    BOMB(0x1F4A3),
+    SPEECH_BALLOON(0x1F4AC),
+    EYE_IN_SPEECH_BUBBLE(0x1F441),
+    LEFT_SPEECH_BUBBLE(0x1F5E8),
+    RIGHT_ANGER_BUBBLE(0x1F5EF),
+    THOUGHT_BALLOON(0x1F4AD),
+    ZZZ(0x1F4A4),
+    WAVING_HAND(0x1F44B),
+    WAVING_HAND_LIGHT_SKIN_TONE(0x1F44B),
+    WAVING_HAND_MEDIUM_LIGHT_SKIN_TONE(0x1F44B),
+    WAVING_HAND_MEDIUM_SKIN_TONE(0x1F44B),
+    WAVING_HAND_MEDIUM_DARK_SKIN_TONE(0x1F44B),
+    WAVING_HAND_DARK_SKIN_TONE(0x1F44B),
+    RAISED_BACK_OF_HAND(0x1F91A),
+    RAISED_BACK_OF_HAND_LIGHT_SKIN_TONE(0x1F91A),
+    RAISED_BACK_OF_HAND_MEDIUM_LIGHT_SKIN_TONE(0x1F91A),
+    RAISED_BACK_OF_HAND_MEDIUM_SKIN_TONE(0x1F91A),
+    RAISED_BACK_OF_HAND_MEDIUM_DARK_SKIN_TONE(0x1F91A),
+    RAISED_BACK_OF_HAND_DARK_SKIN_TONE(0x1F91A),
+    HAND_WITH_FINGERS_SPLAYED(0x1F590),
+    HAND_WITH_FINGERS_SPLAYED_LIGHT_SKIN_TONE(0x1F590),
+    HAND_WITH_FINGERS_SPLAYED_MEDIUM_LIGHT_SKIN_TONE(0x1F590),
+    HAND_WITH_FINGERS_SPLAYED_MEDIUM_SKIN_TONE(0x1F590),
+    HAND_WITH_FINGERS_SPLAYED_MEDIUM_DARK_SKIN_TONE(0x1F590),
+    HAND_WITH_FINGERS_SPLAYED_DARK_SKIN_TONE(0x1F590),
+    RAISED_HAND(0x270B),
+    RAISED_HAND_LIGHT_SKIN_TONE(0x270B),
+    RAISED_HAND_MEDIUM_LIGHT_SKIN_TONE(0x270B),
+    RAISED_HAND_MEDIUM_SKIN_TONE(0x270B),
+    RAISED_HAND_MEDIUM_DARK_SKIN_TONE(0x270B),
+    RAISED_HAND_DARK_SKIN_TONE(0x270B),
+    VULCAN_SALUTE(0x1F596),
+    VULCAN_SALUTE_LIGHT_SKIN_TONE(0x1F596),
+    VULCAN_SALUTE_MEDIUM_LIGHT_SKIN_TONE(0x1F596),
+    VULCAN_SALUTE_MEDIUM_SKIN_TONE(0x1F596),
+    VULCAN_SALUTE_MEDIUM_DARK_SKIN_TONE(0x1F596),
+    VULCAN_SALUTE_DARK_SKIN_TONE(0x1F596),
+    RIGHTWARDS_HAND(0x1FAF1),
+    RIGHTWARDS_HAND_LIGHT_SKIN_TONE(0x1FAF1),
+    RIGHTWARDS_HAND_MEDIUM_LIGHT_SKIN_TONE(0x1FAF1),
+    RIGHTWARDS_HAND_MEDIUM_SKIN_TONE(0x1FAF1),
+    RIGHTWARDS_HAND_MEDIUM_DARK_SKIN_TONE(0x1FAF1),
+    RIGHTWARDS_HAND_DARK_SKIN_TONE(0x1FAF1),
+    LEFTWARDS_HAND(0x1FAF2),
+    LEFTWARDS_HAND_LIGHT_SKIN_TONE(0x1FAF2),
+    LEFTWARDS_HAND_MEDIUM_LIGHT_SKIN_TONE(0x1FAF2),
+    LEFTWARDS_HAND_MEDIUM_SKIN_TONE(0x1FAF2),
+    LEFTWARDS_HAND_MEDIUM_DARK_SKIN_TONE(0x1FAF2),
+    LEFTWARDS_HAND_DARK_SKIN_TONE(0x1FAF2),
+    PALM_DOWN_HAND(0x1FAF3),
+    PALM_DOWN_HAND_LIGHT_SKIN_TONE(0x1FAF3),
+    PALM_DOWN_HAND_MEDIUM_LIGHT_SKIN_TONE(0x1FAF3),
+    PALM_DOWN_HAND_MEDIUM_SKIN_TONE(0x1FAF3),
+    PALM_DOWN_HAND_MEDIUM_DARK_SKIN_TONE(0x1FAF3),
+    PALM_DOWN_HAND_DARK_SKIN_TONE(0x1FAF3),
+    PALM_UP_HAND(0x1FAF4),
+    PALM_UP_HAND_LIGHT_SKIN_TONE(0x1FAF4),
+    PALM_UP_HAND_MEDIUM_LIGHT_SKIN_TONE(0x1FAF4),
+    PALM_UP_HAND_MEDIUM_SKIN_TONE(0x1FAF4),
+    PALM_UP_HAND_MEDIUM_DARK_SKIN_TONE(0x1FAF4),
+    PALM_UP_HAND_DARK_SKIN_TONE(0x1FAF4),
+    OK_HAND(0x1F44C),
+    OK_HAND_LIGHT_SKIN_TONE(0x1F44C),
+    OK_HAND_MEDIUM_LIGHT_SKIN_TONE(0x1F44C),
+    OK_HAND_MEDIUM_SKIN_TONE(0x1F44C),
+    OK_HAND_MEDIUM_DARK_SKIN_TONE(0x1F44C),
+    OK_HAND_DARK_SKIN_TONE(0x1F44C),
+    PINCHED_FINGERS(0x1F90C),
+    PINCHED_FINGERS_LIGHT_SKIN_TONE(0x1F90C),
+    PINCHED_FINGERS_MEDIUM_LIGHT_SKIN_TONE(0x1F90C),
+    PINCHED_FINGERS_MEDIUM_SKIN_TONE(0x1F90C),
+    PINCHED_FINGERS_MEDIUM_DARK_SKIN_TONE(0x1F90C),
+    PINCHED_FINGERS_DARK_SKIN_TONE(0x1F90C),
+    PINCHING_HAND(0x1F90F),
+    PINCHING_HAND_LIGHT_SKIN_TONE(0x1F90F),
+    PINCHING_HAND_MEDIUM_LIGHT_SKIN_TONE(0x1F90F),
+    PINCHING_HAND_MEDIUM_SKIN_TONE(0x1F90F),
+    PINCHING_HAND_MEDIUM_DARK_SKIN_TONE(0x1F90F),
+    PINCHING_HAND_DARK_SKIN_TONE(0x1F90F),
+    VICTORY_HAND(0x270C),
+    VICTORY_HAND_LIGHT_SKIN_TONE(0x270C),
+    VICTORY_HAND_MEDIUM_LIGHT_SKIN_TONE(0x270C),
+    VICTORY_HAND_MEDIUM_SKIN_TONE(0x270C),
+    VICTORY_HAND_MEDIUM_DARK_SKIN_TONE(0x270C),
+    VICTORY_HAND_DARK_SKIN_TONE(0x270C),
+    CROSSED_FINGERS(0x1F91E),
+    CROSSED_FINGERS_LIGHT_SKIN_TONE(0x1F91E),
+    CROSSED_FINGERS_MEDIUM_LIGHT_SKIN_TONE(0x1F91E),
+    CROSSED_FINGERS_MEDIUM_SKIN_TONE(0x1F91E),
+    CROSSED_FINGERS_MEDIUM_DARK_SKIN_TONE(0x1F91E),
+    CROSSED_FINGERS_DARK_SKIN_TONE(0x1F91E),
+    HAND_WITH_INDEX_FINGER_AND_THUMB_CROSSED(0x1FAF0),
+    HAND_WITH_INDEX_FINGER_AND_THUMB_CROSSED_LIGHT_SKIN_TONE(0x1FAF0),
+    HAND_WITH_INDEX_FINGER_AND_THUMB_CROSSED_MEDIUM_LIGHT_SKIN_TONE(0x1FAF0),
+    HAND_WITH_INDEX_FINGER_AND_THUMB_CROSSED_MEDIUM_SKIN_TONE(0x1FAF0),
+    HAND_WITH_INDEX_FINGER_AND_THUMB_CROSSED_MEDIUM_DARK_SKIN_TONE(0x1FAF0),
+    HAND_WITH_INDEX_FINGER_AND_THUMB_CROSSED_DARK_SKIN_TONE(0x1FAF0),
+    LOVE_YOU_GESTURE(0x1F91F),
+    LOVE_YOU_GESTURE_LIGHT_SKIN_TONE(0x1F91F),
+    LOVE_YOU_GESTURE_MEDIUM_LIGHT_SKIN_TONE(0x1F91F),
+    LOVE_YOU_GESTURE_MEDIUM_SKIN_TONE(0x1F91F),
+    LOVE_YOU_GESTURE_MEDIUM_DARK_SKIN_TONE(0x1F91F),
+    LOVE_YOU_GESTURE_DARK_SKIN_TONE(0x1F91F),
+    SIGN_OF_THE_HORNS(0x1F918),
+    SIGN_OF_THE_HORNS_LIGHT_SKIN_TONE(0x1F918),
+    SIGN_OF_THE_HORNS_MEDIUM_LIGHT_SKIN_TONE(0x1F918),
+    SIGN_OF_THE_HORNS_MEDIUM_SKIN_TONE(0x1F918),
+    SIGN_OF_THE_HORNS_MEDIUM_DARK_SKIN_TONE(0x1F918),
+    SIGN_OF_THE_HORNS_DARK_SKIN_TONE(0x1F918),
+    CALL_ME_HAND(0x1F919),
+    CALL_ME_HAND_LIGHT_SKIN_TONE(0x1F919),
+    CALL_ME_HAND_MEDIUM_LIGHT_SKIN_TONE(0x1F919),
+    CALL_ME_HAND_MEDIUM_SKIN_TONE(0x1F919),
+    CALL_ME_HAND_MEDIUM_DARK_SKIN_TONE(0x1F919),
+    CALL_ME_HAND_DARK_SKIN_TONE(0x1F919),
+    BACKHAND_INDEX_POINTING_LEFT(0x1F448),
+    BACKHAND_INDEX_POINTING_LEFT_LIGHT_SKIN_TONE(0x1F448),
+    BACKHAND_INDEX_POINTING_LEFT_MEDIUM_LIGHT_SKIN_TONE(0x1F448),
+    BACKHAND_INDEX_POINTING_LEFT_MEDIUM_SKIN_TONE(0x1F448),
+    BACKHAND_INDEX_POINTING_LEFT_MEDIUM_DARK_SKIN_TONE(0x1F448),
+    BACKHAND_INDEX_POINTING_LEFT_DARK_SKIN_TONE(0x1F448),
+    BACKHAND_INDEX_POINTING_RIGHT(0x1F449),
+    BACKHAND_INDEX_POINTING_RIGHT_LIGHT_SKIN_TONE(0x1F449),
+    BACKHAND_INDEX_POINTING_RIGHT_MEDIUM_LIGHT_SKIN_TONE(0x1F449),
+    BACKHAND_INDEX_POINTING_RIGHT_MEDIUM_SKIN_TONE(0x1F449),
+    BACKHAND_INDEX_POINTING_RIGHT_MEDIUM_DARK_SKIN_TONE(0x1F449),
+    BACKHAND_INDEX_POINTING_RIGHT_DARK_SKIN_TONE(0x1F449),
+    BACKHAND_INDEX_POINTING_UP(0x1F446),
+    BACKHAND_INDEX_POINTING_UP_LIGHT_SKIN_TONE(0x1F446),
+    BACKHAND_INDEX_POINTING_UP_MEDIUM_LIGHT_SKIN_TONE(0x1F446),
+    BACKHAND_INDEX_POINTING_UP_MEDIUM_SKIN_TONE(0x1F446),
+    BACKHAND_INDEX_POINTING_UP_MEDIUM_DARK_SKIN_TONE(0x1F446),
+    BACKHAND_INDEX_POINTING_UP_DARK_SKIN_TONE(0x1F446),
+    MIDDLE_FINGER(0x1F595),
+    MIDDLE_FINGER_LIGHT_SKIN_TONE(0x1F595),
+    MIDDLE_FINGER_MEDIUM_LIGHT_SKIN_TONE(0x1F595),
+    MIDDLE_FINGER_MEDIUM_SKIN_TONE(0x1F595),
+    MIDDLE_FINGER_MEDIUM_DARK_SKIN_TONE(0x1F595),
+    MIDDLE_FINGER_DARK_SKIN_TONE(0x1F595),
+    BACKHAND_INDEX_POINTING_DOWN(0x1F447),
+    BACKHAND_INDEX_POINTING_DOWN_LIGHT_SKIN_TONE(0x1F447),
+    BACKHAND_INDEX_POINTING_DOWN_MEDIUM_LIGHT_SKIN_TONE(0x1F447),
+    BACKHAND_INDEX_POINTING_DOWN_MEDIUM_SKIN_TONE(0x1F447),
+    BACKHAND_INDEX_POINTING_DOWN_MEDIUM_DARK_SKIN_TONE(0x1F447),
+    BACKHAND_INDEX_POINTING_DOWN_DARK_SKIN_TONE(0x1F447),
+    INDEX_POINTING_UP(0x261D),
+    INDEX_POINTING_UP_LIGHT_SKIN_TONE(0x261D),
+    INDEX_POINTING_UP_MEDIUM_LIGHT_SKIN_TONE(0x261D),
+    INDEX_POINTING_UP_MEDIUM_SKIN_TONE(0x261D),
+    INDEX_POINTING_UP_MEDIUM_DARK_SKIN_TONE(0x261D),
+    INDEX_POINTING_UP_DARK_SKIN_TONE(0x261D),
+    INDEX_POINTING_AT_THE_VIEWER(0x1FAF5),
+    INDEX_POINTING_AT_THE_VIEWER_LIGHT_SKIN_TONE(0x1FAF5),
+    INDEX_POINTING_AT_THE_VIEWER_MEDIUM_LIGHT_SKIN_TONE(0x1FAF5),
+    INDEX_POINTING_AT_THE_VIEWER_MEDIUM_SKIN_TONE(0x1FAF5),
+    INDEX_POINTING_AT_THE_VIEWER_MEDIUM_DARK_SKIN_TONE(0x1FAF5),
+    INDEX_POINTING_AT_THE_VIEWER_DARK_SKIN_TONE(0x1FAF5),
+    THUMBS_UP(0x1F44D),
+    THUMBS_UP_LIGHT_SKIN_TONE(0x1F44D),
+    THUMBS_UP_MEDIUM_LIGHT_SKIN_TONE(0x1F44D),
+    THUMBS_UP_MEDIUM_SKIN_TONE(0x1F44D),
+    THUMBS_UP_MEDIUM_DARK_SKIN_TONE(0x1F44D),
+    THUMBS_UP_DARK_SKIN_TONE(0x1F44D),
+    THUMBS_DOWN(0x1F44E),
+    THUMBS_DOWN_LIGHT_SKIN_TONE(0x1F44E),
+    THUMBS_DOWN_MEDIUM_LIGHT_SKIN_TONE(0x1F44E),
+    THUMBS_DOWN_MEDIUM_SKIN_TONE(0x1F44E),
+    THUMBS_DOWN_MEDIUM_DARK_SKIN_TONE(0x1F44E),
+    THUMBS_DOWN_DARK_SKIN_TONE(0x1F44E),
+    RAISED_FIST(0x270A),
+    RAISED_FIST_LIGHT_SKIN_TONE(0x270A),
+    RAISED_FIST_MEDIUM_LIGHT_SKIN_TONE(0x270A),
+    RAISED_FIST_MEDIUM_SKIN_TONE(0x270A),
+    RAISED_FIST_MEDIUM_DARK_SKIN_TONE(0x270A),
+    RAISED_FIST_DARK_SKIN_TONE(0x270A),
+    ONCOMING_FIST(0x1F44A),
+    ONCOMING_FIST_LIGHT_SKIN_TONE(0x1F44A),
+    ONCOMING_FIST_MEDIUM_LIGHT_SKIN_TONE(0x1F44A),
+    ONCOMING_FIST_MEDIUM_SKIN_TONE(0x1F44A),
+    ONCOMING_FIST_MEDIUM_DARK_SKIN_TONE(0x1F44A),
+    ONCOMING_FIST_DARK_SKIN_TONE(0x1F44A),
+    LEFT_FACING_FIST(0x1F91B),
+    LEFT_FACING_FIST_LIGHT_SKIN_TONE(0x1F91B),
+    LEFT_FACING_FIST_MEDIUM_LIGHT_SKIN_TONE(0x1F91B),
+    LEFT_FACING_FIST_MEDIUM_SKIN_TONE(0x1F91B),
+    LEFT_FACING_FIST_MEDIUM_DARK_SKIN_TONE(0x1F91B),
+    LEFT_FACING_FIST_DARK_SKIN_TONE(0x1F91B),
+    RIGHT_FACING_FIST(0x1F91C),
+    RIGHT_FACING_FIST_LIGHT_SKIN_TONE(0x1F91C),
+    RIGHT_FACING_FIST_MEDIUM_LIGHT_SKIN_TONE(0x1F91C),
+    RIGHT_FACING_FIST_MEDIUM_SKIN_TONE(0x1F91C),
+    RIGHT_FACING_FIST_MEDIUM_DARK_SKIN_TONE(0x1F91C),
+    RIGHT_FACING_FIST_DARK_SKIN_TONE(0x1F91C),
+    CLAPPING_HANDS(0x1F44F),
+    CLAPPING_HANDS_LIGHT_SKIN_TONE(0x1F44F),
+    CLAPPING_HANDS_MEDIUM_LIGHT_SKIN_TONE(0x1F44F),
+    CLAPPING_HANDS_MEDIUM_SKIN_TONE(0x1F44F),
+    CLAPPING_HANDS_MEDIUM_DARK_SKIN_TONE(0x1F44F),
+    CLAPPING_HANDS_DARK_SKIN_TONE(0x1F44F),
+    RAISING_HANDS(0x1F64C),
+    RAISING_HANDS_LIGHT_SKIN_TONE(0x1F64C),
+    RAISING_HANDS_MEDIUM_LIGHT_SKIN_TONE(0x1F64C),
+    RAISING_HANDS_MEDIUM_SKIN_TONE(0x1F64C),
+    RAISING_HANDS_MEDIUM_DARK_SKIN_TONE(0x1F64C),
+    RAISING_HANDS_DARK_SKIN_TONE(0x1F64C),
+    HEART_HANDS(0x1FAF6),
+    HEART_HANDS_LIGHT_SKIN_TONE(0x1FAF6),
+    HEART_HANDS_MEDIUM_LIGHT_SKIN_TONE(0x1FAF6),
+    HEART_HANDS_MEDIUM_SKIN_TONE(0x1FAF6),
+    HEART_HANDS_MEDIUM_DARK_SKIN_TONE(0x1FAF6),
+    HEART_HANDS_DARK_SKIN_TONE(0x1FAF6),
+    OPEN_HANDS(0x1F450),
+    OPEN_HANDS_LIGHT_SKIN_TONE(0x1F450),
+    OPEN_HANDS_MEDIUM_LIGHT_SKIN_TONE(0x1F450),
+    OPEN_HANDS_MEDIUM_SKIN_TONE(0x1F450),
+    OPEN_HANDS_MEDIUM_DARK_SKIN_TONE(0x1F450),
+    OPEN_HANDS_DARK_SKIN_TONE(0x1F450),
+    PALMS_UP_TOGETHER(0x1F932),
+    PALMS_UP_TOGETHER_LIGHT_SKIN_TONE(0x1F932),
+    PALMS_UP_TOGETHER_MEDIUM_LIGHT_SKIN_TONE(0x1F932),
+    PALMS_UP_TOGETHER_MEDIUM_SKIN_TONE(0x1F932),
+    PALMS_UP_TOGETHER_MEDIUM_DARK_SKIN_TONE(0x1F932),
+    PALMS_UP_TOGETHER_DARK_SKIN_TONE(0x1F932),
+    HANDSHAKE(0x1F91D),
+    HANDSHAKE_LIGHT_SKIN_TONE(0x1F91D),
+    HANDSHAKE_MEDIUM_LIGHT_SKIN_TONE(0x1F91D),
+    HANDSHAKE_MEDIUM_SKIN_TONE(0x1F91D),
+    HANDSHAKE_MEDIUM_DARK_SKIN_TONE(0x1F91D),
+    HANDSHAKE_DARK_SKIN_TONE(0x1F91D),
+    HANDSHAKE_LIGHT_SKIN_TONECOMMA_MEDIUM_LIGHT_SKIN_TONE(0x1FAF1),
+    HANDSHAKE_LIGHT_SKIN_TONECOMMA_MEDIUM_SKIN_TONE(0x1FAF1),
+    HANDSHAKE_LIGHT_SKIN_TONECOMMA_MEDIUM_DARK_SKIN_TONE(0x1FAF1),
+    HANDSHAKE_LIGHT_SKIN_TONECOMMA_DARK_SKIN_TONE(0x1FAF1),
+    HANDSHAKE_MEDIUM_LIGHT_SKIN_TONECOMMA_LIGHT_SKIN_TONE(0x1FAF1),
+    HANDSHAKE_MEDIUM_LIGHT_SKIN_TONECOMMA_MEDIUM_SKIN_TONE(0x1FAF1),
+    HANDSHAKE_MEDIUM_LIGHT_SKIN_TONECOMMA_MEDIUM_DARK_SKIN_TONE(0x1FAF1),
+    HANDSHAKE_MEDIUM_LIGHT_SKIN_TONECOMMA_DARK_SKIN_TONE(0x1FAF1),
+    HANDSHAKE_MEDIUM_SKIN_TONECOMMA_LIGHT_SKIN_TONE(0x1FAF1),
+    HANDSHAKE_MEDIUM_SKIN_TONECOMMA_MEDIUM_LIGHT_SKIN_TONE(0x1FAF1),
+    HANDSHAKE_MEDIUM_SKIN_TONECOMMA_MEDIUM_DARK_SKIN_TONE(0x1FAF1),
+    HANDSHAKE_MEDIUM_SKIN_TONECOMMA_DARK_SKIN_TONE(0x1FAF1),
+    HANDSHAKE_MEDIUM_DARK_SKIN_TONECOMMA_LIGHT_SKIN_TONE(0x1FAF1),
+    HANDSHAKE_MEDIUM_DARK_SKIN_TONECOMMA_MEDIUM_LIGHT_SKIN_TONE(0x1FAF1),
+    HANDSHAKE_MEDIUM_DARK_SKIN_TONECOMMA_MEDIUM_SKIN_TONE(0x1FAF1),
+    HANDSHAKE_MEDIUM_DARK_SKIN_TONECOMMA_DARK_SKIN_TONE(0x1FAF1),
+    HANDSHAKE_DARK_SKIN_TONECOMMA_LIGHT_SKIN_TONE(0x1FAF1),
+    HANDSHAKE_DARK_SKIN_TONECOMMA_MEDIUM_LIGHT_SKIN_TONE(0x1FAF1),
+    HANDSHAKE_DARK_SKIN_TONECOMMA_MEDIUM_SKIN_TONE(0x1FAF1),
+    HANDSHAKE_DARK_SKIN_TONECOMMA_MEDIUM_DARK_SKIN_TONE(0x1FAF1),
+    FOLDED_HANDS(0x1F64F),
+    FOLDED_HANDS_LIGHT_SKIN_TONE(0x1F64F),
+    FOLDED_HANDS_MEDIUM_LIGHT_SKIN_TONE(0x1F64F),
+    FOLDED_HANDS_MEDIUM_SKIN_TONE(0x1F64F),
+    FOLDED_HANDS_MEDIUM_DARK_SKIN_TONE(0x1F64F),
+    FOLDED_HANDS_DARK_SKIN_TONE(0x1F64F),
+    WRITING_HAND(0x270D),
+    WRITING_HAND_LIGHT_SKIN_TONE(0x270D),
+    WRITING_HAND_MEDIUM_LIGHT_SKIN_TONE(0x270D),
+    WRITING_HAND_MEDIUM_SKIN_TONE(0x270D),
+    WRITING_HAND_MEDIUM_DARK_SKIN_TONE(0x270D),
+    WRITING_HAND_DARK_SKIN_TONE(0x270D),
+    NAIL_POLISH(0x1F485),
+    NAIL_POLISH_LIGHT_SKIN_TONE(0x1F485),
+    NAIL_POLISH_MEDIUM_LIGHT_SKIN_TONE(0x1F485),
+    NAIL_POLISH_MEDIUM_SKIN_TONE(0x1F485),
+    NAIL_POLISH_MEDIUM_DARK_SKIN_TONE(0x1F485),
+    NAIL_POLISH_DARK_SKIN_TONE(0x1F485),
+    SELFIE(0x1F933),
+    SELFIE_LIGHT_SKIN_TONE(0x1F933),
+    SELFIE_MEDIUM_LIGHT_SKIN_TONE(0x1F933),
+    SELFIE_MEDIUM_SKIN_TONE(0x1F933),
+    SELFIE_MEDIUM_DARK_SKIN_TONE(0x1F933),
+    SELFIE_DARK_SKIN_TONE(0x1F933),
+    FLEXED_BICEPS(0x1F4AA),
+    FLEXED_BICEPS_LIGHT_SKIN_TONE(0x1F4AA),
+    FLEXED_BICEPS_MEDIUM_LIGHT_SKIN_TONE(0x1F4AA),
+    FLEXED_BICEPS_MEDIUM_SKIN_TONE(0x1F4AA),
+    FLEXED_BICEPS_MEDIUM_DARK_SKIN_TONE(0x1F4AA),
+    FLEXED_BICEPS_DARK_SKIN_TONE(0x1F4AA),
+    MECHANICAL_ARM(0x1F9BE),
+    MECHANICAL_LEG(0x1F9BF),
+    LEG(0x1F9B5),
+    LEG_LIGHT_SKIN_TONE(0x1F9B5),
+    LEG_MEDIUM_LIGHT_SKIN_TONE(0x1F9B5),
+    LEG_MEDIUM_SKIN_TONE(0x1F9B5),
+    LEG_MEDIUM_DARK_SKIN_TONE(0x1F9B5),
+    LEG_DARK_SKIN_TONE(0x1F9B5),
+    FOOT(0x1F9B6),
+    FOOT_LIGHT_SKIN_TONE(0x1F9B6),
+    FOOT_MEDIUM_LIGHT_SKIN_TONE(0x1F9B6),
+    FOOT_MEDIUM_SKIN_TONE(0x1F9B6),
+    FOOT_MEDIUM_DARK_SKIN_TONE(0x1F9B6),
+    FOOT_DARK_SKIN_TONE(0x1F9B6),
+    EAR(0x1F442),
+    EAR_LIGHT_SKIN_TONE(0x1F442),
+    EAR_MEDIUM_LIGHT_SKIN_TONE(0x1F442),
+    EAR_MEDIUM_SKIN_TONE(0x1F442),
+    EAR_MEDIUM_DARK_SKIN_TONE(0x1F442),
+    EAR_DARK_SKIN_TONE(0x1F442),
+    EAR_WITH_HEARING_AID(0x1F9BB),
+    EAR_WITH_HEARING_AID_LIGHT_SKIN_TONE(0x1F9BB),
+    EAR_WITH_HEARING_AID_MEDIUM_LIGHT_SKIN_TONE(0x1F9BB),
+    EAR_WITH_HEARING_AID_MEDIUM_SKIN_TONE(0x1F9BB),
+    EAR_WITH_HEARING_AID_MEDIUM_DARK_SKIN_TONE(0x1F9BB),
+    EAR_WITH_HEARING_AID_DARK_SKIN_TONE(0x1F9BB),
+    NOSE(0x1F443),
+    NOSE_LIGHT_SKIN_TONE(0x1F443),
+    NOSE_MEDIUM_LIGHT_SKIN_TONE(0x1F443),
+    NOSE_MEDIUM_SKIN_TONE(0x1F443),
+    NOSE_MEDIUM_DARK_SKIN_TONE(0x1F443),
+    NOSE_DARK_SKIN_TONE(0x1F443),
+    BRAIN(0x1F9E0),
+    ANATOMICAL_HEART(0x1FAC0),
+    LUNGS(0x1FAC1),
+    TOOTH(0x1F9B7),
+    BONE(0x1F9B4),
+    EYES(0x1F440),
+    EYE(0x1F441),
+    TONGUE(0x1F445),
+    MOUTH(0x1F444),
+    BITING_LIP(0x1FAE6),
+    BABY(0x1F476),
+    BABY_LIGHT_SKIN_TONE(0x1F476),
+    BABY_MEDIUM_LIGHT_SKIN_TONE(0x1F476),
+    BABY_MEDIUM_SKIN_TONE(0x1F476),
+    BABY_MEDIUM_DARK_SKIN_TONE(0x1F476),
+    BABY_DARK_SKIN_TONE(0x1F476),
+    CHILD(0x1F9D2),
+    CHILD_LIGHT_SKIN_TONE(0x1F9D2),
+    CHILD_MEDIUM_LIGHT_SKIN_TONE(0x1F9D2),
+    CHILD_MEDIUM_SKIN_TONE(0x1F9D2),
+    CHILD_MEDIUM_DARK_SKIN_TONE(0x1F9D2),
+    CHILD_DARK_SKIN_TONE(0x1F9D2),
+    BOY(0x1F466),
+    BOY_LIGHT_SKIN_TONE(0x1F466),
+    BOY_MEDIUM_LIGHT_SKIN_TONE(0x1F466),
+    BOY_MEDIUM_SKIN_TONE(0x1F466),
+    BOY_MEDIUM_DARK_SKIN_TONE(0x1F466),
+    BOY_DARK_SKIN_TONE(0x1F466),
+    GIRL(0x1F467),
+    GIRL_LIGHT_SKIN_TONE(0x1F467),
+    GIRL_MEDIUM_LIGHT_SKIN_TONE(0x1F467),
+    GIRL_MEDIUM_SKIN_TONE(0x1F467),
+    GIRL_MEDIUM_DARK_SKIN_TONE(0x1F467),
+    GIRL_DARK_SKIN_TONE(0x1F467),
+    PERSON(0x1F9D1),
+    PERSON_LIGHT_SKIN_TONE(0x1F9D1),
+    PERSON_MEDIUM_LIGHT_SKIN_TONE(0x1F9D1),
+    PERSON_MEDIUM_SKIN_TONE(0x1F9D1),
+    PERSON_MEDIUM_DARK_SKIN_TONE(0x1F9D1),
+    PERSON_DARK_SKIN_TONE(0x1F9D1),
+    PERSON_BLOND_HAIR(0x1F471),
+    PERSON_LIGHT_SKIN_TONECOMMA_BLOND_HAIR(0x1F471),
+    PERSON_MEDIUM_LIGHT_SKIN_TONECOMMA_BLOND_HAIR(0x1F471),
+    PERSON_MEDIUM_SKIN_TONECOMMA_BLOND_HAIR(0x1F471),
+    PERSON_MEDIUM_DARK_SKIN_TONECOMMA_BLOND_HAIR(0x1F471),
+    PERSON_DARK_SKIN_TONECOMMA_BLOND_HAIR(0x1F471),
+    MAN(0x1F468),
+    MAN_LIGHT_SKIN_TONE(0x1F468),
+    MAN_MEDIUM_LIGHT_SKIN_TONE(0x1F468),
+    MAN_MEDIUM_SKIN_TONE(0x1F468),
+    MAN_MEDIUM_DARK_SKIN_TONE(0x1F468),
+    MAN_DARK_SKIN_TONE(0x1F468),
+    PERSON_BEARD(0x1F9D4),
+    PERSON_LIGHT_SKIN_TONECOMMA_BEARD(0x1F9D4),
+    PERSON_MEDIUM_LIGHT_SKIN_TONECOMMA_BEARD(0x1F9D4),
+    PERSON_MEDIUM_SKIN_TONECOMMA_BEARD(0x1F9D4),
+    PERSON_MEDIUM_DARK_SKIN_TONECOMMA_BEARD(0x1F9D4),
+    PERSON_DARK_SKIN_TONECOMMA_BEARD(0x1F9D4),
+    MAN_BEARD(0x1F9D4),
+    MAN_LIGHT_SKIN_TONECOMMA_BEARD(0x1F9D4),
+    MAN_MEDIUM_LIGHT_SKIN_TONECOMMA_BEARD(0x1F9D4),
+    MAN_MEDIUM_SKIN_TONECOMMA_BEARD(0x1F9D4),
+    MAN_MEDIUM_DARK_SKIN_TONECOMMA_BEARD(0x1F9D4),
+    MAN_DARK_SKIN_TONECOMMA_BEARD(0x1F9D4),
+    WOMAN_BEARD(0x1F9D4),
+    WOMAN_LIGHT_SKIN_TONECOMMA_BEARD(0x1F9D4),
+    WOMAN_MEDIUM_LIGHT_SKIN_TONECOMMA_BEARD(0x1F9D4),
+    WOMAN_MEDIUM_SKIN_TONECOMMA_BEARD(0x1F9D4),
+    WOMAN_MEDIUM_DARK_SKIN_TONECOMMA_BEARD(0x1F9D4),
+    WOMAN_DARK_SKIN_TONECOMMA_BEARD(0x1F9D4),
+    MAN_RED_HAIR(0x1F468),
+    MAN_LIGHT_SKIN_TONECOMMA_RED_HAIR(0x1F468),
+    MAN_MEDIUM_LIGHT_SKIN_TONECOMMA_RED_HAIR(0x1F468),
+    MAN_MEDIUM_SKIN_TONECOMMA_RED_HAIR(0x1F468),
+    MAN_MEDIUM_DARK_SKIN_TONECOMMA_RED_HAIR(0x1F468),
+    MAN_DARK_SKIN_TONECOMMA_RED_HAIR(0x1F468),
+    MAN_CURLY_HAIR(0x1F468),
+    MAN_LIGHT_SKIN_TONECOMMA_CURLY_HAIR(0x1F468),
+    MAN_MEDIUM_LIGHT_SKIN_TONECOMMA_CURLY_HAIR(0x1F468),
+    MAN_MEDIUM_SKIN_TONECOMMA_CURLY_HAIR(0x1F468),
+    MAN_MEDIUM_DARK_SKIN_TONECOMMA_CURLY_HAIR(0x1F468),
+    MAN_DARK_SKIN_TONECOMMA_CURLY_HAIR(0x1F468),
+    MAN_WHITE_HAIR(0x1F468),
+    MAN_LIGHT_SKIN_TONECOMMA_WHITE_HAIR(0x1F468),
+    MAN_MEDIUM_LIGHT_SKIN_TONECOMMA_WHITE_HAIR(0x1F468),
+    MAN_MEDIUM_SKIN_TONECOMMA_WHITE_HAIR(0x1F468),
+    MAN_MEDIUM_DARK_SKIN_TONECOMMA_WHITE_HAIR(0x1F468),
+    MAN_DARK_SKIN_TONECOMMA_WHITE_HAIR(0x1F468),
+    MAN_BALD(0x1F468),
+    MAN_LIGHT_SKIN_TONECOMMA_BALD(0x1F468),
+    MAN_MEDIUM_LIGHT_SKIN_TONECOMMA_BALD(0x1F468),
+    MAN_MEDIUM_SKIN_TONECOMMA_BALD(0x1F468),
+    MAN_MEDIUM_DARK_SKIN_TONECOMMA_BALD(0x1F468),
+    MAN_DARK_SKIN_TONECOMMA_BALD(0x1F468),
+    WOMAN(0x1F469),
+    WOMAN_LIGHT_SKIN_TONE(0x1F469),
+    WOMAN_MEDIUM_LIGHT_SKIN_TONE(0x1F469),
+    WOMAN_MEDIUM_SKIN_TONE(0x1F469),
+    WOMAN_MEDIUM_DARK_SKIN_TONE(0x1F469),
+    WOMAN_DARK_SKIN_TONE(0x1F469),
+    WOMAN_RED_HAIR(0x1F469),
+    WOMAN_LIGHT_SKIN_TONECOMMA_RED_HAIR(0x1F469),
+    WOMAN_MEDIUM_LIGHT_SKIN_TONECOMMA_RED_HAIR(0x1F469),
+    WOMAN_MEDIUM_SKIN_TONECOMMA_RED_HAIR(0x1F469),
+    WOMAN_MEDIUM_DARK_SKIN_TONECOMMA_RED_HAIR(0x1F469),
+    WOMAN_DARK_SKIN_TONECOMMA_RED_HAIR(0x1F469),
+    PERSON_RED_HAIR(0x1F9D1),
+    PERSON_LIGHT_SKIN_TONECOMMA_RED_HAIR(0x1F9D1),
+    PERSON_MEDIUM_LIGHT_SKIN_TONECOMMA_RED_HAIR(0x1F9D1),
+    PERSON_MEDIUM_SKIN_TONECOMMA_RED_HAIR(0x1F9D1),
+    PERSON_MEDIUM_DARK_SKIN_TONECOMMA_RED_HAIR(0x1F9D1),
+    PERSON_DARK_SKIN_TONECOMMA_RED_HAIR(0x1F9D1),
+    WOMAN_CURLY_HAIR(0x1F469),
+    WOMAN_LIGHT_SKIN_TONECOMMA_CURLY_HAIR(0x1F469),
+    WOMAN_MEDIUM_LIGHT_SKIN_TONECOMMA_CURLY_HAIR(0x1F469),
+    WOMAN_MEDIUM_SKIN_TONECOMMA_CURLY_HAIR(0x1F469),
+    WOMAN_MEDIUM_DARK_SKIN_TONECOMMA_CURLY_HAIR(0x1F469),
+    WOMAN_DARK_SKIN_TONECOMMA_CURLY_HAIR(0x1F469),
+    PERSON_CURLY_HAIR(0x1F9D1),
+    PERSON_LIGHT_SKIN_TONECOMMA_CURLY_HAIR(0x1F9D1),
+    PERSON_MEDIUM_LIGHT_SKIN_TONECOMMA_CURLY_HAIR(0x1F9D1),
+    PERSON_MEDIUM_SKIN_TONECOMMA_CURLY_HAIR(0x1F9D1),
+    PERSON_MEDIUM_DARK_SKIN_TONECOMMA_CURLY_HAIR(0x1F9D1),
+    PERSON_DARK_SKIN_TONECOMMA_CURLY_HAIR(0x1F9D1),
+    WOMAN_WHITE_HAIR(0x1F469),
+    WOMAN_LIGHT_SKIN_TONECOMMA_WHITE_HAIR(0x1F469),
+    WOMAN_MEDIUM_LIGHT_SKIN_TONECOMMA_WHITE_HAIR(0x1F469),
+    WOMAN_MEDIUM_SKIN_TONECOMMA_WHITE_HAIR(0x1F469),
+    WOMAN_MEDIUM_DARK_SKIN_TONECOMMA_WHITE_HAIR(0x1F469),
+    WOMAN_DARK_SKIN_TONECOMMA_WHITE_HAIR(0x1F469),
+    PERSON_WHITE_HAIR(0x1F9D1),
+    PERSON_LIGHT_SKIN_TONECOMMA_WHITE_HAIR(0x1F9D1),
+    PERSON_MEDIUM_LIGHT_SKIN_TONECOMMA_WHITE_HAIR(0x1F9D1),
+    PERSON_MEDIUM_SKIN_TONECOMMA_WHITE_HAIR(0x1F9D1),
+    PERSON_MEDIUM_DARK_SKIN_TONECOMMA_WHITE_HAIR(0x1F9D1),
+    PERSON_DARK_SKIN_TONECOMMA_WHITE_HAIR(0x1F9D1),
+    WOMAN_BALD(0x1F469),
+    WOMAN_LIGHT_SKIN_TONECOMMA_BALD(0x1F469),
+    WOMAN_MEDIUM_LIGHT_SKIN_TONECOMMA_BALD(0x1F469),
+    WOMAN_MEDIUM_SKIN_TONECOMMA_BALD(0x1F469),
+    WOMAN_MEDIUM_DARK_SKIN_TONECOMMA_BALD(0x1F469),
+    WOMAN_DARK_SKIN_TONECOMMA_BALD(0x1F469),
+    PERSON_BALD(0x1F9D1),
+    PERSON_LIGHT_SKIN_TONECOMMA_BALD(0x1F9D1),
+    PERSON_MEDIUM_LIGHT_SKIN_TONECOMMA_BALD(0x1F9D1),
+    PERSON_MEDIUM_SKIN_TONECOMMA_BALD(0x1F9D1),
+    PERSON_MEDIUM_DARK_SKIN_TONECOMMA_BALD(0x1F9D1),
+    PERSON_DARK_SKIN_TONECOMMA_BALD(0x1F9D1),
+    WOMAN_BLOND_HAIR(0x1F471),
+    WOMAN_LIGHT_SKIN_TONECOMMA_BLOND_HAIR(0x1F471),
+    WOMAN_MEDIUM_LIGHT_SKIN_TONECOMMA_BLOND_HAIR(0x1F471),
+    WOMAN_MEDIUM_SKIN_TONECOMMA_BLOND_HAIR(0x1F471),
+    WOMAN_MEDIUM_DARK_SKIN_TONECOMMA_BLOND_HAIR(0x1F471),
+    WOMAN_DARK_SKIN_TONECOMMA_BLOND_HAIR(0x1F471),
+    MAN_BLOND_HAIR(0x1F471),
+    MAN_LIGHT_SKIN_TONECOMMA_BLOND_HAIR(0x1F471),
+    MAN_MEDIUM_LIGHT_SKIN_TONECOMMA_BLOND_HAIR(0x1F471),
+    MAN_MEDIUM_SKIN_TONECOMMA_BLOND_HAIR(0x1F471),
+    MAN_MEDIUM_DARK_SKIN_TONECOMMA_BLOND_HAIR(0x1F471),
+    MAN_DARK_SKIN_TONECOMMA_BLOND_HAIR(0x1F471),
+    OLDER_PERSON(0x1F9D3),
+    OLDER_PERSON_LIGHT_SKIN_TONE(0x1F9D3),
+    OLDER_PERSON_MEDIUM_LIGHT_SKIN_TONE(0x1F9D3),
+    OLDER_PERSON_MEDIUM_SKIN_TONE(0x1F9D3),
+    OLDER_PERSON_MEDIUM_DARK_SKIN_TONE(0x1F9D3),
+    OLDER_PERSON_DARK_SKIN_TONE(0x1F9D3),
+    OLD_MAN(0x1F474),
+    OLD_MAN_LIGHT_SKIN_TONE(0x1F474),
+    OLD_MAN_MEDIUM_LIGHT_SKIN_TONE(0x1F474),
+    OLD_MAN_MEDIUM_SKIN_TONE(0x1F474),
+    OLD_MAN_MEDIUM_DARK_SKIN_TONE(0x1F474),
+    OLD_MAN_DARK_SKIN_TONE(0x1F474),
+    OLD_WOMAN(0x1F475),
+    OLD_WOMAN_LIGHT_SKIN_TONE(0x1F475),
+    OLD_WOMAN_MEDIUM_LIGHT_SKIN_TONE(0x1F475),
+    OLD_WOMAN_MEDIUM_SKIN_TONE(0x1F475),
+    OLD_WOMAN_MEDIUM_DARK_SKIN_TONE(0x1F475),
+    OLD_WOMAN_DARK_SKIN_TONE(0x1F475),
+    PERSON_FROWNING(0x1F64D),
+    PERSON_FROWNING_LIGHT_SKIN_TONE(0x1F64D),
+    PERSON_FROWNING_MEDIUM_LIGHT_SKIN_TONE(0x1F64D),
+    PERSON_FROWNING_MEDIUM_SKIN_TONE(0x1F64D),
+    PERSON_FROWNING_MEDIUM_DARK_SKIN_TONE(0x1F64D),
+    PERSON_FROWNING_DARK_SKIN_TONE(0x1F64D),
+    MAN_FROWNING(0x1F64D),
+    MAN_FROWNING_LIGHT_SKIN_TONE(0x1F64D),
+    MAN_FROWNING_MEDIUM_LIGHT_SKIN_TONE(0x1F64D),
+    MAN_FROWNING_MEDIUM_SKIN_TONE(0x1F64D),
+    MAN_FROWNING_MEDIUM_DARK_SKIN_TONE(0x1F64D),
+    MAN_FROWNING_DARK_SKIN_TONE(0x1F64D),
+    WOMAN_FROWNING(0x1F64D),
+    WOMAN_FROWNING_LIGHT_SKIN_TONE(0x1F64D),
+    WOMAN_FROWNING_MEDIUM_LIGHT_SKIN_TONE(0x1F64D),
+    WOMAN_FROWNING_MEDIUM_SKIN_TONE(0x1F64D),
+    WOMAN_FROWNING_MEDIUM_DARK_SKIN_TONE(0x1F64D),
+    WOMAN_FROWNING_DARK_SKIN_TONE(0x1F64D),
+    PERSON_POUTING(0x1F64E),
+    PERSON_POUTING_LIGHT_SKIN_TONE(0x1F64E),
+    PERSON_POUTING_MEDIUM_LIGHT_SKIN_TONE(0x1F64E),
+    PERSON_POUTING_MEDIUM_SKIN_TONE(0x1F64E),
+    PERSON_POUTING_MEDIUM_DARK_SKIN_TONE(0x1F64E),
+    PERSON_POUTING_DARK_SKIN_TONE(0x1F64E),
+    MAN_POUTING(0x1F64E),
+    MAN_POUTING_LIGHT_SKIN_TONE(0x1F64E),
+    MAN_POUTING_MEDIUM_LIGHT_SKIN_TONE(0x1F64E),
+    MAN_POUTING_MEDIUM_SKIN_TONE(0x1F64E),
+    MAN_POUTING_MEDIUM_DARK_SKIN_TONE(0x1F64E),
+    MAN_POUTING_DARK_SKIN_TONE(0x1F64E),
+    WOMAN_POUTING(0x1F64E),
+    WOMAN_POUTING_LIGHT_SKIN_TONE(0x1F64E),
+    WOMAN_POUTING_MEDIUM_LIGHT_SKIN_TONE(0x1F64E),
+    WOMAN_POUTING_MEDIUM_SKIN_TONE(0x1F64E),
+    WOMAN_POUTING_MEDIUM_DARK_SKIN_TONE(0x1F64E),
+    WOMAN_POUTING_DARK_SKIN_TONE(0x1F64E),
+    PERSON_GESTURING_NO(0x1F645),
+    PERSON_GESTURING_NO_LIGHT_SKIN_TONE(0x1F645),
+    PERSON_GESTURING_NO_MEDIUM_LIGHT_SKIN_TONE(0x1F645),
+    PERSON_GESTURING_NO_MEDIUM_SKIN_TONE(0x1F645),
+    PERSON_GESTURING_NO_MEDIUM_DARK_SKIN_TONE(0x1F645),
+    PERSON_GESTURING_NO_DARK_SKIN_TONE(0x1F645),
+    MAN_GESTURING_NO(0x1F645),
+    MAN_GESTURING_NO_LIGHT_SKIN_TONE(0x1F645),
+    MAN_GESTURING_NO_MEDIUM_LIGHT_SKIN_TONE(0x1F645),
+    MAN_GESTURING_NO_MEDIUM_SKIN_TONE(0x1F645),
+    MAN_GESTURING_NO_MEDIUM_DARK_SKIN_TONE(0x1F645),
+    MAN_GESTURING_NO_DARK_SKIN_TONE(0x1F645),
+    WOMAN_GESTURING_NO(0x1F645),
+    WOMAN_GESTURING_NO_LIGHT_SKIN_TONE(0x1F645),
+    WOMAN_GESTURING_NO_MEDIUM_LIGHT_SKIN_TONE(0x1F645),
+    WOMAN_GESTURING_NO_MEDIUM_SKIN_TONE(0x1F645),
+    WOMAN_GESTURING_NO_MEDIUM_DARK_SKIN_TONE(0x1F645),
+    WOMAN_GESTURING_NO_DARK_SKIN_TONE(0x1F645),
+    PERSON_GESTURING_OK(0x1F646),
+    PERSON_GESTURING_OK_LIGHT_SKIN_TONE(0x1F646),
+    PERSON_GESTURING_OK_MEDIUM_LIGHT_SKIN_TONE(0x1F646),
+    PERSON_GESTURING_OK_MEDIUM_SKIN_TONE(0x1F646),
+    PERSON_GESTURING_OK_MEDIUM_DARK_SKIN_TONE(0x1F646),
+    PERSON_GESTURING_OK_DARK_SKIN_TONE(0x1F646),
+    MAN_GESTURING_OK(0x1F646),
+    MAN_GESTURING_OK_LIGHT_SKIN_TONE(0x1F646),
+    MAN_GESTURING_OK_MEDIUM_LIGHT_SKIN_TONE(0x1F646),
+    MAN_GESTURING_OK_MEDIUM_SKIN_TONE(0x1F646),
+    MAN_GESTURING_OK_MEDIUM_DARK_SKIN_TONE(0x1F646),
+    MAN_GESTURING_OK_DARK_SKIN_TONE(0x1F646),
+    WOMAN_GESTURING_OK(0x1F646),
+    WOMAN_GESTURING_OK_LIGHT_SKIN_TONE(0x1F646),
+    WOMAN_GESTURING_OK_MEDIUM_LIGHT_SKIN_TONE(0x1F646),
+    WOMAN_GESTURING_OK_MEDIUM_SKIN_TONE(0x1F646),
+    WOMAN_GESTURING_OK_MEDIUM_DARK_SKIN_TONE(0x1F646),
+    WOMAN_GESTURING_OK_DARK_SKIN_TONE(0x1F646),
+    PERSON_TIPPING_HAND(0x1F481),
+    PERSON_TIPPING_HAND_LIGHT_SKIN_TONE(0x1F481),
+    PERSON_TIPPING_HAND_MEDIUM_LIGHT_SKIN_TONE(0x1F481),
+    PERSON_TIPPING_HAND_MEDIUM_SKIN_TONE(0x1F481),
+    PERSON_TIPPING_HAND_MEDIUM_DARK_SKIN_TONE(0x1F481),
+    PERSON_TIPPING_HAND_DARK_SKIN_TONE(0x1F481),
+    MAN_TIPPING_HAND(0x1F481),
+    MAN_TIPPING_HAND_LIGHT_SKIN_TONE(0x1F481),
+    MAN_TIPPING_HAND_MEDIUM_LIGHT_SKIN_TONE(0x1F481),
+    MAN_TIPPING_HAND_MEDIUM_SKIN_TONE(0x1F481),
+    MAN_TIPPING_HAND_MEDIUM_DARK_SKIN_TONE(0x1F481),
+    MAN_TIPPING_HAND_DARK_SKIN_TONE(0x1F481),
+    WOMAN_TIPPING_HAND(0x1F481),
+    WOMAN_TIPPING_HAND_LIGHT_SKIN_TONE(0x1F481),
+    WOMAN_TIPPING_HAND_MEDIUM_LIGHT_SKIN_TONE(0x1F481),
+    WOMAN_TIPPING_HAND_MEDIUM_SKIN_TONE(0x1F481),
+    WOMAN_TIPPING_HAND_MEDIUM_DARK_SKIN_TONE(0x1F481),
+    WOMAN_TIPPING_HAND_DARK_SKIN_TONE(0x1F481),
+    PERSON_RAISING_HAND(0x1F64B),
+    PERSON_RAISING_HAND_LIGHT_SKIN_TONE(0x1F64B),
+    PERSON_RAISING_HAND_MEDIUM_LIGHT_SKIN_TONE(0x1F64B),
+    PERSON_RAISING_HAND_MEDIUM_SKIN_TONE(0x1F64B),
+    PERSON_RAISING_HAND_MEDIUM_DARK_SKIN_TONE(0x1F64B),
+    PERSON_RAISING_HAND_DARK_SKIN_TONE(0x1F64B),
+    MAN_RAISING_HAND(0x1F64B),
+    MAN_RAISING_HAND_LIGHT_SKIN_TONE(0x1F64B),
+    MAN_RAISING_HAND_MEDIUM_LIGHT_SKIN_TONE(0x1F64B),
+    MAN_RAISING_HAND_MEDIUM_SKIN_TONE(0x1F64B),
+    MAN_RAISING_HAND_MEDIUM_DARK_SKIN_TONE(0x1F64B),
+    MAN_RAISING_HAND_DARK_SKIN_TONE(0x1F64B),
+    WOMAN_RAISING_HAND(0x1F64B),
+    WOMAN_RAISING_HAND_LIGHT_SKIN_TONE(0x1F64B),
+    WOMAN_RAISING_HAND_MEDIUM_LIGHT_SKIN_TONE(0x1F64B),
+    WOMAN_RAISING_HAND_MEDIUM_SKIN_TONE(0x1F64B),
+    WOMAN_RAISING_HAND_MEDIUM_DARK_SKIN_TONE(0x1F64B),
+    WOMAN_RAISING_HAND_DARK_SKIN_TONE(0x1F64B),
+    DEAF_PERSON(0x1F9CF),
+    DEAF_PERSON_LIGHT_SKIN_TONE(0x1F9CF),
+    DEAF_PERSON_MEDIUM_LIGHT_SKIN_TONE(0x1F9CF),
+    DEAF_PERSON_MEDIUM_SKIN_TONE(0x1F9CF),
+    DEAF_PERSON_MEDIUM_DARK_SKIN_TONE(0x1F9CF),
+    DEAF_PERSON_DARK_SKIN_TONE(0x1F9CF),
+    DEAF_MAN(0x1F9CF),
+    DEAF_MAN_LIGHT_SKIN_TONE(0x1F9CF),
+    DEAF_MAN_MEDIUM_LIGHT_SKIN_TONE(0x1F9CF),
+    DEAF_MAN_MEDIUM_SKIN_TONE(0x1F9CF),
+    DEAF_MAN_MEDIUM_DARK_SKIN_TONE(0x1F9CF),
+    DEAF_MAN_DARK_SKIN_TONE(0x1F9CF),
+    DEAF_WOMAN(0x1F9CF),
+    DEAF_WOMAN_LIGHT_SKIN_TONE(0x1F9CF),
+    DEAF_WOMAN_MEDIUM_LIGHT_SKIN_TONE(0x1F9CF),
+    DEAF_WOMAN_MEDIUM_SKIN_TONE(0x1F9CF),
+    DEAF_WOMAN_MEDIUM_DARK_SKIN_TONE(0x1F9CF),
+    DEAF_WOMAN_DARK_SKIN_TONE(0x1F9CF),
+    PERSON_BOWING(0x1F647),
+    PERSON_BOWING_LIGHT_SKIN_TONE(0x1F647),
+    PERSON_BOWING_MEDIUM_LIGHT_SKIN_TONE(0x1F647),
+    PERSON_BOWING_MEDIUM_SKIN_TONE(0x1F647),
+    PERSON_BOWING_MEDIUM_DARK_SKIN_TONE(0x1F647),
+    PERSON_BOWING_DARK_SKIN_TONE(0x1F647),
+    MAN_BOWING(0x1F647),
+    MAN_BOWING_LIGHT_SKIN_TONE(0x1F647),
+    MAN_BOWING_MEDIUM_LIGHT_SKIN_TONE(0x1F647),
+    MAN_BOWING_MEDIUM_SKIN_TONE(0x1F647),
+    MAN_BOWING_MEDIUM_DARK_SKIN_TONE(0x1F647),
+    MAN_BOWING_DARK_SKIN_TONE(0x1F647),
+    WOMAN_BOWING(0x1F647),
+    WOMAN_BOWING_LIGHT_SKIN_TONE(0x1F647),
+    WOMAN_BOWING_MEDIUM_LIGHT_SKIN_TONE(0x1F647),
+    WOMAN_BOWING_MEDIUM_SKIN_TONE(0x1F647),
+    WOMAN_BOWING_MEDIUM_DARK_SKIN_TONE(0x1F647),
+    WOMAN_BOWING_DARK_SKIN_TONE(0x1F647),
+    PERSON_FACEPALMING(0x1F926),
+    PERSON_FACEPALMING_LIGHT_SKIN_TONE(0x1F926),
+    PERSON_FACEPALMING_MEDIUM_LIGHT_SKIN_TONE(0x1F926),
+    PERSON_FACEPALMING_MEDIUM_SKIN_TONE(0x1F926),
+    PERSON_FACEPALMING_MEDIUM_DARK_SKIN_TONE(0x1F926),
+    PERSON_FACEPALMING_DARK_SKIN_TONE(0x1F926),
+    MAN_FACEPALMING(0x1F926),
+    MAN_FACEPALMING_LIGHT_SKIN_TONE(0x1F926),
+    MAN_FACEPALMING_MEDIUM_LIGHT_SKIN_TONE(0x1F926),
+    MAN_FACEPALMING_MEDIUM_SKIN_TONE(0x1F926),
+    MAN_FACEPALMING_MEDIUM_DARK_SKIN_TONE(0x1F926),
+    MAN_FACEPALMING_DARK_SKIN_TONE(0x1F926),
+    WOMAN_FACEPALMING(0x1F926),
+    WOMAN_FACEPALMING_LIGHT_SKIN_TONE(0x1F926),
+    WOMAN_FACEPALMING_MEDIUM_LIGHT_SKIN_TONE(0x1F926),
+    WOMAN_FACEPALMING_MEDIUM_SKIN_TONE(0x1F926),
+    WOMAN_FACEPALMING_MEDIUM_DARK_SKIN_TONE(0x1F926),
+    WOMAN_FACEPALMING_DARK_SKIN_TONE(0x1F926),
+    PERSON_SHRUGGING(0x1F937),
+    PERSON_SHRUGGING_LIGHT_SKIN_TONE(0x1F937),
+    PERSON_SHRUGGING_MEDIUM_LIGHT_SKIN_TONE(0x1F937),
+    PERSON_SHRUGGING_MEDIUM_SKIN_TONE(0x1F937),
+    PERSON_SHRUGGING_MEDIUM_DARK_SKIN_TONE(0x1F937),
+    PERSON_SHRUGGING_DARK_SKIN_TONE(0x1F937),
+    MAN_SHRUGGING(0x1F937),
+    MAN_SHRUGGING_LIGHT_SKIN_TONE(0x1F937),
+    MAN_SHRUGGING_MEDIUM_LIGHT_SKIN_TONE(0x1F937),
+    MAN_SHRUGGING_MEDIUM_SKIN_TONE(0x1F937),
+    MAN_SHRUGGING_MEDIUM_DARK_SKIN_TONE(0x1F937),
+    MAN_SHRUGGING_DARK_SKIN_TONE(0x1F937),
+    WOMAN_SHRUGGING(0x1F937),
+    WOMAN_SHRUGGING_LIGHT_SKIN_TONE(0x1F937),
+    WOMAN_SHRUGGING_MEDIUM_LIGHT_SKIN_TONE(0x1F937),
+    WOMAN_SHRUGGING_MEDIUM_SKIN_TONE(0x1F937),
+    WOMAN_SHRUGGING_MEDIUM_DARK_SKIN_TONE(0x1F937),
+    WOMAN_SHRUGGING_DARK_SKIN_TONE(0x1F937),
+    HEALTH_WORKER(0x1F9D1),
+    HEALTH_WORKER_LIGHT_SKIN_TONE(0x1F9D1),
+    HEALTH_WORKER_MEDIUM_LIGHT_SKIN_TONE(0x1F9D1),
+    HEALTH_WORKER_MEDIUM_SKIN_TONE(0x1F9D1),
+    HEALTH_WORKER_MEDIUM_DARK_SKIN_TONE(0x1F9D1),
+    HEALTH_WORKER_DARK_SKIN_TONE(0x1F9D1),
+    MAN_HEALTH_WORKER(0x1F468),
+    MAN_HEALTH_WORKER_LIGHT_SKIN_TONE(0x1F468),
+    MAN_HEALTH_WORKER_MEDIUM_LIGHT_SKIN_TONE(0x1F468),
+    MAN_HEALTH_WORKER_MEDIUM_SKIN_TONE(0x1F468),
+    MAN_HEALTH_WORKER_MEDIUM_DARK_SKIN_TONE(0x1F468),
+    MAN_HEALTH_WORKER_DARK_SKIN_TONE(0x1F468),
+    WOMAN_HEALTH_WORKER(0x1F469),
+    WOMAN_HEALTH_WORKER_LIGHT_SKIN_TONE(0x1F469),
+    WOMAN_HEALTH_WORKER_MEDIUM_LIGHT_SKIN_TONE(0x1F469),
+    WOMAN_HEALTH_WORKER_MEDIUM_SKIN_TONE(0x1F469),
+    WOMAN_HEALTH_WORKER_MEDIUM_DARK_SKIN_TONE(0x1F469),
+    WOMAN_HEALTH_WORKER_DARK_SKIN_TONE(0x1F469),
+    STUDENT(0x1F9D1),
+    STUDENT_LIGHT_SKIN_TONE(0x1F9D1),
+    STUDENT_MEDIUM_LIGHT_SKIN_TONE(0x1F9D1),
+    STUDENT_MEDIUM_SKIN_TONE(0x1F9D1),
+    STUDENT_MEDIUM_DARK_SKIN_TONE(0x1F9D1),
+    STUDENT_DARK_SKIN_TONE(0x1F9D1),
+    MAN_STUDENT(0x1F468),
+    MAN_STUDENT_LIGHT_SKIN_TONE(0x1F468),
+    MAN_STUDENT_MEDIUM_LIGHT_SKIN_TONE(0x1F468),
+    MAN_STUDENT_MEDIUM_SKIN_TONE(0x1F468),
+    MAN_STUDENT_MEDIUM_DARK_SKIN_TONE(0x1F468),
+    MAN_STUDENT_DARK_SKIN_TONE(0x1F468),
+    WOMAN_STUDENT(0x1F469),
+    WOMAN_STUDENT_LIGHT_SKIN_TONE(0x1F469),
+    WOMAN_STUDENT_MEDIUM_LIGHT_SKIN_TONE(0x1F469),
+    WOMAN_STUDENT_MEDIUM_SKIN_TONE(0x1F469),
+    WOMAN_STUDENT_MEDIUM_DARK_SKIN_TONE(0x1F469),
+    WOMAN_STUDENT_DARK_SKIN_TONE(0x1F469),
+    TEACHER(0x1F9D1),
+    TEACHER_LIGHT_SKIN_TONE(0x1F9D1),
+    TEACHER_MEDIUM_LIGHT_SKIN_TONE(0x1F9D1),
+    TEACHER_MEDIUM_SKIN_TONE(0x1F9D1),
+    TEACHER_MEDIUM_DARK_SKIN_TONE(0x1F9D1),
+    TEACHER_DARK_SKIN_TONE(0x1F9D1),
+    MAN_TEACHER(0x1F468),
+    MAN_TEACHER_LIGHT_SKIN_TONE(0x1F468),
+    MAN_TEACHER_MEDIUM_LIGHT_SKIN_TONE(0x1F468),
+    MAN_TEACHER_MEDIUM_SKIN_TONE(0x1F468),
+    MAN_TEACHER_MEDIUM_DARK_SKIN_TONE(0x1F468),
+    MAN_TEACHER_DARK_SKIN_TONE(0x1F468),
+    WOMAN_TEACHER(0x1F469),
+    WOMAN_TEACHER_LIGHT_SKIN_TONE(0x1F469),
+    WOMAN_TEACHER_MEDIUM_LIGHT_SKIN_TONE(0x1F469),
+    WOMAN_TEACHER_MEDIUM_SKIN_TONE(0x1F469),
+    WOMAN_TEACHER_MEDIUM_DARK_SKIN_TONE(0x1F469),
+    WOMAN_TEACHER_DARK_SKIN_TONE(0x1F469),
+    JUDGE(0x1F9D1),
+    JUDGE_LIGHT_SKIN_TONE(0x1F9D1),
+    JUDGE_MEDIUM_LIGHT_SKIN_TONE(0x1F9D1),
+    JUDGE_MEDIUM_SKIN_TONE(0x1F9D1),
+    JUDGE_MEDIUM_DARK_SKIN_TONE(0x1F9D1),
+    JUDGE_DARK_SKIN_TONE(0x1F9D1),
+    MAN_JUDGE(0x1F468),
+    MAN_JUDGE_LIGHT_SKIN_TONE(0x1F468),
+    MAN_JUDGE_MEDIUM_LIGHT_SKIN_TONE(0x1F468),
+    MAN_JUDGE_MEDIUM_SKIN_TONE(0x1F468),
+    MAN_JUDGE_MEDIUM_DARK_SKIN_TONE(0x1F468),
+    MAN_JUDGE_DARK_SKIN_TONE(0x1F468),
+    WOMAN_JUDGE(0x1F469),
+    WOMAN_JUDGE_LIGHT_SKIN_TONE(0x1F469),
+    WOMAN_JUDGE_MEDIUM_LIGHT_SKIN_TONE(0x1F469),
+    WOMAN_JUDGE_MEDIUM_SKIN_TONE(0x1F469),
+    WOMAN_JUDGE_MEDIUM_DARK_SKIN_TONE(0x1F469),
+    WOMAN_JUDGE_DARK_SKIN_TONE(0x1F469),
+    FARMER(0x1F9D1),
+    FARMER_LIGHT_SKIN_TONE(0x1F9D1),
+    FARMER_MEDIUM_LIGHT_SKIN_TONE(0x1F9D1),
+    FARMER_MEDIUM_SKIN_TONE(0x1F9D1),
+    FARMER_MEDIUM_DARK_SKIN_TONE(0x1F9D1),
+    FARMER_DARK_SKIN_TONE(0x1F9D1),
+    MAN_FARMER(0x1F468),
+    MAN_FARMER_LIGHT_SKIN_TONE(0x1F468),
+    MAN_FARMER_MEDIUM_LIGHT_SKIN_TONE(0x1F468),
+    MAN_FARMER_MEDIUM_SKIN_TONE(0x1F468),
+    MAN_FARMER_MEDIUM_DARK_SKIN_TONE(0x1F468),
+    MAN_FARMER_DARK_SKIN_TONE(0x1F468),
+    WOMAN_FARMER(0x1F469),
+    WOMAN_FARMER_LIGHT_SKIN_TONE(0x1F469),
+    WOMAN_FARMER_MEDIUM_LIGHT_SKIN_TONE(0x1F469),
+    WOMAN_FARMER_MEDIUM_SKIN_TONE(0x1F469),
+    WOMAN_FARMER_MEDIUM_DARK_SKIN_TONE(0x1F469),
+    WOMAN_FARMER_DARK_SKIN_TONE(0x1F469),
+    COOK(0x1F9D1),
+    COOK_LIGHT_SKIN_TONE(0x1F9D1),
+    COOK_MEDIUM_LIGHT_SKIN_TONE(0x1F9D1),
+    COOK_MEDIUM_SKIN_TONE(0x1F9D1),
+    COOK_MEDIUM_DARK_SKIN_TONE(0x1F9D1),
+    COOK_DARK_SKIN_TONE(0x1F9D1),
+    MAN_COOK(0x1F468),
+    MAN_COOK_LIGHT_SKIN_TONE(0x1F468),
+    MAN_COOK_MEDIUM_LIGHT_SKIN_TONE(0x1F468),
+    MAN_COOK_MEDIUM_SKIN_TONE(0x1F468),
+    MAN_COOK_MEDIUM_DARK_SKIN_TONE(0x1F468),
+    MAN_COOK_DARK_SKIN_TONE(0x1F468),
+    WOMAN_COOK(0x1F469),
+    WOMAN_COOK_LIGHT_SKIN_TONE(0x1F469),
+    WOMAN_COOK_MEDIUM_LIGHT_SKIN_TONE(0x1F469),
+    WOMAN_COOK_MEDIUM_SKIN_TONE(0x1F469),
+    WOMAN_COOK_MEDIUM_DARK_SKIN_TONE(0x1F469),
+    WOMAN_COOK_DARK_SKIN_TONE(0x1F469),
+    MECHANIC(0x1F9D1),
+    MECHANIC_LIGHT_SKIN_TONE(0x1F9D1),
+    MECHANIC_MEDIUM_LIGHT_SKIN_TONE(0x1F9D1),
+    MECHANIC_MEDIUM_SKIN_TONE(0x1F9D1),
+    MECHANIC_MEDIUM_DARK_SKIN_TONE(0x1F9D1),
+    MECHANIC_DARK_SKIN_TONE(0x1F9D1),
+    MAN_MECHANIC(0x1F468),
+    MAN_MECHANIC_LIGHT_SKIN_TONE(0x1F468),
+    MAN_MECHANIC_MEDIUM_LIGHT_SKIN_TONE(0x1F468),
+    MAN_MECHANIC_MEDIUM_SKIN_TONE(0x1F468),
+    MAN_MECHANIC_MEDIUM_DARK_SKIN_TONE(0x1F468),
+    MAN_MECHANIC_DARK_SKIN_TONE(0x1F468),
+    WOMAN_MECHANIC(0x1F469),
+    WOMAN_MECHANIC_LIGHT_SKIN_TONE(0x1F469),
+    WOMAN_MECHANIC_MEDIUM_LIGHT_SKIN_TONE(0x1F469),
+    WOMAN_MECHANIC_MEDIUM_SKIN_TONE(0x1F469),
+    WOMAN_MECHANIC_MEDIUM_DARK_SKIN_TONE(0x1F469),
+    WOMAN_MECHANIC_DARK_SKIN_TONE(0x1F469),
+    FACTORY_WORKER(0x1F9D1),
+    FACTORY_WORKER_LIGHT_SKIN_TONE(0x1F9D1),
+    FACTORY_WORKER_MEDIUM_LIGHT_SKIN_TONE(0x1F9D1),
+    FACTORY_WORKER_MEDIUM_SKIN_TONE(0x1F9D1),
+    FACTORY_WORKER_MEDIUM_DARK_SKIN_TONE(0x1F9D1),
+    FACTORY_WORKER_DARK_SKIN_TONE(0x1F9D1),
+    MAN_FACTORY_WORKER(0x1F468),
+    MAN_FACTORY_WORKER_LIGHT_SKIN_TONE(0x1F468),
+    MAN_FACTORY_WORKER_MEDIUM_LIGHT_SKIN_TONE(0x1F468),
+    MAN_FACTORY_WORKER_MEDIUM_SKIN_TONE(0x1F468),
+    MAN_FACTORY_WORKER_MEDIUM_DARK_SKIN_TONE(0x1F468),
+    MAN_FACTORY_WORKER_DARK_SKIN_TONE(0x1F468),
+    WOMAN_FACTORY_WORKER(0x1F469),
+    WOMAN_FACTORY_WORKER_LIGHT_SKIN_TONE(0x1F469),
+    WOMAN_FACTORY_WORKER_MEDIUM_LIGHT_SKIN_TONE(0x1F469),
+    WOMAN_FACTORY_WORKER_MEDIUM_SKIN_TONE(0x1F469),
+    WOMAN_FACTORY_WORKER_MEDIUM_DARK_SKIN_TONE(0x1F469),
+    WOMAN_FACTORY_WORKER_DARK_SKIN_TONE(0x1F469),
+    OFFICE_WORKER(0x1F9D1),
+    OFFICE_WORKER_LIGHT_SKIN_TONE(0x1F9D1),
+    OFFICE_WORKER_MEDIUM_LIGHT_SKIN_TONE(0x1F9D1),
+    OFFICE_WORKER_MEDIUM_SKIN_TONE(0x1F9D1),
+    OFFICE_WORKER_MEDIUM_DARK_SKIN_TONE(0x1F9D1),
+    OFFICE_WORKER_DARK_SKIN_TONE(0x1F9D1),
+    MAN_OFFICE_WORKER(0x1F468),
+    MAN_OFFICE_WORKER_LIGHT_SKIN_TONE(0x1F468),
+    MAN_OFFICE_WORKER_MEDIUM_LIGHT_SKIN_TONE(0x1F468),
+    MAN_OFFICE_WORKER_MEDIUM_SKIN_TONE(0x1F468),
+    MAN_OFFICE_WORKER_MEDIUM_DARK_SKIN_TONE(0x1F468),
+    MAN_OFFICE_WORKER_DARK_SKIN_TONE(0x1F468),
+    WOMAN_OFFICE_WORKER(0x1F469),
+    WOMAN_OFFICE_WORKER_LIGHT_SKIN_TONE(0x1F469),
+    WOMAN_OFFICE_WORKER_MEDIUM_LIGHT_SKIN_TONE(0x1F469),
+    WOMAN_OFFICE_WORKER_MEDIUM_SKIN_TONE(0x1F469),
+    WOMAN_OFFICE_WORKER_MEDIUM_DARK_SKIN_TONE(0x1F469),
+    WOMAN_OFFICE_WORKER_DARK_SKIN_TONE(0x1F469),
+    SCIENTIST(0x1F9D1),
+    SCIENTIST_LIGHT_SKIN_TONE(0x1F9D1),
+    SCIENTIST_MEDIUM_LIGHT_SKIN_TONE(0x1F9D1),
+    SCIENTIST_MEDIUM_SKIN_TONE(0x1F9D1),
+    SCIENTIST_MEDIUM_DARK_SKIN_TONE(0x1F9D1),
+    SCIENTIST_DARK_SKIN_TONE(0x1F9D1),
+    MAN_SCIENTIST(0x1F468),
+    MAN_SCIENTIST_LIGHT_SKIN_TONE(0x1F468),
+    MAN_SCIENTIST_MEDIUM_LIGHT_SKIN_TONE(0x1F468),
+    MAN_SCIENTIST_MEDIUM_SKIN_TONE(0x1F468),
+    MAN_SCIENTIST_MEDIUM_DARK_SKIN_TONE(0x1F468),
+    MAN_SCIENTIST_DARK_SKIN_TONE(0x1F468),
+    WOMAN_SCIENTIST(0x1F469);
+
+    override fun toString(): String {
+        return String(Character.toChars(code))
+    }
+}
