+++ src/main/kotlin/net/purefunc/emoji/Emoji1.kt
@@ -1,0 +1,1011 @@
+package net.purefunc.emoji
+
+enum class Emoji1(
+    private val code: Int,
+) {
+
+    WOMAN_SCIENTIST_LIGHT_SKIN_TONE(0x1F469),
+    WOMAN_SCIENTIST_MEDIUM_LIGHT_SKIN_TONE(0x1F469),
+    WOMAN_SCIENTIST_MEDIUM_SKIN_TONE(0x1F469),
+    WOMAN_SCIENTIST_MEDIUM_DARK_SKIN_TONE(0x1F469),
+    WOMAN_SCIENTIST_DARK_SKIN_TONE(0x1F469),
+    TECHNOLOGIST(0x1F9D1),
+    TECHNOLOGIST_LIGHT_SKIN_TONE(0x1F9D1),
+    TECHNOLOGIST_MEDIUM_LIGHT_SKIN_TONE(0x1F9D1),
+    TECHNOLOGIST_MEDIUM_SKIN_TONE(0x1F9D1),
+    TECHNOLOGIST_MEDIUM_DARK_SKIN_TONE(0x1F9D1),
+    TECHNOLOGIST_DARK_SKIN_TONE(0x1F9D1),
+    MAN_TECHNOLOGIST(0x1F468),
+    MAN_TECHNOLOGIST_LIGHT_SKIN_TONE(0x1F468),
+    MAN_TECHNOLOGIST_MEDIUM_LIGHT_SKIN_TONE(0x1F468),
+    MAN_TECHNOLOGIST_MEDIUM_SKIN_TONE(0x1F468),
+    MAN_TECHNOLOGIST_MEDIUM_DARK_SKIN_TONE(0x1F468),
+    MAN_TECHNOLOGIST_DARK_SKIN_TONE(0x1F468),
+    WOMAN_TECHNOLOGIST(0x1F469),
+    WOMAN_TECHNOLOGIST_LIGHT_SKIN_TONE(0x1F469),
+    WOMAN_TECHNOLOGIST_MEDIUM_LIGHT_SKIN_TONE(0x1F469),
+    WOMAN_TECHNOLOGIST_MEDIUM_SKIN_TONE(0x1F469),
+    WOMAN_TECHNOLOGIST_MEDIUM_DARK_SKIN_TONE(0x1F469),
+    WOMAN_TECHNOLOGIST_DARK_SKIN_TONE(0x1F469),
+    SINGER(0x1F9D1),
+    SINGER_LIGHT_SKIN_TONE(0x1F9D1),
+    SINGER_MEDIUM_LIGHT_SKIN_TONE(0x1F9D1),
+    SINGER_MEDIUM_SKIN_TONE(0x1F9D1),
+    SINGER_MEDIUM_DARK_SKIN_TONE(0x1F9D1),
+    SINGER_DARK_SKIN_TONE(0x1F9D1),
+    MAN_SINGER(0x1F468),
+    MAN_SINGER_LIGHT_SKIN_TONE(0x1F468),
+    MAN_SINGER_MEDIUM_LIGHT_SKIN_TONE(0x1F468),
+    MAN_SINGER_MEDIUM_SKIN_TONE(0x1F468),
+    MAN_SINGER_MEDIUM_DARK_SKIN_TONE(0x1F468),
+    MAN_SINGER_DARK_SKIN_TONE(0x1F468),
+    WOMAN_SINGER(0x1F469),
+    WOMAN_SINGER_LIGHT_SKIN_TONE(0x1F469),
+    WOMAN_SINGER_MEDIUM_LIGHT_SKIN_TONE(0x1F469),
+    WOMAN_SINGER_MEDIUM_SKIN_TONE(0x1F469),
+    WOMAN_SINGER_MEDIUM_DARK_SKIN_TONE(0x1F469),
+    WOMAN_SINGER_DARK_SKIN_TONE(0x1F469),
+    ARTIST(0x1F9D1),
+    ARTIST_LIGHT_SKIN_TONE(0x1F9D1),
+    ARTIST_MEDIUM_LIGHT_SKIN_TONE(0x1F9D1),
+    ARTIST_MEDIUM_SKIN_TONE(0x1F9D1),
+    ARTIST_MEDIUM_DARK_SKIN_TONE(0x1F9D1),
+    ARTIST_DARK_SKIN_TONE(0x1F9D1),
+    MAN_ARTIST(0x1F468),
+    MAN_ARTIST_LIGHT_SKIN_TONE(0x1F468),
+    MAN_ARTIST_MEDIUM_LIGHT_SKIN_TONE(0x1F468),
+    MAN_ARTIST_MEDIUM_SKIN_TONE(0x1F468),
+    MAN_ARTIST_MEDIUM_DARK_SKIN_TONE(0x1F468),
+    MAN_ARTIST_DARK_SKIN_TONE(0x1F468),
+    WOMAN_ARTIST(0x1F469),
+    WOMAN_ARTIST_LIGHT_SKIN_TONE(0x1F469),
+    WOMAN_ARTIST_MEDIUM_LIGHT_SKIN_TONE(0x1F469),
+    WOMAN_ARTIST_MEDIUM_SKIN_TONE(0x1F469),
+    WOMAN_ARTIST_MEDIUM_DARK_SKIN_TONE(0x1F469),
+    WOMAN_ARTIST_DARK_SKIN_TONE(0x1F469),
+    PILOT(0x1F9D1),
+    PILOT_LIGHT_SKIN_TONE(0x1F9D1),
+    PILOT_MEDIUM_LIGHT_SKIN_TONE(0x1F9D1),
+    PILOT_MEDIUM_SKIN_TONE(0x1F9D1),
+    PILOT_MEDIUM_DARK_SKIN_TONE(0x1F9D1),
+    PILOT_DARK_SKIN_TONE(0x1F9D1),
+    MAN_PILOT(0x1F468),
+    MAN_PILOT_LIGHT_SKIN_TONE(0x1F468),
+    MAN_PILOT_MEDIUM_LIGHT_SKIN_TONE(0x1F468),
+    MAN_PILOT_MEDIUM_SKIN_TONE(0x1F468),
+    MAN_PILOT_MEDIUM_DARK_SKIN_TONE(0x1F468),
+    MAN_PILOT_DARK_SKIN_TONE(0x1F468),
+    WOMAN_PILOT(0x1F469),
+    WOMAN_PILOT_LIGHT_SKIN_TONE(0x1F469),
+    WOMAN_PILOT_MEDIUM_LIGHT_SKIN_TONE(0x1F469),
+    WOMAN_PILOT_MEDIUM_SKIN_TONE(0x1F469),
+    WOMAN_PILOT_MEDIUM_DARK_SKIN_TONE(0x1F469),
+    WOMAN_PILOT_DARK_SKIN_TONE(0x1F469),
+    ASTRONAUT(0x1F9D1),
+    ASTRONAUT_LIGHT_SKIN_TONE(0x1F9D1),
+    ASTRONAUT_MEDIUM_LIGHT_SKIN_TONE(0x1F9D1),
+    ASTRONAUT_MEDIUM_SKIN_TONE(0x1F9D1),
+    ASTRONAUT_MEDIUM_DARK_SKIN_TONE(0x1F9D1),
+    ASTRONAUT_DARK_SKIN_TONE(0x1F9D1),
+    MAN_ASTRONAUT(0x1F468),
+    MAN_ASTRONAUT_LIGHT_SKIN_TONE(0x1F468),
+    MAN_ASTRONAUT_MEDIUM_LIGHT_SKIN_TONE(0x1F468),
+    MAN_ASTRONAUT_MEDIUM_SKIN_TONE(0x1F468),
+    MAN_ASTRONAUT_MEDIUM_DARK_SKIN_TONE(0x1F468),
+    MAN_ASTRONAUT_DARK_SKIN_TONE(0x1F468),
+    WOMAN_ASTRONAUT(0x1F469),
+    WOMAN_ASTRONAUT_LIGHT_SKIN_TONE(0x1F469),
+    WOMAN_ASTRONAUT_MEDIUM_LIGHT_SKIN_TONE(0x1F469),
+    WOMAN_ASTRONAUT_MEDIUM_SKIN_TONE(0x1F469),
+    WOMAN_ASTRONAUT_MEDIUM_DARK_SKIN_TONE(0x1F469),
+    WOMAN_ASTRONAUT_DARK_SKIN_TONE(0x1F469),
+    FIREFIGHTER(0x1F9D1),
+    FIREFIGHTER_LIGHT_SKIN_TONE(0x1F9D1),
+    FIREFIGHTER_MEDIUM_LIGHT_SKIN_TONE(0x1F9D1),
+    FIREFIGHTER_MEDIUM_SKIN_TONE(0x1F9D1),
+    FIREFIGHTER_MEDIUM_DARK_SKIN_TONE(0x1F9D1),
+    FIREFIGHTER_DARK_SKIN_TONE(0x1F9D1),
+    MAN_FIREFIGHTER(0x1F468),
+    MAN_FIREFIGHTER_LIGHT_SKIN_TONE(0x1F468),
+    MAN_FIREFIGHTER_MEDIUM_LIGHT_SKIN_TONE(0x1F468),
+    MAN_FIREFIGHTER_MEDIUM_SKIN_TONE(0x1F468),
+    MAN_FIREFIGHTER_MEDIUM_DARK_SKIN_TONE(0x1F468),
+    MAN_FIREFIGHTER_DARK_SKIN_TONE(0x1F468),
+    WOMAN_FIREFIGHTER(0x1F469),
+    WOMAN_FIREFIGHTER_LIGHT_SKIN_TONE(0x1F469),
+    WOMAN_FIREFIGHTER_MEDIUM_LIGHT_SKIN_TONE(0x1F469),
+    WOMAN_FIREFIGHTER_MEDIUM_SKIN_TONE(0x1F469),
+    WOMAN_FIREFIGHTER_MEDIUM_DARK_SKIN_TONE(0x1F469),
+    WOMAN_FIREFIGHTER_DARK_SKIN_TONE(0x1F469),
+    POLICE_OFFICER(0x1F46E),
+    POLICE_OFFICER_LIGHT_SKIN_TONE(0x1F46E),
+    POLICE_OFFICER_MEDIUM_LIGHT_SKIN_TONE(0x1F46E),
+    POLICE_OFFICER_MEDIUM_SKIN_TONE(0x1F46E),
+    POLICE_OFFICER_MEDIUM_DARK_SKIN_TONE(0x1F46E),
+    POLICE_OFFICER_DARK_SKIN_TONE(0x1F46E),
+    MAN_POLICE_OFFICER(0x1F46E),
+    MAN_POLICE_OFFICER_LIGHT_SKIN_TONE(0x1F46E),
+    MAN_POLICE_OFFICER_MEDIUM_LIGHT_SKIN_TONE(0x1F46E),
+    MAN_POLICE_OFFICER_MEDIUM_SKIN_TONE(0x1F46E),
+    MAN_POLICE_OFFICER_MEDIUM_DARK_SKIN_TONE(0x1F46E),
+    MAN_POLICE_OFFICER_DARK_SKIN_TONE(0x1F46E),
+    WOMAN_POLICE_OFFICER(0x1F46E),
+    WOMAN_POLICE_OFFICER_LIGHT_SKIN_TONE(0x1F46E),
+    WOMAN_POLICE_OFFICER_MEDIUM_LIGHT_SKIN_TONE(0x1F46E),
+    WOMAN_POLICE_OFFICER_MEDIUM_SKIN_TONE(0x1F46E),
+    WOMAN_POLICE_OFFICER_MEDIUM_DARK_SKIN_TONE(0x1F46E),
+    WOMAN_POLICE_OFFICER_DARK_SKIN_TONE(0x1F46E),
+    DETECTIVE(0x1F575),
+    DETECTIVE_LIGHT_SKIN_TONE(0x1F575),
+    DETECTIVE_MEDIUM_LIGHT_SKIN_TONE(0x1F575),
+    DETECTIVE_MEDIUM_SKIN_TONE(0x1F575),
+    DETECTIVE_MEDIUM_DARK_SKIN_TONE(0x1F575),
+    DETECTIVE_DARK_SKIN_TONE(0x1F575),
+    MAN_DETECTIVE(0x1F575),
+    MAN_DETECTIVE_LIGHT_SKIN_TONE(0x1F575),
+    MAN_DETECTIVE_MEDIUM_LIGHT_SKIN_TONE(0x1F575),
+    MAN_DETECTIVE_MEDIUM_SKIN_TONE(0x1F575),
+    MAN_DETECTIVE_MEDIUM_DARK_SKIN_TONE(0x1F575),
+    MAN_DETECTIVE_DARK_SKIN_TONE(0x1F575),
+    WOMAN_DETECTIVE(0x1F575),
+    WOMAN_DETECTIVE_LIGHT_SKIN_TONE(0x1F575),
+    WOMAN_DETECTIVE_MEDIUM_LIGHT_SKIN_TONE(0x1F575),
+    WOMAN_DETECTIVE_MEDIUM_SKIN_TONE(0x1F575),
+    WOMAN_DETECTIVE_MEDIUM_DARK_SKIN_TONE(0x1F575),
+    WOMAN_DETECTIVE_DARK_SKIN_TONE(0x1F575),
+    GUARD(0x1F482),
+    GUARD_LIGHT_SKIN_TONE(0x1F482),
+    GUARD_MEDIUM_LIGHT_SKIN_TONE(0x1F482),
+    GUARD_MEDIUM_SKIN_TONE(0x1F482),
+    GUARD_MEDIUM_DARK_SKIN_TONE(0x1F482),
+    GUARD_DARK_SKIN_TONE(0x1F482),
+    MAN_GUARD(0x1F482),
+    MAN_GUARD_LIGHT_SKIN_TONE(0x1F482),
+    MAN_GUARD_MEDIUM_LIGHT_SKIN_TONE(0x1F482),
+    MAN_GUARD_MEDIUM_SKIN_TONE(0x1F482),
+    MAN_GUARD_MEDIUM_DARK_SKIN_TONE(0x1F482),
+    MAN_GUARD_DARK_SKIN_TONE(0x1F482),
+    WOMAN_GUARD(0x1F482),
+    WOMAN_GUARD_LIGHT_SKIN_TONE(0x1F482),
+    WOMAN_GUARD_MEDIUM_LIGHT_SKIN_TONE(0x1F482),
+    WOMAN_GUARD_MEDIUM_SKIN_TONE(0x1F482),
+    WOMAN_GUARD_MEDIUM_DARK_SKIN_TONE(0x1F482),
+    WOMAN_GUARD_DARK_SKIN_TONE(0x1F482),
+    NINJA(0x1F977),
+    NINJA_LIGHT_SKIN_TONE(0x1F977),
+    NINJA_MEDIUM_LIGHT_SKIN_TONE(0x1F977),
+    NINJA_MEDIUM_SKIN_TONE(0x1F977),
+    NINJA_MEDIUM_DARK_SKIN_TONE(0x1F977),
+    NINJA_DARK_SKIN_TONE(0x1F977),
+    CONSTRUCTION_WORKER(0x1F477),
+    CONSTRUCTION_WORKER_LIGHT_SKIN_TONE(0x1F477),
+    CONSTRUCTION_WORKER_MEDIUM_LIGHT_SKIN_TONE(0x1F477),
+    CONSTRUCTION_WORKER_MEDIUM_SKIN_TONE(0x1F477),
+    CONSTRUCTION_WORKER_MEDIUM_DARK_SKIN_TONE(0x1F477),
+    CONSTRUCTION_WORKER_DARK_SKIN_TONE(0x1F477),
+    MAN_CONSTRUCTION_WORKER(0x1F477),
+    MAN_CONSTRUCTION_WORKER_LIGHT_SKIN_TONE(0x1F477),
+    MAN_CONSTRUCTION_WORKER_MEDIUM_LIGHT_SKIN_TONE(0x1F477),
+    MAN_CONSTRUCTION_WORKER_MEDIUM_SKIN_TONE(0x1F477),
+    MAN_CONSTRUCTION_WORKER_MEDIUM_DARK_SKIN_TONE(0x1F477),
+    MAN_CONSTRUCTION_WORKER_DARK_SKIN_TONE(0x1F477),
+    WOMAN_CONSTRUCTION_WORKER(0x1F477),
+    WOMAN_CONSTRUCTION_WORKER_LIGHT_SKIN_TONE(0x1F477),
+    WOMAN_CONSTRUCTION_WORKER_MEDIUM_LIGHT_SKIN_TONE(0x1F477),
+    WOMAN_CONSTRUCTION_WORKER_MEDIUM_SKIN_TONE(0x1F477),
+    WOMAN_CONSTRUCTION_WORKER_MEDIUM_DARK_SKIN_TONE(0x1F477),
+    WOMAN_CONSTRUCTION_WORKER_DARK_SKIN_TONE(0x1F477),
+    PERSON_WITH_CROWN(0x1FAC5),
+    PERSON_WITH_CROWN_LIGHT_SKIN_TONE(0x1FAC5),
+    PERSON_WITH_CROWN_MEDIUM_LIGHT_SKIN_TONE(0x1FAC5),
+    PERSON_WITH_CROWN_MEDIUM_SKIN_TONE(0x1FAC5),
+    PERSON_WITH_CROWN_MEDIUM_DARK_SKIN_TONE(0x1FAC5),
+    PERSON_WITH_CROWN_DARK_SKIN_TONE(0x1FAC5),
+    PRINCE(0x1F934),
+    PRINCE_LIGHT_SKIN_TONE(0x1F934),
+    PRINCE_MEDIUM_LIGHT_SKIN_TONE(0x1F934),
+    PRINCE_MEDIUM_SKIN_TONE(0x1F934),
+    PRINCE_MEDIUM_DARK_SKIN_TONE(0x1F934),
+    PRINCE_DARK_SKIN_TONE(0x1F934),
+    PRINCESS(0x1F478),
+    PRINCESS_LIGHT_SKIN_TONE(0x1F478),
+    PRINCESS_MEDIUM_LIGHT_SKIN_TONE(0x1F478),
+    PRINCESS_MEDIUM_SKIN_TONE(0x1F478),
+    PRINCESS_MEDIUM_DARK_SKIN_TONE(0x1F478),
+    PRINCESS_DARK_SKIN_TONE(0x1F478),
+    PERSON_WEARING_TURBAN(0x1F473),
+    PERSON_WEARING_TURBAN_LIGHT_SKIN_TONE(0x1F473),
+    PERSON_WEARING_TURBAN_MEDIUM_LIGHT_SKIN_TONE(0x1F473),
+    PERSON_WEARING_TURBAN_MEDIUM_SKIN_TONE(0x1F473),
+    PERSON_WEARING_TURBAN_MEDIUM_DARK_SKIN_TONE(0x1F473),
+    PERSON_WEARING_TURBAN_DARK_SKIN_TONE(0x1F473),
+    MAN_WEARING_TURBAN(0x1F473),
+    MAN_WEARING_TURBAN_LIGHT_SKIN_TONE(0x1F473),
+    MAN_WEARING_TURBAN_MEDIUM_LIGHT_SKIN_TONE(0x1F473),
+    MAN_WEARING_TURBAN_MEDIUM_SKIN_TONE(0x1F473),
+    MAN_WEARING_TURBAN_MEDIUM_DARK_SKIN_TONE(0x1F473),
+    MAN_WEARING_TURBAN_DARK_SKIN_TONE(0x1F473),
+    WOMAN_WEARING_TURBAN(0x1F473),
+    WOMAN_WEARING_TURBAN_LIGHT_SKIN_TONE(0x1F473),
+    WOMAN_WEARING_TURBAN_MEDIUM_LIGHT_SKIN_TONE(0x1F473),
+    WOMAN_WEARING_TURBAN_MEDIUM_SKIN_TONE(0x1F473),
+    WOMAN_WEARING_TURBAN_MEDIUM_DARK_SKIN_TONE(0x1F473),
+    WOMAN_WEARING_TURBAN_DARK_SKIN_TONE(0x1F473),
+    PERSON_WITH_SKULLCAP(0x1F472),
+    PERSON_WITH_SKULLCAP_LIGHT_SKIN_TONE(0x1F472),
+    PERSON_WITH_SKULLCAP_MEDIUM_LIGHT_SKIN_TONE(0x1F472),
+    PERSON_WITH_SKULLCAP_MEDIUM_SKIN_TONE(0x1F472),
+    PERSON_WITH_SKULLCAP_MEDIUM_DARK_SKIN_TONE(0x1F472),
+    PERSON_WITH_SKULLCAP_DARK_SKIN_TONE(0x1F472),
+    WOMAN_WITH_HEADSCARF(0x1F9D5),
+    WOMAN_WITH_HEADSCARF_LIGHT_SKIN_TONE(0x1F9D5),
+    WOMAN_WITH_HEADSCARF_MEDIUM_LIGHT_SKIN_TONE(0x1F9D5),
+    WOMAN_WITH_HEADSCARF_MEDIUM_SKIN_TONE(0x1F9D5),
+    WOMAN_WITH_HEADSCARF_MEDIUM_DARK_SKIN_TONE(0x1F9D5),
+    WOMAN_WITH_HEADSCARF_DARK_SKIN_TONE(0x1F9D5),
+    PERSON_IN_TUXEDO(0x1F935),
+    PERSON_IN_TUXEDO_LIGHT_SKIN_TONE(0x1F935),
+    PERSON_IN_TUXEDO_MEDIUM_LIGHT_SKIN_TONE(0x1F935),
+    PERSON_IN_TUXEDO_MEDIUM_SKIN_TONE(0x1F935),
+    PERSON_IN_TUXEDO_MEDIUM_DARK_SKIN_TONE(0x1F935),
+    PERSON_IN_TUXEDO_DARK_SKIN_TONE(0x1F935),
+    MAN_IN_TUXEDO(0x1F935),
+    MAN_IN_TUXEDO_LIGHT_SKIN_TONE(0x1F935),
+    MAN_IN_TUXEDO_MEDIUM_LIGHT_SKIN_TONE(0x1F935),
+    MAN_IN_TUXEDO_MEDIUM_SKIN_TONE(0x1F935),
+    MAN_IN_TUXEDO_MEDIUM_DARK_SKIN_TONE(0x1F935),
+    MAN_IN_TUXEDO_DARK_SKIN_TONE(0x1F935),
+    WOMAN_IN_TUXEDO(0x1F935),
+    WOMAN_IN_TUXEDO_LIGHT_SKIN_TONE(0x1F935),
+    WOMAN_IN_TUXEDO_MEDIUM_LIGHT_SKIN_TONE(0x1F935),
+    WOMAN_IN_TUXEDO_MEDIUM_SKIN_TONE(0x1F935),
+    WOMAN_IN_TUXEDO_MEDIUM_DARK_SKIN_TONE(0x1F935),
+    WOMAN_IN_TUXEDO_DARK_SKIN_TONE(0x1F935),
+    PERSON_WITH_VEIL(0x1F470),
+    PERSON_WITH_VEIL_LIGHT_SKIN_TONE(0x1F470),
+    PERSON_WITH_VEIL_MEDIUM_LIGHT_SKIN_TONE(0x1F470),
+    PERSON_WITH_VEIL_MEDIUM_SKIN_TONE(0x1F470),
+    PERSON_WITH_VEIL_MEDIUM_DARK_SKIN_TONE(0x1F470),
+    PERSON_WITH_VEIL_DARK_SKIN_TONE(0x1F470),
+    MAN_WITH_VEIL(0x1F470),
+    MAN_WITH_VEIL_LIGHT_SKIN_TONE(0x1F470),
+    MAN_WITH_VEIL_MEDIUM_LIGHT_SKIN_TONE(0x1F470),
+    MAN_WITH_VEIL_MEDIUM_SKIN_TONE(0x1F470),
+    MAN_WITH_VEIL_MEDIUM_DARK_SKIN_TONE(0x1F470),
+    MAN_WITH_VEIL_DARK_SKIN_TONE(0x1F470),
+    WOMAN_WITH_VEIL(0x1F470),
+    WOMAN_WITH_VEIL_LIGHT_SKIN_TONE(0x1F470),
+    WOMAN_WITH_VEIL_MEDIUM_LIGHT_SKIN_TONE(0x1F470),
+    WOMAN_WITH_VEIL_MEDIUM_SKIN_TONE(0x1F470),
+    WOMAN_WITH_VEIL_MEDIUM_DARK_SKIN_TONE(0x1F470),
+    WOMAN_WITH_VEIL_DARK_SKIN_TONE(0x1F470),
+    PREGNANT_WOMAN(0x1F930),
+    PREGNANT_WOMAN_LIGHT_SKIN_TONE(0x1F930),
+    PREGNANT_WOMAN_MEDIUM_LIGHT_SKIN_TONE(0x1F930),
+    PREGNANT_WOMAN_MEDIUM_SKIN_TONE(0x1F930),
+    PREGNANT_WOMAN_MEDIUM_DARK_SKIN_TONE(0x1F930),
+    PREGNANT_WOMAN_DARK_SKIN_TONE(0x1F930),
+    PREGNANT_MAN(0x1FAC3),
+    PREGNANT_MAN_LIGHT_SKIN_TONE(0x1FAC3),
+    PREGNANT_MAN_MEDIUM_LIGHT_SKIN_TONE(0x1FAC3),
+    PREGNANT_MAN_MEDIUM_SKIN_TONE(0x1FAC3),
+    PREGNANT_MAN_MEDIUM_DARK_SKIN_TONE(0x1FAC3),
+    PREGNANT_MAN_DARK_SKIN_TONE(0x1FAC3),
+    PREGNANT_PERSON(0x1FAC4),
+    PREGNANT_PERSON_LIGHT_SKIN_TONE(0x1FAC4),
+    PREGNANT_PERSON_MEDIUM_LIGHT_SKIN_TONE(0x1FAC4),
+    PREGNANT_PERSON_MEDIUM_SKIN_TONE(0x1FAC4),
+    PREGNANT_PERSON_MEDIUM_DARK_SKIN_TONE(0x1FAC4),
+    PREGNANT_PERSON_DARK_SKIN_TONE(0x1FAC4),
+    BREAST_FEEDING(0x1F931),
+    BREAST_FEEDING_LIGHT_SKIN_TONE(0x1F931),
+    BREAST_FEEDING_MEDIUM_LIGHT_SKIN_TONE(0x1F931),
+    BREAST_FEEDING_MEDIUM_SKIN_TONE(0x1F931),
+    BREAST_FEEDING_MEDIUM_DARK_SKIN_TONE(0x1F931),
+    BREAST_FEEDING_DARK_SKIN_TONE(0x1F931),
+    WOMAN_FEEDING_BABY(0x1F469),
+    WOMAN_FEEDING_BABY_LIGHT_SKIN_TONE(0x1F469),
+    WOMAN_FEEDING_BABY_MEDIUM_LIGHT_SKIN_TONE(0x1F469),
+    WOMAN_FEEDING_BABY_MEDIUM_SKIN_TONE(0x1F469),
+    WOMAN_FEEDING_BABY_MEDIUM_DARK_SKIN_TONE(0x1F469),
+    WOMAN_FEEDING_BABY_DARK_SKIN_TONE(0x1F469),
+    MAN_FEEDING_BABY(0x1F468),
+    MAN_FEEDING_BABY_LIGHT_SKIN_TONE(0x1F468),
+    MAN_FEEDING_BABY_MEDIUM_LIGHT_SKIN_TONE(0x1F468),
+    MAN_FEEDING_BABY_MEDIUM_SKIN_TONE(0x1F468),
+    MAN_FEEDING_BABY_MEDIUM_DARK_SKIN_TONE(0x1F468),
+    MAN_FEEDING_BABY_DARK_SKIN_TONE(0x1F468),
+    PERSON_FEEDING_BABY(0x1F9D1),
+    PERSON_FEEDING_BABY_LIGHT_SKIN_TONE(0x1F9D1),
+    PERSON_FEEDING_BABY_MEDIUM_LIGHT_SKIN_TONE(0x1F9D1),
+    PERSON_FEEDING_BABY_MEDIUM_SKIN_TONE(0x1F9D1),
+    PERSON_FEEDING_BABY_MEDIUM_DARK_SKIN_TONE(0x1F9D1),
+    PERSON_FEEDING_BABY_DARK_SKIN_TONE(0x1F9D1),
+    BABY_ANGEL(0x1F47C),
+    BABY_ANGEL_LIGHT_SKIN_TONE(0x1F47C),
+    BABY_ANGEL_MEDIUM_LIGHT_SKIN_TONE(0x1F47C),
+    BABY_ANGEL_MEDIUM_SKIN_TONE(0x1F47C),
+    BABY_ANGEL_MEDIUM_DARK_SKIN_TONE(0x1F47C),
+    BABY_ANGEL_DARK_SKIN_TONE(0x1F47C),
+    SANTA_CLAUS(0x1F385),
+    SANTA_CLAUS_LIGHT_SKIN_TONE(0x1F385),
+    SANTA_CLAUS_MEDIUM_LIGHT_SKIN_TONE(0x1F385),
+    SANTA_CLAUS_MEDIUM_SKIN_TONE(0x1F385),
+    SANTA_CLAUS_MEDIUM_DARK_SKIN_TONE(0x1F385),
+    SANTA_CLAUS_DARK_SKIN_TONE(0x1F385),
+    MRS_CLAUS(0x1F936),
+    MRS_CLAUS_LIGHT_SKIN_TONE(0x1F936),
+    MRS_CLAUS_MEDIUM_LIGHT_SKIN_TONE(0x1F936),
+    MRS_CLAUS_MEDIUM_SKIN_TONE(0x1F936),
+    MRS_CLAUS_MEDIUM_DARK_SKIN_TONE(0x1F936),
+    MRS_CLAUS_DARK_SKIN_TONE(0x1F936),
+    MX_CLAUS(0x1F9D1),
+    MX_CLAUS_LIGHT_SKIN_TONE(0x1F9D1),
+    MX_CLAUS_MEDIUM_LIGHT_SKIN_TONE(0x1F9D1),
+    MX_CLAUS_MEDIUM_SKIN_TONE(0x1F9D1),
+    MX_CLAUS_MEDIUM_DARK_SKIN_TONE(0x1F9D1),
+    MX_CLAUS_DARK_SKIN_TONE(0x1F9D1),
+    SUPERHERO(0x1F9B8),
+    SUPERHERO_LIGHT_SKIN_TONE(0x1F9B8),
+    SUPERHERO_MEDIUM_LIGHT_SKIN_TONE(0x1F9B8),
+    SUPERHERO_MEDIUM_SKIN_TONE(0x1F9B8),
+    SUPERHERO_MEDIUM_DARK_SKIN_TONE(0x1F9B8),
+    SUPERHERO_DARK_SKIN_TONE(0x1F9B8),
+    MAN_SUPERHERO(0x1F9B8),
+    MAN_SUPERHERO_LIGHT_SKIN_TONE(0x1F9B8),
+    MAN_SUPERHERO_MEDIUM_LIGHT_SKIN_TONE(0x1F9B8),
+    MAN_SUPERHERO_MEDIUM_SKIN_TONE(0x1F9B8),
+    MAN_SUPERHERO_MEDIUM_DARK_SKIN_TONE(0x1F9B8),
+    MAN_SUPERHERO_DARK_SKIN_TONE(0x1F9B8),
+    WOMAN_SUPERHERO(0x1F9B8),
+    WOMAN_SUPERHERO_LIGHT_SKIN_TONE(0x1F9B8),
+    WOMAN_SUPERHERO_MEDIUM_LIGHT_SKIN_TONE(0x1F9B8),
+    WOMAN_SUPERHERO_MEDIUM_SKIN_TONE(0x1F9B8),
+    WOMAN_SUPERHERO_MEDIUM_DARK_SKIN_TONE(0x1F9B8),
+    WOMAN_SUPERHERO_DARK_SKIN_TONE(0x1F9B8),
+    SUPERVILLAIN(0x1F9B9),
+    SUPERVILLAIN_LIGHT_SKIN_TONE(0x1F9B9),
+    SUPERVILLAIN_MEDIUM_LIGHT_SKIN_TONE(0x1F9B9),
+    SUPERVILLAIN_MEDIUM_SKIN_TONE(0x1F9B9),
+    SUPERVILLAIN_MEDIUM_DARK_SKIN_TONE(0x1F9B9),
+    SUPERVILLAIN_DARK_SKIN_TONE(0x1F9B9),
+    MAN_SUPERVILLAIN(0x1F9B9),
+    MAN_SUPERVILLAIN_LIGHT_SKIN_TONE(0x1F9B9),
+    MAN_SUPERVILLAIN_MEDIUM_LIGHT_SKIN_TONE(0x1F9B9),
+    MAN_SUPERVILLAIN_MEDIUM_SKIN_TONE(0x1F9B9),
+    MAN_SUPERVILLAIN_MEDIUM_DARK_SKIN_TONE(0x1F9B9),
+    MAN_SUPERVILLAIN_DARK_SKIN_TONE(0x1F9B9),
+    WOMAN_SUPERVILLAIN(0x1F9B9),
+    WOMAN_SUPERVILLAIN_LIGHT_SKIN_TONE(0x1F9B9),
+    WOMAN_SUPERVILLAIN_MEDIUM_LIGHT_SKIN_TONE(0x1F9B9),
+    WOMAN_SUPERVILLAIN_MEDIUM_SKIN_TONE(0x1F9B9),
+    WOMAN_SUPERVILLAIN_MEDIUM_DARK_SKIN_TONE(0x1F9B9),
+    WOMAN_SUPERVILLAIN_DARK_SKIN_TONE(0x1F9B9),
+    MAGE(0x1F9D9),
+    MAGE_LIGHT_SKIN_TONE(0x1F9D9),
+    MAGE_MEDIUM_LIGHT_SKIN_TONE(0x1F9D9),
+    MAGE_MEDIUM_SKIN_TONE(0x1F9D9),
+    MAGE_MEDIUM_DARK_SKIN_TONE(0x1F9D9),
+    MAGE_DARK_SKIN_TONE(0x1F9D9),
+    MAN_MAGE(0x1F9D9),
+    MAN_MAGE_LIGHT_SKIN_TONE(0x1F9D9),
+    MAN_MAGE_MEDIUM_LIGHT_SKIN_TONE(0x1F9D9),
+    MAN_MAGE_MEDIUM_SKIN_TONE(0x1F9D9),
+    MAN_MAGE_MEDIUM_DARK_SKIN_TONE(0x1F9D9),
+    MAN_MAGE_DARK_SKIN_TONE(0x1F9D9),
+    WOMAN_MAGE(0x1F9D9),
+    WOMAN_MAGE_LIGHT_SKIN_TONE(0x1F9D9),
+    WOMAN_MAGE_MEDIUM_LIGHT_SKIN_TONE(0x1F9D9),
+    WOMAN_MAGE_MEDIUM_SKIN_TONE(0x1F9D9),
+    WOMAN_MAGE_MEDIUM_DARK_SKIN_TONE(0x1F9D9),
+    WOMAN_MAGE_DARK_SKIN_TONE(0x1F9D9),
+    FAIRY(0x1F9DA),
+    FAIRY_LIGHT_SKIN_TONE(0x1F9DA),
+    FAIRY_MEDIUM_LIGHT_SKIN_TONE(0x1F9DA),
+    FAIRY_MEDIUM_SKIN_TONE(0x1F9DA),
+    FAIRY_MEDIUM_DARK_SKIN_TONE(0x1F9DA),
+    FAIRY_DARK_SKIN_TONE(0x1F9DA),
+    MAN_FAIRY(0x1F9DA),
+    MAN_FAIRY_LIGHT_SKIN_TONE(0x1F9DA),
+    MAN_FAIRY_MEDIUM_LIGHT_SKIN_TONE(0x1F9DA),
+    MAN_FAIRY_MEDIUM_SKIN_TONE(0x1F9DA),
+    MAN_FAIRY_MEDIUM_DARK_SKIN_TONE(0x1F9DA),
+    MAN_FAIRY_DARK_SKIN_TONE(0x1F9DA),
+    WOMAN_FAIRY(0x1F9DA),
+    WOMAN_FAIRY_LIGHT_SKIN_TONE(0x1F9DA),
+    WOMAN_FAIRY_MEDIUM_LIGHT_SKIN_TONE(0x1F9DA),
+    WOMAN_FAIRY_MEDIUM_SKIN_TONE(0x1F9DA),
+    WOMAN_FAIRY_MEDIUM_DARK_SKIN_TONE(0x1F9DA),
+    WOMAN_FAIRY_DARK_SKIN_TONE(0x1F9DA),
+    VAMPIRE(0x1F9DB),
+    VAMPIRE_LIGHT_SKIN_TONE(0x1F9DB),
+    VAMPIRE_MEDIUM_LIGHT_SKIN_TONE(0x1F9DB),
+    VAMPIRE_MEDIUM_SKIN_TONE(0x1F9DB),
+    VAMPIRE_MEDIUM_DARK_SKIN_TONE(0x1F9DB),
+    VAMPIRE_DARK_SKIN_TONE(0x1F9DB),
+    MAN_VAMPIRE(0x1F9DB),
+    MAN_VAMPIRE_LIGHT_SKIN_TONE(0x1F9DB),
+    MAN_VAMPIRE_MEDIUM_LIGHT_SKIN_TONE(0x1F9DB),
+    MAN_VAMPIRE_MEDIUM_SKIN_TONE(0x1F9DB),
+    MAN_VAMPIRE_MEDIUM_DARK_SKIN_TONE(0x1F9DB),
+    MAN_VAMPIRE_DARK_SKIN_TONE(0x1F9DB),
+    WOMAN_VAMPIRE(0x1F9DB),
+    WOMAN_VAMPIRE_LIGHT_SKIN_TONE(0x1F9DB),
+    WOMAN_VAMPIRE_MEDIUM_LIGHT_SKIN_TONE(0x1F9DB),
+    WOMAN_VAMPIRE_MEDIUM_SKIN_TONE(0x1F9DB),
+    WOMAN_VAMPIRE_MEDIUM_DARK_SKIN_TONE(0x1F9DB),
+    WOMAN_VAMPIRE_DARK_SKIN_TONE(0x1F9DB),
+    MERPERSON(0x1F9DC),
+    MERPERSON_LIGHT_SKIN_TONE(0x1F9DC),
+    MERPERSON_MEDIUM_LIGHT_SKIN_TONE(0x1F9DC),
+    MERPERSON_MEDIUM_SKIN_TONE(0x1F9DC),
+    MERPERSON_MEDIUM_DARK_SKIN_TONE(0x1F9DC),
+    MERPERSON_DARK_SKIN_TONE(0x1F9DC),
+    MERMAN(0x1F9DC),
+    MERMAN_LIGHT_SKIN_TONE(0x1F9DC),
+    MERMAN_MEDIUM_LIGHT_SKIN_TONE(0x1F9DC),
+    MERMAN_MEDIUM_SKIN_TONE(0x1F9DC),
+    MERMAN_MEDIUM_DARK_SKIN_TONE(0x1F9DC),
+    MERMAN_DARK_SKIN_TONE(0x1F9DC),
+    MERMAID(0x1F9DC),
+    MERMAID_LIGHT_SKIN_TONE(0x1F9DC),
+    MERMAID_MEDIUM_LIGHT_SKIN_TONE(0x1F9DC),
+    MERMAID_MEDIUM_SKIN_TONE(0x1F9DC),
+    MERMAID_MEDIUM_DARK_SKIN_TONE(0x1F9DC),
+    MERMAID_DARK_SKIN_TONE(0x1F9DC),
+    ELF(0x1F9DD),
+    ELF_LIGHT_SKIN_TONE(0x1F9DD),
+    ELF_MEDIUM_LIGHT_SKIN_TONE(0x1F9DD),
+    ELF_MEDIUM_SKIN_TONE(0x1F9DD),
+    ELF_MEDIUM_DARK_SKIN_TONE(0x1F9DD),
+    ELF_DARK_SKIN_TONE(0x1F9DD),
+    MAN_ELF(0x1F9DD),
+    MAN_ELF_LIGHT_SKIN_TONE(0x1F9DD),
+    MAN_ELF_MEDIUM_LIGHT_SKIN_TONE(0x1F9DD),
+    MAN_ELF_MEDIUM_SKIN_TONE(0x1F9DD),
+    MAN_ELF_MEDIUM_DARK_SKIN_TONE(0x1F9DD),
+    MAN_ELF_DARK_SKIN_TONE(0x1F9DD),
+    WOMAN_ELF(0x1F9DD),
+    WOMAN_ELF_LIGHT_SKIN_TONE(0x1F9DD),
+    WOMAN_ELF_MEDIUM_LIGHT_SKIN_TONE(0x1F9DD),
+    WOMAN_ELF_MEDIUM_SKIN_TONE(0x1F9DD),
+    WOMAN_ELF_MEDIUM_DARK_SKIN_TONE(0x1F9DD),
+    WOMAN_ELF_DARK_SKIN_TONE(0x1F9DD),
+    GENIE(0x1F9DE),
+    MAN_GENIE(0x1F9DE),
+    WOMAN_GENIE(0x1F9DE),
+    ZOMBIE(0x1F9DF),
+    MAN_ZOMBIE(0x1F9DF),
+    WOMAN_ZOMBIE(0x1F9DF),
+    TROLL(0x1F9CC),
+    PERSON_GETTING_MASSAGE(0x1F486),
+    PERSON_GETTING_MASSAGE_LIGHT_SKIN_TONE(0x1F486),
+    PERSON_GETTING_MASSAGE_MEDIUM_LIGHT_SKIN_TONE(0x1F486),
+    PERSON_GETTING_MASSAGE_MEDIUM_SKIN_TONE(0x1F486),
+    PERSON_GETTING_MASSAGE_MEDIUM_DARK_SKIN_TONE(0x1F486),
+    PERSON_GETTING_MASSAGE_DARK_SKIN_TONE(0x1F486),
+    MAN_GETTING_MASSAGE(0x1F486),
+    MAN_GETTING_MASSAGE_LIGHT_SKIN_TONE(0x1F486),
+    MAN_GETTING_MASSAGE_MEDIUM_LIGHT_SKIN_TONE(0x1F486),
+    MAN_GETTING_MASSAGE_MEDIUM_SKIN_TONE(0x1F486),
+    MAN_GETTING_MASSAGE_MEDIUM_DARK_SKIN_TONE(0x1F486),
+    MAN_GETTING_MASSAGE_DARK_SKIN_TONE(0x1F486),
+    WOMAN_GETTING_MASSAGE(0x1F486),
+    WOMAN_GETTING_MASSAGE_LIGHT_SKIN_TONE(0x1F486),
+    WOMAN_GETTING_MASSAGE_MEDIUM_LIGHT_SKIN_TONE(0x1F486),
+    WOMAN_GETTING_MASSAGE_MEDIUM_SKIN_TONE(0x1F486),
+    WOMAN_GETTING_MASSAGE_MEDIUM_DARK_SKIN_TONE(0x1F486),
+    WOMAN_GETTING_MASSAGE_DARK_SKIN_TONE(0x1F486),
+    PERSON_GETTING_HAIRCUT(0x1F487),
+    PERSON_GETTING_HAIRCUT_LIGHT_SKIN_TONE(0x1F487),
+    PERSON_GETTING_HAIRCUT_MEDIUM_LIGHT_SKIN_TONE(0x1F487),
+    PERSON_GETTING_HAIRCUT_MEDIUM_SKIN_TONE(0x1F487),
+    PERSON_GETTING_HAIRCUT_MEDIUM_DARK_SKIN_TONE(0x1F487),
+    PERSON_GETTING_HAIRCUT_DARK_SKIN_TONE(0x1F487),
+    MAN_GETTING_HAIRCUT(0x1F487),
+    MAN_GETTING_HAIRCUT_LIGHT_SKIN_TONE(0x1F487),
+    MAN_GETTING_HAIRCUT_MEDIUM_LIGHT_SKIN_TONE(0x1F487),
+    MAN_GETTING_HAIRCUT_MEDIUM_SKIN_TONE(0x1F487),
+    MAN_GETTING_HAIRCUT_MEDIUM_DARK_SKIN_TONE(0x1F487),
+    MAN_GETTING_HAIRCUT_DARK_SKIN_TONE(0x1F487),
+    WOMAN_GETTING_HAIRCUT(0x1F487),
+    WOMAN_GETTING_HAIRCUT_LIGHT_SKIN_TONE(0x1F487),
+    WOMAN_GETTING_HAIRCUT_MEDIUM_LIGHT_SKIN_TONE(0x1F487),
+    WOMAN_GETTING_HAIRCUT_MEDIUM_SKIN_TONE(0x1F487),
+    WOMAN_GETTING_HAIRCUT_MEDIUM_DARK_SKIN_TONE(0x1F487),
+    WOMAN_GETTING_HAIRCUT_DARK_SKIN_TONE(0x1F487),
+    PERSON_WALKING(0x1F6B6),
+    PERSON_WALKING_LIGHT_SKIN_TONE(0x1F6B6),
+    PERSON_WALKING_MEDIUM_LIGHT_SKIN_TONE(0x1F6B6),
+    PERSON_WALKING_MEDIUM_SKIN_TONE(0x1F6B6),
+    PERSON_WALKING_MEDIUM_DARK_SKIN_TONE(0x1F6B6),
+    PERSON_WALKING_DARK_SKIN_TONE(0x1F6B6),
+    MAN_WALKING(0x1F6B6),
+    MAN_WALKING_LIGHT_SKIN_TONE(0x1F6B6),
+    MAN_WALKING_MEDIUM_LIGHT_SKIN_TONE(0x1F6B6),
+    MAN_WALKING_MEDIUM_SKIN_TONE(0x1F6B6),
+    MAN_WALKING_MEDIUM_DARK_SKIN_TONE(0x1F6B6),
+    MAN_WALKING_DARK_SKIN_TONE(0x1F6B6),
+    WOMAN_WALKING(0x1F6B6),
+    WOMAN_WALKING_LIGHT_SKIN_TONE(0x1F6B6),
+    WOMAN_WALKING_MEDIUM_LIGHT_SKIN_TONE(0x1F6B6),
+    WOMAN_WALKING_MEDIUM_SKIN_TONE(0x1F6B6),
+    WOMAN_WALKING_MEDIUM_DARK_SKIN_TONE(0x1F6B6),
+    WOMAN_WALKING_DARK_SKIN_TONE(0x1F6B6),
+    PERSON_STANDING(0x1F9CD),
+    PERSON_STANDING_LIGHT_SKIN_TONE(0x1F9CD),
+    PERSON_STANDING_MEDIUM_LIGHT_SKIN_TONE(0x1F9CD),
+    PERSON_STANDING_MEDIUM_SKIN_TONE(0x1F9CD),
+    PERSON_STANDING_MEDIUM_DARK_SKIN_TONE(0x1F9CD),
+    PERSON_STANDING_DARK_SKIN_TONE(0x1F9CD),
+    MAN_STANDING(0x1F9CD),
+    MAN_STANDING_LIGHT_SKIN_TONE(0x1F9CD),
+    MAN_STANDING_MEDIUM_LIGHT_SKIN_TONE(0x1F9CD),
+    MAN_STANDING_MEDIUM_SKIN_TONE(0x1F9CD),
+    MAN_STANDING_MEDIUM_DARK_SKIN_TONE(0x1F9CD),
+    MAN_STANDING_DARK_SKIN_TONE(0x1F9CD),
+    WOMAN_STANDING(0x1F9CD),
+    WOMAN_STANDING_LIGHT_SKIN_TONE(0x1F9CD),
+    WOMAN_STANDING_MEDIUM_LIGHT_SKIN_TONE(0x1F9CD),
+    WOMAN_STANDING_MEDIUM_SKIN_TONE(0x1F9CD),
+    WOMAN_STANDING_MEDIUM_DARK_SKIN_TONE(0x1F9CD),
+    WOMAN_STANDING_DARK_SKIN_TONE(0x1F9CD),
+    PERSON_KNEELING(0x1F9CE),
+    PERSON_KNEELING_LIGHT_SKIN_TONE(0x1F9CE),
+    PERSON_KNEELING_MEDIUM_LIGHT_SKIN_TONE(0x1F9CE),
+    PERSON_KNEELING_MEDIUM_SKIN_TONE(0x1F9CE),
+    PERSON_KNEELING_MEDIUM_DARK_SKIN_TONE(0x1F9CE),
+    PERSON_KNEELING_DARK_SKIN_TONE(0x1F9CE),
+    MAN_KNEELING(0x1F9CE),
+    MAN_KNEELING_LIGHT_SKIN_TONE(0x1F9CE),
+    MAN_KNEELING_MEDIUM_LIGHT_SKIN_TONE(0x1F9CE),
+    MAN_KNEELING_MEDIUM_SKIN_TONE(0x1F9CE),
+    MAN_KNEELING_MEDIUM_DARK_SKIN_TONE(0x1F9CE),
+    MAN_KNEELING_DARK_SKIN_TONE(0x1F9CE),
+    WOMAN_KNEELING(0x1F9CE),
+    WOMAN_KNEELING_LIGHT_SKIN_TONE(0x1F9CE),
+    WOMAN_KNEELING_MEDIUM_LIGHT_SKIN_TONE(0x1F9CE),
+    WOMAN_KNEELING_MEDIUM_SKIN_TONE(0x1F9CE),
+    WOMAN_KNEELING_MEDIUM_DARK_SKIN_TONE(0x1F9CE),
+    WOMAN_KNEELING_DARK_SKIN_TONE(0x1F9CE),
+    PERSON_WITH_WHITE_CANE(0x1F9D1),
+    PERSON_WITH_WHITE_CANE_LIGHT_SKIN_TONE(0x1F9D1),
+    PERSON_WITH_WHITE_CANE_MEDIUM_LIGHT_SKIN_TONE(0x1F9D1),
+    PERSON_WITH_WHITE_CANE_MEDIUM_SKIN_TONE(0x1F9D1),
+    PERSON_WITH_WHITE_CANE_MEDIUM_DARK_SKIN_TONE(0x1F9D1),
+    PERSON_WITH_WHITE_CANE_DARK_SKIN_TONE(0x1F9D1),
+    MAN_WITH_WHITE_CANE(0x1F468),
+    MAN_WITH_WHITE_CANE_LIGHT_SKIN_TONE(0x1F468),
+    MAN_WITH_WHITE_CANE_MEDIUM_LIGHT_SKIN_TONE(0x1F468),
+    MAN_WITH_WHITE_CANE_MEDIUM_SKIN_TONE(0x1F468),
+    MAN_WITH_WHITE_CANE_MEDIUM_DARK_SKIN_TONE(0x1F468),
+    MAN_WITH_WHITE_CANE_DARK_SKIN_TONE(0x1F468),
+    WOMAN_WITH_WHITE_CANE(0x1F469),
+    WOMAN_WITH_WHITE_CANE_LIGHT_SKIN_TONE(0x1F469),
+    WOMAN_WITH_WHITE_CANE_MEDIUM_LIGHT_SKIN_TONE(0x1F469),
+    WOMAN_WITH_WHITE_CANE_MEDIUM_SKIN_TONE(0x1F469),
+    WOMAN_WITH_WHITE_CANE_MEDIUM_DARK_SKIN_TONE(0x1F469),
+    WOMAN_WITH_WHITE_CANE_DARK_SKIN_TONE(0x1F469),
+    PERSON_IN_MOTORIZED_WHEELCHAIR(0x1F9D1),
+    PERSON_IN_MOTORIZED_WHEELCHAIR_LIGHT_SKIN_TONE(0x1F9D1),
+    PERSON_IN_MOTORIZED_WHEELCHAIR_MEDIUM_LIGHT_SKIN_TONE(0x1F9D1),
+    PERSON_IN_MOTORIZED_WHEELCHAIR_MEDIUM_SKIN_TONE(0x1F9D1),
+    PERSON_IN_MOTORIZED_WHEELCHAIR_MEDIUM_DARK_SKIN_TONE(0x1F9D1),
+    PERSON_IN_MOTORIZED_WHEELCHAIR_DARK_SKIN_TONE(0x1F9D1),
+    MAN_IN_MOTORIZED_WHEELCHAIR(0x1F468),
+    MAN_IN_MOTORIZED_WHEELCHAIR_LIGHT_SKIN_TONE(0x1F468),
+    MAN_IN_MOTORIZED_WHEELCHAIR_MEDIUM_LIGHT_SKIN_TONE(0x1F468),
+    MAN_IN_MOTORIZED_WHEELCHAIR_MEDIUM_SKIN_TONE(0x1F468),
+    MAN_IN_MOTORIZED_WHEELCHAIR_MEDIUM_DARK_SKIN_TONE(0x1F468),
+    MAN_IN_MOTORIZED_WHEELCHAIR_DARK_SKIN_TONE(0x1F468),
+    WOMAN_IN_MOTORIZED_WHEELCHAIR(0x1F469),
+    WOMAN_IN_MOTORIZED_WHEELCHAIR_LIGHT_SKIN_TONE(0x1F469),
+    WOMAN_IN_MOTORIZED_WHEELCHAIR_MEDIUM_LIGHT_SKIN_TONE(0x1F469),
+    WOMAN_IN_MOTORIZED_WHEELCHAIR_MEDIUM_SKIN_TONE(0x1F469),
+    WOMAN_IN_MOTORIZED_WHEELCHAIR_MEDIUM_DARK_SKIN_TONE(0x1F469),
+    WOMAN_IN_MOTORIZED_WHEELCHAIR_DARK_SKIN_TONE(0x1F469),
+    PERSON_IN_MANUAL_WHEELCHAIR(0x1F9D1),
+    PERSON_IN_MANUAL_WHEELCHAIR_LIGHT_SKIN_TONE(0x1F9D1),
+    PERSON_IN_MANUAL_WHEELCHAIR_MEDIUM_LIGHT_SKIN_TONE(0x1F9D1),
+    PERSON_IN_MANUAL_WHEELCHAIR_MEDIUM_SKIN_TONE(0x1F9D1),
+    PERSON_IN_MANUAL_WHEELCHAIR_MEDIUM_DARK_SKIN_TONE(0x1F9D1),
+    PERSON_IN_MANUAL_WHEELCHAIR_DARK_SKIN_TONE(0x1F9D1),
+    MAN_IN_MANUAL_WHEELCHAIR(0x1F468),
+    MAN_IN_MANUAL_WHEELCHAIR_LIGHT_SKIN_TONE(0x1F468),
+    MAN_IN_MANUAL_WHEELCHAIR_MEDIUM_LIGHT_SKIN_TONE(0x1F468),
+    MAN_IN_MANUAL_WHEELCHAIR_MEDIUM_SKIN_TONE(0x1F468),
+    MAN_IN_MANUAL_WHEELCHAIR_MEDIUM_DARK_SKIN_TONE(0x1F468),
+    MAN_IN_MANUAL_WHEELCHAIR_DARK_SKIN_TONE(0x1F468),
+    WOMAN_IN_MANUAL_WHEELCHAIR(0x1F469),
+    WOMAN_IN_MANUAL_WHEELCHAIR_LIGHT_SKIN_TONE(0x1F469),
+    WOMAN_IN_MANUAL_WHEELCHAIR_MEDIUM_LIGHT_SKIN_TONE(0x1F469),
+    WOMAN_IN_MANUAL_WHEELCHAIR_MEDIUM_SKIN_TONE(0x1F469),
+    WOMAN_IN_MANUAL_WHEELCHAIR_MEDIUM_DARK_SKIN_TONE(0x1F469),
+    WOMAN_IN_MANUAL_WHEELCHAIR_DARK_SKIN_TONE(0x1F469),
+    PERSON_RUNNING(0x1F3C3),
+    PERSON_RUNNING_LIGHT_SKIN_TONE(0x1F3C3),
+    PERSON_RUNNING_MEDIUM_LIGHT_SKIN_TONE(0x1F3C3),
+    PERSON_RUNNING_MEDIUM_SKIN_TONE(0x1F3C3),
+    PERSON_RUNNING_MEDIUM_DARK_SKIN_TONE(0x1F3C3),
+    PERSON_RUNNING_DARK_SKIN_TONE(0x1F3C3),
+    MAN_RUNNING(0x1F3C3),
+    MAN_RUNNING_LIGHT_SKIN_TONE(0x1F3C3),
+    MAN_RUNNING_MEDIUM_LIGHT_SKIN_TONE(0x1F3C3),
+    MAN_RUNNING_MEDIUM_SKIN_TONE(0x1F3C3),
+    MAN_RUNNING_MEDIUM_DARK_SKIN_TONE(0x1F3C3),
+    MAN_RUNNING_DARK_SKIN_TONE(0x1F3C3),
+    WOMAN_RUNNING(0x1F3C3),
+    WOMAN_RUNNING_LIGHT_SKIN_TONE(0x1F3C3),
+    WOMAN_RUNNING_MEDIUM_LIGHT_SKIN_TONE(0x1F3C3),
+    WOMAN_RUNNING_MEDIUM_SKIN_TONE(0x1F3C3),
+    WOMAN_RUNNING_MEDIUM_DARK_SKIN_TONE(0x1F3C3),
+    WOMAN_RUNNING_DARK_SKIN_TONE(0x1F3C3),
+    WOMAN_DANCING(0x1F483),
+    WOMAN_DANCING_LIGHT_SKIN_TONE(0x1F483),
+    WOMAN_DANCING_MEDIUM_LIGHT_SKIN_TONE(0x1F483),
+    WOMAN_DANCING_MEDIUM_SKIN_TONE(0x1F483),
+    WOMAN_DANCING_MEDIUM_DARK_SKIN_TONE(0x1F483),
+    WOMAN_DANCING_DARK_SKIN_TONE(0x1F483),
+    MAN_DANCING(0x1F57A),
+    MAN_DANCING_LIGHT_SKIN_TONE(0x1F57A),
+    MAN_DANCING_MEDIUM_LIGHT_SKIN_TONE(0x1F57A),
+    MAN_DANCING_MEDIUM_SKIN_TONE(0x1F57A),
+    MAN_DANCING_MEDIUM_DARK_SKIN_TONE(0x1F57A),
+    MAN_DANCING_DARK_SKIN_TONE(0x1F57A),
+    PERSON_IN_SUIT_LEVITATING(0x1F574),
+    PERSON_IN_SUIT_LEVITATING_LIGHT_SKIN_TONE(0x1F574),
+    PERSON_IN_SUIT_LEVITATING_MEDIUM_LIGHT_SKIN_TONE(0x1F574),
+    PERSON_IN_SUIT_LEVITATING_MEDIUM_SKIN_TONE(0x1F574),
+    PERSON_IN_SUIT_LEVITATING_MEDIUM_DARK_SKIN_TONE(0x1F574),
+    PERSON_IN_SUIT_LEVITATING_DARK_SKIN_TONE(0x1F574),
+    PEOPLE_WITH_BUNNY_EARS(0x1F46F),
+    MEN_WITH_BUNNY_EARS(0x1F46F),
+    WOMEN_WITH_BUNNY_EARS(0x1F46F),
+    PERSON_IN_STEAMY_ROOM(0x1F9D6),
+    PERSON_IN_STEAMY_ROOM_LIGHT_SKIN_TONE(0x1F9D6),
+    PERSON_IN_STEAMY_ROOM_MEDIUM_LIGHT_SKIN_TONE(0x1F9D6),
+    PERSON_IN_STEAMY_ROOM_MEDIUM_SKIN_TONE(0x1F9D6),
+    PERSON_IN_STEAMY_ROOM_MEDIUM_DARK_SKIN_TONE(0x1F9D6),
+    PERSON_IN_STEAMY_ROOM_DARK_SKIN_TONE(0x1F9D6),
+    MAN_IN_STEAMY_ROOM(0x1F9D6),
+    MAN_IN_STEAMY_ROOM_LIGHT_SKIN_TONE(0x1F9D6),
+    MAN_IN_STEAMY_ROOM_MEDIUM_LIGHT_SKIN_TONE(0x1F9D6),
+    MAN_IN_STEAMY_ROOM_MEDIUM_SKIN_TONE(0x1F9D6),
+    MAN_IN_STEAMY_ROOM_MEDIUM_DARK_SKIN_TONE(0x1F9D6),
+    MAN_IN_STEAMY_ROOM_DARK_SKIN_TONE(0x1F9D6),
+    WOMAN_IN_STEAMY_ROOM(0x1F9D6),
+    WOMAN_IN_STEAMY_ROOM_LIGHT_SKIN_TONE(0x1F9D6),
+    WOMAN_IN_STEAMY_ROOM_MEDIUM_LIGHT_SKIN_TONE(0x1F9D6),
+    WOMAN_IN_STEAMY_ROOM_MEDIUM_SKIN_TONE(0x1F9D6),
+    WOMAN_IN_STEAMY_ROOM_MEDIUM_DARK_SKIN_TONE(0x1F9D6),
+    WOMAN_IN_STEAMY_ROOM_DARK_SKIN_TONE(0x1F9D6),
+    PERSON_CLIMBING(0x1F9D7),
+    PERSON_CLIMBING_LIGHT_SKIN_TONE(0x1F9D7),
+    PERSON_CLIMBING_MEDIUM_LIGHT_SKIN_TONE(0x1F9D7),
+    PERSON_CLIMBING_MEDIUM_SKIN_TONE(0x1F9D7),
+    PERSON_CLIMBING_MEDIUM_DARK_SKIN_TONE(0x1F9D7),
+    PERSON_CLIMBING_DARK_SKIN_TONE(0x1F9D7),
+    MAN_CLIMBING(0x1F9D7),
+    MAN_CLIMBING_LIGHT_SKIN_TONE(0x1F9D7),
+    MAN_CLIMBING_MEDIUM_LIGHT_SKIN_TONE(0x1F9D7),
+    MAN_CLIMBING_MEDIUM_SKIN_TONE(0x1F9D7),
+    MAN_CLIMBING_MEDIUM_DARK_SKIN_TONE(0x1F9D7),
+    MAN_CLIMBING_DARK_SKIN_TONE(0x1F9D7),
+    WOMAN_CLIMBING(0x1F9D7),
+    WOMAN_CLIMBING_LIGHT_SKIN_TONE(0x1F9D7),
+    WOMAN_CLIMBING_MEDIUM_LIGHT_SKIN_TONE(0x1F9D7),
+    WOMAN_CLIMBING_MEDIUM_SKIN_TONE(0x1F9D7),
+    WOMAN_CLIMBING_MEDIUM_DARK_SKIN_TONE(0x1F9D7),
+    WOMAN_CLIMBING_DARK_SKIN_TONE(0x1F9D7),
+    PERSON_FENCING(0x1F93A),
+    HORSE_RACING(0x1F3C7),
+    HORSE_RACING_LIGHT_SKIN_TONE(0x1F3C7),
+    HORSE_RACING_MEDIUM_LIGHT_SKIN_TONE(0x1F3C7),
+    HORSE_RACING_MEDIUM_SKIN_TONE(0x1F3C7),
+    HORSE_RACING_MEDIUM_DARK_SKIN_TONE(0x1F3C7),
+    HORSE_RACING_DARK_SKIN_TONE(0x1F3C7),
+    SKIER(0x26F7),
+    SNOWBOARDER(0x1F3C2),
+    SNOWBOARDER_LIGHT_SKIN_TONE(0x1F3C2),
+    SNOWBOARDER_MEDIUM_LIGHT_SKIN_TONE(0x1F3C2),
+    SNOWBOARDER_MEDIUM_SKIN_TONE(0x1F3C2),
+    SNOWBOARDER_MEDIUM_DARK_SKIN_TONE(0x1F3C2),
+    SNOWBOARDER_DARK_SKIN_TONE(0x1F3C2),
+    PERSON_GOLFING(0x1F3CC),
+    PERSON_GOLFING_LIGHT_SKIN_TONE(0x1F3CC),
+    PERSON_GOLFING_MEDIUM_LIGHT_SKIN_TONE(0x1F3CC),
+    PERSON_GOLFING_MEDIUM_SKIN_TONE(0x1F3CC),
+    PERSON_GOLFING_MEDIUM_DARK_SKIN_TONE(0x1F3CC),
+    PERSON_GOLFING_DARK_SKIN_TONE(0x1F3CC),
+    MAN_GOLFING(0x1F3CC),
+    MAN_GOLFING_LIGHT_SKIN_TONE(0x1F3CC),
+    MAN_GOLFING_MEDIUM_LIGHT_SKIN_TONE(0x1F3CC),
+    MAN_GOLFING_MEDIUM_SKIN_TONE(0x1F3CC),
+    MAN_GOLFING_MEDIUM_DARK_SKIN_TONE(0x1F3CC),
+    MAN_GOLFING_DARK_SKIN_TONE(0x1F3CC),
+    WOMAN_GOLFING(0x1F3CC),
+    WOMAN_GOLFING_LIGHT_SKIN_TONE(0x1F3CC),
+    WOMAN_GOLFING_MEDIUM_LIGHT_SKIN_TONE(0x1F3CC),
+    WOMAN_GOLFING_MEDIUM_SKIN_TONE(0x1F3CC),
+    WOMAN_GOLFING_MEDIUM_DARK_SKIN_TONE(0x1F3CC),
+    WOMAN_GOLFING_DARK_SKIN_TONE(0x1F3CC),
+    PERSON_SURFING(0x1F3C4),
+    PERSON_SURFING_LIGHT_SKIN_TONE(0x1F3C4),
+    PERSON_SURFING_MEDIUM_LIGHT_SKIN_TONE(0x1F3C4),
+    PERSON_SURFING_MEDIUM_SKIN_TONE(0x1F3C4),
+    PERSON_SURFING_MEDIUM_DARK_SKIN_TONE(0x1F3C4),
+    PERSON_SURFING_DARK_SKIN_TONE(0x1F3C4),
+    MAN_SURFING(0x1F3C4),
+    MAN_SURFING_LIGHT_SKIN_TONE(0x1F3C4),
+    MAN_SURFING_MEDIUM_LIGHT_SKIN_TONE(0x1F3C4),
+    MAN_SURFING_MEDIUM_SKIN_TONE(0x1F3C4),
+    MAN_SURFING_MEDIUM_DARK_SKIN_TONE(0x1F3C4),
+    MAN_SURFING_DARK_SKIN_TONE(0x1F3C4),
+    WOMAN_SURFING(0x1F3C4),
+    WOMAN_SURFING_LIGHT_SKIN_TONE(0x1F3C4),
+    WOMAN_SURFING_MEDIUM_LIGHT_SKIN_TONE(0x1F3C4),
+    WOMAN_SURFING_MEDIUM_SKIN_TONE(0x1F3C4),
+    WOMAN_SURFING_MEDIUM_DARK_SKIN_TONE(0x1F3C4),
+    WOMAN_SURFING_DARK_SKIN_TONE(0x1F3C4),
+    PERSON_ROWING_BOAT(0x1F6A3),
+    PERSON_ROWING_BOAT_LIGHT_SKIN_TONE(0x1F6A3),
+    PERSON_ROWING_BOAT_MEDIUM_LIGHT_SKIN_TONE(0x1F6A3),
+    PERSON_ROWING_BOAT_MEDIUM_SKIN_TONE(0x1F6A3),
+    PERSON_ROWING_BOAT_MEDIUM_DARK_SKIN_TONE(0x1F6A3),
+    PERSON_ROWING_BOAT_DARK_SKIN_TONE(0x1F6A3),
+    MAN_ROWING_BOAT(0x1F6A3),
+    MAN_ROWING_BOAT_LIGHT_SKIN_TONE(0x1F6A3),
+    MAN_ROWING_BOAT_MEDIUM_LIGHT_SKIN_TONE(0x1F6A3),
+    MAN_ROWING_BOAT_MEDIUM_SKIN_TONE(0x1F6A3),
+    MAN_ROWING_BOAT_MEDIUM_DARK_SKIN_TONE(0x1F6A3),
+    MAN_ROWING_BOAT_DARK_SKIN_TONE(0x1F6A3),
+    WOMAN_ROWING_BOAT(0x1F6A3),
+    WOMAN_ROWING_BOAT_LIGHT_SKIN_TONE(0x1F6A3),
+    WOMAN_ROWING_BOAT_MEDIUM_LIGHT_SKIN_TONE(0x1F6A3),
+    WOMAN_ROWING_BOAT_MEDIUM_SKIN_TONE(0x1F6A3),
+    WOMAN_ROWING_BOAT_MEDIUM_DARK_SKIN_TONE(0x1F6A3),
+    WOMAN_ROWING_BOAT_DARK_SKIN_TONE(0x1F6A3),
+    PERSON_SWIMMING(0x1F3CA),
+    PERSON_SWIMMING_LIGHT_SKIN_TONE(0x1F3CA),
+    PERSON_SWIMMING_MEDIUM_LIGHT_SKIN_TONE(0x1F3CA),
+    PERSON_SWIMMING_MEDIUM_SKIN_TONE(0x1F3CA),
+    PERSON_SWIMMING_MEDIUM_DARK_SKIN_TONE(0x1F3CA),
+    PERSON_SWIMMING_DARK_SKIN_TONE(0x1F3CA),
+    MAN_SWIMMING(0x1F3CA),
+    MAN_SWIMMING_LIGHT_SKIN_TONE(0x1F3CA),
+    MAN_SWIMMING_MEDIUM_LIGHT_SKIN_TONE(0x1F3CA),
+    MAN_SWIMMING_MEDIUM_SKIN_TONE(0x1F3CA),
+    MAN_SWIMMING_MEDIUM_DARK_SKIN_TONE(0x1F3CA),
+    MAN_SWIMMING_DARK_SKIN_TONE(0x1F3CA),
+    WOMAN_SWIMMING(0x1F3CA),
+    WOMAN_SWIMMING_LIGHT_SKIN_TONE(0x1F3CA),
+    WOMAN_SWIMMING_MEDIUM_LIGHT_SKIN_TONE(0x1F3CA),
+    WOMAN_SWIMMING_MEDIUM_SKIN_TONE(0x1F3CA),
+    WOMAN_SWIMMING_MEDIUM_DARK_SKIN_TONE(0x1F3CA),
+    WOMAN_SWIMMING_DARK_SKIN_TONE(0x1F3CA),
+    PERSON_BOUNCING_BALL(0x26F9),
+    PERSON_BOUNCING_BALL_LIGHT_SKIN_TONE(0x26F9),
+    PERSON_BOUNCING_BALL_MEDIUM_LIGHT_SKIN_TONE(0x26F9),
+    PERSON_BOUNCING_BALL_MEDIUM_SKIN_TONE(0x26F9),
+    PERSON_BOUNCING_BALL_MEDIUM_DARK_SKIN_TONE(0x26F9),
+    PERSON_BOUNCING_BALL_DARK_SKIN_TONE(0x26F9),
+    MAN_BOUNCING_BALL(0x26F9),
+    MAN_BOUNCING_BALL_LIGHT_SKIN_TONE(0x26F9),
+    MAN_BOUNCING_BALL_MEDIUM_LIGHT_SKIN_TONE(0x26F9),
+    MAN_BOUNCING_BALL_MEDIUM_SKIN_TONE(0x26F9),
+    MAN_BOUNCING_BALL_MEDIUM_DARK_SKIN_TONE(0x26F9),
+    MAN_BOUNCING_BALL_DARK_SKIN_TONE(0x26F9),
+    WOMAN_BOUNCING_BALL(0x26F9),
+    WOMAN_BOUNCING_BALL_LIGHT_SKIN_TONE(0x26F9),
+    WOMAN_BOUNCING_BALL_MEDIUM_LIGHT_SKIN_TONE(0x26F9),
+    WOMAN_BOUNCING_BALL_MEDIUM_SKIN_TONE(0x26F9),
+    WOMAN_BOUNCING_BALL_MEDIUM_DARK_SKIN_TONE(0x26F9),
+    WOMAN_BOUNCING_BALL_DARK_SKIN_TONE(0x26F9),
+    PERSON_LIFTING_WEIGHTS(0x1F3CB),
+    PERSON_LIFTING_WEIGHTS_LIGHT_SKIN_TONE(0x1F3CB),
+    PERSON_LIFTING_WEIGHTS_MEDIUM_LIGHT_SKIN_TONE(0x1F3CB),
+    PERSON_LIFTING_WEIGHTS_MEDIUM_SKIN_TONE(0x1F3CB),
+    PERSON_LIFTING_WEIGHTS_MEDIUM_DARK_SKIN_TONE(0x1F3CB),
+    PERSON_LIFTING_WEIGHTS_DARK_SKIN_TONE(0x1F3CB),
+    MAN_LIFTING_WEIGHTS(0x1F3CB),
+    MAN_LIFTING_WEIGHTS_LIGHT_SKIN_TONE(0x1F3CB),
+    MAN_LIFTING_WEIGHTS_MEDIUM_LIGHT_SKIN_TONE(0x1F3CB),
+    MAN_LIFTING_WEIGHTS_MEDIUM_SKIN_TONE(0x1F3CB),
+    MAN_LIFTING_WEIGHTS_MEDIUM_DARK_SKIN_TONE(0x1F3CB),
+    MAN_LIFTING_WEIGHTS_DARK_SKIN_TONE(0x1F3CB),
+    WOMAN_LIFTING_WEIGHTS(0x1F3CB),
+    WOMAN_LIFTING_WEIGHTS_LIGHT_SKIN_TONE(0x1F3CB),
+    WOMAN_LIFTING_WEIGHTS_MEDIUM_LIGHT_SKIN_TONE(0x1F3CB),
+    WOMAN_LIFTING_WEIGHTS_MEDIUM_SKIN_TONE(0x1F3CB),
+    WOMAN_LIFTING_WEIGHTS_MEDIUM_DARK_SKIN_TONE(0x1F3CB),
+    WOMAN_LIFTING_WEIGHTS_DARK_SKIN_TONE(0x1F3CB),
+    PERSON_BIKING(0x1F6B4),
+    PERSON_BIKING_LIGHT_SKIN_TONE(0x1F6B4),
+    PERSON_BIKING_MEDIUM_LIGHT_SKIN_TONE(0x1F6B4),
+    PERSON_BIKING_MEDIUM_SKIN_TONE(0x1F6B4),
+    PERSON_BIKING_MEDIUM_DARK_SKIN_TONE(0x1F6B4),
+    PERSON_BIKING_DARK_SKIN_TONE(0x1F6B4),
+    MAN_BIKING(0x1F6B4),
+    MAN_BIKING_LIGHT_SKIN_TONE(0x1F6B4),
+    MAN_BIKING_MEDIUM_LIGHT_SKIN_TONE(0x1F6B4),
+    MAN_BIKING_MEDIUM_SKIN_TONE(0x1F6B4),
+    MAN_BIKING_MEDIUM_DARK_SKIN_TONE(0x1F6B4),
+    MAN_BIKING_DARK_SKIN_TONE(0x1F6B4),
+    WOMAN_BIKING(0x1F6B4),
+    WOMAN_BIKING_LIGHT_SKIN_TONE(0x1F6B4),
+    WOMAN_BIKING_MEDIUM_LIGHT_SKIN_TONE(0x1F6B4),
+    WOMAN_BIKING_MEDIUM_SKIN_TONE(0x1F6B4),
+    WOMAN_BIKING_MEDIUM_DARK_SKIN_TONE(0x1F6B4),
+    WOMAN_BIKING_DARK_SKIN_TONE(0x1F6B4),
+    PERSON_MOUNTAIN_BIKING(0x1F6B5),
+    PERSON_MOUNTAIN_BIKING_LIGHT_SKIN_TONE(0x1F6B5),
+    PERSON_MOUNTAIN_BIKING_MEDIUM_LIGHT_SKIN_TONE(0x1F6B5),
+    PERSON_MOUNTAIN_BIKING_MEDIUM_SKIN_TONE(0x1F6B5),
+    PERSON_MOUNTAIN_BIKING_MEDIUM_DARK_SKIN_TONE(0x1F6B5),
+    PERSON_MOUNTAIN_BIKING_DARK_SKIN_TONE(0x1F6B5),
+    MAN_MOUNTAIN_BIKING(0x1F6B5),
+    MAN_MOUNTAIN_BIKING_LIGHT_SKIN_TONE(0x1F6B5),
+    MAN_MOUNTAIN_BIKING_MEDIUM_LIGHT_SKIN_TONE(0x1F6B5),
+    MAN_MOUNTAIN_BIKING_MEDIUM_SKIN_TONE(0x1F6B5),
+    MAN_MOUNTAIN_BIKING_MEDIUM_DARK_SKIN_TONE(0x1F6B5),
+    MAN_MOUNTAIN_BIKING_DARK_SKIN_TONE(0x1F6B5),
+    WOMAN_MOUNTAIN_BIKING(0x1F6B5),
+    WOMAN_MOUNTAIN_BIKING_LIGHT_SKIN_TONE(0x1F6B5),
+    WOMAN_MOUNTAIN_BIKING_MEDIUM_LIGHT_SKIN_TONE(0x1F6B5),
+    WOMAN_MOUNTAIN_BIKING_MEDIUM_SKIN_TONE(0x1F6B5),
+    WOMAN_MOUNTAIN_BIKING_MEDIUM_DARK_SKIN_TONE(0x1F6B5),
+    WOMAN_MOUNTAIN_BIKING_DARK_SKIN_TONE(0x1F6B5),
+    PERSON_CARTWHEELING(0x1F938),
+    PERSON_CARTWHEELING_LIGHT_SKIN_TONE(0x1F938),
+    PERSON_CARTWHEELING_MEDIUM_LIGHT_SKIN_TONE(0x1F938),
+    PERSON_CARTWHEELING_MEDIUM_SKIN_TONE(0x1F938),
+    PERSON_CARTWHEELING_MEDIUM_DARK_SKIN_TONE(0x1F938),
+    PERSON_CARTWHEELING_DARK_SKIN_TONE(0x1F938),
+    MAN_CARTWHEELING(0x1F938),
+    MAN_CARTWHEELING_LIGHT_SKIN_TONE(0x1F938),
+    MAN_CARTWHEELING_MEDIUM_LIGHT_SKIN_TONE(0x1F938),
+    MAN_CARTWHEELING_MEDIUM_SKIN_TONE(0x1F938),
+    MAN_CARTWHEELING_MEDIUM_DARK_SKIN_TONE(0x1F938),
+    MAN_CARTWHEELING_DARK_SKIN_TONE(0x1F938),
+    WOMAN_CARTWHEELING(0x1F938),
+    WOMAN_CARTWHEELING_LIGHT_SKIN_TONE(0x1F938),
+    WOMAN_CARTWHEELING_MEDIUM_LIGHT_SKIN_TONE(0x1F938),
+    WOMAN_CARTWHEELING_MEDIUM_SKIN_TONE(0x1F938),
+    WOMAN_CARTWHEELING_MEDIUM_DARK_SKIN_TONE(0x1F938),
+    WOMAN_CARTWHEELING_DARK_SKIN_TONE(0x1F938),
+    PEOPLE_WRESTLING(0x1F93C),
+    MEN_WRESTLING(0x1F93C),
+    WOMEN_WRESTLING(0x1F93C),
+    PERSON_PLAYING_WATER_POLO(0x1F93D),
+    PERSON_PLAYING_WATER_POLO_LIGHT_SKIN_TONE(0x1F93D),
+    PERSON_PLAYING_WATER_POLO_MEDIUM_LIGHT_SKIN_TONE(0x1F93D),
+    PERSON_PLAYING_WATER_POLO_MEDIUM_SKIN_TONE(0x1F93D),
+    PERSON_PLAYING_WATER_POLO_MEDIUM_DARK_SKIN_TONE(0x1F93D),
+    PERSON_PLAYING_WATER_POLO_DARK_SKIN_TONE(0x1F93D),
+    MAN_PLAYING_WATER_POLO(0x1F93D),
+    MAN_PLAYING_WATER_POLO_LIGHT_SKIN_TONE(0x1F93D),
+    MAN_PLAYING_WATER_POLO_MEDIUM_LIGHT_SKIN_TONE(0x1F93D),
+    MAN_PLAYING_WATER_POLO_MEDIUM_SKIN_TONE(0x1F93D),
+    MAN_PLAYING_WATER_POLO_MEDIUM_DARK_SKIN_TONE(0x1F93D),
+    MAN_PLAYING_WATER_POLO_DARK_SKIN_TONE(0x1F93D),
+    WOMAN_PLAYING_WATER_POLO(0x1F93D),
+    WOMAN_PLAYING_WATER_POLO_LIGHT_SKIN_TONE(0x1F93D),
+    WOMAN_PLAYING_WATER_POLO_MEDIUM_LIGHT_SKIN_TONE(0x1F93D),
+    WOMAN_PLAYING_WATER_POLO_MEDIUM_SKIN_TONE(0x1F93D),
+    WOMAN_PLAYING_WATER_POLO_MEDIUM_DARK_SKIN_TONE(0x1F93D),
+    WOMAN_PLAYING_WATER_POLO_DARK_SKIN_TONE(0x1F93D),
+    PERSON_PLAYING_HANDBALL(0x1F93E),
+    PERSON_PLAYING_HANDBALL_LIGHT_SKIN_TONE(0x1F93E),
+    PERSON_PLAYING_HANDBALL_MEDIUM_LIGHT_SKIN_TONE(0x1F93E),
+    PERSON_PLAYING_HANDBALL_MEDIUM_SKIN_TONE(0x1F93E),
+    PERSON_PLAYING_HANDBALL_MEDIUM_DARK_SKIN_TONE(0x1F93E),
+    PERSON_PLAYING_HANDBALL_DARK_SKIN_TONE(0x1F93E),
+    MAN_PLAYING_HANDBALL(0x1F93E),
+    MAN_PLAYING_HANDBALL_LIGHT_SKIN_TONE(0x1F93E),
+    MAN_PLAYING_HANDBALL_MEDIUM_LIGHT_SKIN_TONE(0x1F93E),
+    MAN_PLAYING_HANDBALL_MEDIUM_SKIN_TONE(0x1F93E),
+    MAN_PLAYING_HANDBALL_MEDIUM_DARK_SKIN_TONE(0x1F93E),
+    MAN_PLAYING_HANDBALL_DARK_SKIN_TONE(0x1F93E),
+    WOMAN_PLAYING_HANDBALL(0x1F93E),
+    WOMAN_PLAYING_HANDBALL_LIGHT_SKIN_TONE(0x1F93E),
+    WOMAN_PLAYING_HANDBALL_MEDIUM_LIGHT_SKIN_TONE(0x1F93E),
+    WOMAN_PLAYING_HANDBALL_MEDIUM_SKIN_TONE(0x1F93E),
+    WOMAN_PLAYING_HANDBALL_MEDIUM_DARK_SKIN_TONE(0x1F93E),
+    WOMAN_PLAYING_HANDBALL_DARK_SKIN_TONE(0x1F93E),
+    PERSON_JUGGLING(0x1F939),
+    PERSON_JUGGLING_LIGHT_SKIN_TONE(0x1F939),
+    PERSON_JUGGLING_MEDIUM_LIGHT_SKIN_TONE(0x1F939),
+    PERSON_JUGGLING_MEDIUM_SKIN_TONE(0x1F939),
+    PERSON_JUGGLING_MEDIUM_DARK_SKIN_TONE(0x1F939),
+    PERSON_JUGGLING_DARK_SKIN_TONE(0x1F939),
+    MAN_JUGGLING(0x1F939),
+    MAN_JUGGLING_LIGHT_SKIN_TONE(0x1F939),
+    MAN_JUGGLING_MEDIUM_LIGHT_SKIN_TONE(0x1F939),
+    MAN_JUGGLING_MEDIUM_SKIN_TONE(0x1F939),
+    MAN_JUGGLING_MEDIUM_DARK_SKIN_TONE(0x1F939),
+    MAN_JUGGLING_DARK_SKIN_TONE(0x1F939),
+    WOMAN_JUGGLING(0x1F939),
+    WOMAN_JUGGLING_LIGHT_SKIN_TONE(0x1F939),
+    WOMAN_JUGGLING_MEDIUM_LIGHT_SKIN_TONE(0x1F939),
+    WOMAN_JUGGLING_MEDIUM_SKIN_TONE(0x1F939),
+    WOMAN_JUGGLING_MEDIUM_DARK_SKIN_TONE(0x1F939),
+    WOMAN_JUGGLING_DARK_SKIN_TONE(0x1F939),
+    PERSON_IN_LOTUS_POSITION(0x1F9D8),
+    PERSON_IN_LOTUS_POSITION_LIGHT_SKIN_TONE(0x1F9D8),
+    PERSON_IN_LOTUS_POSITION_MEDIUM_LIGHT_SKIN_TONE(0x1F9D8),
+    PERSON_IN_LOTUS_POSITION_MEDIUM_SKIN_TONE(0x1F9D8),
+    PERSON_IN_LOTUS_POSITION_MEDIUM_DARK_SKIN_TONE(0x1F9D8),
+    PERSON_IN_LOTUS_POSITION_DARK_SKIN_TONE(0x1F9D8),
+    MAN_IN_LOTUS_POSITION(0x1F9D8),
+    MAN_IN_LOTUS_POSITION_LIGHT_SKIN_TONE(0x1F9D8),
+    MAN_IN_LOTUS_POSITION_MEDIUM_LIGHT_SKIN_TONE(0x1F9D8),
+    MAN_IN_LOTUS_POSITION_MEDIUM_SKIN_TONE(0x1F9D8),
+    MAN_IN_LOTUS_POSITION_MEDIUM_DARK_SKIN_TONE(0x1F9D8),
+    MAN_IN_LOTUS_POSITION_DARK_SKIN_TONE(0x1F9D8),
+    WOMAN_IN_LOTUS_POSITION(0x1F9D8),
+    WOMAN_IN_LOTUS_POSITION_LIGHT_SKIN_TONE(0x1F9D8),
+    WOMAN_IN_LOTUS_POSITION_MEDIUM_LIGHT_SKIN_TONE(0x1F9D8),
+    WOMAN_IN_LOTUS_POSITION_MEDIUM_SKIN_TONE(0x1F9D8),
+    WOMAN_IN_LOTUS_POSITION_MEDIUM_DARK_SKIN_TONE(0x1F9D8),
+    WOMAN_IN_LOTUS_POSITION_DARK_SKIN_TONE(0x1F9D8),
+    PERSON_TAKING_BATH(0x1F6C0),
+    PERSON_TAKING_BATH_LIGHT_SKIN_TONE(0x1F6C0),
+    PERSON_TAKING_BATH_MEDIUM_LIGHT_SKIN_TONE(0x1F6C0),
+    PERSON_TAKING_BATH_MEDIUM_SKIN_TONE(0x1F6C0),
+    PERSON_TAKING_BATH_MEDIUM_DARK_SKIN_TONE(0x1F6C0),
+    PERSON_TAKING_BATH_DARK_SKIN_TONE(0x1F6C0),
+    PERSON_IN_BED(0x1F6CC),
+    PERSON_IN_BED_LIGHT_SKIN_TONE(0x1F6CC),
+    PERSON_IN_BED_MEDIUM_LIGHT_SKIN_TONE(0x1F6CC),
+    PERSON_IN_BED_MEDIUM_SKIN_TONE(0x1F6CC),
+    PERSON_IN_BED_MEDIUM_DARK_SKIN_TONE(0x1F6CC),
+    PERSON_IN_BED_DARK_SKIN_TONE(0x1F6CC),
+    PEOPLE_HOLDING_HANDS(0x1F9D1),
+    PEOPLE_HOLDING_HANDS_LIGHT_SKIN_TONE(0x1F9D1),
+    PEOPLE_HOLDING_HANDS_LIGHT_SKIN_TONECOMMA_MEDIUM_LIGHT_SKIN_TONE(0x1F9D1),
+    PEOPLE_HOLDING_HANDS_LIGHT_SKIN_TONECOMMA_MEDIUM_SKIN_TONE(0x1F9D1),
+    PEOPLE_HOLDING_HANDS_LIGHT_SKIN_TONECOMMA_MEDIUM_DARK_SKIN_TONE(0x1F9D1),
+    PEOPLE_HOLDING_HANDS_LIGHT_SKIN_TONECOMMA_DARK_SKIN_TONE(0x1F9D1),
+    PEOPLE_HOLDING_HANDS_MEDIUM_LIGHT_SKIN_TONECOMMA_LIGHT_SKIN_TONE(0x1F9D1),
+    PEOPLE_HOLDING_HANDS_MEDIUM_LIGHT_SKIN_TONE(0x1F9D1),
+    PEOPLE_HOLDING_HANDS_MEDIUM_LIGHT_SKIN_TONECOMMA_MEDIUM_SKIN_TONE(0x1F9D1),
+    PEOPLE_HOLDING_HANDS_MEDIUM_LIGHT_SKIN_TONECOMMA_MEDIUM_DARK_SKIN_TONE(0x1F9D1),
+    PEOPLE_HOLDING_HANDS_MEDIUM_LIGHT_SKIN_TONECOMMA_DARK_SKIN_TONE(0x1F9D1),
+    PEOPLE_HOLDING_HANDS_MEDIUM_SKIN_TONECOMMA_LIGHT_SKIN_TONE(0x1F9D1),
+    PEOPLE_HOLDING_HANDS_MEDIUM_SKIN_TONECOMMA_MEDIUM_LIGHT_SKIN_TONE(0x1F9D1),
+    PEOPLE_HOLDING_HANDS_MEDIUM_SKIN_TONE(0x1F9D1),
+    PEOPLE_HOLDING_HANDS_MEDIUM_SKIN_TONECOMMA_MEDIUM_DARK_SKIN_TONE(0x1F9D1),
+    PEOPLE_HOLDING_HANDS_MEDIUM_SKIN_TONECOMMA_DARK_SKIN_TONE(0x1F9D1),
+    PEOPLE_HOLDING_HANDS_MEDIUM_DARK_SKIN_TONECOMMA_LIGHT_SKIN_TONE(0x1F9D1),
+    PEOPLE_HOLDING_HANDS_MEDIUM_DARK_SKIN_TONECOMMA_MEDIUM_LIGHT_SKIN_TONE(0x1F9D1),
+    PEOPLE_HOLDING_HANDS_MEDIUM_DARK_SKIN_TONECOMMA_MEDIUM_SKIN_TONE(0x1F9D1),
+    PEOPLE_HOLDING_HANDS_MEDIUM_DARK_SKIN_TONE(0x1F9D1),
+    PEOPLE_HOLDING_HANDS_MEDIUM_DARK_SKIN_TONECOMMA_DARK_SKIN_TONE(0x1F9D1),
+    PEOPLE_HOLDING_HANDS_DARK_SKIN_TONECOMMA_LIGHT_SKIN_TONE(0x1F9D1),
+    PEOPLE_HOLDING_HANDS_DARK_SKIN_TONECOMMA_MEDIUM_LIGHT_SKIN_TONE(0x1F9D1),
+    PEOPLE_HOLDING_HANDS_DARK_SKIN_TONECOMMA_MEDIUM_SKIN_TONE(0x1F9D1),
+    PEOPLE_HOLDING_HANDS_DARK_SKIN_TONECOMMA_MEDIUM_DARK_SKIN_TONE(0x1F9D1),
+    PEOPLE_HOLDING_HANDS_DARK_SKIN_TONE(0x1F9D1),
+    WOMEN_HOLDING_HANDS(0x1F46D),
+    WOMEN_HOLDING_HANDS_LIGHT_SKIN_TONE(0x1F46D),
+    WOMEN_HOLDING_HANDS_LIGHT_SKIN_TONECOMMA_MEDIUM_LIGHT_SKIN_TONE(0x1F469),
+    WOMEN_HOLDING_HANDS_LIGHT_SKIN_TONECOMMA_MEDIUM_SKIN_TONE(0x1F469),
+    WOMEN_HOLDING_HANDS_LIGHT_SKIN_TONECOMMA_MEDIUM_DARK_SKIN_TONE(0x1F469),
+    WOMEN_HOLDING_HANDS_LIGHT_SKIN_TONECOMMA_DARK_SKIN_TONE(0x1F469),
+    WOMEN_HOLDING_HANDS_MEDIUM_LIGHT_SKIN_TONECOMMA_LIGHT_SKIN_TONE(0x1F469),
+    WOMEN_HOLDING_HANDS_MEDIUM_LIGHT_SKIN_TONE(0x1F46D),
+    WOMEN_HOLDING_HANDS_MEDIUM_LIGHT_SKIN_TONECOMMA_MEDIUM_SKIN_TONE(0x1F469),
+    WOMEN_HOLDING_HANDS_MEDIUM_LIGHT_SKIN_TONECOMMA_MEDIUM_DARK_SKIN_TONE(0x1F469),
+    WOMEN_HOLDING_HANDS_MEDIUM_LIGHT_SKIN_TONECOMMA_DARK_SKIN_TONE(0x1F469),
+    WOMEN_HOLDING_HANDS_MEDIUM_SKIN_TONECOMMA_LIGHT_SKIN_TONE(0x1F469),
+    WOMEN_HOLDING_HANDS_MEDIUM_SKIN_TONECOMMA_MEDIUM_LIGHT_SKIN_TONE(0x1F469),
+    WOMEN_HOLDING_HANDS_MEDIUM_SKIN_TONE(0x1F46D),
+    WOMEN_HOLDING_HANDS_MEDIUM_SKIN_TONECOMMA_MEDIUM_DARK_SKIN_TONE(0x1F469),
+    WOMEN_HOLDING_HANDS_MEDIUM_SKIN_TONECOMMA_DARK_SKIN_TONE(0x1F469),
+    WOMEN_HOLDING_HANDS_MEDIUM_DARK_SKIN_TONECOMMA_LIGHT_SKIN_TONE(0x1F469),
+    WOMEN_HOLDING_HANDS_MEDIUM_DARK_SKIN_TONECOMMA_MEDIUM_LIGHT_SKIN_TONE(0x1F469);
+
+    override fun toString(): String {
+        return String(Character.toChars(code))
+    }
+}
