+++ src/main/kotlin/net/purefunc/emoji/Emoji2.kt
@@ -1,0 +1,1011 @@
+package net.purefunc.emoji
+
+enum class Emoji2(
+    private val code: Int,
+) {
+
+    WOMEN_HOLDING_HANDS_MEDIUM_DARK_SKIN_TONECOMMA_MEDIUM_SKIN_TONE(0x1F469),
+    WOMEN_HOLDING_HANDS_MEDIUM_DARK_SKIN_TONE(0x1F46D),
+    WOMEN_HOLDING_HANDS_MEDIUM_DARK_SKIN_TONECOMMA_DARK_SKIN_TONE(0x1F469),
+    WOMEN_HOLDING_HANDS_DARK_SKIN_TONECOMMA_LIGHT_SKIN_TONE(0x1F469),
+    WOMEN_HOLDING_HANDS_DARK_SKIN_TONECOMMA_MEDIUM_LIGHT_SKIN_TONE(0x1F469),
+    WOMEN_HOLDING_HANDS_DARK_SKIN_TONECOMMA_MEDIUM_SKIN_TONE(0x1F469),
+    WOMEN_HOLDING_HANDS_DARK_SKIN_TONECOMMA_MEDIUM_DARK_SKIN_TONE(0x1F469),
+    WOMEN_HOLDING_HANDS_DARK_SKIN_TONE(0x1F46D),
+    WOMAN_AND_MAN_HOLDING_HANDS(0x1F46B),
+    WOMAN_AND_MAN_HOLDING_HANDS_LIGHT_SKIN_TONE(0x1F46B),
+    WOMAN_AND_MAN_HOLDING_HANDS_LIGHT_SKIN_TONECOMMA_MEDIUM_LIGHT_SKIN_TONE(0x1F469),
+    WOMAN_AND_MAN_HOLDING_HANDS_LIGHT_SKIN_TONECOMMA_MEDIUM_SKIN_TONE(0x1F469),
+    WOMAN_AND_MAN_HOLDING_HANDS_LIGHT_SKIN_TONECOMMA_MEDIUM_DARK_SKIN_TONE(0x1F469),
+    WOMAN_AND_MAN_HOLDING_HANDS_LIGHT_SKIN_TONECOMMA_DARK_SKIN_TONE(0x1F469),
+    WOMAN_AND_MAN_HOLDING_HANDS_MEDIUM_LIGHT_SKIN_TONECOMMA_LIGHT_SKIN_TONE(0x1F469),
+    WOMAN_AND_MAN_HOLDING_HANDS_MEDIUM_LIGHT_SKIN_TONE(0x1F46B),
+    WOMAN_AND_MAN_HOLDING_HANDS_MEDIUM_LIGHT_SKIN_TONECOMMA_MEDIUM_SKIN_TONE(0x1F469),
+    WOMAN_AND_MAN_HOLDING_HANDS_MEDIUM_LIGHT_SKIN_TONECOMMA_MEDIUM_DARK_SKIN_TONE(0x1F469),
+    WOMAN_AND_MAN_HOLDING_HANDS_MEDIUM_LIGHT_SKIN_TONECOMMA_DARK_SKIN_TONE(0x1F469),
+    WOMAN_AND_MAN_HOLDING_HANDS_MEDIUM_SKIN_TONECOMMA_LIGHT_SKIN_TONE(0x1F469),
+    WOMAN_AND_MAN_HOLDING_HANDS_MEDIUM_SKIN_TONECOMMA_MEDIUM_LIGHT_SKIN_TONE(0x1F469),
+    WOMAN_AND_MAN_HOLDING_HANDS_MEDIUM_SKIN_TONE(0x1F46B),
+    WOMAN_AND_MAN_HOLDING_HANDS_MEDIUM_SKIN_TONECOMMA_MEDIUM_DARK_SKIN_TONE(0x1F469),
+    WOMAN_AND_MAN_HOLDING_HANDS_MEDIUM_SKIN_TONECOMMA_DARK_SKIN_TONE(0x1F469),
+    WOMAN_AND_MAN_HOLDING_HANDS_MEDIUM_DARK_SKIN_TONECOMMA_LIGHT_SKIN_TONE(0x1F469),
+    WOMAN_AND_MAN_HOLDING_HANDS_MEDIUM_DARK_SKIN_TONECOMMA_MEDIUM_LIGHT_SKIN_TONE(0x1F469),
+    WOMAN_AND_MAN_HOLDING_HANDS_MEDIUM_DARK_SKIN_TONECOMMA_MEDIUM_SKIN_TONE(0x1F469),
+    WOMAN_AND_MAN_HOLDING_HANDS_MEDIUM_DARK_SKIN_TONE(0x1F46B),
+    WOMAN_AND_MAN_HOLDING_HANDS_MEDIUM_DARK_SKIN_TONECOMMA_DARK_SKIN_TONE(0x1F469),
+    WOMAN_AND_MAN_HOLDING_HANDS_DARK_SKIN_TONECOMMA_LIGHT_SKIN_TONE(0x1F469),
+    WOMAN_AND_MAN_HOLDING_HANDS_DARK_SKIN_TONECOMMA_MEDIUM_LIGHT_SKIN_TONE(0x1F469),
+    WOMAN_AND_MAN_HOLDING_HANDS_DARK_SKIN_TONECOMMA_MEDIUM_SKIN_TONE(0x1F469),
+    WOMAN_AND_MAN_HOLDING_HANDS_DARK_SKIN_TONECOMMA_MEDIUM_DARK_SKIN_TONE(0x1F469),
+    WOMAN_AND_MAN_HOLDING_HANDS_DARK_SKIN_TONE(0x1F46B),
+    MEN_HOLDING_HANDS(0x1F46C),
+    MEN_HOLDING_HANDS_LIGHT_SKIN_TONE(0x1F46C),
+    MEN_HOLDING_HANDS_LIGHT_SKIN_TONECOMMA_MEDIUM_LIGHT_SKIN_TONE(0x1F468),
+    MEN_HOLDING_HANDS_LIGHT_SKIN_TONECOMMA_MEDIUM_SKIN_TONE(0x1F468),
+    MEN_HOLDING_HANDS_LIGHT_SKIN_TONECOMMA_MEDIUM_DARK_SKIN_TONE(0x1F468),
+    MEN_HOLDING_HANDS_LIGHT_SKIN_TONECOMMA_DARK_SKIN_TONE(0x1F468),
+    MEN_HOLDING_HANDS_MEDIUM_LIGHT_SKIN_TONECOMMA_LIGHT_SKIN_TONE(0x1F468),
+    MEN_HOLDING_HANDS_MEDIUM_LIGHT_SKIN_TONE(0x1F46C),
+    MEN_HOLDING_HANDS_MEDIUM_LIGHT_SKIN_TONECOMMA_MEDIUM_SKIN_TONE(0x1F468),
+    MEN_HOLDING_HANDS_MEDIUM_LIGHT_SKIN_TONECOMMA_MEDIUM_DARK_SKIN_TONE(0x1F468),
+    MEN_HOLDING_HANDS_MEDIUM_LIGHT_SKIN_TONECOMMA_DARK_SKIN_TONE(0x1F468),
+    MEN_HOLDING_HANDS_MEDIUM_SKIN_TONECOMMA_LIGHT_SKIN_TONE(0x1F468),
+    MEN_HOLDING_HANDS_MEDIUM_SKIN_TONECOMMA_MEDIUM_LIGHT_SKIN_TONE(0x1F468),
+    MEN_HOLDING_HANDS_MEDIUM_SKIN_TONE(0x1F46C),
+    MEN_HOLDING_HANDS_MEDIUM_SKIN_TONECOMMA_MEDIUM_DARK_SKIN_TONE(0x1F468),
+    MEN_HOLDING_HANDS_MEDIUM_SKIN_TONECOMMA_DARK_SKIN_TONE(0x1F468),
+    MEN_HOLDING_HANDS_MEDIUM_DARK_SKIN_TONECOMMA_LIGHT_SKIN_TONE(0x1F468),
+    MEN_HOLDING_HANDS_MEDIUM_DARK_SKIN_TONECOMMA_MEDIUM_LIGHT_SKIN_TONE(0x1F468),
+    MEN_HOLDING_HANDS_MEDIUM_DARK_SKIN_TONECOMMA_MEDIUM_SKIN_TONE(0x1F468),
+    MEN_HOLDING_HANDS_MEDIUM_DARK_SKIN_TONE(0x1F46C),
+    MEN_HOLDING_HANDS_MEDIUM_DARK_SKIN_TONECOMMA_DARK_SKIN_TONE(0x1F468),
+    MEN_HOLDING_HANDS_DARK_SKIN_TONECOMMA_LIGHT_SKIN_TONE(0x1F468),
+    MEN_HOLDING_HANDS_DARK_SKIN_TONECOMMA_MEDIUM_LIGHT_SKIN_TONE(0x1F468),
+    MEN_HOLDING_HANDS_DARK_SKIN_TONECOMMA_MEDIUM_SKIN_TONE(0x1F468),
+    MEN_HOLDING_HANDS_DARK_SKIN_TONECOMMA_MEDIUM_DARK_SKIN_TONE(0x1F468),
+    MEN_HOLDING_HANDS_DARK_SKIN_TONE(0x1F46C),
+    KISS(0x1F48F),
+    KISS_LIGHT_SKIN_TONE(0x1F48F),
+    KISS_MEDIUM_LIGHT_SKIN_TONE(0x1F48F),
+    KISS_MEDIUM_SKIN_TONE(0x1F48F),
+    KISS_MEDIUM_DARK_SKIN_TONE(0x1F48F),
+    KISS_DARK_SKIN_TONE(0x1F48F),
+    KISS_PERSONCOMMA_PERSONCOMMA_LIGHT_SKIN_TONECOMMA_MEDIUM_LIGHT_SKIN_TONE(0x1F9D1),
+    KISS_PERSONCOMMA_PERSONCOMMA_LIGHT_SKIN_TONECOMMA_MEDIUM_SKIN_TONE(0x1F9D1),
+    KISS_PERSONCOMMA_PERSONCOMMA_LIGHT_SKIN_TONECOMMA_MEDIUM_DARK_SKIN_TONE(0x1F9D1),
+    KISS_PERSONCOMMA_PERSONCOMMA_LIGHT_SKIN_TONECOMMA_DARK_SKIN_TONE(0x1F9D1),
+    KISS_PERSONCOMMA_PERSONCOMMA_MEDIUM_LIGHT_SKIN_TONECOMMA_LIGHT_SKIN_TONE(0x1F9D1),
+    KISS_PERSONCOMMA_PERSONCOMMA_MEDIUM_LIGHT_SKIN_TONECOMMA_MEDIUM_SKIN_TONE(0x1F9D1),
+    KISS_PERSONCOMMA_PERSONCOMMA_MEDIUM_LIGHT_SKIN_TONECOMMA_MEDIUM_DARK_SKIN_TONE(0x1F9D1),
+    KISS_PERSONCOMMA_PERSONCOMMA_MEDIUM_LIGHT_SKIN_TONECOMMA_DARK_SKIN_TONE(0x1F9D1),
+    KISS_PERSONCOMMA_PERSONCOMMA_MEDIUM_SKIN_TONECOMMA_LIGHT_SKIN_TONE(0x1F9D1),
+    KISS_PERSONCOMMA_PERSONCOMMA_MEDIUM_SKIN_TONECOMMA_MEDIUM_LIGHT_SKIN_TONE(0x1F9D1),
+    KISS_PERSONCOMMA_PERSONCOMMA_MEDIUM_SKIN_TONECOMMA_MEDIUM_DARK_SKIN_TONE(0x1F9D1),
+    KISS_PERSONCOMMA_PERSONCOMMA_MEDIUM_SKIN_TONECOMMA_DARK_SKIN_TONE(0x1F9D1),
+    KISS_PERSONCOMMA_PERSONCOMMA_MEDIUM_DARK_SKIN_TONECOMMA_LIGHT_SKIN_TONE(0x1F9D1),
+    KISS_PERSONCOMMA_PERSONCOMMA_MEDIUM_DARK_SKIN_TONECOMMA_MEDIUM_LIGHT_SKIN_TONE(0x1F9D1),
+    KISS_PERSONCOMMA_PERSONCOMMA_MEDIUM_DARK_SKIN_TONECOMMA_MEDIUM_SKIN_TONE(0x1F9D1),
+    KISS_PERSONCOMMA_PERSONCOMMA_MEDIUM_DARK_SKIN_TONECOMMA_DARK_SKIN_TONE(0x1F9D1),
+    KISS_PERSONCOMMA_PERSONCOMMA_DARK_SKIN_TONECOMMA_LIGHT_SKIN_TONE(0x1F9D1),
+    KISS_PERSONCOMMA_PERSONCOMMA_DARK_SKIN_TONECOMMA_MEDIUM_LIGHT_SKIN_TONE(0x1F9D1),
+    KISS_PERSONCOMMA_PERSONCOMMA_DARK_SKIN_TONECOMMA_MEDIUM_SKIN_TONE(0x1F9D1),
+    KISS_PERSONCOMMA_PERSONCOMMA_DARK_SKIN_TONECOMMA_MEDIUM_DARK_SKIN_TONE(0x1F9D1),
+    KISS_WOMANCOMMA_MAN(0x1F469),
+    KISS_WOMANCOMMA_MANCOMMA_LIGHT_SKIN_TONE(0x1F469),
+    KISS_WOMANCOMMA_MANCOMMA_LIGHT_SKIN_TONECOMMA_MEDIUM_LIGHT_SKIN_TONE(0x1F469),
+    KISS_WOMANCOMMA_MANCOMMA_LIGHT_SKIN_TONECOMMA_MEDIUM_SKIN_TONE(0x1F469),
+    KISS_WOMANCOMMA_MANCOMMA_LIGHT_SKIN_TONECOMMA_MEDIUM_DARK_SKIN_TONE(0x1F469),
+    KISS_WOMANCOMMA_MANCOMMA_LIGHT_SKIN_TONECOMMA_DARK_SKIN_TONE(0x1F469),
+    KISS_WOMANCOMMA_MANCOMMA_MEDIUM_LIGHT_SKIN_TONECOMMA_LIGHT_SKIN_TONE(0x1F469),
+    KISS_WOMANCOMMA_MANCOMMA_MEDIUM_LIGHT_SKIN_TONE(0x1F469),
+    KISS_WOMANCOMMA_MANCOMMA_MEDIUM_LIGHT_SKIN_TONECOMMA_MEDIUM_SKIN_TONE(0x1F469),
+    KISS_WOMANCOMMA_MANCOMMA_MEDIUM_LIGHT_SKIN_TONECOMMA_MEDIUM_DARK_SKIN_TONE(0x1F469),
+    KISS_WOMANCOMMA_MANCOMMA_MEDIUM_LIGHT_SKIN_TONECOMMA_DARK_SKIN_TONE(0x1F469),
+    KISS_WOMANCOMMA_MANCOMMA_MEDIUM_SKIN_TONECOMMA_LIGHT_SKIN_TONE(0x1F469),
+    KISS_WOMANCOMMA_MANCOMMA_MEDIUM_SKIN_TONECOMMA_MEDIUM_LIGHT_SKIN_TONE(0x1F469),
+    KISS_WOMANCOMMA_MANCOMMA_MEDIUM_SKIN_TONE(0x1F469),
+    KISS_WOMANCOMMA_MANCOMMA_MEDIUM_SKIN_TONECOMMA_MEDIUM_DARK_SKIN_TONE(0x1F469),
+    KISS_WOMANCOMMA_MANCOMMA_MEDIUM_SKIN_TONECOMMA_DARK_SKIN_TONE(0x1F469),
+    KISS_WOMANCOMMA_MANCOMMA_MEDIUM_DARK_SKIN_TONECOMMA_LIGHT_SKIN_TONE(0x1F469),
+    KISS_WOMANCOMMA_MANCOMMA_MEDIUM_DARK_SKIN_TONECOMMA_MEDIUM_LIGHT_SKIN_TONE(0x1F469),
+    KISS_WOMANCOMMA_MANCOMMA_MEDIUM_DARK_SKIN_TONECOMMA_MEDIUM_SKIN_TONE(0x1F469),
+    KISS_WOMANCOMMA_MANCOMMA_MEDIUM_DARK_SKIN_TONE(0x1F469),
+    KISS_WOMANCOMMA_MANCOMMA_MEDIUM_DARK_SKIN_TONECOMMA_DARK_SKIN_TONE(0x1F469),
+    KISS_WOMANCOMMA_MANCOMMA_DARK_SKIN_TONECOMMA_LIGHT_SKIN_TONE(0x1F469),
+    KISS_WOMANCOMMA_MANCOMMA_DARK_SKIN_TONECOMMA_MEDIUM_LIGHT_SKIN_TONE(0x1F469),
+    KISS_WOMANCOMMA_MANCOMMA_DARK_SKIN_TONECOMMA_MEDIUM_SKIN_TONE(0x1F469),
+    KISS_WOMANCOMMA_MANCOMMA_DARK_SKIN_TONECOMMA_MEDIUM_DARK_SKIN_TONE(0x1F469),
+    KISS_WOMANCOMMA_MANCOMMA_DARK_SKIN_TONE(0x1F469),
+    KISS_MANCOMMA_MAN(0x1F468),
+    KISS_MANCOMMA_MANCOMMA_LIGHT_SKIN_TONE(0x1F468),
+    KISS_MANCOMMA_MANCOMMA_LIGHT_SKIN_TONECOMMA_MEDIUM_LIGHT_SKIN_TONE(0x1F468),
+    KISS_MANCOMMA_MANCOMMA_LIGHT_SKIN_TONECOMMA_MEDIUM_SKIN_TONE(0x1F468),
+    KISS_MANCOMMA_MANCOMMA_LIGHT_SKIN_TONECOMMA_MEDIUM_DARK_SKIN_TONE(0x1F468),
+    KISS_MANCOMMA_MANCOMMA_LIGHT_SKIN_TONECOMMA_DARK_SKIN_TONE(0x1F468),
+    KISS_MANCOMMA_MANCOMMA_MEDIUM_LIGHT_SKIN_TONECOMMA_LIGHT_SKIN_TONE(0x1F468),
+    KISS_MANCOMMA_MANCOMMA_MEDIUM_LIGHT_SKIN_TONE(0x1F468),
+    KISS_MANCOMMA_MANCOMMA_MEDIUM_LIGHT_SKIN_TONECOMMA_MEDIUM_SKIN_TONE(0x1F468),
+    KISS_MANCOMMA_MANCOMMA_MEDIUM_LIGHT_SKIN_TONECOMMA_MEDIUM_DARK_SKIN_TONE(0x1F468),
+    KISS_MANCOMMA_MANCOMMA_MEDIUM_LIGHT_SKIN_TONECOMMA_DARK_SKIN_TONE(0x1F468),
+    KISS_MANCOMMA_MANCOMMA_MEDIUM_SKIN_TONECOMMA_LIGHT_SKIN_TONE(0x1F468),
+    KISS_MANCOMMA_MANCOMMA_MEDIUM_SKIN_TONECOMMA_MEDIUM_LIGHT_SKIN_TONE(0x1F468),
+    KISS_MANCOMMA_MANCOMMA_MEDIUM_SKIN_TONE(0x1F468),
+    KISS_MANCOMMA_MANCOMMA_MEDIUM_SKIN_TONECOMMA_MEDIUM_DARK_SKIN_TONE(0x1F468),
+    KISS_MANCOMMA_MANCOMMA_MEDIUM_SKIN_TONECOMMA_DARK_SKIN_TONE(0x1F468),
+    KISS_MANCOMMA_MANCOMMA_MEDIUM_DARK_SKIN_TONECOMMA_LIGHT_SKIN_TONE(0x1F468),
+    KISS_MANCOMMA_MANCOMMA_MEDIUM_DARK_SKIN_TONECOMMA_MEDIUM_LIGHT_SKIN_TONE(0x1F468),
+    KISS_MANCOMMA_MANCOMMA_MEDIUM_DARK_SKIN_TONECOMMA_MEDIUM_SKIN_TONE(0x1F468),
+    KISS_MANCOMMA_MANCOMMA_MEDIUM_DARK_SKIN_TONE(0x1F468),
+    KISS_MANCOMMA_MANCOMMA_MEDIUM_DARK_SKIN_TONECOMMA_DARK_SKIN_TONE(0x1F468),
+    KISS_MANCOMMA_MANCOMMA_DARK_SKIN_TONECOMMA_LIGHT_SKIN_TONE(0x1F468),
+    KISS_MANCOMMA_MANCOMMA_DARK_SKIN_TONECOMMA_MEDIUM_LIGHT_SKIN_TONE(0x1F468),
+    KISS_MANCOMMA_MANCOMMA_DARK_SKIN_TONECOMMA_MEDIUM_SKIN_TONE(0x1F468),
+    KISS_MANCOMMA_MANCOMMA_DARK_SKIN_TONECOMMA_MEDIUM_DARK_SKIN_TONE(0x1F468),
+    KISS_MANCOMMA_MANCOMMA_DARK_SKIN_TONE(0x1F468),
+    KISS_WOMANCOMMA_WOMAN(0x1F469),
+    KISS_WOMANCOMMA_WOMANCOMMA_LIGHT_SKIN_TONE(0x1F469),
+    KISS_WOMANCOMMA_WOMANCOMMA_LIGHT_SKIN_TONECOMMA_MEDIUM_LIGHT_SKIN_TONE(0x1F469),
+    KISS_WOMANCOMMA_WOMANCOMMA_LIGHT_SKIN_TONECOMMA_MEDIUM_SKIN_TONE(0x1F469),
+    KISS_WOMANCOMMA_WOMANCOMMA_LIGHT_SKIN_TONECOMMA_MEDIUM_DARK_SKIN_TONE(0x1F469),
+    KISS_WOMANCOMMA_WOMANCOMMA_LIGHT_SKIN_TONECOMMA_DARK_SKIN_TONE(0x1F469),
+    KISS_WOMANCOMMA_WOMANCOMMA_MEDIUM_LIGHT_SKIN_TONECOMMA_LIGHT_SKIN_TONE(0x1F469),
+    KISS_WOMANCOMMA_WOMANCOMMA_MEDIUM_LIGHT_SKIN_TONE(0x1F469),
+    KISS_WOMANCOMMA_WOMANCOMMA_MEDIUM_LIGHT_SKIN_TONECOMMA_MEDIUM_SKIN_TONE(0x1F469),
+    KISS_WOMANCOMMA_WOMANCOMMA_MEDIUM_LIGHT_SKIN_TONECOMMA_MEDIUM_DARK_SKIN_TONE(0x1F469),
+    KISS_WOMANCOMMA_WOMANCOMMA_MEDIUM_LIGHT_SKIN_TONECOMMA_DARK_SKIN_TONE(0x1F469),
+    KISS_WOMANCOMMA_WOMANCOMMA_MEDIUM_SKIN_TONECOMMA_LIGHT_SKIN_TONE(0x1F469),
+    KISS_WOMANCOMMA_WOMANCOMMA_MEDIUM_SKIN_TONECOMMA_MEDIUM_LIGHT_SKIN_TONE(0x1F469),
+    KISS_WOMANCOMMA_WOMANCOMMA_MEDIUM_SKIN_TONE(0x1F469),
+    KISS_WOMANCOMMA_WOMANCOMMA_MEDIUM_SKIN_TONECOMMA_MEDIUM_DARK_SKIN_TONE(0x1F469),
+    KISS_WOMANCOMMA_WOMANCOMMA_MEDIUM_SKIN_TONECOMMA_DARK_SKIN_TONE(0x1F469),
+    KISS_WOMANCOMMA_WOMANCOMMA_MEDIUM_DARK_SKIN_TONECOMMA_LIGHT_SKIN_TONE(0x1F469),
+    KISS_WOMANCOMMA_WOMANCOMMA_MEDIUM_DARK_SKIN_TONECOMMA_MEDIUM_LIGHT_SKIN_TONE(0x1F469),
+    KISS_WOMANCOMMA_WOMANCOMMA_MEDIUM_DARK_SKIN_TONECOMMA_MEDIUM_SKIN_TONE(0x1F469),
+    KISS_WOMANCOMMA_WOMANCOMMA_MEDIUM_DARK_SKIN_TONE(0x1F469),
+    KISS_WOMANCOMMA_WOMANCOMMA_MEDIUM_DARK_SKIN_TONECOMMA_DARK_SKIN_TONE(0x1F469),
+    KISS_WOMANCOMMA_WOMANCOMMA_DARK_SKIN_TONECOMMA_LIGHT_SKIN_TONE(0x1F469),
+    KISS_WOMANCOMMA_WOMANCOMMA_DARK_SKIN_TONECOMMA_MEDIUM_LIGHT_SKIN_TONE(0x1F469),
+    KISS_WOMANCOMMA_WOMANCOMMA_DARK_SKIN_TONECOMMA_MEDIUM_SKIN_TONE(0x1F469),
+    KISS_WOMANCOMMA_WOMANCOMMA_DARK_SKIN_TONECOMMA_MEDIUM_DARK_SKIN_TONE(0x1F469),
+    KISS_WOMANCOMMA_WOMANCOMMA_DARK_SKIN_TONE(0x1F469),
+    COUPLE_WITH_HEART(0x1F491),
+    COUPLE_WITH_HEART_LIGHT_SKIN_TONE(0x1F491),
+    COUPLE_WITH_HEART_MEDIUM_LIGHT_SKIN_TONE(0x1F491),
+    COUPLE_WITH_HEART_MEDIUM_SKIN_TONE(0x1F491),
+    COUPLE_WITH_HEART_MEDIUM_DARK_SKIN_TONE(0x1F491),
+    COUPLE_WITH_HEART_DARK_SKIN_TONE(0x1F491),
+    COUPLE_WITH_HEART_PERSONCOMMA_PERSONCOMMA_LIGHT_SKIN_TONECOMMA_MEDIUM_LIGHT_SKIN_TONE(0x1F9D1),
+    COUPLE_WITH_HEART_PERSONCOMMA_PERSONCOMMA_LIGHT_SKIN_TONECOMMA_MEDIUM_SKIN_TONE(0x1F9D1),
+    COUPLE_WITH_HEART_PERSONCOMMA_PERSONCOMMA_LIGHT_SKIN_TONECOMMA_MEDIUM_DARK_SKIN_TONE(0x1F9D1),
+    COUPLE_WITH_HEART_PERSONCOMMA_PERSONCOMMA_LIGHT_SKIN_TONECOMMA_DARK_SKIN_TONE(0x1F9D1),
+    COUPLE_WITH_HEART_PERSONCOMMA_PERSONCOMMA_MEDIUM_LIGHT_SKIN_TONECOMMA_LIGHT_SKIN_TONE(0x1F9D1),
+    COUPLE_WITH_HEART_PERSONCOMMA_PERSONCOMMA_MEDIUM_LIGHT_SKIN_TONECOMMA_MEDIUM_SKIN_TONE(0x1F9D1),
+    COUPLE_WITH_HEART_PERSONCOMMA_PERSONCOMMA_MEDIUM_LIGHT_SKIN_TONECOMMA_MEDIUM_DARK_SKIN_TONE(0x1F9D1),
+    COUPLE_WITH_HEART_PERSONCOMMA_PERSONCOMMA_MEDIUM_LIGHT_SKIN_TONECOMMA_DARK_SKIN_TONE(0x1F9D1),
+    COUPLE_WITH_HEART_PERSONCOMMA_PERSONCOMMA_MEDIUM_SKIN_TONECOMMA_LIGHT_SKIN_TONE(0x1F9D1),
+    COUPLE_WITH_HEART_PERSONCOMMA_PERSONCOMMA_MEDIUM_SKIN_TONECOMMA_MEDIUM_LIGHT_SKIN_TONE(0x1F9D1),
+    COUPLE_WITH_HEART_PERSONCOMMA_PERSONCOMMA_MEDIUM_SKIN_TONECOMMA_MEDIUM_DARK_SKIN_TONE(0x1F9D1),
+    COUPLE_WITH_HEART_PERSONCOMMA_PERSONCOMMA_MEDIUM_SKIN_TONECOMMA_DARK_SKIN_TONE(0x1F9D1),
+    COUPLE_WITH_HEART_PERSONCOMMA_PERSONCOMMA_MEDIUM_DARK_SKIN_TONECOMMA_LIGHT_SKIN_TONE(0x1F9D1),
+    COUPLE_WITH_HEART_PERSONCOMMA_PERSONCOMMA_MEDIUM_DARK_SKIN_TONECOMMA_MEDIUM_LIGHT_SKIN_TONE(0x1F9D1),
+    COUPLE_WITH_HEART_PERSONCOMMA_PERSONCOMMA_MEDIUM_DARK_SKIN_TONECOMMA_MEDIUM_SKIN_TONE(0x1F9D1),
+    COUPLE_WITH_HEART_PERSONCOMMA_PERSONCOMMA_MEDIUM_DARK_SKIN_TONECOMMA_DARK_SKIN_TONE(0x1F9D1),
+    COUPLE_WITH_HEART_PERSONCOMMA_PERSONCOMMA_DARK_SKIN_TONECOMMA_LIGHT_SKIN_TONE(0x1F9D1),
+    COUPLE_WITH_HEART_PERSONCOMMA_PERSONCOMMA_DARK_SKIN_TONECOMMA_MEDIUM_LIGHT_SKIN_TONE(0x1F9D1),
+    COUPLE_WITH_HEART_PERSONCOMMA_PERSONCOMMA_DARK_SKIN_TONECOMMA_MEDIUM_SKIN_TONE(0x1F9D1),
+    COUPLE_WITH_HEART_PERSONCOMMA_PERSONCOMMA_DARK_SKIN_TONECOMMA_MEDIUM_DARK_SKIN_TONE(0x1F9D1),
+    COUPLE_WITH_HEART_WOMANCOMMA_MAN(0x1F469),
+    COUPLE_WITH_HEART_WOMANCOMMA_MANCOMMA_LIGHT_SKIN_TONE(0x1F469),
+    COUPLE_WITH_HEART_WOMANCOMMA_MANCOMMA_LIGHT_SKIN_TONECOMMA_MEDIUM_LIGHT_SKIN_TONE(0x1F469),
+    COUPLE_WITH_HEART_WOMANCOMMA_MANCOMMA_LIGHT_SKIN_TONECOMMA_MEDIUM_SKIN_TONE(0x1F469),
+    COUPLE_WITH_HEART_WOMANCOMMA_MANCOMMA_LIGHT_SKIN_TONECOMMA_MEDIUM_DARK_SKIN_TONE(0x1F469),
+    COUPLE_WITH_HEART_WOMANCOMMA_MANCOMMA_LIGHT_SKIN_TONECOMMA_DARK_SKIN_TONE(0x1F469),
+    COUPLE_WITH_HEART_WOMANCOMMA_MANCOMMA_MEDIUM_LIGHT_SKIN_TONECOMMA_LIGHT_SKIN_TONE(0x1F469),
+    COUPLE_WITH_HEART_WOMANCOMMA_MANCOMMA_MEDIUM_LIGHT_SKIN_TONE(0x1F469),
+    COUPLE_WITH_HEART_WOMANCOMMA_MANCOMMA_MEDIUM_LIGHT_SKIN_TONECOMMA_MEDIUM_SKIN_TONE(0x1F469),
+    COUPLE_WITH_HEART_WOMANCOMMA_MANCOMMA_MEDIUM_LIGHT_SKIN_TONECOMMA_MEDIUM_DARK_SKIN_TONE(0x1F469),
+    COUPLE_WITH_HEART_WOMANCOMMA_MANCOMMA_MEDIUM_LIGHT_SKIN_TONECOMMA_DARK_SKIN_TONE(0x1F469),
+    COUPLE_WITH_HEART_WOMANCOMMA_MANCOMMA_MEDIUM_SKIN_TONECOMMA_LIGHT_SKIN_TONE(0x1F469),
+    COUPLE_WITH_HEART_WOMANCOMMA_MANCOMMA_MEDIUM_SKIN_TONECOMMA_MEDIUM_LIGHT_SKIN_TONE(0x1F469),
+    COUPLE_WITH_HEART_WOMANCOMMA_MANCOMMA_MEDIUM_SKIN_TONE(0x1F469),
+    COUPLE_WITH_HEART_WOMANCOMMA_MANCOMMA_MEDIUM_SKIN_TONECOMMA_MEDIUM_DARK_SKIN_TONE(0x1F469),
+    COUPLE_WITH_HEART_WOMANCOMMA_MANCOMMA_MEDIUM_SKIN_TONECOMMA_DARK_SKIN_TONE(0x1F469),
+    COUPLE_WITH_HEART_WOMANCOMMA_MANCOMMA_MEDIUM_DARK_SKIN_TONECOMMA_LIGHT_SKIN_TONE(0x1F469),
+    COUPLE_WITH_HEART_WOMANCOMMA_MANCOMMA_MEDIUM_DARK_SKIN_TONECOMMA_MEDIUM_LIGHT_SKIN_TONE(0x1F469),
+    COUPLE_WITH_HEART_WOMANCOMMA_MANCOMMA_MEDIUM_DARK_SKIN_TONECOMMA_MEDIUM_SKIN_TONE(0x1F469),
+    COUPLE_WITH_HEART_WOMANCOMMA_MANCOMMA_MEDIUM_DARK_SKIN_TONE(0x1F469),
+    COUPLE_WITH_HEART_WOMANCOMMA_MANCOMMA_MEDIUM_DARK_SKIN_TONECOMMA_DARK_SKIN_TONE(0x1F469),
+    COUPLE_WITH_HEART_WOMANCOMMA_MANCOMMA_DARK_SKIN_TONECOMMA_LIGHT_SKIN_TONE(0x1F469),
+    COUPLE_WITH_HEART_WOMANCOMMA_MANCOMMA_DARK_SKIN_TONECOMMA_MEDIUM_LIGHT_SKIN_TONE(0x1F469),
+    COUPLE_WITH_HEART_WOMANCOMMA_MANCOMMA_DARK_SKIN_TONECOMMA_MEDIUM_SKIN_TONE(0x1F469),
+    COUPLE_WITH_HEART_WOMANCOMMA_MANCOMMA_DARK_SKIN_TONECOMMA_MEDIUM_DARK_SKIN_TONE(0x1F469),
+    COUPLE_WITH_HEART_WOMANCOMMA_MANCOMMA_DARK_SKIN_TONE(0x1F469),
+    COUPLE_WITH_HEART_MANCOMMA_MAN(0x1F468),
+    COUPLE_WITH_HEART_MANCOMMA_MANCOMMA_LIGHT_SKIN_TONE(0x1F468),
+    COUPLE_WITH_HEART_MANCOMMA_MANCOMMA_LIGHT_SKIN_TONECOMMA_MEDIUM_LIGHT_SKIN_TONE(0x1F468),
+    COUPLE_WITH_HEART_MANCOMMA_MANCOMMA_LIGHT_SKIN_TONECOMMA_MEDIUM_SKIN_TONE(0x1F468),
+    COUPLE_WITH_HEART_MANCOMMA_MANCOMMA_LIGHT_SKIN_TONECOMMA_MEDIUM_DARK_SKIN_TONE(0x1F468),
+    COUPLE_WITH_HEART_MANCOMMA_MANCOMMA_LIGHT_SKIN_TONECOMMA_DARK_SKIN_TONE(0x1F468),
+    COUPLE_WITH_HEART_MANCOMMA_MANCOMMA_MEDIUM_LIGHT_SKIN_TONECOMMA_LIGHT_SKIN_TONE(0x1F468),
+    COUPLE_WITH_HEART_MANCOMMA_MANCOMMA_MEDIUM_LIGHT_SKIN_TONE(0x1F468),
+    COUPLE_WITH_HEART_MANCOMMA_MANCOMMA_MEDIUM_LIGHT_SKIN_TONECOMMA_MEDIUM_SKIN_TONE(0x1F468),
+    COUPLE_WITH_HEART_MANCOMMA_MANCOMMA_MEDIUM_LIGHT_SKIN_TONECOMMA_MEDIUM_DARK_SKIN_TONE(0x1F468),
+    COUPLE_WITH_HEART_MANCOMMA_MANCOMMA_MEDIUM_LIGHT_SKIN_TONECOMMA_DARK_SKIN_TONE(0x1F468),
+    COUPLE_WITH_HEART_MANCOMMA_MANCOMMA_MEDIUM_SKIN_TONECOMMA_LIGHT_SKIN_TONE(0x1F468),
+    COUPLE_WITH_HEART_MANCOMMA_MANCOMMA_MEDIUM_SKIN_TONECOMMA_MEDIUM_LIGHT_SKIN_TONE(0x1F468),
+    COUPLE_WITH_HEART_MANCOMMA_MANCOMMA_MEDIUM_SKIN_TONE(0x1F468),
+    COUPLE_WITH_HEART_MANCOMMA_MANCOMMA_MEDIUM_SKIN_TONECOMMA_MEDIUM_DARK_SKIN_TONE(0x1F468),
+    COUPLE_WITH_HEART_MANCOMMA_MANCOMMA_MEDIUM_SKIN_TONECOMMA_DARK_SKIN_TONE(0x1F468),
+    COUPLE_WITH_HEART_MANCOMMA_MANCOMMA_MEDIUM_DARK_SKIN_TONECOMMA_LIGHT_SKIN_TONE(0x1F468),
+    COUPLE_WITH_HEART_MANCOMMA_MANCOMMA_MEDIUM_DARK_SKIN_TONECOMMA_MEDIUM_LIGHT_SKIN_TONE(0x1F468),
+    COUPLE_WITH_HEART_MANCOMMA_MANCOMMA_MEDIUM_DARK_SKIN_TONECOMMA_MEDIUM_SKIN_TONE(0x1F468),
+    COUPLE_WITH_HEART_MANCOMMA_MANCOMMA_MEDIUM_DARK_SKIN_TONE(0x1F468),
+    COUPLE_WITH_HEART_MANCOMMA_MANCOMMA_MEDIUM_DARK_SKIN_TONECOMMA_DARK_SKIN_TONE(0x1F468),
+    COUPLE_WITH_HEART_MANCOMMA_MANCOMMA_DARK_SKIN_TONECOMMA_LIGHT_SKIN_TONE(0x1F468),
+    COUPLE_WITH_HEART_MANCOMMA_MANCOMMA_DARK_SKIN_TONECOMMA_MEDIUM_LIGHT_SKIN_TONE(0x1F468),
+    COUPLE_WITH_HEART_MANCOMMA_MANCOMMA_DARK_SKIN_TONECOMMA_MEDIUM_SKIN_TONE(0x1F468),
+    COUPLE_WITH_HEART_MANCOMMA_MANCOMMA_DARK_SKIN_TONECOMMA_MEDIUM_DARK_SKIN_TONE(0x1F468),
+    COUPLE_WITH_HEART_MANCOMMA_MANCOMMA_DARK_SKIN_TONE(0x1F468),
+    COUPLE_WITH_HEART_WOMANCOMMA_WOMAN(0x1F469),
+    COUPLE_WITH_HEART_WOMANCOMMA_WOMANCOMMA_LIGHT_SKIN_TONE(0x1F469),
+    COUPLE_WITH_HEART_WOMANCOMMA_WOMANCOMMA_LIGHT_SKIN_TONECOMMA_MEDIUM_LIGHT_SKIN_TONE(0x1F469),
+    COUPLE_WITH_HEART_WOMANCOMMA_WOMANCOMMA_LIGHT_SKIN_TONECOMMA_MEDIUM_SKIN_TONE(0x1F469),
+    COUPLE_WITH_HEART_WOMANCOMMA_WOMANCOMMA_LIGHT_SKIN_TONECOMMA_MEDIUM_DARK_SKIN_TONE(0x1F469),
+    COUPLE_WITH_HEART_WOMANCOMMA_WOMANCOMMA_LIGHT_SKIN_TONECOMMA_DARK_SKIN_TONE(0x1F469),
+    COUPLE_WITH_HEART_WOMANCOMMA_WOMANCOMMA_MEDIUM_LIGHT_SKIN_TONECOMMA_LIGHT_SKIN_TONE(0x1F469),
+    COUPLE_WITH_HEART_WOMANCOMMA_WOMANCOMMA_MEDIUM_LIGHT_SKIN_TONE(0x1F469),
+    COUPLE_WITH_HEART_WOMANCOMMA_WOMANCOMMA_MEDIUM_LIGHT_SKIN_TONECOMMA_MEDIUM_SKIN_TONE(0x1F469),
+    COUPLE_WITH_HEART_WOMANCOMMA_WOMANCOMMA_MEDIUM_LIGHT_SKIN_TONECOMMA_MEDIUM_DARK_SKIN_TONE(0x1F469),
+    COUPLE_WITH_HEART_WOMANCOMMA_WOMANCOMMA_MEDIUM_LIGHT_SKIN_TONECOMMA_DARK_SKIN_TONE(0x1F469),
+    COUPLE_WITH_HEART_WOMANCOMMA_WOMANCOMMA_MEDIUM_SKIN_TONECOMMA_LIGHT_SKIN_TONE(0x1F469),
+    COUPLE_WITH_HEART_WOMANCOMMA_WOMANCOMMA_MEDIUM_SKIN_TONECOMMA_MEDIUM_LIGHT_SKIN_TONE(0x1F469),
+    COUPLE_WITH_HEART_WOMANCOMMA_WOMANCOMMA_MEDIUM_SKIN_TONE(0x1F469),
+    COUPLE_WITH_HEART_WOMANCOMMA_WOMANCOMMA_MEDIUM_SKIN_TONECOMMA_MEDIUM_DARK_SKIN_TONE(0x1F469),
+    COUPLE_WITH_HEART_WOMANCOMMA_WOMANCOMMA_MEDIUM_SKIN_TONECOMMA_DARK_SKIN_TONE(0x1F469),
+    COUPLE_WITH_HEART_WOMANCOMMA_WOMANCOMMA_MEDIUM_DARK_SKIN_TONECOMMA_LIGHT_SKIN_TONE(0x1F469),
+    COUPLE_WITH_HEART_WOMANCOMMA_WOMANCOMMA_MEDIUM_DARK_SKIN_TONECOMMA_MEDIUM_LIGHT_SKIN_TONE(0x1F469),
+    COUPLE_WITH_HEART_WOMANCOMMA_WOMANCOMMA_MEDIUM_DARK_SKIN_TONECOMMA_MEDIUM_SKIN_TONE(0x1F469),
+    COUPLE_WITH_HEART_WOMANCOMMA_WOMANCOMMA_MEDIUM_DARK_SKIN_TONE(0x1F469),
+    COUPLE_WITH_HEART_WOMANCOMMA_WOMANCOMMA_MEDIUM_DARK_SKIN_TONECOMMA_DARK_SKIN_TONE(0x1F469),
+    COUPLE_WITH_HEART_WOMANCOMMA_WOMANCOMMA_DARK_SKIN_TONECOMMA_LIGHT_SKIN_TONE(0x1F469),
+    COUPLE_WITH_HEART_WOMANCOMMA_WOMANCOMMA_DARK_SKIN_TONECOMMA_MEDIUM_LIGHT_SKIN_TONE(0x1F469),
+    COUPLE_WITH_HEART_WOMANCOMMA_WOMANCOMMA_DARK_SKIN_TONECOMMA_MEDIUM_SKIN_TONE(0x1F469),
+    COUPLE_WITH_HEART_WOMANCOMMA_WOMANCOMMA_DARK_SKIN_TONECOMMA_MEDIUM_DARK_SKIN_TONE(0x1F469),
+    COUPLE_WITH_HEART_WOMANCOMMA_WOMANCOMMA_DARK_SKIN_TONE(0x1F469),
+    FAMILY(0x1F46A),
+    FAMILY_MANCOMMA_WOMANCOMMA_BOY(0x1F468),
+    FAMILY_MANCOMMA_WOMANCOMMA_GIRL(0x1F468),
+    FAMILY_MANCOMMA_WOMANCOMMA_GIRLCOMMA_BOY(0x1F468),
+    FAMILY_MANCOMMA_WOMANCOMMA_BOYCOMMA_BOY(0x1F468),
+    FAMILY_MANCOMMA_WOMANCOMMA_GIRLCOMMA_GIRL(0x1F468),
+    FAMILY_MANCOMMA_MANCOMMA_BOY(0x1F468),
+    FAMILY_MANCOMMA_MANCOMMA_GIRL(0x1F468),
+    FAMILY_MANCOMMA_MANCOMMA_GIRLCOMMA_BOY(0x1F468),
+    FAMILY_MANCOMMA_MANCOMMA_BOYCOMMA_BOY(0x1F468),
+    FAMILY_MANCOMMA_MANCOMMA_GIRLCOMMA_GIRL(0x1F468),
+    FAMILY_WOMANCOMMA_WOMANCOMMA_BOY(0x1F469),
+    FAMILY_WOMANCOMMA_WOMANCOMMA_GIRL(0x1F469),
+    FAMILY_WOMANCOMMA_WOMANCOMMA_GIRLCOMMA_BOY(0x1F469),
+    FAMILY_WOMANCOMMA_WOMANCOMMA_BOYCOMMA_BOY(0x1F469),
+    FAMILY_WOMANCOMMA_WOMANCOMMA_GIRLCOMMA_GIRL(0x1F469),
+    FAMILY_MANCOMMA_BOY(0x1F468),
+    FAMILY_MANCOMMA_BOYCOMMA_BOY(0x1F468),
+    FAMILY_MANCOMMA_GIRL(0x1F468),
+    FAMILY_MANCOMMA_GIRLCOMMA_BOY(0x1F468),
+    FAMILY_MANCOMMA_GIRLCOMMA_GIRL(0x1F468),
+    FAMILY_WOMANCOMMA_BOY(0x1F469),
+    FAMILY_WOMANCOMMA_BOYCOMMA_BOY(0x1F469),
+    FAMILY_WOMANCOMMA_GIRL(0x1F469),
+    FAMILY_WOMANCOMMA_GIRLCOMMA_BOY(0x1F469),
+    FAMILY_WOMANCOMMA_GIRLCOMMA_GIRL(0x1F469),
+    SPEAKING_HEAD(0x1F5E3),
+    BUST_IN_SILHOUETTE(0x1F464),
+    BUSTS_IN_SILHOUETTE(0x1F465),
+    PEOPLE_HUGGING(0x1FAC2),
+    FOOTPRINTS(0x1F463),
+    MONKEY_FACE(0x1F435),
+    MONKEY(0x1F412),
+    GORILLA(0x1F98D),
+    ORANGUTAN(0x1F9A7),
+    DOG_FACE(0x1F436),
+    DOG(0x1F415),
+    GUIDE_DOG(0x1F9AE),
+    SERVICE_DOG(0x1F415),
+    POODLE(0x1F429),
+    WOLF(0x1F43A),
+    FOX(0x1F98A),
+    RACCOON(0x1F99D),
+    CAT_FACE(0x1F431),
+    CAT(0x1F408),
+    BLACK_CAT(0x1F408),
+    LION(0x1F981),
+    TIGER_FACE(0x1F42F),
+    TIGER(0x1F405),
+    LEOPARD(0x1F406),
+    HORSE_FACE(0x1F434),
+    HORSE(0x1F40E),
+    UNICORN(0x1F984),
+    ZEBRA(0x1F993),
+    DEER(0x1F98C),
+    BISON(0x1F9AC),
+    COW_FACE(0x1F42E),
+    OX(0x1F402),
+    WATER_BUFFALO(0x1F403),
+    COW(0x1F404),
+    PIG_FACE(0x1F437),
+    PIG(0x1F416),
+    BOAR(0x1F417),
+    PIG_NOSE(0x1F43D),
+    RAM(0x1F40F),
+    EWE(0x1F411),
+    GOAT(0x1F410),
+    CAMEL(0x1F42A),
+    TWO_HUMP_CAMEL(0x1F42B),
+    LLAMA(0x1F999),
+    GIRAFFE(0x1F992),
+    ELEPHANT(0x1F418),
+    MAMMOTH(0x1F9A3),
+    RHINOCEROS(0x1F98F),
+    HIPPOPOTAMUS(0x1F99B),
+    MOUSE_FACE(0x1F42D),
+    MOUSE(0x1F401),
+    RAT(0x1F400),
+    HAMSTER(0x1F439),
+    RABBIT_FACE(0x1F430),
+    RABBIT(0x1F407),
+    CHIPMUNK(0x1F43F),
+    BEAVER(0x1F9AB),
+    HEDGEHOG(0x1F994),
+    BAT(0x1F987),
+    BEAR(0x1F43B),
+    POLAR_BEAR(0x1F43B),
+    KOALA(0x1F428),
+    PANDA(0x1F43C),
+    SLOTH(0x1F9A5),
+    OTTER(0x1F9A6),
+    SKUNK(0x1F9A8),
+    KANGAROO(0x1F998),
+    BADGER(0x1F9A1),
+    PAW_PRINTS(0x1F43E),
+    TURKEY(0x1F983),
+    CHICKEN(0x1F414),
+    ROOSTER(0x1F413),
+    HATCHING_CHICK(0x1F423),
+    BABY_CHICK(0x1F424),
+    FRONT_FACING_BABY_CHICK(0x1F425),
+    BIRD(0x1F426),
+    PENGUIN(0x1F427),
+    DOVE(0x1F54A),
+    EAGLE(0x1F985),
+    DUCK(0x1F986),
+    SWAN(0x1F9A2),
+    OWL(0x1F989),
+    DODO(0x1F9A4),
+    FEATHER(0x1FAB6),
+    FLAMINGO(0x1F9A9),
+    PEACOCK(0x1F99A),
+    PARROT(0x1F99C),
+    FROG(0x1F438),
+    CROCODILE(0x1F40A),
+    TURTLE(0x1F422),
+    LIZARD(0x1F98E),
+    SNAKE(0x1F40D),
+    DRAGON_FACE(0x1F432),
+    DRAGON(0x1F409),
+    SAUROPOD(0x1F995),
+    T_REX(0x1F996),
+    SPOUTING_WHALE(0x1F433),
+    WHALE(0x1F40B),
+    DOLPHIN(0x1F42C),
+    SEAL(0x1F9AD),
+    FISH(0x1F41F),
+    TROPICAL_FISH(0x1F420),
+    BLOWFISH(0x1F421),
+    SHARK(0x1F988),
+    OCTOPUS(0x1F419),
+    SPIRAL_SHELL(0x1F41A),
+    CORAL(0x1FAB8),
+    SNAIL(0x1F40C),
+    BUTTERFLY(0x1F98B),
+    BUG(0x1F41B),
+    ANT(0x1F41C),
+    HONEYBEE(0x1F41D),
+    BEETLE(0x1FAB2),
+    LADY_BEETLE(0x1F41E),
+    CRICKET(0x1F997),
+    COCKROACH(0x1FAB3),
+    SPIDER(0x1F577),
+    SPIDER_WEB(0x1F578),
+    SCORPION(0x1F982),
+    MOSQUITO(0x1F99F),
+    FLY(0x1FAB0),
+    WORM(0x1FAB1),
+    MICROBE(0x1F9A0),
+    BOUQUET(0x1F490),
+    CHERRY_BLOSSOM(0x1F338),
+    WHITE_FLOWER(0x1F4AE),
+    LOTUS(0x1FAB7),
+    ROSETTE(0x1F3F5),
+    ROSE(0x1F339),
+    WILTED_FLOWER(0x1F940),
+    HIBISCUS(0x1F33A),
+    SUNFLOWER(0x1F33B),
+    BLOSSOM(0x1F33C),
+    TULIP(0x1F337),
+    SEEDLING(0x1F331),
+    POTTED_PLANT(0x1FAB4),
+    EVERGREEN_TREE(0x1F332),
+    DECIDUOUS_TREE(0x1F333),
+    PALM_TREE(0x1F334),
+    CACTUS(0x1F335),
+    SHEAF_OF_RICE(0x1F33E),
+    HERB(0x1F33F),
+    SHAMROCK(0x2618),
+    FOUR_LEAF_CLOVER(0x1F340),
+    MAPLE_LEAF(0x1F341),
+    FALLEN_LEAF(0x1F342),
+    LEAF_FLUTTERING_IN_WIND(0x1F343),
+    EMPTY_NEST(0x1FAB9),
+    NEST_WITH_EGGS(0x1FABA),
+    GRAPES(0x1F347),
+    MELON(0x1F348),
+    WATERMELON(0x1F349),
+    TANGERINE(0x1F34A),
+    LEMON(0x1F34B),
+    BANANA(0x1F34C),
+    PINEAPPLE(0x1F34D),
+    MANGO(0x1F96D),
+    RED_APPLE(0x1F34E),
+    GREEN_APPLE(0x1F34F),
+    PEAR(0x1F350),
+    PEACH(0x1F351),
+    CHERRIES(0x1F352),
+    STRAWBERRY(0x1F353),
+    BLUEBERRIES(0x1FAD0),
+    KIWI_FRUIT(0x1F95D),
+    TOMATO(0x1F345),
+    OLIVE(0x1FAD2),
+    COCONUT(0x1F965),
+    AVOCADO(0x1F951),
+    EGGPLANT(0x1F346),
+    POTATO(0x1F954),
+    CARROT(0x1F955),
+    EAR_OF_CORN(0x1F33D),
+    HOT_PEPPER(0x1F336),
+    BELL_PEPPER(0x1FAD1),
+    CUCUMBER(0x1F952),
+    LEAFY_GREEN(0x1F96C),
+    BROCCOLI(0x1F966),
+    GARLIC(0x1F9C4),
+    ONION(0x1F9C5),
+    MUSHROOM(0x1F344),
+    PEANUTS(0x1F95C),
+    BEANS(0x1FAD8),
+    CHESTNUT(0x1F330),
+    BREAD(0x1F35E),
+    CROISSANT(0x1F950),
+    BAGUETTE_BREAD(0x1F956),
+    FLATBREAD(0x1FAD3),
+    PRETZEL(0x1F968),
+    BAGEL(0x1F96F),
+    PANCAKES(0x1F95E),
+    WAFFLE(0x1F9C7),
+    CHEESE_WEDGE(0x1F9C0),
+    MEAT_ON_BONE(0x1F356),
+    POULTRY_LEG(0x1F357),
+    CUT_OF_MEAT(0x1F969),
+    BACON(0x1F953),
+    HAMBURGER(0x1F354),
+    FRENCH_FRIES(0x1F35F),
+    PIZZA(0x1F355),
+    HOT_DOG(0x1F32D),
+    SANDWICH(0x1F96A),
+    TACO(0x1F32E),
+    BURRITO(0x1F32F),
+    TAMALE(0x1FAD4),
+    STUFFED_FLATBREAD(0x1F959),
+    FALAFEL(0x1F9C6),
+    EGG(0x1F95A),
+    COOKING(0x1F373),
+    SHALLOW_PAN_OF_FOOD(0x1F958),
+    POT_OF_FOOD(0x1F372),
+    FONDUE(0x1FAD5),
+    BOWL_WITH_SPOON(0x1F963),
+    GREEN_SALAD(0x1F957),
+    POPCORN(0x1F37F),
+    BUTTER(0x1F9C8),
+    SALT(0x1F9C2),
+    CANNED_FOOD(0x1F96B),
+    BENTO_BOX(0x1F371),
+    RICE_CRACKER(0x1F358),
+    RICE_BALL(0x1F359),
+    COOKED_RICE(0x1F35A),
+    CURRY_RICE(0x1F35B),
+    STEAMING_BOWL(0x1F35C),
+    SPAGHETTI(0x1F35D),
+    ROASTED_SWEET_POTATO(0x1F360),
+    ODEN(0x1F362),
+    SUSHI(0x1F363),
+    FRIED_SHRIMP(0x1F364),
+    FISH_CAKE_WITH_SWIRL(0x1F365),
+    MOON_CAKE(0x1F96E),
+    DANGO(0x1F361),
+    DUMPLING(0x1F95F),
+    FORTUNE_COOKIE(0x1F960),
+    TAKEOUT_BOX(0x1F961),
+    CRAB(0x1F980),
+    LOBSTER(0x1F99E),
+    SHRIMP(0x1F990),
+    SQUID(0x1F991),
+    OYSTER(0x1F9AA),
+    SOFT_ICE_CREAM(0x1F366),
+    SHAVED_ICE(0x1F367),
+    ICE_CREAM(0x1F368),
+    DOUGHNUT(0x1F369),
+    COOKIE(0x1F36A),
+    BIRTHDAY_CAKE(0x1F382),
+    SHORTCAKE(0x1F370),
+    CUPCAKE(0x1F9C1),
+    PIE(0x1F967),
+    CHOCOLATE_BAR(0x1F36B),
+    CANDY(0x1F36C),
+    LOLLIPOP(0x1F36D),
+    CUSTARD(0x1F36E),
+    HONEY_POT(0x1F36F),
+    BABY_BOTTLE(0x1F37C),
+    GLASS_OF_MILK(0x1F95B),
+    HOT_BEVERAGE(0x2615),
+    TEAPOT(0x1FAD6),
+    TEACUP_WITHOUT_HANDLE(0x1F375),
+    SAKE(0x1F376),
+    BOTTLE_WITH_POPPING_CORK(0x1F37E),
+    WINE_GLASS(0x1F377),
+    COCKTAIL_GLASS(0x1F378),
+    TROPICAL_DRINK(0x1F379),
+    BEER_MUG(0x1F37A),
+    CLINKING_BEER_MUGS(0x1F37B),
+    CLINKING_GLASSES(0x1F942),
+    TUMBLER_GLASS(0x1F943),
+    POURING_LIQUID(0x1FAD7),
+    CUP_WITH_STRAW(0x1F964),
+    BUBBLE_TEA(0x1F9CB),
+    BEVERAGE_BOX(0x1F9C3),
+    MATE(0x1F9C9),
+    ICE(0x1F9CA),
+    CHOPSTICKS(0x1F962),
+    FORK_AND_KNIFE_WITH_PLATE(0x1F37D),
+    FORK_AND_KNIFE(0x1F374),
+    SPOON(0x1F944),
+    KITCHEN_KNIFE(0x1F52A),
+    JAR(0x1FAD9),
+    AMPHORA(0x1F3FA),
+    GLOBE_SHOWING_EUROPE_AFRICA(0x1F30D),
+    GLOBE_SHOWING_AMERICAS(0x1F30E),
+    GLOBE_SHOWING_ASIA_AUSTRALIA(0x1F30F),
+    GLOBE_WITH_MERIDIANS(0x1F310),
+    WORLD_MAP(0x1F5FA),
+    MAP_OF_JAPAN(0x1F5FE),
+    COMPASS(0x1F9ED),
+    SNOW_CAPPED_MOUNTAIN(0x1F3D4),
+    MOUNTAIN(0x26F0),
+    VOLCANO(0x1F30B),
+    MOUNT_FUJI(0x1F5FB),
+    CAMPING(0x1F3D5),
+    BEACH_WITH_UMBRELLA(0x1F3D6),
+    DESERT(0x1F3DC),
+    DESERT_ISLAND(0x1F3DD),
+    NATIONAL_PARK(0x1F3DE),
+    STADIUM(0x1F3DF),
+    CLASSICAL_BUILDING(0x1F3DB),
+    BUILDING_CONSTRUCTION(0x1F3D7),
+    BRICK(0x1F9F1),
+    ROCK(0x1FAA8),
+    WOOD(0x1FAB5),
+    HUT(0x1F6D6),
+    HOUSES(0x1F3D8),
+    DERELICT_HOUSE(0x1F3DA),
+    HOUSE(0x1F3E0),
+    HOUSE_WITH_GARDEN(0x1F3E1),
+    OFFICE_BUILDING(0x1F3E2),
+    JAPANESE_POST_OFFICE(0x1F3E3),
+    POST_OFFICE(0x1F3E4),
+    HOSPITAL(0x1F3E5),
+    BANK(0x1F3E6),
+    HOTEL(0x1F3E8),
+    LOVE_HOTEL(0x1F3E9),
+    CONVENIENCE_STORE(0x1F3EA),
+    SCHOOL(0x1F3EB),
+    DEPARTMENT_STORE(0x1F3EC),
+    FACTORY(0x1F3ED),
+    JAPANESE_CASTLE(0x1F3EF),
+    CASTLE(0x1F3F0),
+    WEDDING(0x1F492),
+    TOKYO_TOWER(0x1F5FC),
+    STATUE_OF_LIBERTY(0x1F5FD),
+    CHURCH(0x26EA),
+    MOSQUE(0x1F54C),
+    HINDU_TEMPLE(0x1F6D5),
+    SYNAGOGUE(0x1F54D),
+    SHINTO_SHRINE(0x26E9),
+    KAABA(0x1F54B),
+    FOUNTAIN(0x26F2),
+    TENT(0x26FA),
+    FOGGY(0x1F301),
+    NIGHT_WITH_STARS(0x1F303),
+    CITYSCAPE(0x1F3D9),
+    SUNRISE_OVER_MOUNTAINS(0x1F304),
+    SUNRISE(0x1F305),
+    CITYSCAPE_AT_DUSK(0x1F306),
+    SUNSET(0x1F307),
+    BRIDGE_AT_NIGHT(0x1F309),
+    HOT_SPRINGS(0x2668),
+    CAROUSEL_HORSE(0x1F3A0),
+    PLAYGROUND_SLIDE(0x1F6DD),
+    FERRIS_WHEEL(0x1F3A1),
+    ROLLER_COASTER(0x1F3A2),
+    BARBER_POLE(0x1F488),
+    CIRCUS_TENT(0x1F3AA),
+    LOCOMOTIVE(0x1F682),
+    RAILWAY_CAR(0x1F683),
+    HIGH_SPEED_TRAIN(0x1F684),
+    BULLET_TRAIN(0x1F685),
+    TRAIN(0x1F686),
+    METRO(0x1F687),
+    LIGHT_RAIL(0x1F688),
+    STATION(0x1F689),
+    TRAM(0x1F68A),
+    MONORAIL(0x1F69D),
+    MOUNTAIN_RAILWAY(0x1F69E),
+    TRAM_CAR(0x1F68B),
+    BUS(0x1F68C),
+    ONCOMING_BUS(0x1F68D),
+    TROLLEYBUS(0x1F68E),
+    MINIBUS(0x1F690),
+    AMBULANCE(0x1F691),
+    FIRE_ENGINE(0x1F692),
+    POLICE_CAR(0x1F693),
+    ONCOMING_POLICE_CAR(0x1F694),
+    TAXI(0x1F695),
+    ONCOMING_TAXI(0x1F696),
+    AUTOMOBILE(0x1F697),
+    ONCOMING_AUTOMOBILE(0x1F698),
+    SPORT_UTILITY_VEHICLE(0x1F699),
+    PICKUP_TRUCK(0x1F6FB),
+    DELIVERY_TRUCK(0x1F69A),
+    ARTICULATED_LORRY(0x1F69B),
+    TRACTOR(0x1F69C),
+    RACING_CAR(0x1F3CE),
+    MOTORCYCLE(0x1F3CD),
+    MOTOR_SCOOTER(0x1F6F5),
+    MANUAL_WHEELCHAIR(0x1F9BD),
+    MOTORIZED_WHEELCHAIR(0x1F9BC),
+    AUTO_RICKSHAW(0x1F6FA),
+    BICYCLE(0x1F6B2),
+    KICK_SCOOTER(0x1F6F4),
+    SKATEBOARD(0x1F6F9),
+    ROLLER_SKATE(0x1F6FC),
+    BUS_STOP(0x1F68F),
+    MOTORWAY(0x1F6E3),
+    RAILWAY_TRACK(0x1F6E4),
+    OIL_DRUM(0x1F6E2),
+    FUEL_PUMP(0x26FD),
+    WHEEL(0x1F6DE),
+    POLICE_CAR_LIGHT(0x1F6A8),
+    HORIZONTAL_TRAFFIC_LIGHT(0x1F6A5),
+    VERTICAL_TRAFFIC_LIGHT(0x1F6A6),
+    STOP_SIGN(0x1F6D1),
+    CONSTRUCTION(0x1F6A7),
+    ANCHOR(0x2693),
+    RING_BUOY(0x1F6DF),
+    SAILBOAT(0x26F5),
+    CANOE(0x1F6F6),
+    SPEEDBOAT(0x1F6A4),
+    PASSENGER_SHIP(0x1F6F3),
+    FERRY(0x26F4),
+    MOTOR_BOAT(0x1F6E5),
+    SHIP(0x1F6A2),
+    AIRPLANE(0x2708),
+    SMALL_AIRPLANE(0x1F6E9),
+    AIRPLANE_DEPARTURE(0x1F6EB),
+    AIRPLANE_ARRIVAL(0x1F6EC),
+    PARACHUTE(0x1FA82),
+    SEAT(0x1F4BA),
+    HELICOPTER(0x1F681),
+    SUSPENSION_RAILWAY(0x1F69F),
+    MOUNTAIN_CABLEWAY(0x1F6A0),
+    AERIAL_TRAMWAY(0x1F6A1),
+    SATELLITE(0x1F6F0),
+    ROCKET(0x1F680),
+    FLYING_SAUCER(0x1F6F8),
+    BELLHOP_BELL(0x1F6CE),
+    LUGGAGE(0x1F9F3),
+    HOURGLASS_DONE(0x231B),
+    HOURGLASS_NOT_DONE(0x23F3),
+    WATCH(0x231A),
+    ALARM_CLOCK(0x23F0),
+    STOPWATCH(0x23F1),
+    TIMER_CLOCK(0x23F2),
+    MANTELPIECE_CLOCK(0x1F570),
+    TWELVE_OCLOCK(0x1F55B),
+    TWELVE_THIRTY(0x1F567),
+    ONE_OCLOCK(0x1F550),
+    ONE_THIRTY(0x1F55C),
+    TWO_OCLOCK(0x1F551),
+    TWO_THIRTY(0x1F55D),
+    THREE_OCLOCK(0x1F552),
+    THREE_THIRTY(0x1F55E),
+    FOUR_OCLOCK(0x1F553),
+    FOUR_THIRTY(0x1F55F),
+    FIVE_OCLOCK(0x1F554),
+    FIVE_THIRTY(0x1F560),
+    SIX_OCLOCK(0x1F555),
+    SIX_THIRTY(0x1F561),
+    SEVEN_OCLOCK(0x1F556),
+    SEVEN_THIRTY(0x1F562),
+    EIGHT_OCLOCK(0x1F557),
+    EIGHT_THIRTY(0x1F563),
+    NINE_OCLOCK(0x1F558),
+    NINE_THIRTY(0x1F564),
+    TEN_OCLOCK(0x1F559),
+    TEN_THIRTY(0x1F565),
+    ELEVEN_OCLOCK(0x1F55A),
+    ELEVEN_THIRTY(0x1F566),
+    NEW_MOON(0x1F311),
+    WAXING_CRESCENT_MOON(0x1F312),
+    FIRST_QUARTER_MOON(0x1F313),
+    WAXING_GIBBOUS_MOON(0x1F314),
+    FULL_MOON(0x1F315),
+    WANING_GIBBOUS_MOON(0x1F316),
+    LAST_QUARTER_MOON(0x1F317),
+    WANING_CRESCENT_MOON(0x1F318),
+    CRESCENT_MOON(0x1F319),
+    NEW_MOON_FACE(0x1F31A),
+    FIRST_QUARTER_MOON_FACE(0x1F31B),
+    LAST_QUARTER_MOON_FACE(0x1F31C),
+    THERMOMETER(0x1F321),
+    SUN(0x2600),
+    FULL_MOON_FACE(0x1F31D),
+    SUN_WITH_FACE(0x1F31E),
+    RINGED_PLANET(0x1FA90),
+    STAR(0x2B50),
+    GLOWING_STAR(0x1F31F),
+    SHOOTING_STAR(0x1F320),
+    MILKY_WAY(0x1F30C),
+    CLOUD(0x2601),
+    SUN_BEHIND_CLOUD(0x26C5),
+    CLOUD_WITH_LIGHTNING_AND_RAIN(0x26C8),
+    SUN_BEHIND_SMALL_CLOUD(0x1F324),
+    SUN_BEHIND_LARGE_CLOUD(0x1F325),
+    SUN_BEHIND_RAIN_CLOUD(0x1F326),
+    CLOUD_WITH_RAIN(0x1F327),
+    CLOUD_WITH_SNOW(0x1F328),
+    CLOUD_WITH_LIGHTNING(0x1F329),
+    TORNADO(0x1F32A),
+    FOG(0x1F32B),
+    WIND_FACE(0x1F32C),
+    CYCLONE(0x1F300),
+    RAINBOW(0x1F308),
+    CLOSED_UMBRELLA(0x1F302),
+    UMBRELLA(0x2602),
+    UMBRELLA_WITH_RAIN_DROPS(0x2614),
+    UMBRELLA_ON_GROUND(0x26F1),
+    HIGH_VOLTAGE(0x26A1),
+    SNOWFLAKE(0x2744),
+    SNOWMAN(0x2603),
+    SNOWMAN_WITHOUT_SNOW(0x26C4),
+    COMET(0x2604),
+    FIRE(0x1F525),
+    DROPLET(0x1F4A7),
+    WATER_WAVE(0x1F30A),
+    JACK_O_LANTERN(0x1F383),
+    CHRISTMAS_TREE(0x1F384),
+    FIREWORKS(0x1F386),
+    SPARKLER(0x1F387),
+    FIRECRACKER(0x1F9E8),
+    SPARKLES(0x2728),
+    BALLOON(0x1F388),
+    PARTY_POPPER(0x1F389),
+    CONFETTI_BALL(0x1F38A),
+    TANABATA_TREE(0x1F38B),
+    PINE_DECORATION(0x1F38D),
+    JAPANESE_DOLLS(0x1F38E),
+    CARP_STREAMER(0x1F38F),
+    WIND_CHIME(0x1F390),
+    MOON_VIEWING_CEREMONY(0x1F391),
+    RED_ENVELOPE(0x1F9E7),
+    RIBBON(0x1F380),
+    WRAPPED_GIFT(0x1F381),
+    REMINDER_RIBBON(0x1F397),
+    ADMISSION_TICKETS(0x1F39F),
+    TICKET(0x1F3AB),
+    MILITARY_MEDAL(0x1F396),
+    TROPHY(0x1F3C6),
+    SPORTS_MEDAL(0x1F3C5),
+    FIRST_PLACE_MEDAL(0x1F947),
+    SECOND_PLACE_MEDAL(0x1F948),
+    THIRD_PLACE_MEDAL(0x1F949),
+    SOCCER_BALL(0x26BD),
+    BASEBALL(0x26BE),
+    SOFTBALL(0x1F94E),
+    BASKETBALL(0x1F3C0),
+    VOLLEYBALL(0x1F3D0),
+    AMERICAN_FOOTBALL(0x1F3C8),
+    RUGBY_FOOTBALL(0x1F3C9),
+    TENNIS(0x1F3BE),
+    FLYING_DISC(0x1F94F),
+    BOWLING(0x1F3B3),
+    CRICKET_GAME(0x1F3CF),
+    FIELD_HOCKEY(0x1F3D1),
+    ICE_HOCKEY(0x1F3D2),
+    LACROSSE(0x1F94D),
+    PING_PONG(0x1F3D3),
+    BADMINTON(0x1F3F8),
+    BOXING_GLOVE(0x1F94A),
+    MARTIAL_ARTS_UNIFORM(0x1F94B),
+    GOAL_NET(0x1F945),
+    FLAG_IN_HOLE(0x26F3),
+    ICE_SKATE(0x26F8),
+    FISHING_POLE(0x1F3A3),
+    DIVING_MASK(0x1F93F),
+    RUNNING_SHIRT(0x1F3BD),
+    SKIS(0x1F3BF),
+    SLED(0x1F6F7),
+    CURLING_STONE(0x1F94C),
+    BULLSEYE(0x1F3AF),
+    YO_YO(0x1FA80),
+    KITE(0x1FA81),
+    POOL_8_BALL(0x1F3B1),
+    CRYSTAL_BALL(0x1F52E),
+    MAGIC_WAND(0x1FA84),
+    NAZAR_AMULET(0x1F9FF),
+    HAMSA(0x1FAAC),
+    VIDEO_GAME(0x1F3AE),
+    JOYSTICK(0x1F579),
+    SLOT_MACHINE(0x1F3B0),
+    GAME_DIE(0x1F3B2),
+    PUZZLE_PIECE(0x1F9E9),
+    TEDDY_BEAR(0x1F9F8),
+    PIÑATA(0x1FA85),
+    MIRROR_BALL(0x1FAA9),
+    NESTING_DOLLS(0x1FA86),
+    SPADE_SUIT(0x2660),
+    HEART_SUIT(0x2665),
+    DIAMOND_SUIT(0x2666),
+    CLUB_SUIT(0x2663),
+    CHESS_PAWN(0x265F),
+    JOKER(0x1F0CF),
+    MAHJONG_RED_DRAGON(0x1F004),
+    FLOWER_PLAYING_CARDS(0x1F3B4),
+    PERFORMING_ARTS(0x1F3AD),
+    FRAMED_PICTURE(0x1F5BC),
+    ARTIST_PALETTE(0x1F3A8),
+    THREAD(0x1F9F5),
+    SEWING_NEEDLE(0x1FAA1),
+    YARN(0x1F9F6),
+    KNOT(0x1FAA2),
+    GLASSES(0x1F453),
+    SUNGLASSES(0x1F576),
+    GOGGLES(0x1F97D),
+    LAB_COAT(0x1F97C),
+    SAFETY_VEST(0x1F9BA),
+    NECKTIE(0x1F454),
+    T_SHIRT(0x1F455),
+    JEANS(0x1F456),
+    SCARF(0x1F9E3),
+    GLOVES(0x1F9E4),
+    COAT(0x1F9E5),
+    SOCKS(0x1F9E6),
+    DRESS(0x1F457),
+    KIMONO(0x1F458),
+    SARI(0x1F97B),
+    ONE_PIECE_SWIMSUIT(0x1FA71),
+    BRIEFS(0x1FA72),
+    SHORTS(0x1FA73),
+    BIKINI(0x1F459),
+    WOMANS_CLOTHES(0x1F45A),
+    PURSE(0x1F45B),
+    HANDBAG(0x1F45C),
+    CLUTCH_BAG(0x1F45D),
+    SHOPPING_BAGS(0x1F6CD),
+    BACKPACK(0x1F392),
+    THONG_SANDAL(0x1FA74),
+    MANS_SHOE(0x1F45E),
+    RUNNING_SHOE(0x1F45F),
+    HIKING_BOOT(0x1F97E),
+    FLAT_SHOE(0x1F97F),
+    HIGH_HEELED_SHOE(0x1F460),
+    WOMANS_SANDAL(0x1F461),
+    BALLET_SHOES(0x1FA70),
+    WOMANS_BOOT(0x1F462),
+    CROWN(0x1F451),
+    WOMANS_HAT(0x1F452),
+    TOP_HAT(0x1F3A9),
+    GRADUATION_CAP(0x1F393),
+    BILLED_CAP(0x1F9E2),
+    MILITARY_HELMET(0x1FA96),
+    RESCUE_WORKERS_HELMET(0x26D1),
+    PRAYER_BEADS(0x1F4FF),
+    LIPSTICK(0x1F484),
+    RING(0x1F48D),
+    GEM_STONE(0x1F48E),
+    MUTED_SPEAKER(0x1F507),
+    SPEAKER_LOW_VOLUME(0x1F508),
+    SPEAKER_MEDIUM_VOLUME(0x1F509),
+    SPEAKER_HIGH_VOLUME(0x1F50A),
+    LOUDSPEAKER(0x1F4E2),
+    MEGAPHONE(0x1F4E3),
+    POSTAL_HORN(0x1F4EF),
+    BELL(0x1F514),
+    BELL_WITH_SLASH(0x1F515),
+    MUSICAL_SCORE(0x1F3BC),
+    MUSICAL_NOTE(0x1F3B5),
+    MUSICAL_NOTES(0x1F3B6),
+    STUDIO_MICROPHONE(0x1F399),
+    LEVEL_SLIDER(0x1F39A),
+    CONTROL_KNOBS(0x1F39B),
+    MICROPHONE(0x1F3A4),
+    HEADPHONE(0x1F3A7),
+    RADIO(0x1F4FB),
+    SAXOPHONE(0x1F3B7),
+    ACCORDION(0x1FA97),
+    GUITAR(0x1F3B8),
+    MUSICAL_KEYBOARD(0x1F3B9),
+    TRUMPET(0x1F3BA),
+    VIOLIN(0x1F3BB),
+    BANJO(0x1FA95),
+    DRUM(0x1F941),
+    LONG_DRUM(0x1FA98),
+    MOBILE_PHONE(0x1F4F1),
+    MOBILE_PHONE_WITH_ARROW(0x1F4F2),
+    TELEPHONE(0x260E),
+    TELEPHONE_RECEIVER(0x1F4DE),
+    PAGER(0x1F4DF),
+    FAX_MACHINE(0x1F4E0),
+    BATTERY(0x1F50B),
+    LOW_BATTERY(0x1FAAB),
+    ELECTRIC_PLUG(0x1F50C),
+    LAPTOP(0x1F4BB),
+    DESKTOP_COMPUTER(0x1F5A5),
+    PRINTER(0x1F5A8),
+    KEYBOARD(0x2328),
+    COMPUTER_MOUSE(0x1F5B1),
+    TRACKBALL(0x1F5B2),
+    COMPUTER_DISK(0x1F4BD),
+    FLOPPY_DISK(0x1F4BE),
+    OPTICAL_DISK(0x1F4BF),
+    DVD(0x1F4C0),
+    ABACUS(0x1F9EE),
+    MOVIE_CAMERA(0x1F3A5),
+    FILM_FRAMES(0x1F39E),
+    FILM_PROJECTOR(0x1F4FD),
+    CLAPPER_BOARD(0x1F3AC),
+    TELEVISION(0x1F4FA),
+    CAMERA(0x1F4F7),
+    CAMERA_WITH_FLASH(0x1F4F8),
+    VIDEO_CAMERA(0x1F4F9),
+    VIDEOCASSETTE(0x1F4FC),
+    MAGNIFYING_GLASS_TILTED_LEFT(0x1F50D),
+    MAGNIFYING_GLASS_TILTED_RIGHT(0x1F50E),
+    CANDLE(0x1F56F),
+    LIGHT_BULB(0x1F4A1),
+    FLASHLIGHT(0x1F526),
+    RED_PAPER_LANTERN(0x1F3EE),
+    DIYA_LAMP(0x1FA94),
+    NOTEBOOK_WITH_DECORATIVE_COVER(0x1F4D4),
+    CLOSED_BOOK(0x1F4D5),
+    OPEN_BOOK(0x1F4D6),
+    GREEN_BOOK(0x1F4D7),
+    BLUE_BOOK(0x1F4D8),
+    ORANGE_BOOK(0x1F4D9),
+    BOOKS(0x1F4DA),
+    NOTEBOOK(0x1F4D3),
+    LEDGER(0x1F4D2),
+    PAGE_WITH_CURL(0x1F4C3),
+    SCROLL(0x1F4DC),
+    PAGE_FACING_UP(0x1F4C4),
+    NEWSPAPER(0x1F4F0);
+
+    override fun toString(): String {
+        return String(Character.toChars(code))
+    }
+}
