+++ src/main/kotlin/net/purefunc/emoji/Emoji3.kt
@@ -1,0 +1,635 @@
+package net.purefunc.emoji
+
+enum class Emoji3(
+    private val code: Int,
+) {
+
+    ROLLED_UP_NEWSPAPER(0x1F5DE),
+    BOOKMARK_TABS(0x1F4D1),
+    BOOKMARK(0x1F516),
+    LABEL(0x1F3F7),
+    MONEY_BAG(0x1F4B0),
+    COIN(0x1FA99),
+    YEN_BANKNOTE(0x1F4B4),
+    DOLLAR_BANKNOTE(0x1F4B5),
+    EURO_BANKNOTE(0x1F4B6),
+    POUND_BANKNOTE(0x1F4B7),
+    MONEY_WITH_WINGS(0x1F4B8),
+    CREDIT_CARD(0x1F4B3),
+    RECEIPT(0x1F9FE),
+    CHART_INCREASING_WITH_YEN(0x1F4B9),
+    ENVELOPE(0x2709),
+    E_MAIL(0x1F4E7),
+    INCOMING_ENVELOPE(0x1F4E8),
+    ENVELOPE_WITH_ARROW(0x1F4E9),
+    OUTBOX_TRAY(0x1F4E4),
+    INBOX_TRAY(0x1F4E5),
+    PACKAGES(0x1F4E6),
+    CLOSED_MAILBOX_WITH_RAISED_FLAG(0x1F4EB),
+    CLOSED_MAILBOX_WITH_LOWERED_FLAG(0x1F4EA),
+    OPEN_MAILBOX_WITH_RAISED_FLAG(0x1F4EC),
+    OPEN_MAILBOX_WITH_LOWERED_FLAG(0x1F4ED),
+    POSTBOX(0x1F4EE),
+    BALLOT_BOX_WITH_BALLOT(0x1F5F3),
+    PENCIL(0x270F),
+    BLACK_NIB(0x2712),
+    FOUNTAIN_PEN(0x1F58B),
+    PEN(0x1F58A),
+    PAINTBRUSH(0x1F58C),
+    CRAYON(0x1F58D),
+    MEMO(0x1F4DD),
+    BRIEFCASE(0x1F4BC),
+    FILE_FOLDER(0x1F4C1),
+    OPEN_FILE_FOLDER(0x1F4C2),
+    CARD_INDEX_DIVIDERS(0x1F5C2),
+    CALENDAR(0x1F4C5),
+    TEAR_OFF_CALENDAR(0x1F4C6),
+    SPIRAL_NOTEPAD(0x1F5D2),
+    SPIRAL_CALENDAR(0x1F5D3),
+    CARD_INDEX(0x1F4C7),
+    CHART_INCREASING(0x1F4C8),
+    CHART_DECREASING(0x1F4C9),
+    BAR_CHART(0x1F4CA),
+    CLIPBOARD(0x1F4CB),
+    PUSHPIN(0x1F4CC),
+    ROUND_PUSHPIN(0x1F4CD),
+    PAPERCLIP(0x1F4CE),
+    LINKED_PAPERCLIPS(0x1F587),
+    STRAIGHT_RULER(0x1F4CF),
+    TRIANGULAR_RULER(0x1F4D0),
+    SCISSORS(0x2702),
+    CARD_FILE_BOX(0x1F5C3),
+    FILE_CABINET(0x1F5C4),
+    WASTEBASKET(0x1F5D1),
+    LOCKED(0x1F512),
+    UNLOCKED(0x1F513),
+    LOCKED_WITH_PEN(0x1F50F),
+    LOCKED_WITH_KEY(0x1F510),
+    KEY(0x1F511),
+    OLD_KEY(0x1F5DD),
+    HAMMER(0x1F528),
+    AXE(0x1FA93),
+    PICK(0x26CF),
+    HAMMER_AND_PICK(0x2692),
+    HAMMER_AND_WRENCH(0x1F6E0),
+    DAGGER(0x1F5E1),
+    CROSSED_SWORDS(0x2694),
+    WATER_PISTOL(0x1F52B),
+    BOOMERANG(0x1FA83),
+    BOW_AND_ARROW(0x1F3F9),
+    SHIELD(0x1F6E1),
+    CARPENTRY_SAW(0x1FA9A),
+    WRENCH(0x1F527),
+    SCREWDRIVER(0x1FA9B),
+    NUT_AND_BOLT(0x1F529),
+    GEAR(0x2699),
+    CLAMP(0x1F5DC),
+    BALANCE_SCALE(0x2696),
+    WHITE_CANE(0x1F9AF),
+    LINK(0x1F517),
+    CHAINS(0x26D3),
+    HOOK(0x1FA9D),
+    TOOLBOX(0x1F9F0),
+    MAGNET(0x1F9F2),
+    LADDER(0x1FA9C),
+    ALEMBIC(0x2697),
+    TEST_TUBE(0x1F9EA),
+    PETRI_DISH(0x1F9EB),
+    DNA(0x1F9EC),
+    MICROSCOPE(0x1F52C),
+    TELESCOPE(0x1F52D),
+    SATELLITE_ANTENNA(0x1F4E1),
+    SYRINGE(0x1F489),
+    DROP_OF_BLOOD(0x1FA78),
+    PILL(0x1F48A),
+    ADHESIVE_BANDAGE(0x1FA79),
+    CRUTCH(0x1FA7C),
+    STETHOSCOPE(0x1FA7A),
+    X_RAY(0x1FA7B),
+    DOOR(0x1F6AA),
+    ELEVATOR(0x1F6D7),
+    MIRROR(0x1FA9E),
+    WINDOW(0x1FA9F),
+    BED(0x1F6CF),
+    COUCH_AND_LAMP(0x1F6CB),
+    CHAIR(0x1FA91),
+    TOILET(0x1F6BD),
+    PLUNGER(0x1FAA0),
+    SHOWER(0x1F6BF),
+    BATHTUB(0x1F6C1),
+    MOUSE_TRAP(0x1FAA4),
+    RAZOR(0x1FA92),
+    LOTION_BOTTLE(0x1F9F4),
+    SAFETY_PIN(0x1F9F7),
+    BROOM(0x1F9F9),
+    BASKET(0x1F9FA),
+    ROLL_OF_PAPER(0x1F9FB),
+    BUCKET(0x1FAA3),
+    SOAP(0x1F9FC),
+    BUBBLES(0x1FAE7),
+    TOOTHBRUSH(0x1FAA5),
+    SPONGE(0x1F9FD),
+    FIRE_EXTINGUISHER(0x1F9EF),
+    SHOPPING_CART(0x1F6D2),
+    CIGARETTE(0x1F6AC),
+    COFFIN(0x26B0),
+    HEADSTONE(0x1FAA6),
+    FUNERAL_URN(0x26B1),
+    MOAI(0x1F5FF),
+    PLACARD(0x1FAA7),
+    IDENTIFICATION_CARD(0x1FAAA),
+    ATM_SIGN(0x1F3E7),
+    LITTER_IN_BIN_SIGN(0x1F6AE),
+    POTABLE_WATER(0x1F6B0),
+    WHEELCHAIR_SYMBOL(0x267F),
+    MENS_ROOM(0x1F6B9),
+    WOMENS_ROOM(0x1F6BA),
+    RESTROOM(0x1F6BB),
+    BABY_SYMBOL(0x1F6BC),
+    WATER_CLOSET(0x1F6BE),
+    PASSPORT_CONTROL(0x1F6C2),
+    CUSTOMS(0x1F6C3),
+    BAGGAGE_CLAIM(0x1F6C4),
+    LEFT_LUGGAGE(0x1F6C5),
+    WARNING(0x26A0),
+    CHILDREN_CROSSING(0x1F6B8),
+    NO_ENTRY(0x26D4),
+    PROHIBITED(0x1F6AB),
+    NO_BICYCLES(0x1F6B3),
+    NO_SMOKING(0x1F6AD),
+    NO_LITTERING(0x1F6AF),
+    NON_POTABLE_WATER(0x1F6B1),
+    NO_PEDESTRIANS(0x1F6B7),
+    NO_MOBILE_PHONES(0x1F4F5),
+    NO_ONE_UNDER_EIGHTEEN(0x1F51E),
+    RADIOACTIVE(0x2622),
+    BIOHAZARD(0x2623),
+    UP_ARROW(0x2B06),
+    UP_RIGHT_ARROW(0x2197),
+    RIGHT_ARROW(0x27A1),
+    DOWN_RIGHT_ARROW(0x2198),
+    DOWN_ARROW(0x2B07),
+    DOWN_LEFT_ARROW(0x2199),
+    LEFT_ARROW(0x2B05),
+    UP_LEFT_ARROW(0x2196),
+    UP_DOWN_ARROW(0x2195),
+    LEFT_RIGHT_ARROW(0x2194),
+    RIGHT_ARROW_CURVING_LEFT(0x21A9),
+    LEFT_ARROW_CURVING_RIGHT(0x21AA),
+    RIGHT_ARROW_CURVING_UP(0x2934),
+    RIGHT_ARROW_CURVING_DOWN(0x2935),
+    CLOCKWISE_VERTICAL_ARROWS(0x1F503),
+    COUNTERCLOCKWISE_ARROWS_BUTTON(0x1F504),
+    BACK_ARROW(0x1F519),
+    END_ARROW(0x1F51A),
+    ON_ARROW(0x1F51B),
+    SOON_ARROW(0x1F51C),
+    TOP_ARROW(0x1F51D),
+    PLACE_OF_WORSHIP(0x1F6D0),
+    ATOM_SYMBOL(0x269B),
+    OM(0x1F549),
+    STAR_OF_DAVID(0x2721),
+    WHEEL_OF_DHARMA(0x2638),
+    YIN_YANG(0x262F),
+    LATIN_CROSS(0x271D),
+    ORTHODOX_CROSS(0x2626),
+    STAR_AND_CRESCENT(0x262A),
+    PEACE_SYMBOL(0x262E),
+    MENORAH(0x1F54E),
+    DOTTED_SIX_POINTED_STAR(0x1F52F),
+    ARIES(0x2648),
+    TAURUS(0x2649),
+    GEMINI(0x264A),
+    CANCER(0x264B),
+    LEO(0x264C),
+    VIRGO(0x264D),
+    LIBRA(0x264E),
+    SCORPIO(0x264F),
+    SAGITTARIUS(0x2650),
+    CAPRICORN(0x2651),
+    AQUARIUS(0x2652),
+    PISCES(0x2653),
+    OPHIUCHUS(0x26CE),
+    SHUFFLE_TRACKS_BUTTON(0x1F500),
+    REPEAT_BUTTON(0x1F501),
+    REPEAT_SINGLE_BUTTON(0x1F502),
+    PLAY_BUTTON(0x25B6),
+    FAST_FORWARD_BUTTON(0x23E9),
+    NEXT_TRACK_BUTTON(0x23ED),
+    PLAY_OR_PAUSE_BUTTON(0x23EF),
+    REVERSE_BUTTON(0x25C0),
+    FAST_REVERSE_BUTTON(0x23EA),
+    LAST_TRACK_BUTTON(0x23EE),
+    UPWARDS_BUTTON(0x1F53C),
+    FAST_UP_BUTTON(0x23EB),
+    DOWNWARDS_BUTTON(0x1F53D),
+    FAST_DOWN_BUTTON(0x23EC),
+    PAUSE_BUTTON(0x23F8),
+    STOP_BUTTON(0x23F9),
+    RECORD_BUTTON(0x23FA),
+    EJECT_BUTTON(0x23CF),
+    CINEMA(0x1F3A6),
+    DIM_BUTTON(0x1F505),
+    BRIGHT_BUTTON(0x1F506),
+    ANTENNA_BARS(0x1F4F6),
+    VIBRATION_MODE(0x1F4F3),
+    MOBILE_PHONE_OFF(0x1F4F4),
+    FEMALE_SIGN(0x2640),
+    MALE_SIGN(0x2642),
+    TRANSGENDER_SYMBOL(0x26A7),
+    MULTIPLY(0x2716),
+    PLUS(0x2795),
+    MINUS(0x2796),
+    DIVIDE(0x2797),
+    HEAVY_EQUALS_SIGN(0x1F7F0),
+    INFINITY(0x267E),
+    DOUBLE_EXCLAMATION_MARK(0x203C),
+    EXCLAMATION_QUESTION_MARK(0x2049),
+    RED_QUESTION_MARK(0x2753),
+    WHITE_QUESTION_MARK(0x2754),
+    WHITE_EXCLAMATION_MARK(0x2755),
+    RED_EXCLAMATION_MARK(0x2757),
+    WAVY_DASH(0x3030),
+    CURRENCY_EXCHANGE(0x1F4B1),
+    HEAVY_DOLLAR_SIGN(0x1F4B2),
+    MEDICAL_SYMBOL(0x2695),
+    RECYCLING_SYMBOL(0x267B),
+    FLEUR_DE_LIS(0x269C),
+    TRIDENT_EMBLEM(0x1F531),
+    NAME_BADGE(0x1F4DB),
+    JAPANESE_SYMBOL_FOR_BEGINNER(0x1F530),
+    HOLLOW_RED_CIRCLE(0x2B55),
+    CHECK_MARK_BUTTON(0x2705),
+    CHECK_BOX_WITH_CHECK(0x2611),
+    CHECK_MARK(0x2714),
+    CROSS_MARK(0x274C),
+    CROSS_MARK_BUTTON(0x274E),
+    CURLY_LOOP(0x27B0),
+    DOUBLE_CURLY_LOOP(0x27BF),
+    PART_ALTERNATION_MARK(0x303D),
+    EIGHT_SPOKED_ASTERISK(0x2733),
+    EIGHT_POINTED_STAR(0x2734),
+    SPARKLE(0x2747),
+    COPYRIGHT(0x00A9),
+    REGISTERED(0x00AE),
+    TRADE_MARK(0x2122),
+    KEYCAP_HASH(0x0023),
+    KEYCAP_STAR(0x002A),
+    KEYCAP_0(0x0030),
+    KEYCAP_1(0x0031),
+    KEYCAP_2(0x0032),
+    KEYCAP_3(0x0033),
+    KEYCAP_4(0x0034),
+    KEYCAP_5(0x0035),
+    KEYCAP_6(0x0036),
+    KEYCAP_7(0x0037),
+    KEYCAP_8(0x0038),
+    KEYCAP_9(0x0039),
+    KEYCAP_10(0x1F51F),
+    INPUT_LATIN_UPPERCASE(0x1F520),
+    INPUT_LATIN_LOWERCASE(0x1F521),
+    INPUT_NUMBERS(0x1F522),
+    INPUT_SYMBOLS(0x1F523),
+    INPUT_LATIN_LETTERS(0x1F524),
+    A_BUTTON_BLOOD_TYPE(0x1F170),
+    AB_BUTTON_BLOOD_TYPE(0x1F18E),
+    B_BUTTON_BLOOD_TYPE(0x1F171),
+    CL_BUTTON(0x1F191),
+    COOL_BUTTON(0x1F192),
+    FREE_BUTTON(0x1F193),
+    INFORMATION(0x2139),
+    ID_BUTTON(0x1F194),
+    CIRCLED_M(0x24C2),
+    NEW_BUTTON(0x1F195),
+    NG_BUTTON(0x1F196),
+    O_BUTTON_BLOOD_TYPE(0x1F17E),
+    OK_BUTTON(0x1F197),
+    P_BUTTON(0x1F17F),
+    SOS_BUTTON(0x1F198),
+    UP_BUTTON(0x1F199),
+    VS_BUTTON(0x1F19A),
+    JAPANESE_HERE_BUTTON(0x1F201),
+    JAPANESE_SERVICE_CHARGE_BUTTON(0x1F202),
+    JAPANESE_MONTHLY_AMOUNT_BUTTON(0x1F237),
+    JAPANESE_NOT_FREE_OF_CHARGE_BUTTON(0x1F236),
+    JAPANESE_RESERVED_BUTTON(0x1F22F),
+    JAPANESE_BARGAIN_BUTTON(0x1F250),
+    JAPANESE_DISCOUNT_BUTTON(0x1F239),
+    JAPANESE_FREE_OF_CHARGE_BUTTON(0x1F21A),
+    JAPANESE_PROHIBITED_BUTTON(0x1F232),
+    JAPANESE_ACCEPTABLE_BUTTON(0x1F251),
+    JAPANESE_APPLICATION_BUTTON(0x1F238),
+    JAPANESE_PASSING_GRADE_BUTTON(0x1F234),
+    JAPANESE_VACANCY_BUTTON(0x1F233),
+    JAPANESE_CONGRATULATIONS_BUTTON(0x3297),
+    JAPANESE_SECRET_BUTTON(0x3299),
+    JAPANESE_OPEN_FOR_BUSINESS_BUTTON(0x1F23A),
+    JAPANESE_NO_VACANCY_BUTTON(0x1F235),
+    RED_CIRCLE(0x1F534),
+    ORANGE_CIRCLE(0x1F7E0),
+    YELLOW_CIRCLE(0x1F7E1),
+    GREEN_CIRCLE(0x1F7E2),
+    BLUE_CIRCLE(0x1F535),
+    PURPLE_CIRCLE(0x1F7E3),
+    BROWN_CIRCLE(0x1F7E4),
+    BLACK_CIRCLE(0x26AB),
+    WHITE_CIRCLE(0x26AA),
+    RED_SQUARE(0x1F7E5),
+    ORANGE_SQUARE(0x1F7E7),
+    YELLOW_SQUARE(0x1F7E8),
+    GREEN_SQUARE(0x1F7E9),
+    BLUE_SQUARE(0x1F7E6),
+    PURPLE_SQUARE(0x1F7EA),
+    BROWN_SQUARE(0x1F7EB),
+    BLACK_LARGE_SQUARE(0x2B1B),
+    WHITE_LARGE_SQUARE(0x2B1C),
+    BLACK_MEDIUM_SQUARE(0x25FC),
+    WHITE_MEDIUM_SQUARE(0x25FB),
+    BLACK_MEDIUM_SMALL_SQUARE(0x25FE),
+    WHITE_MEDIUM_SMALL_SQUARE(0x25FD),
+    BLACK_SMALL_SQUARE(0x25AA),
+    WHITE_SMALL_SQUARE(0x25AB),
+    LARGE_ORANGE_DIAMOND(0x1F536),
+    LARGE_BLUE_DIAMOND(0x1F537),
+    SMALL_ORANGE_DIAMOND(0x1F538),
+    SMALL_BLUE_DIAMOND(0x1F539),
+    RED_TRIANGLE_POINTED_UP(0x1F53A),
+    RED_TRIANGLE_POINTED_DOWN(0x1F53B),
+    DIAMOND_WITH_A_DOT(0x1F4A0),
+    RADIO_BUTTON(0x1F518),
+    WHITE_SQUARE_BUTTON(0x1F533),
+    BLACK_SQUARE_BUTTON(0x1F532),
+    CHEQUERED_FLAG(0x1F3C1),
+    TRIANGULAR_FLAG(0x1F6A9),
+    CROSSED_FLAGS(0x1F38C),
+    BLACK_FLAG(0x1F3F4),
+    WHITE_FLAG(0x1F3F3),
+    RAINBOW_FLAG(0x1F3F3),
+    TRANSGENDER_FLAG(0x1F3F3),
+    PIRATE_FLAG(0x1F3F4),
+    FLAG_ASCENSION_ISLAND(0x1F1E6),
+    FLAG_ANDORRA(0x1F1E6),
+    FLAG_UNITED_ARAB_EMIRATES(0x1F1E6),
+    FLAG_AFGHANISTAN(0x1F1E6),
+    FLAG_ANTIGUA_AND_BARBUDA(0x1F1E6),
+    FLAG_ANGUILLA(0x1F1E6),
+    FLAG_ALBANIA(0x1F1E6),
+    FLAG_ARMENIA(0x1F1E6),
+    FLAG_ANGOLA(0x1F1E6),
+    FLAG_ANTARCTICA(0x1F1E6),
+    FLAG_ARGENTINA(0x1F1E6),
+    FLAG_AMERICAN_SAMOA(0x1F1E6),
+    FLAG_AUSTRIA(0x1F1E6),
+    FLAG_AUSTRALIA(0x1F1E6),
+    FLAG_ARUBA(0x1F1E6),
+    FLAG_ÅLAND_ISLANDS(0x1F1E6),
+    FLAG_AZERBAIJAN(0x1F1E6),
+    FLAG_BOSNIA_AND_HERZEGOVINA(0x1F1E7),
+    FLAG_BARBADOS(0x1F1E7),
+    FLAG_BANGLADESH(0x1F1E7),
+    FLAG_BELGIUM(0x1F1E7),
+    FLAG_BURKINA_FASO(0x1F1E7),
+    FLAG_BULGARIA(0x1F1E7),
+    FLAG_BAHRAIN(0x1F1E7),
+    FLAG_BURUNDI(0x1F1E7),
+    FLAG_BENIN(0x1F1E7),
+    FLAG_ST_BARTHÉLEMY(0x1F1E7),
+    FLAG_BERMUDA(0x1F1E7),
+    FLAG_BRUNEI(0x1F1E7),
+    FLAG_BOLIVIA(0x1F1E7),
+    FLAG_CARIBBEAN_NETHERLANDS(0x1F1E7),
+    FLAG_BRAZIL(0x1F1E7),
+    FLAG_BAHAMAS(0x1F1E7),
+    FLAG_BHUTAN(0x1F1E7),
+    FLAG_BOUVET_ISLAND(0x1F1E7),
+    FLAG_BOTSWANA(0x1F1E7),
+    FLAG_BELARUS(0x1F1E7),
+    FLAG_BELIZE(0x1F1E7),
+    FLAG_CANADA(0x1F1E8),
+    FLAG_COCOS_KEELING_ISLANDS(0x1F1E8),
+    FLAG_CONGO___KINSHASA(0x1F1E8),
+    FLAG_CENTRAL_AFRICAN_REPUBLIC(0x1F1E8),
+    FLAG_CONGO___BRAZZAVILLE(0x1F1E8),
+    FLAG_SWITZERLAND(0x1F1E8),
+    FLAG_CÔTE_DIVOIRE(0x1F1E8),
+    FLAG_COOK_ISLANDS(0x1F1E8),
+    FLAG_CHILE(0x1F1E8),
+    FLAG_CAMEROON(0x1F1E8),
+    FLAG_CHINA(0x1F1E8),
+    FLAG_COLOMBIA(0x1F1E8),
+    FLAG_CLIPPERTON_ISLAND(0x1F1E8),
+    FLAG_COSTA_RICA(0x1F1E8),
+    FLAG_CUBA(0x1F1E8),
+    FLAG_CAPE_VERDE(0x1F1E8),
+    FLAG_CURAÇAO(0x1F1E8),
+    FLAG_CHRISTMAS_ISLAND(0x1F1E8),
+    FLAG_CYPRUS(0x1F1E8),
+    FLAG_CZECHIA(0x1F1E8),
+    FLAG_GERMANY(0x1F1E9),
+    FLAG_DIEGO_GARCIA(0x1F1E9),
+    FLAG_DJIBOUTI(0x1F1E9),
+    FLAG_DENMARK(0x1F1E9),
+    FLAG_DOMINICA(0x1F1E9),
+    FLAG_DOMINICAN_REPUBLIC(0x1F1E9),
+    FLAG_ALGERIA(0x1F1E9),
+    FLAG_CEUTA_AND_MELILLA(0x1F1EA),
+    FLAG_ECUADOR(0x1F1EA),
+    FLAG_ESTONIA(0x1F1EA),
+    FLAG_EGYPT(0x1F1EA),
+    FLAG_WESTERN_SAHARA(0x1F1EA),
+    FLAG_ERITREA(0x1F1EA),
+    FLAG_SPAIN(0x1F1EA),
+    FLAG_ETHIOPIA(0x1F1EA),
+    FLAG_EUROPEAN_UNION(0x1F1EA),
+    FLAG_FINLAND(0x1F1EB),
+    FLAG_FIJI(0x1F1EB),
+    FLAG_FALKLAND_ISLANDS(0x1F1EB),
+    FLAG_MICRONESIA(0x1F1EB),
+    FLAG_FAROE_ISLANDS(0x1F1EB),
+    FLAG_FRANCE(0x1F1EB),
+    FLAG_GABON(0x1F1EC),
+    FLAG_UNITED_KINGDOM(0x1F1EC),
+    FLAG_GRENADA(0x1F1EC),
+    FLAG_GEORGIA(0x1F1EC),
+    FLAG_FRENCH_GUIANA(0x1F1EC),
+    FLAG_GUERNSEY(0x1F1EC),
+    FLAG_GHANA(0x1F1EC),
+    FLAG_GIBRALTAR(0x1F1EC),
+    FLAG_GREENLAND(0x1F1EC),
+    FLAG_GAMBIA(0x1F1EC),
+    FLAG_GUINEA(0x1F1EC),
+    FLAG_GUADELOUPE(0x1F1EC),
+    FLAG_EQUATORIAL_GUINEA(0x1F1EC),
+    FLAG_GREECE(0x1F1EC),
+    FLAG_SOUTH_GEORGIA_AND_SOUTH_SANDWICH_ISLANDS(0x1F1EC),
+    FLAG_GUATEMALA(0x1F1EC),
+    FLAG_GUAM(0x1F1EC),
+    FLAG_GUINEA_BISSAU(0x1F1EC),
+    FLAG_GUYANA(0x1F1EC),
+    FLAG_HONG_KONG_SAR_CHINA(0x1F1ED),
+    FLAG_HEARD_AND_MCDONALD_ISLANDS(0x1F1ED),
+    FLAG_HONDURAS(0x1F1ED),
+    FLAG_CROATIA(0x1F1ED),
+    FLAG_HAITI(0x1F1ED),
+    FLAG_HUNGARY(0x1F1ED),
+    FLAG_CANARY_ISLANDS(0x1F1EE),
+    FLAG_INDONESIA(0x1F1EE),
+    FLAG_IRELAND(0x1F1EE),
+    FLAG_ISRAEL(0x1F1EE),
+    FLAG_ISLE_OF_MAN(0x1F1EE),
+    FLAG_INDIA(0x1F1EE),
+    FLAG_BRITISH_INDIAN_OCEAN_TERRITORY(0x1F1EE),
+    FLAG_IRAQ(0x1F1EE),
+    FLAG_IRAN(0x1F1EE),
+    FLAG_ICELAND(0x1F1EE),
+    FLAG_ITALY(0x1F1EE),
+    FLAG_JERSEY(0x1F1EF),
+    FLAG_JAMAICA(0x1F1EF),
+    FLAG_JORDAN(0x1F1EF),
+    FLAG_JAPAN(0x1F1EF),
+    FLAG_KENYA(0x1F1F0),
+    FLAG_KYRGYZSTAN(0x1F1F0),
+    FLAG_CAMBODIA(0x1F1F0),
+    FLAG_KIRIBATI(0x1F1F0),
+    FLAG_COMOROS(0x1F1F0),
+    FLAG_ST_KITTS_AND_NEVIS(0x1F1F0),
+    FLAG_NORTH_KOREA(0x1F1F0),
+    FLAG_SOUTH_KOREA(0x1F1F0),
+    FLAG_KUWAIT(0x1F1F0),
+    FLAG_CAYMAN_ISLANDS(0x1F1F0),
+    FLAG_KAZAKHSTAN(0x1F1F0),
+    FLAG_LAOS(0x1F1F1),
+    FLAG_LEBANON(0x1F1F1),
+    FLAG_ST_LUCIA(0x1F1F1),
+    FLAG_LIECHTENSTEIN(0x1F1F1),
+    FLAG_SRI_LANKA(0x1F1F1),
+    FLAG_LIBERIA(0x1F1F1),
+    FLAG_LESOTHO(0x1F1F1),
+    FLAG_LITHUANIA(0x1F1F1),
+    FLAG_LUXEMBOURG(0x1F1F1),
+    FLAG_LATVIA(0x1F1F1),
+    FLAG_LIBYA(0x1F1F1),
+    FLAG_MOROCCO(0x1F1F2),
+    FLAG_MONACO(0x1F1F2),
+    FLAG_MOLDOVA(0x1F1F2),
+    FLAG_MONTENEGRO(0x1F1F2),
+    FLAG_ST_MARTIN(0x1F1F2),
+    FLAG_MADAGASCAR(0x1F1F2),
+    FLAG_MARSHALL_ISLANDS(0x1F1F2),
+    FLAG_NORTH_MACEDONIA(0x1F1F2),
+    FLAG_MALI(0x1F1F2),
+    FLAG_MYANMAR_BURMA(0x1F1F2),
+    FLAG_MONGOLIA(0x1F1F2),
+    FLAG_MACAO_SAR_CHINA(0x1F1F2),
+    FLAG_NORTHERN_MARIANA_ISLANDS(0x1F1F2),
+    FLAG_MARTINIQUE(0x1F1F2),
+    FLAG_MAURITANIA(0x1F1F2),
+    FLAG_MONTSERRAT(0x1F1F2),
+    FLAG_MALTA(0x1F1F2),
+    FLAG_MAURITIUS(0x1F1F2),
+    FLAG_MALDIVES(0x1F1F2),
+    FLAG_MALAWI(0x1F1F2),
+    FLAG_MEXICO(0x1F1F2),
+    FLAG_MALAYSIA(0x1F1F2),
+    FLAG_MOZAMBIQUE(0x1F1F2),
+    FLAG_NAMIBIA(0x1F1F3),
+    FLAG_NEW_CALEDONIA(0x1F1F3),
+    FLAG_NIGER(0x1F1F3),
+    FLAG_NORFOLK_ISLAND(0x1F1F3),
+    FLAG_NIGERIA(0x1F1F3),
+    FLAG_NICARAGUA(0x1F1F3),
+    FLAG_NETHERLANDS(0x1F1F3),
+    FLAG_NORWAY(0x1F1F3),
+    FLAG_NEPAL(0x1F1F3),
+    FLAG_NAURU(0x1F1F3),
+    FLAG_NIUE(0x1F1F3),
+    FLAG_NEW_ZEALAND(0x1F1F3),
+    FLAG_OMAN(0x1F1F4),
+    FLAG_PANAMA(0x1F1F5),
+    FLAG_PERU(0x1F1F5),
+    FLAG_FRENCH_POLYNESIA(0x1F1F5),
+    FLAG_PAPUA_NEW_GUINEA(0x1F1F5),
+    FLAG_PHILIPPINES(0x1F1F5),
+    FLAG_PAKISTAN(0x1F1F5),
+    FLAG_POLAND(0x1F1F5),
+    FLAG_ST_PIERRE_AND_MIQUELON(0x1F1F5),
+    FLAG_PITCAIRN_ISLANDS(0x1F1F5),
+    FLAG_PUERTO_RICO(0x1F1F5),
+    FLAG_PALESTINIAN_TERRITORIES(0x1F1F5),
+    FLAG_PORTUGAL(0x1F1F5),
+    FLAG_PALAU(0x1F1F5),
+    FLAG_PARAGUAY(0x1F1F5),
+    FLAG_QATAR(0x1F1F6),
+    FLAG_RÉUNION(0x1F1F7),
+    FLAG_ROMANIA(0x1F1F7),
+    FLAG_SERBIA(0x1F1F7),
+    FLAG_RUSSIA(0x1F1F7),
+    FLAG_RWANDA(0x1F1F7),
+    FLAG_SAUDI_ARABIA(0x1F1F8),
+    FLAG_SOLOMON_ISLANDS(0x1F1F8),
+    FLAG_SEYCHELLES(0x1F1F8),
+    FLAG_SUDAN(0x1F1F8),
+    FLAG_SWEDEN(0x1F1F8),
+    FLAG_SINGAPORE(0x1F1F8),
+    FLAG_ST_HELENA(0x1F1F8),
+    FLAG_SLOVENIA(0x1F1F8),
+    FLAG_SVALBARD_AND_JAN_MAYEN(0x1F1F8),
+    FLAG_SLOVAKIA(0x1F1F8),
+    FLAG_SIERRA_LEONE(0x1F1F8),
+    FLAG_SAN_MARINO(0x1F1F8),
+    FLAG_SENEGAL(0x1F1F8),
+    FLAG_SOMALIA(0x1F1F8),
+    FLAG_SURINAME(0x1F1F8),
+    FLAG_SOUTH_SUDAN(0x1F1F8),
+    FLAG_SÃO_TOMÉ_AND_PRÍNCIPE(0x1F1F8),
+    FLAG_EL_SALVADOR(0x1F1F8),
+    FLAG_SINT_MAARTEN(0x1F1F8),
+    FLAG_SYRIA(0x1F1F8),
+    FLAG_ESWATINI(0x1F1F8),
+    FLAG_TRISTAN_DA_CUNHA(0x1F1F9),
+    FLAG_TURKS_AND_CAICOS_ISLANDS(0x1F1F9),
+    FLAG_CHAD(0x1F1F9),
+    FLAG_FRENCH_SOUTHERN_TERRITORIES(0x1F1F9),
+    FLAG_TOGO(0x1F1F9),
+    FLAG_THAILAND(0x1F1F9),
+    FLAG_TAJIKISTAN(0x1F1F9),
+    FLAG_TOKELAU(0x1F1F9),
+    FLAG_TIMOR_LESTE(0x1F1F9),
+    FLAG_TURKMENISTAN(0x1F1F9),
+    FLAG_TUNISIA(0x1F1F9),
+    FLAG_TONGA(0x1F1F9),
+    FLAG_TURKEY(0x1F1F9),
+    FLAG_TRINIDAD_AND_TOBAGO(0x1F1F9),
+    FLAG_TUVALU(0x1F1F9),
+    FLAG_TAIWAN(0x1F1F9),
+    FLAG_TANZANIA(0x1F1F9),
+    FLAG_UKRAINE(0x1F1FA),
+    FLAG_UGANDA(0x1F1FA),
+    FLAG_US_OUTLYING_ISLANDS(0x1F1FA),
+    FLAG_UNITED_NATIONS(0x1F1FA),
+    FLAG_UNITED_STATES(0x1F1FA),
+    FLAG_URUGUAY(0x1F1FA),
+    FLAG_UZBEKISTAN(0x1F1FA),
+    FLAG_VATICAN_CITY(0x1F1FB),
+    FLAG_ST_VINCENT_AND_GRENADINES(0x1F1FB),
+    FLAG_VENEZUELA(0x1F1FB),
+    FLAG_BRITISH_VIRGIN_ISLANDS(0x1F1FB),
+    FLAG_US_VIRGIN_ISLANDS(0x1F1FB),
+    FLAG_VIETNAM(0x1F1FB),
+    FLAG_VANUATU(0x1F1FB),
+    FLAG_WALLIS_AND_FUTUNA(0x1F1FC),
+    FLAG_SAMOA(0x1F1FC),
+    FLAG_KOSOVO(0x1F1FD),
+    FLAG_YEMEN(0x1F1FE),
+    FLAG_MAYOTTE(0x1F1FE),
+    FLAG_SOUTH_AFRICA(0x1F1FF),
+    FLAG_ZAMBIA(0x1F1FF),
+    FLAG_ZIMBABWE(0x1F1FF),
+    FLAG_ENGLAND(0x1F3F4),
+    FLAG_SCOTLAND(0x1F3F4),
+    FLAG_WALES(0x1F3F4);
+
+    override fun toString(): String {
+        return String(Character.toChars(code))
+    }
+}
